### 2024-12-03 #7552 @lieres
The paper is located here:

https://github.com/cadet/CADET-Core/tree/paper/JOSS_2024/paper
