+++ src/libcadet/model/LumpedRateModelWithPoresDG2D.cpp
@@ -1,0 +1,1773 @@
+// =============================================================================
+//  CADET
+//
+//  Copyright © 2008-present: The CADET-Core Authors
+//            Please see the AUTHORS.md file.
+//
+//  All rights reserved. This program and the accompanying materials
+//  are made available under the terms of the GNU Public License v3.0 (or, at
+//  your option, any later version) which accompanies this distribution, and
+//  is available at http://www.gnu.org/licenses/gpl.html
+// =============================================================================
+
+#include "model/LumpedRateModelWithPoresDG2D.hpp"
+#include "BindingModelFactory.hpp"
+#include "ParamReaderHelper.hpp"
+#include "ParamReaderScopes.hpp"
+#include "cadet/Exceptions.hpp"
+#include "cadet/ExternalFunction.hpp"
+#include "cadet/SolutionRecorder.hpp"
+#include "ConfigurationHelper.hpp"
+#include "model/BindingModel.hpp"
+#include "model/ReactionModel.hpp"
+#include "model/parts/BindingCellKernel.hpp"
+#include "SimulationTypes.hpp"
+#include "linalg/Norms.hpp"
+
+#include "AdUtils.hpp"
+#include "SensParamUtil.hpp"
+
+#include "LoggingUtils.hpp"
+#include "Logging.hpp"
+
+#include <algorithm>
+#include <functional>
+#include <numeric>
+
+#include "ParallelSupport.hpp"
+
+namespace
+{
+
+cadet::model::MultiplexMode readAndRegisterMultiplexParam(cadet::IParameterProvider& paramProvider, std::unordered_map<cadet::ParameterId, cadet::active*>& parameters, std::vector<cadet::active>& values, const std::string& name, unsigned int nAxial, unsigned int nRad, unsigned int nParType, cadet::UnitOpIdx uoi)
+{
+	cadet::model::MultiplexMode mode = cadet::model::MultiplexMode::Independent;
+	readScalarParameterOrArray(values, paramProvider, name, 1);
+	if (paramProvider.exists(name + "_MULTIPLEX"))
+	{
+		const int modeConfig = paramProvider.getInt(name + "_MULTIPLEX");
+		if (modeConfig == 0)
+		{
+			mode = cadet::model::MultiplexMode::Independent;
+			if (values.size() != nParType)
+				throw cadet::InvalidParameterException("Number of elements in field " + name + " inconsistent with " + name + "_MULTIPLEX (should be " + std::to_string(nParType) + ")");
+		}
+		else if (modeConfig == 1)
+		{
+			mode = cadet::model::MultiplexMode::Radial;
+			if (values.size() != nRad * nParType)
+				throw cadet::InvalidParameterException("Number of elements in field " + name + " inconsistent with " + name + "_MULTIPLEX (should be " + std::to_string(nRad * nParType) + ")");
+		}
+		else if (modeConfig == 2)
+		{
+			mode = cadet::model::MultiplexMode::Axial;
+			if (values.size() != nAxial * nParType)
+				throw cadet::InvalidParameterException("Number of elements in field " + name + " inconsistent with " + name + "_MULTIPLEX (should be " + std::to_string(nAxial * nParType) + ")");
+		}
+		else if (modeConfig == 3)
+		{
+			mode = cadet::model::MultiplexMode::AxialRadial;
+			if (values.size() != nAxial * nRad * nParType)
[…]
