+++ src/libcadet/model/LumpedRateModelWithPoresDG2D-LinearSolver.cpp
@@ -1,0 +1,231 @@
+// =============================================================================
+//  CADET
+//
+//  Copyright © 2008-present: The CADET-Core Authors
+//            Please see the AUTHORS.md file.
+//
+//  All rights reserved. This program and the accompanying materials
+//  are made available under the terms of the GNU Public License v3.0 (or, at
+//  your option, any later version) which accompanies this distribution, and
+//  is available at http://www.gnu.org/licenses/gpl.html
+// =============================================================================
+
+#include "model/LumpedRateModelWithPoresDG2D.hpp"
+#include "model/BindingModel.hpp"
+#include "model/parts/BindingCellKernel.hpp"
+#include "linalg/DenseMatrix.hpp"
+#include "AdUtils.hpp"
+
+#include <algorithm>
+#include <functional>
+
+#include "LoggingUtils.hpp"
+#include "Logging.hpp"
+
+namespace cadet
+{
+
+namespace model
+{
+/**
+ * @brief Computes the solution of the linear system involving the system Jacobian
+ * @details The system \f[ \left( \frac{\partial F}{\partial y} + \alpha \frac{\partial F}{\partial \dot{y}} \right) x = b \f]
+ *          has to be solved. The right hand side \f$ b \f$ is given by @p rhs, the Jacobians are evaluated at the
+ *          point \f$(y, \dot{y})\f$ given by @p y and @p yDot. The residual @p res at this point, \f$ F(t, y, \dot{y}) \f$,
+ *          may help with this. Error weights (see IDAS guide) are given in @p weight. The solution is returned in @p rhs.
+ *
+ *          The full Jacobian @f$ J = \left( \frac{\partial F}{\partial y} + \alpha \frac{\partial F}{\partial \dot{y}} \right) @f$ is given by
+ *          @f[ \begin{align}
+				J =
+				\left[\begin{array}{c|ccc|c}
+					 J_0 & & & & J_{0,f} \\
+					\hline
+					 & J_1 & & & J_{1,f} \\
+					 & & \ddots & & \vdots \\
+					 & & & J_{N_z} & J_{N_z,f} \\
+					\hline
+					 J_{f,0} & J_{f,1} & \dots & J_{f,N_z} & J_f
+				\end{array}\right].
+			\end{align} @f]
+ *          By decomposing the Jacobian @f$ J @f$ into @f$ J = LU @f$, we get
+ *          @f[ \begin{align}
+				L &= \left[\begin{array}{c|ccc|c}
+					  J_0     &         &        &           &   \\
+					\hline
+							  & J_1     &        &           &   \\
+							  &         & \ddots &           &   \\
+							  &         &        & J_{N_z}   &   \\
+					\hline
+					  J_{f,0} & J_{f,1} & \dots & J_{f,N_z} & I
+				\end{array}\right], \\
+				U &= \left[\begin{array}{c|ccc|c}
+					  I &   &        &   & J_0^{-1} \, J_{0,f}       \\
+					\hline
+						& I &        &   & J_1^{-1} \, J_{1,f}       \\
+						&   & \ddots &   & \vdots                    \\
+						&   &        & I & J_{N_z}^{-1} \, J_{N_z,f} \\
+					\hline
+						&   &        &   & S
+				\end{array}\right].
+			\end{align} @f]
+ *          Here, the Schur-complement @f$ S @f$ is given by
+ *          @f[ \begin{align}
+				S = J_f - J_{f,0} \, J_0^{-1} \, J_{0,f} - \sum_{p=1}^{N_z}{J_{f,p} \, J_p^{-1} \, J_{p,f}}.
+			\end{align} @f]
[…]
