+++ test/GeneralRateModelDG.cpp
@@ -23,7 +23,7 @@
 #include "Utils.hpp"
 #include "common/Driver.hpp"
 
-TEST_CASE("GRM_DG LWE forward vs backward flow", "[GRM],[DG],[Simulation],[CI]")
+TEST_CASE("GRM_DG LWE forward vs backward flow", "[GRM],[DG],[DG1D],[Simulation],[CI]")
 {
 	cadet::test::column::DGparams disc;
 
@@ -35,7 +35,7 @@ TEST_CASE("GRM_DG LWE forward vs backward flow", "[GRM],[DG],[Simulation],[CI]")
 	}
 }
 
-TEST_CASE("GRM_DG linear pulse vs analytic solution", "[GRM],[DG],[Simulation],[Analytic],[CI]")
+TEST_CASE("GRM_DG linear pulse vs analytic solution", "[GRM],[DG],[DG1D],[Simulation],[Analytic],[CI]")
 {
 	cadet::test::column::DGparams disc;
 	cadet::test::column::testAnalyticBenchmark("GENERAL_RATE_MODEL", "/data/grm-pulseBenchmark.data", true, true, disc, 6e-5, 1e-7);
@@ -44,7 +44,7 @@ TEST_CASE("GRM_DG linear pulse vs analytic solution", "[GRM],[DG],[Simulation],[
 	cadet::test::column::testAnalyticBenchmark("GENERAL_RATE_MODEL", "/data/grm-pulseBenchmark.data", false, false, disc, 6e-5, 1e-7);
 }
 
-TEST_CASE("GRM_DG non-binding linear pulse vs analytic solution", "[GRM],[DG],[Simulation],[Analytic],[NonBinding],[CI]")
+TEST_CASE("GRM_DG non-binding linear pulse vs analytic solution", "[GRM],[DG],[DG1D],[Simulation],[Analytic],[NonBinding],[CI]")
 {
 	cadet::test::column::DGparams disc;
 	cadet::test::column::testAnalyticNonBindingBenchmark("GENERAL_RATE_MODEL", "/data/grm-nonBinding.data", true, disc, 6e-5, 1e-7);
@@ -52,7 +52,7 @@ TEST_CASE("GRM_DG non-binding linear pulse vs analytic solution", "[GRM],[DG],[S
 }
 
 // todo FIX (scheitert bei backward flow jacobian vs AD)
-//TEST_CASE("GRM_DG Jacobian forward vs backward flow", "[GRM],[DG],[UnitOp],[Residual],[Jacobian],[AD],[todo]")
+//TEST_CASE("GRM_DG Jacobian forward vs backward flow", "[GRM],[DG],[DG1D],[UnitOp],[Residual],[Jacobian],[AD],[todo]")
 //{
 //	cadet::test::column::DGparams disc;
 //
@@ -64,7 +64,7 @@ TEST_CASE("GRM_DG non-binding linear pulse vs analytic solution", "[GRM],[DG],[S
 //	}
 //}
 
-TEST_CASE("GRM_DG numerical Benchmark with parameter sensitivities for linear case", "[GRM],[DG],[Simulation],[Reference],[Sensitivity]") // todo CI flag: currently only runs locally but fails on server
+TEST_CASE("GRM_DG numerical Benchmark with parameter sensitivities for linear case", "[GRM],[DG],[DG1D],[Simulation],[Reference],[Sensitivity]") // todo CI flag: currently only runs locally but fails on server
 {
 	const std::string& modelFilePath = std::string("/data/model_GRM_dynLin_1comp_benchmark1.json");
 	const std::string& refFilePath = std::string("/data/ref_GRM_dynLin_1comp_sensbenchmark1_cDG_P3Z8_GSM_parP3parZ1.h5");
[…]
