+++ src/libcadet/model/LumpedRateModelWithPoresDG2D-InitialConditions.cpp
@@ -1,0 +1,1418 @@
+// =============================================================================
+//  CADET
+//
+//  Copyright © 2008-present: The CADET-Core Authors
+//            Please see the AUTHORS.md file.
+//
+//  All rights reserved. This program and the accompanying materials
+//  are made available under the terms of the GNU Public License v3.0 (or, at
+//  your option, any later version) which accompanies this distribution, and
+//  is available at http://www.gnu.org/licenses/gpl.html
+// =============================================================================
+
+#include "model/LumpedRateModelWithPoresDG2D.hpp"
+#include "model/BindingModel.hpp"
+#include "linalg/DenseMatrix.hpp"
+#include "linalg/Subset.hpp"
+#include "ParamReaderHelper.hpp"
+#include "AdUtils.hpp"
+#include "model/parts/BindingCellKernel.hpp"
+#include "SimulationTypes.hpp"
+#include "SensParamUtil.hpp"
+
+#include <algorithm>
+#include <functional>
+
+#include "LoggingUtils.hpp"
+#include "Logging.hpp"
+
+#include "ParallelSupport.hpp"
+
+namespace cadet
+{
+
+namespace model
+{
+
+int LumpedRateModelWithPoresDG2D::multiplexInitialConditions(const cadet::ParameterId& pId, unsigned int adDirection, double adValue)
+{
+	if (pId.name == hashString("INIT_C") && (pId.section == SectionIndep) && (pId.boundState == BoundStateIndep) && (pId.particleType == ParTypeIndep) && (pId.component != CompIndep))
+	{
+		if ((pId.reaction == ReactionIndep) && _singleRadiusInitC)
+		{
+			_sensParams.insert(&_initC[pId.component]);
+			for (unsigned int r = 0; r < _disc.radNPoints; ++r)
+				_initC[r * _disc.nComp + pId.component].setADValue(adDirection, adValue);
+
+			return 1;
+		}
+		else if ((pId.reaction != ReactionIndep) && !_singleRadiusInitC)
+		{
+			_sensParams.insert(&_initC[pId.reaction * _disc.nComp + pId.component]);
+			_initC[pId.reaction * _disc.nComp + pId.component].setADValue(adDirection, adValue);
+			return 1;
+		}
+
+		return -1;
+	}
+	else if (pId.name == hashString("INIT_C"))
+		return -1;
+
+	if (_singleBinding)
+	{
+		if (_singleRadiusInitCp)
+		{
+			if ((pId.name == hashString("INIT_CP")) && (pId.section == SectionIndep) && (pId.boundState == BoundStateIndep) && (pId.particleType == ParTypeIndep) && (pId.component != CompIndep) && (pId.reaction == ReactionIndep))
+			{
+				_sensParams.insert(&_initCp[pId.component]);
+				for (unsigned int r = 0; r < _disc.radNPoints; ++r)
+				{
+					for (unsigned int t = 0; t < _disc.nParType; ++t)
+						_initCp[r * _disc.nComp * _disc.nParType + t * _disc.nComp + pId.component].setADValue(adDirection, adValue);
+				}
+				return 1;
+			}
+			else if (pId.name == hashString("INIT_CP"))
+				return -1;
+		}
+		else
+		{
+			if ((pId.name == hashString("INIT_CP")) && (pId.section == SectionIndep) && (pId.boundState == BoundStateIndep) && (pId.particleType == ParTypeIndep) && (pId.component != CompIndep) && (pId.reaction != ReactionIndep))
+			{
[…]

[…]
