### 2024-12-04 b002d84
Bump version to 5.0.2
