### 2024-12-04 84d0837
Move release checklist to issue template
Release checklist is still inlcuded in the dev guide documentation
