### 2024-12-04 52cb55b
Update release guide
