+++ doc/modelling/unit_operations/2d_lumped_rate_model_with_pores.rst
@@ -1,0 +1,92 @@
+.. 2d_lumped_rate_model_with_pores_model:
+
+Two Dimensional Lumped rate model with pores (LRMP2D)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The general rate model as introduced in Section :ref:`general_rate_model_model` assumes homogeneity in the cross sections of the column.
+This allows to consider transport along the axial dimension only.
+However, due to packing irregularity and inhomogeneous flow at the inlet (i.e., frits), this assumption may be a crude approximation.
+This model can be improved by introducing a radial coordinate :math:`\rho \in [0, R]`, where :math:`R` is the column radius, in the interstitial volume Eq. :eq:`ModelColumn`:
+
+.. math::
+   :label: ModelColumn2D
+
+   	\varepsilon_c \frac{\partial c^\ell_i}{\partial t} = &-\varepsilon_c u \frac{\partial c^\ell_i}{\partial z} + \varepsilon_c D_{\text{ax},i} \frac{\partial^2 c^\ell_i}{\partial z^2} + \frac{1}{\rho} \frac{\partial}{\partial \rho} \left( \rho D_{\text{rad},i} \frac{\partial}{\partial \rho} \left( \varepsilon_c c^\ell_i \right) \right) \\ 
+    &- \left(1 - \varepsilon_c\right) \sum_j d_j \frac{ 3 k_{f,j,i} }{r_{p,j}} \left[ c^\ell_i - c^p_{j,i}(\cdot, \cdot, \cdot, r_{p,j}) \right] + \varepsilon_c f_{\text{react},i}^\ell\left(c^\ell\right). 
+
+Here, 
+
+  - :math:`c^\ell_i\colon \left[0, T_{\text{end}}\right] \times [0, L] \times [0, R] \rightarrow \mathbb{R}^{\geq 0}`,
+  - :math:`c^p_{j,i}\colon \left[0, T_{\text{end}}\right] \times [0, L] \times [0, R] \times [r_{c,j}, r_{p,j}] \rightarrow \mathbb{R}^{\geq 0}`, and 
+  - :math:`c^s_{j,i,m_{j,i}}\colon \left[0, T_{\text{end}}\right] \times [0, L] \times [0, R] \times [r_{c,j}, r_{p,j}] \rightarrow \mathbb{R}^{\geq 0}` 
+  
+depend on :math:`\rho`.
+Additionally, the porosity :math:`\varepsilon_c`, axial dispersion coefficient :math:`D_{\text{ax},i}`, radial dispersion coefficient :math:`D_{\text{rad},i}`, and interstitial velocity :math:`u` may depend on :math:`\rho`.
+
+The dependence of the parameters on :math:`\rho` is not arbitrary.
+For simplicity, it is assumed that the parameters are piecewise constant, that is, the range :math:`[0, R]` is divided into disjoint zones in which all parameters are constant.
+These zones are used for radial discretization and can be supplied to the simulator.
+Continuous dependence of the parameters can be realized by piecewise constant approximation.
+
+The Danckwerts boundary conditions at the column in- and outlet, Eq. :eq:`BCInlet` and :eq:`BCOutlet`, are modified to account for the radial coordinate:
+
+.. math::
+    :label: BCInlet2D 
+
+    \begin{aligned}
[…]
