### 2024-12-04 bf687ff
Read parameter input per radial element instead of node
