+++ src/libcadet/model/parts/TwoDimensionalConvectionDispersionOperatorDG.cpp
@@ -30,10 +30,10 @@ using namespace Eigen;
 namespace
 {
 
-cadet::model::MultiplexMode readAndRegisterMultiplexParam(cadet::IParameterProvider& paramProvider, std::unordered_map<cadet::ParameterId, cadet::active*>& parameters, std::vector<cadet::active>& values, const std::string& name, unsigned int nComp, unsigned int radNPoints, cadet::UnitOpIdx uoi)
+cadet::model::MultiplexMode readAndRegisterMultiplexParam(cadet::IParameterProvider& paramProvider, std::unordered_map<cadet::ParameterId, cadet::active*>& parameters, std::vector<cadet::active>& values, const std::string& name, unsigned int nComp, unsigned int radNElem, cadet::UnitOpIdx uoi)
 {
 	cadet::model::MultiplexMode mode = cadet::model::MultiplexMode::Independent;
-	readParameterMatrix(values, paramProvider, name, nComp * radNPoints, 1);
+	readParameterMatrix(values, paramProvider, name, nComp * radNElem, 1);
 	unsigned int nSec = 1;
 	if (paramProvider.exists(name + "_MULTIPLEX"))
 	{
@@ -47,8 +47,8 @@ cadet::model::MultiplexMode readAndRegisterMultiplexParam(cadet::IParameterProvi
 		else if (modeConfig == 1)
 		{
 			mode = cadet::model::MultiplexMode::Radial;
-			if (values.size() != radNPoints)
-				throw cadet::InvalidParameterException("Number of elements in field " + name + " inconsistent with " + name + "_MULTIPLEX (should be " + std::to_string(radNPoints) + ")");
+			if (values.size() != radNElem)
+				throw cadet::InvalidParameterException("Number of elements in field " + name + " inconsistent with " + name + "_MULTIPLEX (should be " + std::to_string(radNElem) + ")");
 		}
 		else if (modeConfig == 2)
 		{
@@ -59,8 +59,8 @@ cadet::model::MultiplexMode readAndRegisterMultiplexParam(cadet::IParameterProvi
 		else if (modeConfig == 3)
 		{
 			mode = cadet::model::MultiplexMode::ComponentRadial;
-			if (values.size() != nComp * radNPoints)
-				throw cadet::InvalidParameterException("Number of elements in field " + name + " inconsistent with " + name + "_MULTIPLEX (should be " + std::to_string(nComp * radNPoints) + ")");
+			if (values.size() != nComp * radNElem)
+				throw cadet::InvalidParameterException("Number of elements in field " + name + " inconsistent with " + name + "_MULTIPLEX (should be " + std::to_string(nComp * radNElem) + ")");
 		}
 		else if (modeConfig == 4)
 		{
@@ -70,10 +70,10 @@ cadet::model::MultiplexMode readAndRegisterMultiplexParam(cadet::IParameterProvi
 		else if (modeConfig == 5)
 		{
 			mode = cadet::model::MultiplexMode::RadialSection;
-			if (values.size() % radNPoints != 0)
-				throw cadet::InvalidParameterException("Number of elements in field " + name + " is not a positive multiple of radNPoints (" + std::to_string(radNPoints) + ")");
+			if (values.size() % radNElem != 0)
+				throw cadet::InvalidParameterException("Number of elements in field " + name + " is not a positive multiple of radNElem (" + std::to_string(radNElem) + ")");
 
-			nSec = values.size() / radNPoints;
+			nSec = values.size() / radNElem;
 		}
 		else if (modeConfig == 6)
 		{
[…]
