+++ src/libcadet/model/LumpedRateModelWithPoresDG2D.cpp
@@ -449,6 +449,7 @@ bool LumpedRateModelWithPoresDG2D::configureModelDiscretization(IParameterProvid
 
 	_disc.axNPoints = _convDispOp.axNPoints();
 	_disc.radNPoints = _convDispOp.radNPoints();
+	_disc.radNElem = _convDispOp.radNElem();
 	_disc.nBulkPoints = _disc.axNPoints * _disc.radNPoints;
 
 	// Precompute offsets of particle type DOFs
@@ -626,11 +627,11 @@ bool LumpedRateModelWithPoresDG2D::configure(IParameterProvider& paramProvider)
 
 	// Let PAR_TYPE_VOLFRAC default to 1.0 for backwards compatibility
 	if (paramProvider.exists("PAR_TYPE_VOLFRAC"))
-		_parTypeVolFracMode = readAndRegisterMultiplexParam(paramProvider, _parameters, _parTypeVolFrac, "PAR_TYPE_VOLFRAC", _disc.axNPoints, _disc.radNPoints, _disc.nParType, _unitOpIdx);
+		_parTypeVolFracMode = readAndRegisterMultiplexParam(paramProvider, _parameters, _parTypeVolFrac, "PAR_TYPE_VOLFRAC", _disc.axNPoints, _disc.radNElem, _disc.nParType, _unitOpIdx);
 	else
 	{
 		// Only one particle type present
-		_parTypeVolFrac.resize(_disc.axNPoints * _disc.radNPoints, 1.0);
+		_parTypeVolFrac.resize(_disc.axNPoints * _disc.radNElem, 1.0);
 		_parTypeVolFracMode = MultiplexMode::Independent;
 	}
 
@@ -644,12 +645,12 @@ bool LumpedRateModelWithPoresDG2D::configure(IParameterProvider& paramProvider)
 		throw InvalidParameterException("Number of elements in field PAR_POROSITY does not match number of particle types");
 
 	// Check that particle volume fractions sum to 1.0
-	for (unsigned int i = 0; i < _disc.axNPoints * _disc.radNPoints; ++i)
+	for (unsigned int i = 0; i < _disc.axNPoints * _disc.radNElem; ++i)
 	{
 		const double volFracSum = std::accumulate(_parTypeVolFrac.begin() + i * _disc.nParType, _parTypeVolFrac.begin() + (i+1) * _disc.nParType, 0.0,
 			[](double a, const active& b) -> double { return a + static_cast<double>(b); });
 		if (std::abs(1.0 - volFracSum) > 1e-10)
-			throw InvalidParameterException("Sum of field PAR_TYPE_VOLFRAC differs from 1.0 (is " + std::to_string(volFracSum) + ") in axial cell " + std::to_string(i / _disc.radNPoints) + " radial cell " + std::to_string(i % _disc.radNPoints));
+			throw InvalidParameterException("Sum of field PAR_TYPE_VOLFRAC differs from 1.0 (is " + std::to_string(volFracSum) + ") in axial element " + std::to_string(i / _disc.radNElem) + " radial element " + std::to_string(i % _disc.radNElem));
 	}
 
 	// Read vectorial parameters (which may also be section dependent; transport)
@@ -671,7 +672,7 @@ bool LumpedRateModelWithPoresDG2D::configure(IParameterProvider& paramProvider)
 
 	// Register initial conditions parameters
[…]
