+++ doc/interface/unit_operations/index.rst
@@ -11,8 +11,9 @@ Unit Operations
     general_rate_model 
     lumped_rate_model_with_pores
     lumped_rate_model_without_pores
-    2d_general_rate_model
     cstr
+    2d_general_rate_model
+    2d_lumped_rate_model_with_pores
     radial_flow_models
     multi_channel_transport_model
     ../population_balance_model
