### 2024-12-07 1b4ae4f
Set `isParticleLumped` to true in LRMP
