### 2024-12-07 7ce9cd5
Check singleton dimensions before returning coordinates
jo
