+++ src/libcadet/api/CAPIv1.cpp
@@ -917,10 +917,19 @@ namespace v1
 			return cdtDataNotStored;
 		}
 
+		bool keepAxialSingletonDimension = unitRec->keepBulkSingletonDim();
+
 		if (nCoords)
 			*nCoords = unitRec->numAxialCells();
+
+		if (nCoords && *nCoords == 1 && !keepAxialSingletonDimension)
+		{
+			return cdtDataNotStored;
+		}
+
 		if (data)
 			*data = unitRec->primaryCoordinates();
+
 		return cdtOK;
 	}
 
@@ -941,6 +950,7 @@ namespace v1
 			*nCoords = unitRec->numRadialCells();
 		if (data)
 			*data = unitRec->secondaryCoordinates();
+
 		return cdtOK;
 	}
 
@@ -957,14 +967,23 @@ namespace v1
 			return cdtDataNotStored;
 		}
 
+		bool keepParticleSingletonDimension = unitRec->keepParticleSingletonDim();
+
+		if (nCoords)
+			*nCoords = unitRec->numParticleShells(parType);
+
+		if (nCoords && *nCoords == 1 && !keepParticleSingletonDimension)
+		{
+			return cdtDataNotStored;
+		}
+
 		int offset = 0;
 		for (int i = 0; i < parType; ++i)
 			offset += unitRec->numParticleShells(i);
 
-		if (nCoords)
-			*nCoords = unitRec->numParticleShells(parType);
 		if (data)
 			*data = unitRec->particleCoordinates() + offset;
+
 		return cdtOK;
 	}
 
@@ -985,6 +1004,7 @@ namespace v1
 			*nTime = sysRec->numDataPoints();
 		if (time)
 			*time = sysRec->time();
+
 		return cdtOK;
 	}
 
