### 2024-12-09 a833ff7
Fix old CSTR interface support
This commit fixes the calculation of constant solid volume from the old
interface fields, which are supported for backwards compatibility.
