### 2024-12-09 e21fb9a
add moity matrix to cstr
