### 2024-12-09 a16e8f8
Change 2D DG parameter sizes to radNElem
