### 2024-12-09 9e5fbb7
Add radial dependence of parameters to 2D DG discretization
* Introduce element-wise radial zone dependence of parameters: velocity,
dispersion, porosity, particle type volume fractions, etc

* Move radial dispersion into numerical flux

* Component dependent radial/axial dispersion

* Fix computation of velocity from flow rates and add consistency check
for element wise constant velocities

* Add numerical reference test
