+++ test/LumpedRateModelWithPoresDG2D.cpp
@@ -18,7 +18,7 @@
 #include "Utils.hpp"
 #include "JsonTestModels.hpp"
 
-void test2DLRMPJacobian(const std::string relModelFilePath, const int maxAxElem, const int maxRadElem, const int axPolyDeg = 0, const int radPolyDeg = 0)
+void test2DLRMPJacobian(const std::string relModelFilePath, const int maxAxElem, const int maxRadElem, const int axPolyDeg = 0, const int radPolyDeg = 0, const int minAxElem = 1, const int minRadElem = 1)
 {
 	cadet::JsonParameterProvider jpp = cadet::test::column::getReferenceFile(relModelFilePath);
 
@@ -60,16 +60,16 @@ void test2DLRMPJacobian(const std::string relModelFilePath, const int maxAxElem,
 
 	// This test might run out of memory due to the required AD directions:
 	// (axPolyDeg + 1) * axNElem * (radPolyDeg + 1) * radNElem * (nComp + nParType * (nComp + nBound))
-	for (int zElem = 1; zElem <= maxAxElem; zElem++)
+	for (int zElem = minAxElem; zElem <= maxAxElem; zElem++)
 	{
-		for (int rElem = 1; rElem <= maxRadElem; rElem++)
+		for (int rElem = minRadElem; rElem <= maxRadElem; rElem++)
 		{
 			jpp.pushScope("discretization");
 			jpp.set("AX_NELEM", zElem);
 			jpp.set("RAD_NELEM", rElem);
 			jpp.popScope();
 
-			cadet::test::column::testJacobianAD(jpp, 1e10, std::numeric_limits<float>::epsilon() * 100.0, &flowRate[0]); // @todo figure out why FD Jacobian pattern comparison doesnt work but AD Jacobian comparison does
+			cadet::test::column::testJacobianAD(jpp, 1e10, std::numeric_limits<float>::epsilon() * 100.0, &flowRate[0]);
 		}
 	}
 }
@@ -94,7 +94,7 @@ TEST_CASE("LRMP2D transport Jacobian", "[LRMP2D],[DG],[DG2D],[UnitOp],[Jacobian]
 	test2DLRMPJacobian(relModelFilePath, 4, 4, 1, 1);
 }
 
-TEST_CASE("LRMP2D transport Jacobian, full test", "[LRMP2D],[DG],[DG2D],[UnitOp],[Jacobian],[ReleaseCI]")
+TEST_CASE("LRMP2D transport Jacobian with radially constant parameters", "[LRMP2D],[DG],[DG2D],[UnitOp],[Jacobian],[ReleaseCI]")
 {
 	const std::string relModelFilePath = std::string("/data/model_LRMP2D_bulkTransport_1comp.json");
 
@@ -102,6 +102,26 @@ TEST_CASE("LRMP2D transport Jacobian, full test", "[LRMP2D],[DG],[DG2D],[UnitOp]
 	// inletDof + (axPolyDeg + 1) * axNElem * (radPolyDeg + 1) * radNElem * (nComp + nParType * (nComp + nBound))
 	// result here is 14radPoints + 588 pure dofs (21axPoints*14radPoints) * (1 + 1)
 	test2DLRMPJacobian(relModelFilePath, 7, 7, 2, 1);
+}
+
+TEST_CASE("LRMP2D transport Jacobian with radially variable parameters", "[LRMP2D],[DG],[DG2D],[UnitOp],[Jacobian],[CI]")
+{
+	const std::string relModelFilePath = std::string("/data/model_LRMP2D_bulkTransportRadVar_1comp.json");
+
+	// Required AD directions:
[…]
