### 2024-12-09 60064e1
Apply velocity per radial zone
