### 2024-12-11 e018e97
Update installation.rst in the documentation
