### 2024-12-11 3b6d6f0
Add exception for non-parallelized code when NTHREADS > 1
