### 2024-12-11 43e9f38
Add exception for non-parallelized code when NTHREADS > 1
