### 2024-12-11 0cfeefc
Update issue templates
