### 2024-12-12 6c112e8
Update release checklist
