+++ .github/ISSUE_TEMPLATE/release_checklist.md
@@ -10,7 +10,7 @@ CADET-Core release
 CADET-Core releases follow the *semantic versioning system*, which is documented [here](https://semver.org/).
 
 **CADET-Core is released following the so-called *GitLab flow*:**
-In GitLab flow, feature and hotfix branches contain work for new features and bug fixes which will be merged back into the master branch immediately when they’re finished, reviewed, and approved. The master branch is ready to be deployed, but not necessarily the source of truth for a new release: Each release has an associated release branch that is based off the master branch. Changes that are intended to be released are cherry-picked from master into the release branch. The release is made on the release branch. If desired, some release branches can be kept and maintained, which is currently not planned for the CADET-Core project.
+In GitLab flow, feature and hotfix branches contain work for new features and bug fixes which will be merged back into the master branch immediately when they’re finished, reviewed, and approved. The master branch is ready to be deployed, but not necessarily the source of truth for a new release: Each release has an associated release branch that is based off the master branch. Changes that are intended to be released are cherry-picked from master onto the release branch. The release is made on the release branch. Some release branches can be kept and maintained, e.g. branch `v5.0.X` to support the last major release.
 
 
 **To make a release, the tasks in the following checklist have to be executed sequentially.**
@@ -19,34 +19,27 @@ In GitLab flow, feature and hotfix branches contain work for new features and bu
 Release on github
 -----------------
 
-- [ ] Run the release tests:
-
-  - [ ] The release tests contain extensive testing that is not included in our CI, such as EOC tests.
-    Running these tests might take a while and this should be done on the server.
-  - [ ] Some tests are implemented in CADET-Core with a `ReleaseCI` flag. These tests need to pass and can be run by executing the `cd.yml` workflow.
-  - [ ] More tests are implemented in Python, the code can be found in [CADET-Verification](https://github.com/cadet/CADET-Verification)
-  Compare the results with the previous run.
-  The release process can only be continued if the results are reasonable.
-
-- [ ] Run performance benchmarks:
-
-  - [ ] If numerical algorithms were refactored or if performance-critical infrastructure was changed, you should run performance benchmarks to compare the latest release with the planned new one.
-    To this end, you can refer to the performance benchmark templates in CADET-Reference, e.g. the [benchmark for the modified Newton method](https://jugit.fz-juelich.de/IBG-1/ModSim/cadet/cadet-reference/-/tree/benchmark_modified_newton?ref_type=heads)
-
 - [ ] Open a PR and create a version bump commit `Bump version to vX.X.X`:
   
-  - [ ] Update of the version number in the `version.txt`, `zenodo.json`, `cadet.hpp` and cadet.doxyfile, compare to last `bump version` commit
-  - [ ] Update of the authors list if needed: CONTRIBUTING.md and zenodo.json
-  - [ ] Update of the copyright (years)
-  - [ ] Update of the file format if needed
+  - [ ] Update of the version number in the `version.txt`, `zenodo.json` (two places), `cadet.hpp` and `cadet.doxyfile`, compare to last `bump version` commit
+  - [ ] Update of the authors list if needed: `CONTRIBUTING.md` and `zenodo.json`
+  - [ ] Update of the file format if needed, in the `driver.hpp`
 
-- [ ] Merge the PR into master.
+- [ ] Name the commit *Bump version to vX.X.X* and merge the PR into `master`.
 
-- [ ] Create a Release branch off of master and cherry-pick the desired commits onto it.
+- [ ] Create a Release branch `vX.X.X` off of `master` and cherry-pick the desired commits onto it.
+
[…]
