### 2024-12-12 c3fa773
Bump version to v5.0.3
