+++ include/common/Driver.hpp
@@ -664,7 +664,7 @@ public:
 		writer.scalar("TIME_SIM", _sim->lastSimulationDuration());
 
 		if (!writer.exists("FILE_FORMAT"))
-			writer.scalar("FILE_FORMAT", 40000);
+			writer.scalar("FILE_FORMAT", 50000);
 
 		writer.popGroup();
 	}
