### 2024-12-12 0700a3c
Update conda reference (#341)
* Update conda link in README.rst

* Remove link to anaconda and mention of mamba

* Update installation.rst
