### 2024-12-12 98aa72d
Update conda reference (#341)
* Update conda link in README.rst

* Remove link to anaconda and mention of mamba

* Update installation.rst
