+++ README.rst
@@ -29,11 +29,7 @@ CADET-Core can be installed via conda from the ``conda-forge`` channel.
 
 ``conda install -c conda-forge cadet``
 
-This requires a working `conda installation <https://docs.anaconda.com/anaconda/install/index.html>`_.
-
-Optionally, use `mamba <https://github.com/mamba-org/mamba>`_ which uses a faster dependency solver than ``conda``.
-
-``mamba install -c conda-forge cadet``
+This requires a working `conda installation <https://github.com/conda-forge/miniforge>`_.
 
 `Additional information <https://cadet.github.io/master/getting_started/installation>`_ and a `tutorial <https://cadet.github.io/master/getting_started/tutorials/breakthrough>`_ are available to guide you through the installation and the first steps of using CADET.
 
