+++ doc/literature.bib
@@ -507,3 +507,15 @@ abstract = {The presence of a liquid stationary phase in liquid–liquid chromat
   publisher = {Elsevier},
   doi = {https://doi.org/10.1016/j.compchemeng.2024.108612}
 }
+@article{sips1948,
+	title = {On the {Structure} of a {Catalyst} {Surface}},
+	author = {Sips, Robert},
+	year = {1948},
+	journal = {The Journal of Chemical Physics},
+	pages = {490--495},
+	volume = {16},
+	issn = {0021-9606},
+	doi = {https://doi.org/10.1063/1.1746922},
+	number = {5},
+	month = may
+}
