### 2024-12-13 bb77858
Fix manual trigger of CI and CD workflows
