### 2024-12-13 9ade5b7
Fix manual trigger of CI and CD workflows (#349)
