### 2024-12-15 3ea837f
Update .github/ISSUE_TEMPLATE/release_checklist.md
Co-authored-by: Johannes Schmölder <20299934+schmoelder@users.noreply.github.com>
