### 2024-12-15 268f280
Update release_checklist.md
