### 2024-12-15 a309f1f
Remove radial quadrature code
