+++ src/libcadet/model/parts/TwoDimensionalConvectionDispersionOperatorDG.cpp
@@ -471,72 +471,6 @@ TwoDimensionalConvectionDispersionOperatorDG::~TwoDimensionalConvectionDispersio
 	delete[] _jacRadDispersion;
 }
 
-MatrixXd TwoDimensionalConvectionDispersionOperatorDG::calcTildeMr(const unsigned int elemIdx, const active* const dispersion)
-{
-	MatrixXd tildeMr = MatrixXd::Zero(_radNNodes, _qNNodes);
-	MatrixXd ell = MatrixXd::Zero(_radNNodes, _qNNodes);
-
-	for (unsigned int j = 0; j < _radNNodes; j++)
-		ell.row(j) = dgtoolbox::evalLagrangeBasis(j, _radNodes, _qNodes);
-
-	for (unsigned int j = 0; j < _radNNodes; j++)
-	{
-		for (unsigned int k = 0; k < _qNNodes; k++)
-		{
-			tildeMr(j, k) = static_cast<double>(dgtoolbox::mapRefToPhys<active>(_radDelta, elemIdx, _qNodes[k])) * static_cast<double>(dispersion[elemIdx * _qNNodes + k]) * ell(j, k) * _qWeights[k];
-		}
-	}
-
-	return tildeMr;
-}
-
-MatrixXd TwoDimensionalConvectionDispersionOperatorDG::calcTildeMrDash(const unsigned int elemIdx, const active * const dispersion)
-{
-	MatrixXd ellEll = MatrixXd::Zero(_radNNodes, _qNNodes);
-	MatrixXd tildeMrDash = MatrixXd::Zero(_radNNodes, _radNNodes * _qNNodes);
-
-	for (unsigned int j = 0; j < _radNNodes; j++)
-		ellEll.row(j) = dgtoolbox::evalLagrangeBasis(j, _radNodes, _qNodes);
-
-	for (unsigned int j = 0; j < _radNNodes; j++)
-	{
-		for (unsigned int m = 0; m < _radNNodes; m++)
-		{
-			for (unsigned int k = 0; k < _qNNodes; k++)
-			{
-				tildeMrDash(j, m * _qNNodes + k) = static_cast<double>(dgtoolbox::mapRefToPhys<active>(_radDelta, elemIdx, _qNodes[k])) * static_cast<double>(dispersion[elemIdx * _qNNodes + k]) * ellEll(m, k) * ellEll(j, k) * _qWeights[k];
-			}
-		}
-	}
-
-	return tildeMrDash;
-}
-
-MatrixXd TwoDimensionalConvectionDispersionOperatorDG::calcTildeSrDash(const unsigned int elemIdx, const active* const dispersion)
-{
-	MatrixXd ellEll = MatrixXd::Zero(_radNNodes, _qNNodes);
-	MatrixXd tildeSrDash = MatrixXd::Zero(_radNNodes, _radNNodes * _qNNodes);
-
-	for (unsigned int j = 0; j < _radNNodes; j++)
-		ellEll.row(j) = dgtoolbox::evalLagrangeBasis(j, _radNodes, _qNodes);
-
-	MatrixXd tildeDr = dgtoolbox::derivativeMatrix(_quadratureOrder, _qNodes);
-	MatrixXd tildeSr = _qWeights.asDiagonal() * tildeDr;
-
-	for (unsigned int j = 0; j < _radNNodes; j++)
-	{
-		for (unsigned int m = 0; m < _radNNodes; m++)
-		{
-			for (unsigned int k = 0; k < _qNNodes; k++)
-			{
[…]
