### 2024-12-15 2b2ab8d
Move dispersion and mapping to main equation
