### 2024-12-15 9f6d324
Reset number of AD directions in reference benchmark test
