+++ test/LumpedRateModelWithPoresDG2D.cpp
@@ -118,7 +118,7 @@ TEST_CASE("LRMP2D sensitivity Jacobians", "[LRMP2D],[UnitOp],[Sensitivity],[CI]"
 {
 	cadet::JsonParameterProvider jpp = createColumnWithTwoCompLinearBinding("LUMPED_RATE_MODEL_WITH_PORES_2D", "DG");
 
-	cadet::test::column::testFwdSensJacobians(jpp, 1e-4, 6e-7);
+	cadet::test::column::testFwdSensJacobians(jpp, 1e-4, 1e-6);
 }
 
 TEST_CASE("LRMP2D consistent initialization with linear binding", "[LRMP2D],[ConsistentInit],[CI]")
