+++ src/libcadet/model/parts/TwoDimensionalConvectionDispersionOperatorDG.cpp
@@ -995,9 +995,8 @@ int TwoDimensionalConvectionDispersionOperatorDG::residualImpl(const IModel& mod
 					else
 					{
 						Eigen::Map<const Vector<StateType, Dynamic>, 0, InnerStride<Dynamic>> GRnextNode(reinterpret_cast<StateType*>(&_radAuxStateG[0]) + auxElemOffset + auxRadElemStride, _axNNodes, InnerStride<Dynamic>(auxAxNodeStride));
-						_gStarDispR.col(1) = 0.5 * static_cast<ParamType>(curRadialDispersion[rEidx * _nComp + comp]) * (_Gr.col(_radPolyDeg) + GRnextNode);
+						_gStarDispR.col(1) = 0.5 * (_Gr.col(_radPolyDeg) + GRnextNode) * 0.5 * (static_cast<ParamType>(curRadialDispersion[rEidx * _nComp + comp]) + static_cast<ParamType>(curRadialDispersion[(rEidx + 1) * _nComp + comp])) * 0.5 * (static_cast<ParamType>(_colPorosities[rEidx]) + static_cast<ParamType>(_colPorosities[rEidx + 1]));
 					}
-
 				}
 
 				// axial dispersion (central) flux and axial convection (upwind) flux
@@ -1046,9 +1045,11 @@ int TwoDimensionalConvectionDispersionOperatorDG::residualImpl(const IModel& mod
 							)
 						);
 
+				const ParamType porosityFluxFactor = 1.0 / static_cast<ParamType>(_colPorosities[rEidx]);
+
 				// Radial dispersion
 				_Res -= 2.0 / static_cast<ParamType>(_radDelta[rEidx]) * 2.0 / static_cast<ParamType>(_radDelta[rEidx]) * (
-					_gStarDispR * _radLiftMCyl[rEidx].template cast<ResidualType>()
+					_gStarDispR * porosityFluxFactor * _radLiftMCyl[rEidx].template cast<ResidualType>()
 					- static_cast<ParamType>(curRadialDispersion[rEidx * _nComp + comp]) * _Gr.template cast<ResidualType>() * _SrCyl[rEidx].template cast<ResidualType>()
 					) * _invTransMrCyl[rEidx].template cast<ResidualType>();
 
@@ -1265,7 +1266,7 @@ bool TwoDimensionalConvectionDispersionOperatorDG::computeConvDispJacobianBlocks
 
 		/* axial dispersion block */
 		const active* const curAxialDispersion = getSectionDependentSlice(_axialDispersion, _radNElem * _nComp, 0);
-		const double axDisp = static_cast<double>(curAxialDispersion[0]);
+		const double axDisp = static_cast<double>(curAxialDispersion[rElem * _nComp]);
 
 		MatrixXd gStarZ = MatrixXd::Zero(Np, 5 * Np);
 
@@ -1309,7 +1310,8 @@ bool TwoDimensionalConvectionDispersionOperatorDG::computeConvDispJacobianBlocks
 
 		/* radial dispersion block */
 		const active* const curRadialDispersion = getSectionDependentSlice(_radialDispersion, _radNElem * _nComp, 0);
-		const double radDisp = static_cast<double>(curRadialDispersion[0]);
+		const double radDisp = static_cast<double>(curRadialDispersion[rElem * _nComp]);
+		const double curPorosity = static_cast<double>(_colPorosities[rElem]);
 
[…]
