+++ src/libcadet/model/parts/TwoDimensionalConvectionDispersionOperatorDG.cpp
@@ -1266,46 +1266,48 @@ bool TwoDimensionalConvectionDispersionOperatorDG::computeConvDispJacobianBlocks
 
 		/* axial dispersion block */
 		const active* const curAxialDispersion = getSectionDependentSlice(_axialDispersion, _radNElem * _nComp, 0);
-		const double axDisp = static_cast<double>(curAxialDispersion[rElem * _nComp]);
 
 		MatrixXd gStarZ = MatrixXd::Zero(Np, 5 * Np);
 
 		for (int i = 0; i < uAxElem; i++)
 		{
-			// first, we need to create numerical flux block gStarZ
-			// note: we have three unique auxiliary block indices and need to find the indices wrt the currently considered element
-			int auxIdx;
-			if (i == 0)
-				auxIdx = 0; // note that if _axNElem == 0, the corresponding auxiliary block is also stored at index 0
-			else
-				auxIdx = i == uAxElem - 1 ? 2 : 1;
-
-			if (i > 0) // left boundary condition -> zero
+			for (int comp = 0; comp < _nComp; comp++)
 			{
-				const int leftAuxIdx = (i == 1) ? 0 : 1; // if the left neighbour is the left boundary element, set index to 0, else to 1
-				gStarZ.block(0, 0, _radNNodes, 3 * Np) += GzDer[leftAuxIdx].block(Np - _radNNodes, 0, _radNNodes, 3 * Np);
-				gStarZ.block(0, Np, _radNNodes, 3 * Np) += GzDer[auxIdx].block(0, 0, _radNNodes, 3 * Np);
+				// first, we need to create numerical flux block gStarZ
+				// note: we have three unique auxiliary block indices and need to find the indices wrt the currently considered element
+				int auxIdx;
+				if (i == 0)
+					auxIdx = 0; // note that if _axNElem == 0, the corresponding auxiliary block is also stored at index 0
+				else
+					auxIdx = i == uAxElem - 1 ? 2 : 1;
+
+				if (i > 0) // left boundary condition -> zero
+				{
+					const int leftAuxIdx = (i == 1) ? 0 : 1; // if the left neighbour is the left boundary element, set index to 0, else to 1
+					gStarZ.block(0, 0, _radNNodes, 3 * Np) += GzDer[leftAuxIdx].block(Np - _radNNodes, 0, _radNNodes, 3 * Np);
+					gStarZ.block(0, Np, _radNNodes, 3 * Np) += GzDer[auxIdx].block(0, 0, _radNNodes, 3 * Np);
+				}
+
+				if (i != uAxElem - 1) // right boundary condition -> zero
+				{
+					// if the right neighbour is the right boundary element, set index to 2, else to 1
+					const int rightAuxIdx = (i + 1 == uAxElem - 1) ? 2 : 1;
+					gStarZ.block(Np - _radNNodes, Np, _radNNodes, 3 * Np) += GzDer[auxIdx].block(Np - _radNNodes, 0, _radNNodes, 3 * Np);
+					gStarZ.block(Np - _radNNodes, 2 * Np, _radNNodes, 3 * Np) += GzDer[rightAuxIdx].block(0, 0, _radNNodes, 3 * Np);
+				}
+
+				gStarZ *= 0.5;
+
+				_jacAxDispersion[rElem * uAxElem + i] = MzKronMrCylInv * 2.0 / deltaZ * (BzKronMrCyl * gStarZ);
+				_jacAxDispersion[rElem * uAxElem + i].block(0, Np, Np, 3 * Np) -= MzKronMrCylInv * 2.0 / deltaZ * (SzTKronMrCyl * GzDer[auxIdx]);
+
+				// filter numerical noise
[…]
