### 2024-12-16 9d27780
Component dependent radial dispersion
