+++ src/libcadet/model/parts/TwoDimensionalConvectionDispersionOperatorDG.cpp
@@ -532,7 +532,7 @@ void TwoDimensionalConvectionDispersionOperatorDG::initializeDG()
 		// Jacobian blocks
 		_jacConvection = new MatrixXd[_radNElem];
 		_jacAxDispersion = new MatrixXd[_radNElem * uAxElem];
-		_jacRadDispersion = new MatrixXd[_radNElem];
+		_jacRadDispersion = new MatrixXd[_nComp * _radNElem];
 	}
 
 	for (int rElem = 0; rElem < _radNElem; rElem++)
@@ -540,7 +540,9 @@ void TwoDimensionalConvectionDispersionOperatorDG::initializeDG()
 		_jacConvection[rElem] = MatrixXd::Zero(_elemNPoints, 5 * _elemNPoints);
 		for (int i = 0; i < uAxElem; i++)
 			_jacAxDispersion[rElem * uAxElem + i] = MatrixXd::Zero(_elemNPoints, 5 * _elemNPoints);
-		_jacRadDispersion[rElem] = MatrixXd::Zero(_elemNPoints, 5 * _elemNPoints);
+
+		for (int comp = 0; comp < _nComp; comp++)
+			_jacRadDispersion[rElem * _nComp + comp] = MatrixXd::Zero(_elemNPoints, 5 * _elemNPoints);
 	}
 }
 
@@ -1271,101 +1273,105 @@ bool TwoDimensionalConvectionDispersionOperatorDG::computeConvDispJacobianBlocks
 
 		for (int i = 0; i < uAxElem; i++)
 		{
-				// first, we need to create numerical flux block gStarZ
-				// note: we have three unique auxiliary block indices and need to find the indices wrt the currently considered element
-				int auxIdx;
-				if (i == 0)
-					auxIdx = 0; // note that if _axNElem == 0, the corresponding auxiliary block is also stored at index 0
-				else
-					auxIdx = i == uAxElem - 1 ? 2 : 1;
+			// first, we need to create numerical flux block gStarZ
+			// note: we have three unique auxiliary block indices and need to find the indices wrt the currently considered element
+			int auxIdx;
+			if (i == 0)
+				auxIdx = 0; // note that if _axNElem == 0, the corresponding auxiliary block is also stored at index 0
+			else
+				auxIdx = i == uAxElem - 1 ? 2 : 1;
 
-				if (i > 0) // left boundary condition -> zero
-				{
-					const int leftAuxIdx = (i == 1) ? 0 : 1; // if the left neighbour is the left boundary element, set index to 0, else to 1
-					gStarZ.block(0, 0, _radNNodes, 3 * Np) += GzDer[leftAuxIdx].block(Np - _radNNodes, 0, _radNNodes, 3 * Np);
-					gStarZ.block(0, Np, _radNNodes, 3 * Np) += GzDer[auxIdx].block(0, 0, _radNNodes, 3 * Np);
-				}
-
-				if (i != uAxElem - 1) // right boundary condition -> zero
-				{
-					// if the right neighbour is the right boundary element, set index to 2, else to 1
-					const int rightAuxIdx = (i + 1 == uAxElem - 1) ? 2 : 1;
-					gStarZ.block(Np - _radNNodes, Np, _radNNodes, 3 * Np) += GzDer[auxIdx].block(Np - _radNNodes, 0, _radNNodes, 3 * Np);
-					gStarZ.block(Np - _radNNodes, 2 * Np, _radNNodes, 3 * Np) += GzDer[rightAuxIdx].block(0, 0, _radNNodes, 3 * Np);
-				}
-
-				gStarZ *= 0.5;
-
[…]
