### 2024-12-16 1c62d6b
Change to aggregation code as starting point
