+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -1,15 +1,3 @@
-// =============================================================================
-//  CADET
-//  
-//  Copyright © 2008-present: 2008-present: The CADET-Core Authors
-//            Please see the AUTHORS.md file.
-//  
-//  All rights reserved. This program and the accompanying materials
-//  are made available under the terms of the GNU Public License v3.0 (or, at
-//  your option, any later version) which accompanies this distribution, and
-//  is available at http://www.gnu.org/licenses/gpl.html
-// =============================================================================
-
 #include "model/reaction/ReactionModelBase.hpp"
 #include "model/ModelUtils.hpp"
 #include "cadet/Exceptions.hpp"
@@ -28,1932 +16,445 @@
 
 namespace cadet
 {
-
-	namespace model
+namespace model
+{
+namespace detail
+{
+	struct AggCoefficients
 	{
-		/**
-         * @brief Flux reconstruction scheme coefficients. See @cite Zhang2024_PBM_partI
-         */
-		namespace detail
+		std::pair<std::vector<unsigned short int>, std::vector<unsigned short int>> index;  // a pair of vectors to store combo of j and k
+		std::vector<unsigned int> index_tracker;           // the idx can be a large number  
+		std::vector<short int> i_index;
+
+		active sum_volume_cube_root = 0.0;
+		unsigned int count = 0;
+
+		// the sizes of j_index and k_index are not known a priori
+		AggCoefficients(const std::vector<active>& binCenters, const std::vector<active>& bins, const std::vector<active>& binSizes)
+			: index(), index_tracker(binCenters.size() + 1), i_index(binCenters.size()* binCenters.size())
 		{
-			struct Weno3Coefficients
-			{
-				std::vector<active> q_1_right_coeff;
-				std::vector<active> q_0_right_coeff;
-				std::vector<active> C_right_coeff;
-				std::vector<active> IS_0_coeff;
-				std::vector<active> IS_1_coeff;
+			// make the first element of index_tracker 0
+			index_tracker[0] = 0;
 
-				Weno3Coefficients(const std::vector<active>& binSizes)
-					: q_1_right_coeff(binSizes.size() - 1), q_0_right_coeff(binSizes.size() - 1), C_right_coeff(binSizes.size() - 1), IS_0_coeff(binSizes.size() - 1), IS_1_coeff(binSizes.size() - 1)
+			// loop over daughter particle i
+			for (unsigned int i = 0; i < binCenters.size(); ++i)
+			{
+				// source term, find lij in set A
+				for (unsigned short int j = 0; j < i; ++j)
 				{
-					// calculate the coefficients and store them, the first element is empty
-					for (int i = 1; i + 1 < binSizes.size(); ++i)
+					for (unsigned short int k = j; k < i + 1; ++k)
 					{
-						const active delta_sum = binSizes[i] + binSizes[i - 1] + binSizes[i + 1];
-						const active delta_left_sum = binSizes[i] + binSizes[i - 1];
[…]
