### 2024-12-16 be95fdc
Fix the handling of short int
