### 2024-12-17 dd5ed37
Add fragmentation part
