+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -20,9 +20,46 @@ namespace model
 {
 namespace detail
 {
+	enum class CrystallizationMode : int
+	{
+		/**
+		 * Population mass balance equation
+		 */
+		PurePBM,
+
+		/**
+		 * Aggregation equation
+		 */
+		PureAggregation,
+
+		/**
+		 * Breakage equation
+		 */
+		PureBreakage,
+
+		/**
+		 * Combined aggregation and breakage equation
+		 */
+		AggregationBreakage
+	};
+
+	struct ModeFlags
+	{
+		bool hasMassBalance = false;
+		bool hasAggregation = false;
+		bool hasFragmentation = false;
+	};
+
+	std::unordered_map<CrystallizationMode, ModeFlags> modeFlagMap = {
+		{CrystallizationMode::PurePBM, {true, false, false}},
+		{CrystallizationMode::PureAggregation, {true, true, false}},
+		{CrystallizationMode::PureBreakage, {false, false, true}},
+		{CrystallizationMode::AggregationBreakage, {true, true, true}}
+	};
+
 	struct AggCoefficients
 	{
-		std::pair<std::vector<unsigned short int>, std::vector<unsigned short int>> index;  // a pair of vectors to store combo of j and k
+		std::pair<std::vector<unsigned short int>, std::vector<unsigned short int>> index;  // a pair of vectors to store combínation of j and k
 		std::vector<unsigned int> index_tracker;           // the idx can be a large number  
 		std::vector<short int> i_index;
 
@@ -74,16 +111,68 @@ namespace detail
 			}
 		}
 	};
+
+	struct FragCoefficients
+	{
+		// define breakage-related local parameters
+		std::vector<active> Upsilon_source;
+		std::vector<active> Upsilon_sink;
+
+		// constructor
+		FragCoefficients(const std::vector<active>& binCenters, const std::vector<active>& bins, const active& breakageKernelGamma)
+			: Upsilon_source(binCenters.size()), Upsilon_sink(binCenters.size())
+		{
+			const active N_j = breakageKernelGamma / (breakageKernelGamma - 1.0);
+
+			active Upsilon_birth_sum = 0.0;
+			active Upsilon_death_sum = 0.0;
+			active b_integral_birth = 0.0;
+
+			// calculate upsilon and store it
+			for (int i = 0; i < binCenters.size(); ++i)
+			{
+				// reset the sum for each i
+				Upsilon_birth_sum = 0.0;
+				Upsilon_death_sum = 0.0;
+				const active x_i_3 = binCenters[i] * binCenters[i] * binCenters[i];
+				for (int j = 0; j < i + 1; ++j)
+				{
+					const active x_j_3 = binCenters[j] * binCenters[j] * binCenters[j];
+					if (cadet_likely(i != j))
+					{
+						b_integral_birth = N_j * (pow(bins[j + 1], 3.0 * breakageKernelGamma - 3.0) - pow(bins[j], 3.0 * breakageKernelGamma - 3.0)) / pow(binCenters[i], 3.0 * breakageKernelGamma - 3.0);
+						Upsilon_birth_sum += b_integral_birth * (x_i_3 - x_j_3);
+						Upsilon_death_sum += b_integral_birth * x_j_3;
+					}
+					else
+					{
+						b_integral_birth = N_j * (pow(binCenters[j], 3.0 * breakageKernelGamma - 3.0) - pow(bins[j], 3.0 * breakageKernelGamma - 3.0)) / pow(binCenters[i], 3.0 * breakageKernelGamma - 3.0);
+						Upsilon_death_sum += b_integral_birth * x_j_3;
+					}
+				}
[…]
