### 2024-12-17 08b565c
Fix code format and add CrystallizationMode flags
