+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -33,14 +33,29 @@ namespace detail
 		PureAggregation,
 
 		/**
-		 * Breakage equation
+		 * Fragmentation equation
 		 */
-		PureBreakage,
+		PureFragmentation,
 
 		/**
-		 * Combined aggregation and breakage equation
+		 * Combined mass balance and aggregation equation
 		 */
-		AggregationBreakage
+		PBMAggregation,
+
+		/**
+		 * Combined mass balance and fragmentation equation
+		 */
+		PBMFragmentation,
+
+		/**
+		 * Combined aggregation and fragmentation equation
+		 */
+		AggregationFragmentation,
+
+		/**
+		 * Combined mass balance and aggregation and fragmentation equation
+		 */
+		PBMAggregationFragmentation
 	};
 
 	struct ModeFlags
@@ -52,26 +67,29 @@ namespace detail
 
 	std::unordered_map<CrystallizationMode, ModeFlags> modeFlagMap = {
 		{CrystallizationMode::PurePBM, {true, false, false}},
-		{CrystallizationMode::PureAggregation, {true, true, false}},
-		{CrystallizationMode::PureBreakage, {false, false, true}},
-		{CrystallizationMode::AggregationBreakage, {true, true, true}}
+		{CrystallizationMode::PureAggregation, {false, true, false}},
+		{CrystallizationMode::PureFragmentation, {false, false, true}},
+		{CrystallizationMode::PBMAggregation, {true, true, false}},
+		{CrystallizationMode::PBMFragmentation, {true, false, true}},
+		{CrystallizationMode::AggregationFragmentation, {false, true, true}},
+		{CrystallizationMode::PBMAggregationFragmentation, {true, true, true}}
 	};
 
 	struct AggCoefficients
 	{
 		std::pair<std::vector<unsigned short int>, std::vector<unsigned short int>> index;  // a pair of vectors to store combínation of j and k
-		std::vector<unsigned int> index_tracker;           // the idx can be a large number  
-		std::vector<short int> i_index;
+		std::vector<unsigned int> indexTracker;           // the idx can be a large number  
+		std::vector<short int> iIndex;
 
-		active sum_volume_cube_root = 0.0;
+		active sumVolumeCubeRoot = 0.0;
 		unsigned int count = 0;
 
 		// the sizes of j_index and k_index are not known a priori
 		AggCoefficients(const std::vector<active>& binCenters, const std::vector<active>& bins, const std::vector<active>& binSizes)
-			: index(), index_tracker(binCenters.size() + 1), i_index(binCenters.size()* binCenters.size())
+			: index(), indexTracker(binCenters.size() + 1), iIndex(binCenters.size()* binCenters.size())
 		{
-			// make the first element of index_tracker 0
-			index_tracker[0] = 0;
+			// make the first element of indexTracker 0
+			indexTracker[0] = 0;
 
 			// loop over daughter particle i
 			for (unsigned int i = 0; i < binCenters.size(); ++i)
@@ -81,8 +99,8 @@ namespace detail
 				{
 					for (int k = j; k < i + 1; ++k)
 					{
-						sum_volume_cube_root = pow(binCenters[k] * binCenters[k] * binCenters[k] + binCenters[j] * binCenters[j] * binCenters[j], 1.0 / 3.0);
[…]
