+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -1,3 +1,15 @@
+// =============================================================================
+//  CADET
+//  
+//  Copyright © 2008-present: The CADET-Core Authors
+//            Please see the AUTHORS.md file.
+//  
+//  All rights reserved. This program and the accompanying materials
+//  are made available under the terms of the GNU Public License v3.0 (or, at
+//  your option, any later version) which accompanies this distribution, and
+//  is available at http://www.gnu.org/licenses/gpl.html
+// =============================================================================
+
 #include "model/reaction/ReactionModelBase.hpp"
 #include "model/ModelUtils.hpp"
 #include "cadet/Exceptions.hpp"
@@ -20,41 +32,44 @@ namespace model
 {
 namespace detail
 {
+	/**
+	* Crystallization mode specification to define which terms are included
+	 */
 	enum class CrystallizationMode : int
 	{
 		/**
 		 * Population mass balance equation
-		 */
+		*/
 		PurePBM,
 
 		/**
 		 * Aggregation equation
-		 */
+		*/
 		PureAggregation,
 
 		/**
 		 * Fragmentation equation
-		 */
+		*/
 		PureFragmentation,
 
 		/**
 		 * Combined mass balance and aggregation equation
-		 */
+		*/
 		PBMAggregation,
 
 		/**
 		 * Combined mass balance and fragmentation equation
-		 */
+		*/
 		PBMFragmentation,
 
 		/**
 		 * Combined aggregation and fragmentation equation
-		 */
+		*/
 		AggregationFragmentation,
 
 		/**
 		 * Combined mass balance and aggregation and fragmentation equation
-		 */
+		*/
 		PBMAggregationFragmentation
 	};
 
@@ -75,6 +90,154 @@ namespace detail
 		{CrystallizationMode::PBMAggregationFragmentation, {true, true, true}}
 	};
 
+	/**
+	 * Flux reconstruction scheme coefficients.For details see @cite Zhang2024
+	*/
+	struct Weno3Coefficients
+	{
+		std::vector<active> q_1_right_coeff;
+		std::vector<active> q_0_right_coeff;
+		std::vector<active> C_right_coeff;
+		std::vector<active> IS_0_coeff;
+		std::vector<active> IS_1_coeff;
+
+		Weno3Coefficients(const std::vector<active>& binSizes)
+			: q_1_right_coeff(binSizes.size() - 1), q_0_right_coeff(binSizes.size() - 1), C_right_coeff(binSizes.size() - 1), IS_0_coeff(binSizes.size() - 1), IS_1_coeff(binSizes.size() - 1)
+		{
+			// calculate the coefficients and store them, the first element is empty
+			for (int i = 1; i + 1 < binSizes.size(); ++i)
+			{
+				const active delta_sum = binSizes[i] + binSizes[i - 1] + binSizes[i + 1];
+				const active delta_left_sum = binSizes[i] + binSizes[i - 1];
+				const active delta_right_sum = binSizes[i] + binSizes[i + 1];
[…]
