### 2024-12-18 cae8f9b
Update release checklist (#347)
