### 2024-12-18 0c8b37c
wip source out PBM kernels
