### 2024-12-18 84f01e9
Remove unnecessary StateParam type
