+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -137,6 +137,83 @@ namespace detail
 		}
 	};
 
+	template <typename ParamType>
+	struct HRParams
+	{
+		ParamType v_g;
+		const ParamType k_g_times_s_g;
+		const ParamType B_0;
+		ParamType r_x_i = 0.0;
+		ParamType phi = 0.0;
+		ParamType F_i = 0.0;
+
+		HRParams(ParamType v_g, ParamType k_g_times_s_g, ParamType B_0,
+			ParamType r_x_i = 0.0, ParamType phi = 0.0, ParamType F_i = 0.0)
+			: v_g(v_g), k_g_times_s_g(k_g_times_s_g), B_0(B_0),
+			r_x_i(r_x_i), phi(phi), F_i(F_i) {}
+	};
+
+	template <typename ParamType>
+	struct WENO3Params
+	{
+		ParamType v_g;
+		const ParamType k_g_times_s_g;
+		const ParamType B_0;
+		ParamType IS_0 = 0.0;
+		ParamType IS_1 = 0.0;
+		ParamType alpha_0 = 0.0;
+		ParamType alpha_1 = 0.0;
+		ParamType W_0 = 0.0;
+		ParamType W_1 = 0.0;
+		ParamType q_0 = 0.0;
+		ParamType q_1 = 0.0;
+
+		WENO3Params(ParamType v_g, ParamType k_g_times_s_g, ParamType B_0,
+			ParamType IS_0 = 0.0, ParamType IS_1 = 0.0,
+			ParamType alpha_0 = 0.0, ParamType alpha_1 = 0.0,
+			ParamType W_0 = 0.0, ParamType W_1 = 0.0,
+			ParamType q_0 = 0.0, ParamType q_1 = 0.0)
+			: v_g(v_g), k_g_times_s_g(k_g_times_s_g), B_0(B_0),
+			IS_0(IS_0), IS_1(IS_1),
+			alpha_0(alpha_0), alpha_1(alpha_1),
+			W_0(W_0), W_1(W_1),
+			q_0(q_0), q_1(q_1) {}
+	};
+
+
+	template <typename ParamType>
+	struct WENO5Params
+	{
+		ParamType v_g;
+		const ParamType k_g_times_s_g;
+		const ParamType B_0;
+		ParamType IS_0 = 0.0;
+		ParamType IS_1 = 0.0;
+		ParamType IS_2 = 0.0;
+		ParamType alpha_0 = 0.0;
+		ParamType alpha_1 = 0.0;
+		ParamType alpha_2 = 0.0;
+		ParamType W_0 = 0.0;
+		ParamType W_1 = 0.0;
+		ParamType W_2 = 0.0;
+		ParamType q_0 = 0.0;
+		ParamType q_1 = 0.0;
+		ParamType q_2 = 0.0;
+
+		WENO5Params(ParamType v_g, ParamType k_g_times_s_g, ParamType B_0,
+			ParamType IS_0 = 0.0, ParamType IS_1 = 0.0, ParamType IS_2 = 0.0,
+			ParamType alpha_0 = 0.0, ParamType alpha_1 = 0.0, ParamType alpha_2 = 0.0,
+			ParamType W_0 = 0.0, ParamType W_1 = 0.0, ParamType W_2 = 0.0,
+			ParamType q_0 = 0.0, ParamType q_1 = 0.0, ParamType q_2 = 0.0)
+			: v_g(v_g), k_g_times_s_g(k_g_times_s_g), B_0(B_0),
+			IS_0(IS_0), IS_1(IS_1), IS_2(IS_2),
+			alpha_0(alpha_0), alpha_1(alpha_1), alpha_2(alpha_2),
+			W_0(W_0), W_1(W_1), W_2(W_2),
[…]
