+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -879,107 +879,95 @@ protected:
 	template <typename StateType, typename ResidualType, typename ParamType, typename FactorType>
 	void WENO35Kernel(StateType const* yCrystal, ResidualType* resCrystal, const FactorType& factor, detail::WENO5Params<ResidualType>& WENO5Params, const int i) const
 	{
-		//ResidualType IS_0 = 0.0;
-		//ResidualType IS_1 = 0.0;
-		//ResidualType IS_2 = 0.0;
-		//ResidualType alpha_0 = 0.0;
-		//ResidualType alpha_1 = 0.0;
-		//ResidualType alpha_2 = 0.0;
-		//ResidualType W_0 = 0.0;
-		//ResidualType W_1 = 0.0;
-		//ResidualType W_2 = 0.0;
-		//ResidualType q_0 = 0.0;
-		//ResidualType q_1 = 0.0;
-		//ResidualType q_2 = 0.0;
-		
 		for (int i = 0; i < _nBins; ++i)
 		{
 			if (cadet_likely((i > 2) && (i + 2 < _nBins)))
 			{
 				// flux through the left face 
-				resCrystal[i] += factor / static_cast<ParamType>(_binSizes[i]) * v_g * (W_0 * q_0 + W_1 * q_1 + W_2 * q_2) - factor * static_cast<ParamType>(_growthDispersionRate) / static_cast<ParamType>(_binSizes[i]) * (yCrystal[i] - yCrystal[i - 1]) / static_cast<ParamType>(_binCenterDists[i - 1]);
+				resCrystal[i] += factor / static_cast<ParamType>(_binSizes[i]) * WENO5Params.v_g * (WENO5Params.W_0 * WENO5Params.q_0 + WENO5Params.W_1 * WENO5Params.q_1 + WENO5Params.W_2 * WENO5Params.q_2) - factor * static_cast<ParamType>(_growthDispersionRate) / static_cast<ParamType>(_binSizes[i]) * (yCrystal[i] - yCrystal[i - 1]) / static_cast<ParamType>(_binCenterDists[i - 1]);
 				// flux through the right face
 				WENO5Params.IS_0 = static_cast<ParamType>(_weno5->IS_0_coeff_1[i]) * (yCrystal[i + 2] - yCrystal[i + 1]) * (yCrystal[i + 2] - yCrystal[i + 1]) + static_cast<ParamType>(_weno5->IS_0_coeff_2[i]) * (yCrystal[i + 2] - yCrystal[i + 1]) * (yCrystal[i] - yCrystal[i + 1]) + static_cast<ParamType>(_weno5->IS_0_coeff_3[i]) * (yCrystal[i] - yCrystal[i + 1]) * (yCrystal[i] - yCrystal[i + 1]);
-				IS_1 = static_cast<ParamType>(_weno5->IS_1_coeff_1[i]) * (yCrystal[i - 1] - yCrystal[i]) * (yCrystal[i - 1] - yCrystal[i]) + static_cast<ParamType>(_weno5->IS_1_coeff_2[i]) * (yCrystal[i + 1] - yCrystal[i]) * (yCrystal[i - 1] - yCrystal[i]) + static_cast<ParamType>(_weno5->IS_1_coeff_3[i]) * (yCrystal[i + 1] - yCrystal[i]) * (yCrystal[i + 1] - yCrystal[i]);
[…]
