### 2024-12-19 b88b5ee
Fix utf-8 character
