### 2024-12-19 a1995e7
Fix error checking for nBin
