### 2024-12-19 1d6bfd9
Move crystallization mode configuration to configureModelDiscretization
