### 2024-12-19 d8a0c43
Fix nBin check
