### 2024-12-23 fa2079a
Fix bin coords
