### 2024-12-23 f2e95e0
Move growth reconstruction to same Kernel
