+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -91,6 +91,90 @@ namespace detail
 	};
 
 	/**
+	 * Numerical reconstruction type for solving the PBM
+	*/
+	enum class PBMReconstruction : int
+	{
+		/**
+		 * Upwind scheme
+		*/
+		Upwind,
+
+		/**
+		 * High-resolution Koren-type reconstruction
+		*/
+		HRKoren,
+
+		/**
+		 * WENO23 reconstruction
+		*/
+		WENO23,
+
+		/**
+		 * WENO35 reconstruction
+		*/
+		WENO35
+	};
+
+	template <typename ParamType>
+	struct ReconstructionParams
+	{
+		ParamType v_g;
+		const ParamType k_g_times_s_g;
+		const ParamType B_0;
+
+		// WENO
+		ParamType IS_0 = 0.0;
+		ParamType IS_1 = 0.0;
+		ParamType IS_2 = 0.0;
+		ParamType alpha_0 = 0.0;
+		ParamType alpha_1 = 0.0;
+		ParamType alpha_2 = 0.0;
+		ParamType W_0 = 0.0;
+		ParamType W_1 = 0.0;
+		ParamType W_2 = 0.0;
+		ParamType q_0 = 0.0;
+		ParamType q_1 = 0.0;
+		ParamType q_2 = 0.0;
+
+		// HRKoren
+		ParamType r_x_i = 0.0;
+		ParamType phi = 0.0;
+		ParamType F_i = 0.0;
+
+		ReconstructionParams(ParamType v_g, ParamType k_g_times_s_g, ParamType B_0,
+			ParamType IS_0 = 0.0, ParamType IS_1 = 0.0, ParamType IS_2 = 0.0,
+			ParamType alpha_0 = 0.0, ParamType alpha_1 = 0.0, ParamType alpha_2 = 0.0,
+			ParamType W_0 = 0.0, ParamType W_1 = 0.0, ParamType W_2 = 0.0,
+			ParamType q_0 = 0.0, ParamType q_1 = 0.0, ParamType q_2 = 0.0,
+			ParamType r_x_i = 0.0, ParamType phi = 0.0, ParamType F_i = 0.0)
+			: v_g(v_g), k_g_times_s_g(k_g_times_s_g), B_0(B_0),
+			IS_0(IS_0), IS_1(IS_1), IS_2(IS_2),
+			alpha_0(alpha_0), alpha_1(alpha_1), alpha_2(alpha_2),
+			W_0(W_0), W_1(W_1), W_2(W_2),
+			q_0(q_0), q_1(q_1), q_2(q_2),
+			r_x_i(r_x_i), phi(phi), F_i(F_i) {}
+	};
+
+	struct HRKorenCoefficients
+	{
+		std::vector<active> A_coeff;
+		std::vector<active> R_coeff;
+
+		HRKorenCoefficients(const std::vector<active>& binSizes)
+			: A_coeff(binSizes.size() - 1), R_coeff(binSizes.size() - 1)
+		{
+			// calculate the coefficients
+			for (int i = 1; i + 1 < binSizes.size(); ++i)
+			{
+				const active delta_xi_xip1 = binSizes[i] + binSizes[i + 1];
+				A_coeff[i] = delta_xi_xip1 / (binSizes[i] + binSizes[i - 1]);
+				R_coeff[i] = delta_xi_xip1 / binSizes[i];
+			}
+		}
+	};
+
+	/**
 	 * Flux reconstruction scheme coefficients.For details see @cite Zhang2024
 	*/
 	struct Weno3Coefficients
@@ -118,101 +202,6 @@ namespace detail
 			}
 		}
 	};
-
-	struct HRCoefficients
-	{
-		std::vector<active> A_coeff;
[…]
