### 2024-12-24 a290864
Residual: Move PBM, aggregation and fragmentation to the same loop
