### 2024-12-24 92bf32b
Jacobian: Move PBM, aggregation an fragmentation to same loop
