+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -1191,9 +1191,1134 @@ protected:
 		return residualLiquidImpl<StateType, ResidualType, ParamType, double>(t, secIdx, colPos, yLiquid, resLiquid, factor, workSpace);
 	}
 
+	struct JacobianParams
+	{
+		const double vG_factor;
+		const double dBp_dc;
+		const double dBp_dceq;
+		const double dBs_dc;
+		const double dBs_dceq;
+		const double dvG_dc_factor;
+		const double dvG_dceq_factor;
+		const double dBs_dni_factor;
+
+		// HR Koren specific parameters
+		const double epsilon = 1e-10;
+		double r_x_i = 0.0;
+		double phi = 0.0;
+		double F_i = 0.0;
+		double dvG_dc_right = 0.0; // also used in weno schemes
+		double dvG_dceq_right = 0.0;
+		double vg_right = 0.0;
+
+		double r_cluster = 0.0;
+		double r_square_cluster = 0.0;
+		double A_cluster = 0.0;
+		double Ar_cluster = 0.0;
+		double ni_difference = 0.0;
+
+		double dFi_wrt_nim1 = 0.0;
+		double dFi_wrt_ni = 0.0;
+		double dFi_wrt_nip1 = 0.0;
+
+		// WENO23 related coefficients
+		double IS_0_right = 0.0;
+		double IS_1_right = 0.0;
+		double alpha_0_right = 0.0;
+		double alpha_1_right = 0.0;
+		double W_0_right = 0.0;
+		double W_1_right = 0.0;
+		double q_0_right = 0.0;
+		double q_1_right = 0.0;
+
+		//double dvG_dc_right = 0.0; // also used in HRKoren and WENO35 schemes
+		//double dvG_dceq_right = 0.0;
+		//double vg_right = 0.0;
+
+		// Jacobian related coefficients
+		double four_w1_B1_right = 0.0;
+		double four_w0_B0_right = 0.0;
+		double four_w0_w1_B1_right = 0.0;
+		double four_w0_w0_B0_right = 0.0;
+		double four_w1_w1_B1_right = 0.0;
+		double four_w0_w1_B0_right = 0.0;
+
+		double dw0_right_dni_m1 = 0.0;
+		double dw0_right_dni = 0.0;
+		double dw0_right_dni_p1 = 0.0;
+
+		double dw1_right_dni_m1 = 0.0;
+		double dw1_right_dni = 0.0;
+		double dw1_right_dni_p1 = 0.0;
+
+		double dq0_right_dni = 0.0;
+		double dq0_right_dni_p1 = 0.0;
+
+		double dq1_right_dni_m1 = 0.0;
+		double dq1_right_dni = 0.0;
+
+		// WENO35 boundary WENO23 related coefficients, for bin N_x-2
+		double IS_0_weno3_right = 0.0;
+		double IS_1_weno3_right = 0.0;
+		double alpha_0_weno3_right = 0.0;
+		double alpha_1_weno3_right = 0.0;
+		double W_0_weno3_right = 0.0;
+		double W_1_weno3_right = 0.0;
+		double q_0_weno3_right = 0.0;
+		double q_1_weno3_right = 0.0;
+
+		// WENO35 boundary WENO23 jacobian related coefficients, for bin N_x-2
+		double four_w0_B0_weno3_right = 0.0;
+		double four_w1_B1_weno3_right = 0.0;
+		double four_w0_w0_B0_weno3_right = 0.0;
+		double four_w0_w1_B1_weno3_right = 0.0;
+		double four_w0_w1_B0_weno3_right = 0.0;
[…]
