### 2025-01-03 1016cbd
Delete unused multithreading macro
The macro CADET_PARFOR_END extends to a closed bracket when the project
is build with multithreading enabled, which causes issues when the
corresponding opening bracket is not defined.
This commit deletes unused CADET_PARFOR_END statements.
