### 2025-01-04 534d36b
Add parallelization (optional) of particle blocks in 2DLRMP
