### 2025-01-04 cce8fc9
Add parallelization (optional) of particle blocks in 2DLRMP
