+++ doc/interface/binding/extended_mobile_phase_modulator_langmuir.rst
@@ -73,6 +73,6 @@ For information on model equations, refer to :ref:`extended_mobile_phase_modulat
 
 **Unit:** :math:`m_{MP}^{3} mol^{-1}`
 
-===================  =========================  =========================================
-**Type:** double     **Range:** :math:`\ge 0`   **Length:** NCOMP
-===================  =========================  =========================================
+===================  ==============================  =========================================
+**Type:** double     **Range:** :math:`\mathbb{R}`   **Length:** NCOMP
+===================  ==============================  =========================================
