+++ doc/interface/binding/mobile_phase_modulator_langmuir.rst
@@ -60,6 +60,6 @@ For information on model equations, refer to :ref:`mobile_phase_modulator_langmu
 
 **Unit:** :math:`m_{MP}^{3} mol^{-1}`
 
-===================  =========================  =========================================
-**Type:** double     **Range:** :math:`\ge 0`   **Length:** NCOMP
-===================  =========================  =========================================
+===================  ==============================  =========================================
+**Type:** double     **Range:** :math:`\mathbb{R}`   **Length:** NCOMP
+===================  ==============================  =========================================
