### 2025-01-14 ddc85d0
Add Multicomponent Freundlich LDF isotherm. (#259)
Co-authored-by: Ronald Jäpel <ronald.jaepel@gmail.com>
