+++ doc/interface/binding/index.rst
@@ -68,6 +68,7 @@ This group also takes precedence over a possibly existing ``/input/model/unit_XX
     saska
     self_association
     freundlich_ldf
+    multi_component_ldf_freundlich
     hic_water_on_hydrophobic_surfaces
     hic_constant_water_activity
     multi_component_colloidal
