### 2025-01-14 05c443a
Add Multicomponent Freundlich LDF isotherm. (#259)
Co-authored-by: Ronald Jäpel <ronald.jaepel@gmail.com>
