+++ doc/modelling/binding/freundlich_ldf.rst
@@ -14,7 +14,7 @@ This variant of the model is based on the linear driving force approximation (se
 No interaction between the components is considered when the model has multiple components. 
 One of the limitation of this isotherm is the first order Jacobian :math:`\left(\frac{dq^*}{dc_p}\right)` tends to infinity as :math:`c_{p} \rightarrow 0` for :math:`n>1`.
 To address this issue an approximation of isotherm is considered near the origin.
-This approximation matches the isotherm in such a way that  :math:`q=0` at :math:`c_p=0` and also matches the first derivative of the istotherm at :math:`c_p = \varepsilon`, where :math:`\varepsilon` is a very small number, for example :math:`1e-14`.
+This approximation matches the isotherm in such a way that  :math:`q=0` at :math:`c_p=0` and also matches the first derivative of the isotherm at :math:`c_p = \varepsilon`, where :math:`\varepsilon` is a very small number, for example :math:`1e-14`.
 The form of approximation and its derivative is given below for :math:`c_p < \varepsilon` and :math:`n>1`:
 
 .. math::
@@ -38,7 +38,7 @@ where :math:`\alpha_0=0` and :math:`\alpha_1` and :math:`\alpha_2` are determine
 		\alpha_2 = \frac{1-n}{n}k_f \varepsilon^{\frac{1-2 n}{n}}
 	\end{aligned}
 
-This approximation can be used for any pore phase cocentration :math:`c_p < \varepsilon` given :math:`n>1`.
-For the case, when :math:`n \le 1` no special treament near the origin is required.
+This approximation can be used for any pore phase concentration :math:`c_p < \varepsilon` given :math:`n>1`.
+For the case, when :math:`n \le 1` no special treatment near the origin is required.
 For more information on model parameters required to define in CADET file format, see :ref:`freundlich_ldf_config`.
 
