+++ doc/modelling/binding/index.rst
@@ -277,6 +277,11 @@ The models also differ in whether a mobile phase modifier (e.g., salt) is suppor
      - x
      - ✓
      - x
+   * - :ref:`multi_component_ldf_freundlich_model`
+     - ✓
+     - ×
+     - ✓
+     - ×
 
 
 .. toctree::
