### 2025-01-15 e054d22
Fix sign check in FD Jacobian pattern test and adjust tolerances accordingly
Co-authored-by: Jan Breuer <jan.breuer@fz-juelich.de>
