### 2025-01-15 17698ee
Fix sign check in FD Jacobian pattern test and adjust tolerances accordingly
Co-authored-by: Jan Breuer <jan.breuer@fz-juelich.de>
