### 2025-01-15 f2aaf00
Fix sign check in FD Jacobian pattern test and adjust tolerances accordingly
Co-authored-by: Jan Breuer <jan.breuer@fz-juelich.de>
