### 2025-01-15 6a4367e
Restart implementation
