+++ src/libcadet/model/reaction/MassActionLawReaction.cpp
@@ -25,7 +25,9 @@
 #include <unordered_map>
 #include <string>
 #include <vector>
+#include <Eigen/Dense>
 
+using namespace Eigen;
 /*<codegen>
 {
 	"name": "MassActionLawParamHandler",
@@ -312,10 +314,16 @@ public:
 	virtual void setExternalFunctions(IExternalFunction** extFuns, unsigned int size) { _paramHandler.setExternalFunctions(extFuns, size); }
 	virtual bool dependsOnTime() const CADET_NOEXCEPT { return ParamHandler_t::dependsOnTime(); }
 	virtual bool requiresWorkspace() const CADET_NOEXCEPT { return true; }
+	bool hasQuasiStationaryReactionsLiquid() { return true; }
+	virtual bool hasDynamicReactions() const CADET_NOEXCEPT { return true; }
+	virtual int const* reactionQuasiStationarity() const CADET_NOEXCEPT { return _reactionQuasistationarity.data(); }
+	
 	virtual unsigned int workspaceSize(unsigned int nComp, unsigned int totalNumBoundStates, unsigned int const* nBoundStates) const CADET_NOEXCEPT
 	{
 		return _paramHandler.cacheSize(maxNumReactions(), nComp, totalNumBoundStates) + std::max(maxNumReactions() * sizeof(active), 2 * (_nComp + totalNumBoundStates) * sizeof(double));
 	}
+
+
 
 	virtual bool configureModelDiscretization(IParameterProvider& paramProvider, unsigned int nComp, unsigned int const* nBound, unsigned int const* boundOffset)
 	{
@@ -371,6 +379,93 @@ public:
 
 		return true;
 	}
+	
+	//void fillConservedMoietiesBulk(Eigen::MatrixXd M, std::vector<int>QSReaction, std::vector<int>QSComponent) {
+	//	ConservedMoietiesLiquid(M, QSReaction,QSComponent);
+	//}
+	/*
+	 * @brief Calculates the conserved moieties based on the stoichiometric matrix and reaction quasistationarity.
+	 * @param S The stoichiometric matrix.
+	 * @param _reactionQuasistationarity The reaction quasistationarity vector.
+	 * @return The conserved moieties matrix.
+	 */
+	virtual void fillConservedMoietiesBulk(Eigen::MatrixXd& M, std::vector<int>& mapQSReac, std::vector<int>& _QsCompBulk)
+	{
+
+		//1. get stoichmetic matrix with only reaction in quasi stationary
+		// S dim -> ncomp x nreac
+
+		int nQS = std::count(mapQSReac.begin(), mapQSReac.end(), 1);
+		// Count the number of entries with value 1 in _reactionQuasistationarity
+
+
+		Eigen::MatrixXd QSS(_stoichiometryBulk.rows(), nQS);
+		int rowIndex = 0;
+
+		// Fülle die Spalten basierend auf _reactionQuasistationarity
+		for (int i = 0; i < _stoichiometryBulk.columns(); ++i)
+		{
+			const double* rowData = reinterpret_cast<const double*>(_stoichiometryBulk.data() + i * _stoichiometryBulk.columns());
+			Eigen::Map<const Eigen::VectorXd> Srow(rowData, _stoichiometryBulk.columns());
+			QSS.row(rowIndex++) = Srow;
+		}
+
+		//Remove zero rows from QSS
+		int nQScomp = 0; // Number of quasi stationary active components
+		for (int i = 0; i < QSS.rows(); ++i)
+		{
[…]
