+++ src/libcadet/model/reaction/DummyReaction.cpp
@@ -99,6 +99,12 @@ public:
 
 	virtual unsigned int numReactionsLiquid() const CADET_NOEXCEPT { return 0; }
 	virtual unsigned int numReactionsCombined() const CADET_NOEXCEPT { return 0; }
+	
+	void fillConservedMoietiesBulk(Eigen::MatrixXd& M, std::vector<int>& QSReaction, std::vector<int>& QSComponent) {}
+	bool hasQuasiStationaryReactionsLiquid() { return false;}
+	
+	virtual int quasiStationaryFlux(double t, unsigned int secIdx, const ColumnPosition& colPos, double const* y,
+		active* fluxes, std::vector<int>& mapQSReac, LinearBufferAllocator workSpace) const {return 0;}
 
 protected:
 };

[…]
