### 2025-01-15 53ca6ab
Update Freundlich isotherm linearization and tests
Extend linearization range into negative numbers to improve stability
Extend tests
Update documentation
