+++ doc/modelling/binding/freundlich_ldf.rst
@@ -13,9 +13,10 @@ This variant of the model is based on the linear driving force approximation (se
 
 No interaction between the components is considered when the model has multiple components. 
 One of the limitation of this isotherm is the first order Jacobian :math:`\left(\frac{dq^*}{dc_p}\right)` tends to infinity as :math:`c_{p} \rightarrow 0` for :math:`n>1`.
-To address this issue an approximation of isotherm is considered near the origin.
-This approximation matches the isotherm in such a way that  :math:`q=0` at :math:`c_p=0` and also matches the first derivative of the isotherm at :math:`c_p = \varepsilon`, where :math:`\varepsilon` is a very small number, for example :math:`1e-14`.
-The form of approximation and its derivative is given below for :math:`c_p < \varepsilon` and :math:`n>1`:
+Additionally, the isotherm is undefined for :math:`c_{p} < 0` if :math:`\frac{1}{n_i}` can be expressed as :math:`\frac{p}{q}` with :math:`p,q \in \mathbb{N}` where :math:`q` is an even number.
+To address these issues an approximation of the isotherm is considered below a threshold concentration :math:`c_p < \varepsilon`.
+This approximation matches the isotherm in such a way that  :math:`q=0` at :math:`c_p=0` and also matches the value and the first derivative of the isotherm at :math:`c_p = \varepsilon`, where :math:`\varepsilon` is a very small number, for example :math:`1e-14`.
+The form of approximation and its derivative is given below for :math:`c_p < \varepsilon`:
 
 .. math::
 
@@ -39,6 +40,5 @@ where :math:`\alpha_0=0` and :math:`\alpha_1` and :math:`\alpha_2` are determine
 	\end{aligned}
 
 This approximation can be used for any pore phase concentration :math:`c_p < \varepsilon` given :math:`n>1`.
-For the case, when :math:`n \le 1` no special treatment near the origin is required.
 For more information on model parameters required to define in CADET file format, see :ref:`freundlich_ldf_config`.
 
