### 2025-01-16 d5428be
Add binding model submodule tests to CI
Fix FD pattern sign tolerance in Fruendlich LDF test

In the test settings, remove EXT_ prefix from constant binding
parameters, fixing the configuration of external binding parameter
handler
