+++ test/BindingModels.cpp
@@ -273,7 +273,7 @@ CADET_BINDINGTEST("MOBILE_PHASE_MODULATOR", "EXT_MOBILE_PHASE_MODULATOR", (1,1,1
 	        "EXT_MPM_BETA_T": [0.0, 1.5, 2.0],
 	        "EXT_MPM_BETA_TT": [0.0, 0.0, 0.0],
 	        "EXT_MPM_BETA_TTT": [0.0, 0.0, 0.0],
-			"EXT_MPM_LINEAR_THRESHOLD": 1e-10
+			"MPM_LINEAR_THRESHOLD": 1e-10
 	)json", \
 	R"json( "EXT_MPM_KA": [0.0, 0.0, 0.0, 0.0],
 	        "EXT_MPM_KA_T": [0.0, 1.14, 1.0, 2.0],
@@ -295,7 +295,7 @@ CADET_BINDINGTEST("MOBILE_PHASE_MODULATOR", "EXT_MOBILE_PHASE_MODULATOR", (1,1,1
 	        "EXT_MPM_BETA_T": [0.0, 1.5, 0.0, 2.0],
 	        "EXT_MPM_BETA_TT": [0.0, 0.0, 0.0, 0.0],
 	        "EXT_MPM_BETA_TTT": [0.0, 0.0, 0.0, 0.0],
-			"EXT_MPM_LINEAR_THRESHOLD": 1e-10
+			"MPM_LINEAR_THRESHOLD": 1e-10
 	)json", \
 	1e-10, 1e-10, CADET_NONBINDING_LIQUIDPHASE_COMP_UNUSED, CADET_COMPARE_BINDING_VS_NONBINDING)
 	
@@ -334,7 +334,7 @@ CADET_BINDINGTEST("MOBILE_PHASE_MODULATOR", "EXT_MOBILE_PHASE_MODULATOR", (1,1,1
 	        "EXT_MPM_BETA_T": [0.0, 1.5, 2.0],
 	        "EXT_MPM_BETA_TT": [0.0, 0.0, 0.0],
 	        "EXT_MPM_BETA_TTT": [0.0, 0.0, 0.0],
-			"EXT_MPM_LINEAR_THRESHOLD": 1e-10
+			"MPM_LINEAR_THRESHOLD": 1e-10
 	)json", \
 	R"json( "EXT_MPM_KA": [0.0, 0.0, 0.0, 0.0],
 	        "EXT_MPM_KA_T": [0.0, 1.14, 1.0, 2.0],
@@ -356,7 +356,7 @@ CADET_BINDINGTEST("MOBILE_PHASE_MODULATOR", "EXT_MOBILE_PHASE_MODULATOR", (1,1,1
 	        "EXT_MPM_BETA_T": [0.0, 1.5, 0.0, 2.0],
 	        "EXT_MPM_BETA_TT": [0.0, 0.0, 0.0, 0.0],
 	        "EXT_MPM_BETA_TTT": [0.0, 0.0, 0.0, 0.0],
-			"EXT_MPM_LINEAR_THRESHOLD": 1e-10
+			"MPM_LINEAR_THRESHOLD": 1e-10
 	)json", \
 	1e-10, 1e-10, CADET_NONBINDING_LIQUIDPHASE_COMP_UNUSED, CADET_COMPARE_BINDING_VS_NONBINDING)
 
@@ -1002,7 +1002,7 @@ CADET_BINDINGTEST("MULTI_COMPONENT_LDF_FREUNDLICH", "EXT_MULTI_COMPONENT_LDF_FRE
 	        "EXT_MCLDFFRL_A_T": [3.0, 2.2, 1.5, 0.5],
[…]
