### 2025-01-16 c3b2f30
Add CI tag to binding tests
