+++ src/libcadet/model/StirredTankModel.cpp
@@ -81,7 +81,8 @@ CSTRModel::~CSTRModel() CADET_NOEXCEPT
 	delete[] _offsetParType;
 
 	delete[] _dynReactionBulk;
-	//delete[] _temp; -> Jan warum funktioniert das nicht?
+	delete[] _temp;
+	delete[] _temp2;
 }
 
 unsigned int CSTRModel::numDofs() const CADET_NOEXCEPT
@@ -241,19 +242,20 @@ bool CSTRModel::configureModelDiscretization(IParameterProvider& paramProvider, 
 		if (_dynReactionBulk->usesParamProviderInDiscretizationConfig())
 			paramProvider.popScope();
 		
-		if (true) //paramProvider.exists("QUASI_STATIONARY_REACTION_BULK")
+		if (paramProvider.exists("QUASI_STATIONARY_REACTION_BULK"))
 		{
-			//_qsReacBulk = paramProvider.getIntArray("QUASI_STATIONARY_REACTION_BULK");
+			_qsReacBulk = paramProvider.getIntArray("QUASI_STATIONARY_REACTION_BULK");
 			_qsReacBulk = {1};
 			_nQsReacBulk = _qsReacBulk.size();
-			//_temp = new active[_nComp];
+			_temp = new active[_nComp];
+			_temp2 = new double[_nQsReacBulk];
 
 			_nMoitiesBulk = (_nComp + _totalBound) - _nQsReacBulk;
 			_MconvMoityBulk = Eigen::MatrixXd::Zero(_nMoitiesBulk, _nComp); // matrix for conserved moities
 		}
 		else
 		{
-			_QsCompBulk.clear();
+			_QsCompBulk.clear();	
 			_qsReacBulk.clear();
 			_nMoitiesBulk = 0;
 			_nQsReacBulk = 0;
@@ -382,10 +384,9 @@ bool CSTRModel::configure(IParameterProvider& paramProvider)
 		dynReactionConfSuccess = _dynReactionBulk->configure(paramProvider, _unitOpIdx, ParTypeIndep);
 		paramProvider.popScope();
 
-		if (true) //paramProvider.exists("QUASI_STATIONARY_REACTION_BULK")
+		if (paramProvider.exists("QUASI_STATIONARY_REACTION_BULK"))
 		{
 			_dynReactionBulk->fillConservedMoietiesBulk(_MconvMoityBulk, _qsReacBulk, _QsCompBulk); // fill conserved moities matrix
-			int a = 0;
 		}
 	}
 
@@ -718,7 +719,7 @@ void CSTRModel::consistentInitialState(const SimulationTime& simTime, double* co
 					if (!qsMask[_nComp + bndIdx])
 						continue;
 
-					conservedQuants[idx] += factor * c[_nComp + bndIdx];
+					conservedQuants[idx] += factor * c[_nComp + bndIdx]; 
 				}
 			}
 
@@ -1341,70 +1342,64 @@ int CSTRModel::residualImpl(double t, unsigned int secIdx, StateType const* cons
 	// Reactions in liquid phase
 	const ColumnPosition colPos{0.0, 0.0, 0.0};
 
+	LinearBufferAllocator subAlloc = tlmAlloc.manageRemainingMemory(); 
+
+	BufferedArray<ResidualType> flux = subAlloc.array<ResidualType>(_nComp);
+	std::fill_n(static_cast<ResidualType*>(flux), _nComp, 0.0);
+
 	if (_dynReactionBulk && (_nQsReacBulk > 0))
 	{
-		LinearBufferAllocator subAlloc = tlmAlloc.manageRemainingMemory(); // todo nachgucken ob das wirklich so geht
-
[…]
