+++ src/libcadet/model/reaction/MassActionLawReaction.cpp
@@ -403,7 +403,7 @@ public:
 		int rowIndex = 0;
 
 		// Fülle die Spalten basierend auf _reactionQuasistationarity
-		for (int i = 0; i < _stoichiometryBulk.columns(); ++i)
+		for (int i = 0; i < _stoichiometryBulk.rows(); ++i)
 		{
 			const double* rowData = reinterpret_cast<const double*>(_stoichiometryBulk.data() + i * _stoichiometryBulk.columns());
 			Eigen::Map<const Eigen::VectorXd> Srow(rowData, _stoichiometryBulk.columns());
@@ -646,18 +646,15 @@ protected:
 		return true;
 	}
 
-	int singleFlux(int r, active const* y, typename DoubleActivePromoter<active, double>::type flow, typename ParamHandler_t::ParamsHandle const p)
+	double singleFlux(int r, double const* y, double kFwdBulk_r, double kBwdBulk_r)
 	{
-		typedef typename DoubleActivePromoter<active, double>::type flux_t;
-
-		flux_t fwd = rateConstantOrZero(static_cast<typename DoubleActiveDemoter<flux_t, active>::type>(p->kFwdBulk[r]), r, _expBulkFwd, _nComp);
+		double fwd = rateConstantOrZero(kFwdBulk_r, r, _expBulkFwd, _nComp);
 		for (int c = 0; c < _nComp; ++c)
 		{
 			if (_expBulkFwd.native(c, r) != 0.0)
 			{
 				if (static_cast<double>(y[c]) > 0.0)
-					fwd *= pow(static_cast<typename DoubleActiveDemoter<flux_t, active>::type>(y[c]),
-						static_cast<typename DoubleActiveDemoter<flux_t, active>::type>(_expBulkFwd.native(c, r)));
+					fwd *= pow(static_cast<double>(y[c]), static_cast<double>(_expBulkFwd.native(c, r)));
 				else
 				{
 					fwd *= 0.0;
@@ -666,14 +663,13 @@ protected:
 			}
 		}
 
-		flux_t bwd = rateConstantOrZero(static_cast<typename DoubleActiveDemoter<flux_t, active>::type>(p->kBwdBulk[r]), r, _expBulkBwd, _nComp);
+		double bwd = rateConstantOrZero(kBwdBulk_r, r, _expBulkBwd, _nComp);
 		for (int c = 0; c < _nComp; ++c)
 		{
 			if (_expBulkBwd.native(c, r) != 0.0)
 			{
 				if (static_cast<double>(y[c]) > 0.0)
-					bwd *= pow(static_cast<typename DoubleActiveDemoter<flux_t, active>::type>(y[c]),
-						static_cast<typename DoubleActiveDemoter<flux_t, active>::type>(_expBulkBwd.native(c, r)));
+					bwd *= pow(static_cast<double>(y[c]), static_cast<double>(_expBulkBwd.native(c, r)));
 				else
 				{
 					bwd *= 0.0;
@@ -681,26 +677,26 @@ protected:
 				}
 			}
 		}
-
-		flow = fwd - bwd;
-		return 0;
+		return fwd - bwd;
 	}
 
+	virtual int quasiStationaryFlux(double t, unsigned int secIdx, const ColumnPosition& colPos, active const* y,
+		Eigen::Map<Eigen::VectorXd> fluxes, std::vector<int> mapQSReac, LinearBufferAllocator workSpace) { return 0; }
 
 	virtual int quasiStationaryFlux(double t, unsigned int secIdx, const ColumnPosition& colPos, double const* y,
-		active* fluxes, std::vector<int>& mapQSReac, LinearBufferAllocator workSpace) const
[…]
