+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -548,7 +548,7 @@ public:
 			_u = paramProvider.getDouble("CRY_U");
 			_parameters[makeParamId(hashString("CRY_U"), unitOpIdx, CompIndep, ParTypeIndep, BoundStateIndep, ReactionIndep, SectionIndep)] = &_u;
 
-			// clear scheme coefficients
+			// clear scheme coefficients // todo delete jacparams
 			if (_HR)
 			{
 				delete _HR;
@@ -568,11 +568,24 @@ public:
 			}
 
 			if (_growthSchemeOrder == detail::PBMReconstruction::HRKoren)
+			{
 				_HR = new detail::HRKorenCoefficients(_binSizes);
+				jacParams = new JacobianParamsHRKoren;
+			}
 			else if (_growthSchemeOrder == detail::PBMReconstruction::WENO23)
+			{
 				_weno3 = new detail::Weno3Coefficients(_binSizes);
+				jacParams = new JacobianParamsWENO23;
+			}
 			else if (_growthSchemeOrder == detail::PBMReconstruction::WENO35)
+			{
 				_weno5 = new detail::Weno5Coefficients(_binSizes);
+				jacParams = new JacobianParamsWENO35;
+			}
+			else
+			{
+				jacParams = new JacobianParamsBase;
+			}
 		}
 		if (_useAgg)
 		{
@@ -682,7 +695,276 @@ public:
 
 protected:
 
+	struct JacobianParamsBase
+	{
+		// Constant parameters used in upwind scheme. Also included in parameters for HRKoren and WENO23 and WENO35 schemes
+		double vG_factor;
+		double dBp_dc;
+		double dBp_dceq;
+		double dBs_dc;
+		double dBs_dceq;
+		double dvG_dc_factor;
+		double dvG_dceq_factor;
+		double dBs_dni_factor;
+	};
+
+	struct JacobianParamsHRKoren : public JacobianParamsBase
+	{
+		// HR Koren specific parameters
+		const double epsilon = 1e-10;
+		double r_x_i = 0.0;
+		double phi = 0.0;
+		double F_i = 0.0;
+		double dvG_dc_right = 0.0;
+		double dvG_dceq_right = 0.0;
+		double vg_right = 0.0;
+
+		double r_cluster = 0.0;
+		double r_square_cluster = 0.0;
+		double A_cluster = 0.0;
+		double Ar_cluster = 0.0;
+		double ni_difference = 0.0;
+
+		double dFi_wrt_nim1 = 0.0;
+		double dFi_wrt_ni = 0.0;
+		double dFi_wrt_nip1 = 0.0;
+	};
+
+	struct JacobianParamsWENO23 : public JacobianParamsBase
+	{
+		// WENO23 related coefficients
+		double dvG_dc_right = 0.0;
+		double dvG_dceq_right = 0.0;
+		double vg_right = 0.0;
+		double IS_0_right = 0.0;
+		double IS_1_right = 0.0;
+		double alpha_0_right = 0.0;
+		double alpha_1_right = 0.0;
+		double W_0_right = 0.0;
+		double W_1_right = 0.0;
+		double q_0_right = 0.0;
+		double q_1_right = 0.0;
+
+		// Jacobian related coefficients
+		double four_w1_B1_right = 0.0;
+		double four_w0_B0_right = 0.0;
+		double four_w0_w1_B1_right = 0.0;
+		double four_w0_w0_B0_right = 0.0;
[…]
