### 2025-01-21 f3eda0c
Move configuration of crys_mode to configure() function
