+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -457,6 +457,28 @@ public:
 
 	virtual bool configure(IParameterProvider& paramProvider, UnitOpIdx unitOpIdx, ParticleTypeIdx parTypeIdx)
 	{
+		if (paramProvider.exists("CRY_MODE"))
+			_mode = static_cast<detail::CrystallizationMode>(paramProvider.getInt("CRY_MODE"));
+		else
+			_mode = detail::CrystallizationMode::PurePBM; // For backwards compatibility
+
+		_usePBM = detail::modeFlagMap[_mode].hasMassBalance;
+		_useAgg = detail::modeFlagMap[_mode].hasAggregation;
+		_useFrag = detail::modeFlagMap[_mode].hasFragmentation;
+
+		if (_usePBM)
+		{
+			// Comp 0 is substrate, last comp is equilibrium
+			_nBins = _nComp - 2;
+
+			if (_nBins < 1)
+				throw InvalidParameterException("Expected at least 3 components (got " + std::to_string(_nComp) + ")");
+		}
+		else
+		{
+			_nBins = _nComp;
+		}
+
 		readScalarParameterOrArray(_bins, paramProvider, "CRY_BINS", 1);
 
 		if (_bins.size() != _nBins + 1)
@@ -588,28 +610,6 @@ public:
 	virtual bool configureModelDiscretization(IParameterProvider& paramProvider, unsigned int nComp, unsigned int const* nBound, unsigned int const* boundOffset)
 	{
 		_nComp = nComp;
-
-		if (paramProvider.exists("CRYSTALLIZATION_MODE"))
-			_mode = static_cast<detail::CrystallizationMode>(paramProvider.getInt("CRYSTALLIZATION_MODE"));
-		else
-			_mode = detail::CrystallizationMode::PurePBM; // For backwards compatibility
-
-		_usePBM = detail::modeFlagMap[_mode].hasMassBalance;
-		_useAgg = detail::modeFlagMap[_mode].hasAggregation;
-		_useFrag = detail::modeFlagMap[_mode].hasFragmentation;
-
-		if (_usePBM)
-		{
-			// Comp 0 is substrate, last comp is equilibrium
-			_nBins = _nComp - 2;
-
-			if (_nBins < 1)
-				throw InvalidParameterException("Expected at least 3 components (got " + std::to_string(_nComp) + ")");
-		}
-		else
-		{
-			_nBins = _nComp;
-		}
 
 		return true;
 	}
