### 2025-01-21 9de400e
Add exceptions for wrong crystallization mode
