### 2025-01-22 fceefa7
Add a guide for binding tests
