### 2025-01-22 e670722
Add template to reconstruction structs
