+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -364,17 +364,17 @@ public:
 
 			if (_growthSchemeOrder == detail::PBMReconstruction::HRKoren)
 			{
-				_reconstruction = new HRKorenParams(_binSizes, 0.0, 0.0, 0.0);
+				_reconstruction = new HRKorenParams<active>(_binSizes, 0.0, 0.0, 0.0);
 				_jacParams = new JacobianParamsHRKoren;
 			}
 			else if (_growthSchemeOrder == detail::PBMReconstruction::WENO23)
 			{
-				_reconstruction = new WENO23Params(_binSizes);
+				_reconstruction = new WENO23Params<active>(_binSizes);
 				_jacParams = new JacobianParamsWENO23;
 			}
 			else if (_growthSchemeOrder == detail::PBMReconstruction::WENO35)
 			{
-				_reconstruction = new WENO35Params(_binSizes);
+				_reconstruction = new WENO35Params<active>(_binSizes);
 				_jacParams = new JacobianParamsWENO35;
 			}
 			else
@@ -490,22 +490,31 @@ public:
 
 protected:
 
+	template<typename ParamType>
 	struct ReconstructionParams
 	{
-		active v_g = 0.0;
-		active k_g_times_s_g = 0.0;
-		active B_0 = 0.0;
+		ParamType v_g = 0.0;
+		ParamType k_g_times_s_g = 0.0;
+		ParamType B_0 = 0.0;
 
-		ReconstructionParams(active v_g = 0.0, active k_g_times_s_g = 0.0, active B_0 = 0.0) : v_g(v_g), k_g_times_s_g(k_g_times_s_g), B_0(B_0) {};
+		ReconstructionParams(ParamType v_g = 0.0, ParamType k_g_times_s_g = 0.0, ParamType B_0 = 0.0) : v_g(v_g), k_g_times_s_g(k_g_times_s_g), B_0(B_0) {};
 
-		void updateParams(active k_g_times_s_g_, active B_0_)
+		void updateParams(ParamType k_g_times_s_g_, ParamType B_0_)
 		{
 			k_g_times_s_g = k_g_times_s_g_;
 			B_0 = B_0_;
 		}
+
+		// Create a double view of this struct
+		template <typename U>
+		ReconstructionParams<U>& as() const {
+			//static_assert(std::is_same_v<U, double>, "Only double views are supported");
+			return reinterpret_cast<ReconstructionParams<U>&>(*this);
+		}
 	};
 
-	struct HRKorenParams : public ReconstructionParams
+	template<typename ParamType>
+	struct HRKorenParams : public ReconstructionParams<ParamType>
 	{
 		active r_x_i = 0.0;
 		active phi = 0.0;
@@ -530,32 +539,33 @@ protected:
 	/**
 	 * Flux reconstruction scheme coefficients.For details see @cite Zhang2024
 	*/
-	struct WENO23Params : public ReconstructionParams
+	template<typename ParamType>
+	struct WENO23Params : public ReconstructionParams<ParamType>
 	{
-		std::vector<active> q_1_right_coeff;
-		std::vector<active> q_0_right_coeff;
-		std::vector<active> C_right_coeff;
-		std::vector<active> IS_0_coeff;
-		std::vector<active> IS_1_coeff;
+		std::vector<ParamType> q_1_right_coeff;
+		std::vector<ParamType> q_0_right_coeff;
+		std::vector<ParamType> C_right_coeff;
+		std::vector<ParamType> IS_0_coeff;
[…]
