### 2025-01-22 dfe87c0
Add template view for reconstruction parameter structs
