+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -364,7 +364,7 @@ public:
 
 			if (_growthSchemeOrder == detail::PBMReconstruction::HRKoren)
 			{
-				_reconstruction = new HRKorenParams<active>(_binSizes, 0.0, 0.0, 0.0);
+				_reconstruction = new HRKorenParams<active>(_binSizes);
 				_jacParams = new JacobianParamsHRKoren;
 			}
 			else if (_growthSchemeOrder == detail::PBMReconstruction::WENO23)
@@ -507,32 +507,51 @@ protected:
 
 		// Create a double view of this struct
 		template <typename U>
-		ReconstructionParams<U>& as() const {
+		ReconstructionParams<U>& as() {
 			//static_assert(std::is_same_v<U, double>, "Only double views are supported");
 			return reinterpret_cast<ReconstructionParams<U>&>(*this);
+		}
+
+		// Overload for const objects
+		template <typename U>
+		const ReconstructionParams<U>& as() const {
+			//static_assert(std::is_same_v<U, double>, "Only double views are supported");
+			return reinterpret_cast<const ReconstructionParams<U>&>(*this);
 		}
 	};
 
 	template<typename ParamType>
 	struct HRKorenParams : public ReconstructionParams<ParamType>
 	{
-		active r_x_i = 0.0;
-		active phi = 0.0;
-		active F_i = 0.0;
+		ParamType r_x_i = 0.0;
+		ParamType phi = 0.0;
+		ParamType F_i = 0.0;
 
-		std::vector<active> A_coeff;
-		std::vector<active> R_coeff;
+		std::vector<ParamType> A_coeff;
+		std::vector<ParamType> R_coeff;
 
-		HRKorenParams(const std::vector<active>& binSizes, active r_x_i = 0.0, active phi = 0.0, active F_i = 0.0)
+		HRKorenParams(const std::vector<active>& binSizes)
 			: r_x_i(r_x_i), phi(phi), F_i(F_i), A_coeff(binSizes.size() - 1), R_coeff(binSizes.size() - 1)
 		{
 			// calculate the coefficients
 			for (int i = 1; i + 1 < binSizes.size(); ++i)
 			{
-				const active delta_xi_xip1 = binSizes[i] + binSizes[i + 1];
-				A_coeff[i] = delta_xi_xip1 / (binSizes[i] + binSizes[i - 1]);
-				R_coeff[i] = delta_xi_xip1 / binSizes[i];
+				const ParamType delta_xi_xip1 = static_cast<ParamType>(binSizes[i] + binSizes[i + 1]);
+				A_coeff[i] = delta_xi_xip1 / (static_cast<ParamType>(binSizes[i] + binSizes[i - 1]));
+				R_coeff[i] = delta_xi_xip1 / static_cast<ParamType>(binSizes[i]);
 			}
+		}
+
+		// Create a double view of this struct
+		template <typename U>
+		HRKorenParams<U>& as() {
+			return reinterpret_cast<HRKorenParams<U>&>(*this);
+		}
+
+		// Overload for const objects
+		template <typename U>
+		const HRKorenParams<U>& as() const {
+			return reinterpret_cast<const HRKorenParams<U>&>(*this);
 		}
 	};
 
@@ -557,21 +576,33 @@ protected:
 		ParamType q_0 = 0.0;
 		ParamType q_1 = 0.0;
 
-		WENO23Params(const std::vector<ParamType>& binSizes)
+		WENO23Params(const std::vector<active>& binSizes)
 			: q_1_right_coeff(binSizes.size() - 1), q_0_right_coeff(binSizes.size() - 1), C_right_coeff(binSizes.size() - 1), IS_0_coeff(binSizes.size() - 1), IS_1_coeff(binSizes.size() - 1)
 		{
[…]
