### 2025-01-22 7d01269
delete unnecessary type casts
