+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -1091,10 +1091,10 @@ protected:
 				// Flux through left face, modified van Leer flux limiter
 				resCrystal[i] += factor / static_cast<ParamType>(_binSizes[i]) * HRKoren.v_g * HRKoren.F_i - factor * static_cast<ParamType>(_growthDispersionRate) / static_cast<ParamType>(_binSizes[i]) * (yCrystal[i] - yCrystal[i - 1]) / static_cast<ParamType>(_binCenterDists[i - 1]);
 				// Flux through left face, modified van Leer flux limiter, update r_x_i, R_i, and growth rate
-				HRKoren.r_x_i = static_cast<ParamType>(HRKoren.A_coeff[i]) * (yCrystal[i] - yCrystal[i - 1] + 1e-10) / (yCrystal[i + 1] - yCrystal[i] + 1e-10);
+				HRKoren.r_x_i = (HRKoren.A_coeff[i]) * (yCrystal[i] - yCrystal[i - 1] + 1e-10) / (yCrystal[i + 1] - yCrystal[i] + 1e-10);
 				if (cadet_likely(HRKoren.r_x_i > 0))
 				{
-					HRKoren.phi = HRKoren.r_x_i / (static_cast<ParamType>(HRKoren.R_coeff[i]) - 1.0 + HRKoren.r_x_i);
+					HRKoren.phi = HRKoren.r_x_i / ((HRKoren.R_coeff[i]) - 1.0 + HRKoren.r_x_i);
 				}
 				else
 				{
@@ -1109,10 +1109,10 @@ protected:
 				// upwind to F_{1/2} and the diffusion
 				resCrystal[i] += factor / static_cast<ParamType>(_binSizes[i]) * HRKoren.v_g * yCrystal[i - 1] - factor * static_cast<ParamType>(_growthDispersionRate) / static_cast<ParamType>(_binSizes[i]) * (yCrystal[i] - yCrystal[i - 1]) / static_cast<ParamType>(_binCenterDists[i - 1]);
 				// HR scheme applied to F_{1+1/2}
-				HRKoren.r_x_i = static_cast<ParamType>(HRKoren.A_coeff[i]) * (yCrystal[i] - yCrystal[i - 1] + 1e-10) / (yCrystal[i + 1] - yCrystal[i] + 1e-10);
+				HRKoren.r_x_i = (HRKoren.A_coeff[i]) * (yCrystal[i] - yCrystal[i - 1] + 1e-10) / (yCrystal[i + 1] - yCrystal[i] + 1e-10);
 				if (cadet_likely(HRKoren.r_x_i > 0))
 				{
-					HRKoren.phi = HRKoren.r_x_i / (static_cast<ParamType>(HRKoren.R_coeff[i]) - 1.0 + HRKoren.r_x_i);
+					HRKoren.phi = HRKoren.r_x_i / ((HRKoren.R_coeff[i]) - 1.0 + HRKoren.r_x_i);
 				}
 				else
 				{
@@ -1149,14 +1149,14 @@ protected:
 				// flux through left face. W_0, q_0, W_1 and q_1 are coming from the right face of bin (i-1).
[…]
