### 2025-01-22 dd2129d
Change crystallization mode flag enum to bitmap
