+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -25,6 +25,7 @@
 #include <string>
 #include <iterator>
 #include <vector>
+#include <bitset>
 
 namespace cadet
 {
@@ -32,62 +33,22 @@ namespace model
 {
 namespace detail
 {
-	/**
-	* Crystallization mode specification to define which terms are included
-	 */
-	enum class CrystallizationMode : int
-	{
-		/**
-		 * Population mass balance equation
-		*/
-		PurePBM,
+	class CrystallizationMode {
+	public:
+		bool hasPBM()    const { return state_[0]; };
+		bool hasAggregation()  const { return state_[1]; };
+		bool hasFragmentation()  const { return state_[2]; };
+		unsigned int getMode() const { return state_.to_ulong(); };
 
-		/**
-		 * Aggregation equation
-		*/
-		PureAggregation,
+		void SetMode(unsigned int mode)
+		{
+			if (mode > 7 || mode < 1)
+				throw InvalidParameterException("Crystallization mode specified as " + std::to_string(mode) + ", which is not defined.");
 
-		/**
-		 * Fragmentation equation
-		*/
-		PureFragmentation,
-
-		/**
-		 * Combined mass balance and aggregation equation
-		*/
-		PBMAggregation,
-
-		/**
-		 * Combined mass balance and fragmentation equation
-		*/
-		PBMFragmentation,
-
-		/**
-		 * Combined aggregation and fragmentation equation
-		*/
-		AggregationFragmentation,
-
-		/**
-		 * Combined mass balance and aggregation and fragmentation equation
-		*/
-		PBMAggregationFragmentation
-	};
-
-	struct ModeFlags
-	{
-		bool hasMassBalance = false;
-		bool hasAggregation = false;
-		bool hasFragmentation = false;
-	};
-
-	std::unordered_map<CrystallizationMode, ModeFlags> modeFlagMap = {
-		{CrystallizationMode::PurePBM, {true, false, false}},
-		{CrystallizationMode::PureAggregation, {false, true, false}},
-		{CrystallizationMode::PureFragmentation, {false, false, true}},
-		{CrystallizationMode::PBMAggregation, {true, true, false}},
-		{CrystallizationMode::PBMFragmentation, {true, false, true}},
-		{CrystallizationMode::AggregationFragmentation, {false, true, true}},
-		{CrystallizationMode::PBMAggregationFragmentation, {true, true, true}}
+			state_ = std::bitset<3>(mode);
+		}
+	private:
+		std::bitset<3> state_;
 	};
 
 	/**
@@ -263,18 +224,14 @@ public:
 	virtual bool configure(IParameterProvider& paramProvider, UnitOpIdx unitOpIdx, ParticleTypeIdx parTypeIdx)
 	{
 		if (paramProvider.exists("CRY_MODE"))
-			_mode = static_cast<detail::CrystallizationMode>(paramProvider.getInt("CRY_MODE"));
+			_mode.SetMode(paramProvider.getInt("CRY_MODE"));
 		else
-			_mode = detail::CrystallizationMode::PurePBM; // For backwards compatibility
+			_mode.SetMode(1); // For backwards compatibility
 
-		_usePBM = detail::modeFlagMap[_mode].hasMassBalance;
-		_useAgg = detail::modeFlagMap[_mode].hasAggregation;
-		_useFrag = detail::modeFlagMap[_mode].hasFragmentation;
-
-		if (_usePBM)
+		if (_mode.hasPBM())
 		{
 			if (!paramProvider.exists("CRY_NUCLEI_MASS_DENSITY"))
[…]
