### 2025-01-22 60ec59c
Move reconstruction update call to PBM scope
