### 2025-01-24 bbb83bf
Progress of the week: add functionality for consisitent intialisation
consisitent intitalisation does not work for singualr jacobian
