### 2025-01-24 0668b1b
Add functionality for consisitent intialisation
