+++ src/libcadet/model/StirredTankModel.cpp
@@ -242,15 +242,15 @@ bool CSTRModel::configureModelDiscretization(IParameterProvider& paramProvider, 
 		if (_dynReactionBulk->usesParamProviderInDiscretizationConfig())
 			paramProvider.popScope();
 		
-		if (paramProvider.exists("QUASI_STATIONARY_REACTION_BULK"))
+		if (true)//(paramProvider.exists("QUASI_STATIONARY_REACTION_BULK"))
 		{
-			_qsReacBulk = paramProvider.getIntArray("QUASI_STATIONARY_REACTION_BULK");
+			//_qsReacBulk = paramProvider.getIntArray("QUASI_STATIONARY_REACTION_BULK");
 			_qsReacBulk = {1};
 			_nQsReacBulk = _qsReacBulk.size();
 			_temp = new active[_nComp];
 			_temp2 = new double[_nQsReacBulk];
 
-			_nMoitiesBulk = (_nComp + _totalBound) - _nQsReacBulk;
+			_nMoitiesBulk = 1; // Achtung nurs fuers debugging
 			_MconvMoityBulk = Eigen::MatrixXd::Zero(_nMoitiesBulk, _nComp); // matrix for conserved moities
 		}
 		else
@@ -384,7 +384,7 @@ bool CSTRModel::configure(IParameterProvider& paramProvider)
 		dynReactionConfSuccess = _dynReactionBulk->configure(paramProvider, _unitOpIdx, ParTypeIndep);
 		paramProvider.popScope();
 
-		if (paramProvider.exists("QUASI_STATIONARY_REACTION_BULK"))
+		if (true)//(paramProvider.exists("QUASI_STATIONARY_REACTION_BULK"))
 		{
 			_dynReactionBulk->fillConservedMoietiesBulk(_MconvMoityBulk, _qsReacBulk, _QsCompBulk); // fill conserved moities matrix
 		}
@@ -451,6 +451,25 @@ unsigned int CSTRModel::threadLocalMemorySize() const CADET_NOEXCEPT
 	lms.add<double>(_totalBound);
 	lms.commit();
 
+	if (_nQsReacBulk > 0)
+	{
+		// Memory for leanConsistentInitialTimeDerivative()
+		lms.add<int>(_nComp + _totalBound);
+		lms.add<double>(_nComp + _totalBound);
+		lms.add<double>(_nComp);
+		lms.add<double>(_nComp + _totalBound);
+		lms.add<double>(numDofs());
+		lms.add<double>(_nonlinearSolver->workspaceSize(_nComp + _totalBound) * sizeof(double));
+		lms.addBlock(resImplSize);
+		lms.commit();
+		
+		// Memory for leanConsistentInitialState()
+		lms.add<int>(_totalBound);
+		lms.add<double>(_nComp + 1);
+		lms.add<double>(_nComp );
+		lms.add<double>(_totalBound);
+		lms.commit();
+	}
 	return lms.bufferSize();
 }
 
@@ -476,6 +495,8 @@ void CSTRModel::notifyDiscontinuousSectionTransition(double t, unsigned int secI
 	{
 		_curFlowRateFilter = _flowRateFilter[0];
 	}
+
+	_curSecIdx = secIdx;
 }
 
 void CSTRModel::reportSolution(ISolutionRecorder& recorder, double const* const solution) const
@@ -1075,11 +1096,14 @@ void CSTRModel::leanConsistentInitialState(const SimulationTime& simTime, double
 	// It is assumed that the bound states are (approximately) correctly initialized.
 	// Thus, only the liquid phase has to be initialized
 
-			double* const c = vecStateY + _nComp;
-			const double vLiquid = c[_nComp];
[…]
