### 2025-01-31 0b3aaa4
edded leanConsistenInitalisationTimeDerivative, fixed some bugs
