+++ src/libcadet/model/StirredTankModel.cpp
@@ -1264,9 +1264,13 @@ void CSTRModel::leanConsistentInitialState(const SimulationTime& simTime, double
 					for (unsigned int MoityIdx = 0; MoityIdx < _nMoitiesBulk; ++MoityIdx)
 					{
 						double dotProduct = 0.0;
-						for (unsigned int i = 0; i < _MconvMoityBulk.cols(); ++i) // hier Optimierung durch Vermeidung von 0 Zeilen in MconvMoityBulk
-							mat.native(rIdx,i) = _MconvMoityBulk(MoityIdx, i);
-						rIdx++;
+						for (unsigned int i = 0; i < _MconvMoityBulk.cols(); ++i)
+						{// hier Optimierung durch Vermeidung von 0 Zeilen in MconvMoityBulk
+							if (!mask.mask[i])
+								continue;
+							mat.native(rIdx, i) = _MconvMoityBulk(MoityIdx, i);
+							rIdx++;
+						}
 					}
 					return true;
 				};
@@ -1293,9 +1297,13 @@ void CSTRModel::leanConsistentInitialState(const SimulationTime& simTime, double
 				{
 					double dotProduct = 0.0;
 					for (unsigned int j = 0; j < _MconvMoityBulk.cols(); ++j) // hier Optimierung durch Vermeidung von 0 Zeilen in MconvMoityBulk
+					{	
+						if(!mask.mask[j])
+							continue;
 						dotProduct += _MconvMoityBulk(MoityIdx, j) * x[j];
-					r[rIdx] = dotProduct  - conservedQuants[rIdx];
-					rIdx++;
+						r[rIdx] = dotProduct - conservedQuants[rIdx];
+						rIdx++;
+					}
 				}
 
 				std::cout << "Residual: " << std::endl;
@@ -1570,7 +1578,7 @@ int CSTRModel::residualImpl(double t, unsigned int secIdx, StateType const* cons
 		resC[i] += -flowIn * cIn[i] + flowOut * c[i];
 	}
 	
-	if (wantJac)
+	if (wantJac && _nQsReacBulk == 0)
 	{
 		_jac.setAll(0.0);
 
@@ -1599,6 +1607,7 @@ int CSTRModel::residualImpl(double t, unsigned int secIdx, StateType const* cons
 
 	if (_dynReactionBulk && (_nQsReacBulk > 0))
 	{
+		_jac.setAll(0.0);
 
 		Eigen::Map<Eigen::Vector<ResidualType, Eigen::Dynamic>> resCMoities(reinterpret_cast<ResidualType*>(_temp), _nComp);
 		resCMoities.setZero();
@@ -1624,15 +1633,15 @@ int CSTRModel::residualImpl(double t, unsigned int secIdx, StateType const* cons
 					
 					if (wantJac && i == state)
 					{
-						_jac.native(state, state) *= _MconvMoityBulk(MoityIdx, i); // dF_{ci}/dcj = v_liquidDot + F_out  
+						_jac.native(state, state) += _MconvMoityBulk(MoityIdx, i) * (static_cast<double>(vDot) + static_cast<double>(flowOut)); // dF_{ci}/dcj = v_liquidDot + F_out  
 						if (cadet_likely(yDot))
 						{
-							_jac.native(i, _nComp + _totalBound) *= _MconvMoityBulk(MoityIdx, i); // dF/dvliquid = cDot 
+							_jac.native(i, _nComp + _totalBound) += _MconvMoityBulk(MoityIdx, i) * cDot[i]; // dF/dvliquid = cDot 
 						}
 					}
 					if (wantJac && state != i) 
 					{
[…]
