+++ src/libcadet/model/StirredTankModel.cpp
@@ -242,26 +242,16 @@ bool CSTRModel::configureModelDiscretization(IParameterProvider& paramProvider, 
 		if (_dynReactionBulk->usesParamProviderInDiscretizationConfig())
 			paramProvider.popScope();
 		
-		if (true)//(paramProvider.exists("QUASI_STATIONARY_REACTION_BULK"))
-		{
-			//_qsReacBulk = paramProvider.getIntArray("QUASI_STATIONARY_REACTION_BULK");
-			_qsReacBulk = {1};
-			_nQsReacBulk = _qsReacBulk.size();
+
+
 			_temp = new active[_nComp];
-			_temp2 = new double[_nQsReacBulk];
+			_temp2 = new double[_nComp];
 
-			_nMoitiesBulk = 1; // Achtung nurs fuers debugging
-			_MconvMoityBulk = Eigen::MatrixXd::Zero(_nMoitiesBulk, _nComp); // matrix for conserved moities
-		}
-		else
-		{
+			_MconvMoityBulk = Eigen::MatrixXd::Zero(0,0); // matrix for conserved moities
 			_QsCompBulk.clear();	
-			_qsReacBulk.clear();
 			_nMoitiesBulk = 0;
-			_nQsReacBulk = 0;
-			_MconvMoityBulk = Eigen::MatrixXd::Zero(0, 0);
-
-		}
+			_nqsReactionBulk = 0;
+			_qsReactionBulk = nullptr;
 	}
 
 	_dynReaction = std::vector<IDynamicReactionModel*>(_nParType, nullptr);
@@ -384,9 +374,15 @@ bool CSTRModel::configure(IParameterProvider& paramProvider)
 		dynReactionConfSuccess = _dynReactionBulk->configure(paramProvider, _unitOpIdx, ParTypeIndep);
 		paramProvider.popScope();
 
-		if (true)//(paramProvider.exists("QUASI_STATIONARY_REACTION_BULK"))
-		{
-			_dynReactionBulk->fillConservedMoietiesBulk(_MconvMoityBulk, _qsReacBulk, _QsCompBulk); // fill conserved moities matrix
+		// Get quasi stationary reactions information vector
+		_qsReactionBulk = _dynReactionBulk->reactionQuasiStationarity();
+
+		if (_qsReactionBulk != nullptr)
+		{	
+			// resize  Conserved moities matrix
+			
+			_dynReactionBulk->fillConservedMoietiesBulk(_MconvMoityBulk, _nqsReactionBulk, _QsCompBulk); // fill conserved moities matrix
+			_nMoitiesBulk = _MconvMoityBulk.rows();
 		}
 	}
 
@@ -451,7 +447,7 @@ unsigned int CSTRModel::threadLocalMemorySize() const CADET_NOEXCEPT
 	lms.add<double>(_totalBound);
 	lms.commit();
 
-	if (_nQsReacBulk > 0)
+	if (_nqsReactionBulk > 0)
 	{
 		// Memory for leanConsistentInitialTimeDerivative()
 		lms.add<int>(_nComp + _totalBound);
@@ -1100,7 +1096,7 @@ void CSTRModel::leanConsistentInitialState(const SimulationTime& simTime, double
 	const double vLiquid = c[_nComp];
 	const double vSolid = static_cast<double>(_constSolidVolume);
 	
-	if(_nQsReacBulk == 0)
+	if(_nqsReactionBulk == 0)
 		return;
 	// Check if volume is 0
 	if (vLiquid == 0.0)
[…]
