### 2025-02-04 778a447
Add interface for quasi stationary reaction in MAL
little changes
