+++ src/libcadet/model/reaction/MassActionLawReaction.cpp
@@ -26,6 +26,7 @@
 #include <string>
 #include <vector>
 #include <Eigen/Dense>
+#include <iostream>
 
 using namespace Eigen;
 /*<codegen>
@@ -341,6 +342,27 @@ public:
 			_expBulkFwd.resize(nComp, nReactions);
 			_expBulkBwd.resize(nComp, nReactions);
 		}
+		if (paramProvider.exists("IS_KINETIC"))
+		{
+			_reactionQuasistationarity.resize(_stoichiometryBulk.columns(), false);
+
+			const std::vector<int> vecKin = paramProvider.getIntArray("IS_KINETIC");
+			if (vecKin.size() == 1)
+			{
+				// Treat an array with a single element as scalar
+				std::fill(_reactionQuasistationarity.begin(), _reactionQuasistationarity.end(), !static_cast<bool>(vecKin[0]));
+			}
+			else if (vecKin.size() < _reactionQuasistationarity.size())
+			{
+				// Error on too few elements
+				throw InvalidParameterException("IS_KINETIC has to have at least " + std::to_string(_reactionQuasistationarity.size()) + " elements");
+			}
+			else
+			{
+				// Copy what we need (ignore excess values)
+				std::transform(vecKin.begin(), vecKin.begin() + _reactionQuasistationarity.size(), _reactionQuasistationarity.begin(), [](int val) { return !static_cast<bool>(val); });
+			}
+		}
 
 		if (!nBound || !boundOffset)
 			return true;
@@ -377,6 +399,7 @@ public:
 			_expSolidBwdLiquid.resize(nComp, nReactions);
 		}
 
+
 		return true;
 	}
 	
@@ -389,18 +412,18 @@ public:
 	 * @param _reactionQuasistationarity The reaction quasistationarity vector.
 	 * @return The conserved moieties matrix.
 	 */
-	virtual void fillConservedMoietiesBulk(Eigen::MatrixXd& M, std::vector<int>& mapQSReac, std::vector<int>& _QsCompBulk)
+	virtual void fillConservedMoietiesBulk(Eigen::MatrixXd& M, unsigned int& numQSReac, std::vector<int>& _QsCompBulk)
 	{
 
 		//1. get stoichmetic matrix with only reaction in quasi stationary
 		// S dim -> ncomp x nreac
-
-		int nQS = std::count(mapQSReac.begin(), mapQSReac.end(), 1);
+		//_reactionQuasistationarity = {1};
+		numQSReac = std::count(_reactionQuasistationarity.begin(), _reactionQuasistationarity.end(), true);
 		// Count the number of entries with value 1 in _reactionQuasistationarity
-
-
-		Eigen::MatrixXd QSS(_stoichiometryBulk.rows(), nQS);
-
+		
+		M.resize(numQSReac,_nComp);
+		Eigen::MatrixXd QSS(_stoichiometryBulk.rows(), numQSReac);
+		
 
 		// Fülle die Spalten basierend auf _reactionQuasistationarity
 		/*for (int i = 0; i < _stoichiometryBulk.rows(); ++i)
@@ -413,16 +436,13 @@ public:
 
 		for (int i = 0; i < _stoichiometryBulk.rows(); ++i)
 		{
+			int rowIndex = 0;
 			for(int j = 0; j < _stoichiometryBulk.columns(); j++)
 			{
-				int rowIndex = 0;
-				for (int j = 0; j < _stoichiometryBulk.columns(); j++)
-				{
-					if (mapQSReac[j] == 0)
+					if (_reactionQuasistationarity[j] == 0)
 						continue;
[…]
