### 2025-02-06 #7552 @hannahlanzrath
Hello @jedbrown,

Thank you! We are very excited to move forward with the revision.

We are not yet familiar with the JOSS reviewers, but we have identified a few potential candidates from the reviewer search who have expertise in Modeling and Simulation in Chemical or Biotechnology Process Engineering, as well as C++ development and scientific computing:

[Reviewer 335](https://reviewers.joss.theoj.org/reviewers/335)
[Reviewer 1238](https://reviewers.joss.theoj.org/reviewers/1238)
[Reviewer 1349](https://reviewers.joss.theoj.org/reviewers/1349)
[Reviewer 1767](https://reviewers.joss.theoj.org/reviewers/1767)
