### 2025-02-13 be69201
add characterisation map for dynamic, conserv and algebraic state entries
