+++ src/libcadet/model/StirredTankModel.cpp
@@ -30,7 +30,7 @@
 
 #include <algorithm>
 #include <functional>
-
+#include <bitset>
 #include <Eigen/Dense>
 
 #include<iostream>
@@ -243,7 +243,6 @@ bool CSTRModel::configureModelDiscretization(IParameterProvider& paramProvider, 
 			paramProvider.popScope();
 		
 
-
 			_temp = new active[_nComp];
 			_temp2 = new double[_nComp];
 
@@ -377,10 +376,32 @@ bool CSTRModel::configure(IParameterProvider& paramProvider)
 		// Get quasi stationary reactions information vector
 		_qsReactionBulk = _dynReactionBulk->reactionQuasiStationarity();
 
-		if (_qsReactionBulk != nullptr)
+		if (false)
 		{	
 			_dynReactionBulk->fillConservedMoietiesBulk(_MconvMoityBulk, _nqsReactionBulk, _QsCompBulk); // fill conserved moities matrix
 			_nMoitiesBulk = _MconvMoityBulk.rows();
+			
+			int mIdx = 0;
+			for (int state = 0; state < _nComp; state++)
+			{
+				std::bitset<3> c;
+				if (_QsCompBulk[state] == 0) // state is dynamic 
+				{
+					c.set(0);
+					stateMap.push_back(c);
+				}
+				else if(mIdx < _nMoitiesBulk) // state is dynamic and calculated with conserved moities
+				{
+					c.set(1);
+					stateMap.push_back(c);
+					mIdx++;
+				}
+				else if(mIdx >= _nMoitiesBulk) // state algebraic 
+				{
+					c.set(2);
+					stateMap.push_back(c);
+				}
+			}
 		}
 	}
 
@@ -1161,8 +1182,16 @@ void CSTRModel::leanConsistentInitialState(const SimulationTime& simTime, double
 		LinearBufferAllocator tlmAlloc = threadLocalMem.get();
 		BufferedArray<int> qsMask = tlmAlloc.array<int>(_nComp);
 		
+		for(int state = 0 ; state < _nComp; state++)
+		{	
+			if(stateMap[state].test(1) || stateMap[state].test(2))
+				qsMask[state] = 1;
+			else
+				qsMask[state] = 0;
+		}
+		
 		// Mark components with quasi-stationary reactions partition
-		qsMask.copyFromVector(_QsCompBulk);
+		//qsMask.copyFromVector(_QsCompBulk);
 
 		const linalg::ConstMaskArray mask{ static_cast<int*>(qsMask), static_cast<int>(_nComp) };
 		const int probSize = linalg::numMaskActive(mask);
@@ -1179,8 +1208,9 @@ void CSTRModel::leanConsistentInitialState(const SimulationTime& simTime, double
 		for (unsigned int MoityIdx = 0; MoityIdx < _nMoitiesBulk; ++MoityIdx)
 		{
 			double dotProduct = 0.0;
-			for (unsigned int i = 0; i < _MconvMoityBulk.cols(); ++i) 
+			for (unsigned int i = 0; i < _MconvMoityBulk.cols(); ++i)
 				dotProduct += _MconvMoityBulk(MoityIdx, i) * c[i];
+			
 			conservedQuants[MoityIdx] = dotProduct;
 		}
 	
@@ -1255,25 +1285,27 @@ void CSTRModel::leanConsistentInitialState(const SimulationTime& simTime, double
 					mat.setAll(0.0);
 					linalg::copyMatrixSubset(_jac, mask, mask, mat);
 					// Replace upper part with conservation relations
[…]
