### 2025-02-13 88c4ece
Add characterisation map
