+++ src/libcadet/model/reaction/MassActionLawReaction.cpp
@@ -346,23 +346,32 @@ public:
 		{
 			_reactionQuasistationarity.resize(_stoichiometryBulk.columns(), false);
 
-			const std::vector<int> vecKin = paramProvider.getIntArray("IS_KINETIC");
-			if (vecKin.size() == 1)
-			{
-				// Treat an array with a single element as scalar
-				std::fill(_reactionQuasistationarity.begin(), _reactionQuasistationarity.end(), !static_cast<bool>(vecKin[0]));
-			}
-			else if (vecKin.size() < _reactionQuasistationarity.size())
-			{
-				// Error on too few elements
-				throw InvalidParameterException("IS_KINETIC has to have at least " + std::to_string(_reactionQuasistationarity.size()) + " elements");
+			if (paramProvider.isArray("IS_KINETIC")) {
+				const std::vector<int> vecKin = paramProvider.getIntArray("IS_KINETIC");
+				int numqsReaction = std::count(vecKin.begin(), vecKin.end(), 1);
+				if (vecKin.size() == 1)
+				{
+					// Treat an array with a single element as scalar
+					std::fill(_reactionQuasistationarity.begin(), _reactionQuasistationarity.end(), !static_cast<bool>(vecKin[0]));
+				}
+				else if (vecKin.size() < _reactionQuasistationarity.size())
+				{
+					// Error on too few elements
+					throw InvalidParameterException("IS_KINETIC has to have at least " + std::to_string(_reactionQuasistationarity.size()) + " elements");
+				}
+				else
+				{
+					// Copy what we need (ignore excess values)
+					std::transform(vecKin.begin(), vecKin.begin() + _reactionQuasistationarity.size(), _reactionQuasistationarity.begin(), [](int val) { return !static_cast<bool>(val); });
+				}
 			}
 			else
 			{
-				// Copy what we need (ignore excess values)
-				std::transform(vecKin.begin(), vecKin.begin() + _reactionQuasistationarity.size(), _reactionQuasistationarity.begin(), [](int val) { return !static_cast<bool>(val); });
+				const bool kineticBinding = paramProvider.getInt("IS_KINETIC");
+				std::fill(_reactionQuasistationarity.begin(), _reactionQuasistationarity.end(), !kineticBinding);
 			}
 		}
+		
 
 		if (!nBound || !boundOffset)
 			return true;
@@ -417,23 +426,19 @@ public:
 
 		//1. get stoichmetic matrix with only reaction in quasi stationary
 		// S dim -> ncomp x nreac
-		//_reactionQuasistationarity = {1};
-		numQSReac = std::count(_reactionQuasistationarity.begin(), _reactionQuasistationarity.end(), true);
+
 		// Count the number of entries with value 1 in _reactionQuasistationarity
+		numQSReac = std::count(_reactionQuasistationarity.begin(), _reactionQuasistationarity.end(), true);
 		
+		if (numQSReac == 0)
+			return;
+
 		M.resize(numQSReac,_nComp);
-		Eigen::MatrixXd QSS(_stoichiometryBulk.rows(), numQSReac);
+
+		// QSS quasistationary stoichiometry matrix contains only coloums with qs reactions
+		Eigen::MatrixXd QSS(_stoichiometryBulk.rows(), numQSReac); 
 		
-
-		// Fülle die Spalten basierend auf _reactionQuasistationarity
[…]
