### 2025-02-14 18d78da
add state track system to differ between moities and dynamic states
