+++ src/libcadet/model/StirredTankModel.cpp
@@ -376,30 +376,32 @@ bool CSTRModel::configure(IParameterProvider& paramProvider)
 		// Get quasi stationary reactions information vector
 		_qsReactionBulk = _dynReactionBulk->reactionQuasiStationarity();
 
-		if (false)
+		if (_qsReactionBulk != nullptr)
 		{	
 			_dynReactionBulk->fillConservedMoietiesBulk(_MconvMoityBulk, _nqsReactionBulk, _QsCompBulk); // fill conserved moities matrix
 			_nMoitiesBulk = _MconvMoityBulk.rows();
-			
-			int mIdx = 0;
-			for (int state = 0; state < _nComp; state++)
+			if (_nMoitiesBulk != 0)
 			{
-				std::bitset<3> c;
-				if (_QsCompBulk[state] == 0) // state is dynamic 
+				int mIdx = 0;
+				for (int state = 0; state < _nComp; state++)
 				{
-					c.set(0);
-					stateMap.push_back(c);
-				}
-				else if(mIdx < _nMoitiesBulk) // state is dynamic and calculated with conserved moities
-				{
-					c.set(1);
-					stateMap.push_back(c);
-					mIdx++;
-				}
-				else if(mIdx >= _nMoitiesBulk) // state algebraic 
-				{
-					c.set(2);
-					stateMap.push_back(c);
+					std::bitset<3> c;
+					if (_QsCompBulk[state] == 0) // state is dynamic 
+					{
+						c.set(0);
+						stateMap.push_back(c);
+					}
+					else if (mIdx < _nMoitiesBulk) // state is dynamic and calculated with conserved moities
+					{
+						c.set(1);
+						stateMap.push_back(c);
+						mIdx++;
+					}
+					else if (mIdx >= _nMoitiesBulk) // state algebraic 
+					{
+						c.set(2);
+						stateMap.push_back(c);
+					}
 				}
 			}
 		}
@@ -1294,14 +1296,18 @@ void CSTRModel::leanConsistentInitialState(const SimulationTime& simTime, double
 						{
 							double dotProduct = 0.0;
 							int j = 0;
-							for (unsigned int i = 0; i < _MconvMoityBulk.cols(); ++i) {
+							for (unsigned int i = 0; i < _MconvMoityBulk.cols(); ++i) 
+							{
 								if (_QsCompBulk[i]==0)
 									continue;
 								mat.native(rIdx, j) = _MconvMoityBulk(MoityIdx, i);
 								j++;
 							}
 							rIdx++;
+							MoityIdx++;
 						}
+						else if (stateMap[state].test(2))
+							rIdx++;
 					}
 					return true;
 				};
@@ -1329,12 +1335,20 @@ void CSTRModel::leanConsistentInitialState(const SimulationTime& simTime, double
 					if (stateMap[state].test(1))
 					{
 						double dotProduct = 0.0;
+						int jIdx = 0;
 						for (unsigned int j = 0; j < _MconvMoityBulk.cols(); ++j)
-							dotProduct += _MconvMoityBulk(MoityIdx, j) * x[j];
+						{
+							if (_QsCompBulk[j] == 0)
+								continue;
+							dotProduct += _MconvMoityBulk(MoityIdx, j) * x[jIdx];
+							jIdx++;
+						}
 						r[rIdx] = dotProduct - conservedQuants[MoityIdx];
 						MoityIdx++;
 						rIdx++;
 					}
+					else if(stateMap[state].test(2))
+						rIdx++;
 				}
 				
[…]
