+++ src/libcadet/model/reaction/MassActionLawReaction.cpp
@@ -346,29 +346,32 @@ public:
 		{
 			_reactionQuasistationarity.resize(_stoichiometryBulk.columns(), false);
 
-			const std::vector<int> vecKin = paramProvider.getIntArray("IS_KINETIC");
-			int numqsReaction = std::count(vecKin.begin(), vecKin.end(), 1);
-			if (numqsReaction == 0)
-			{
-				_reactionQuasistationarity = {};
+			if (paramProvider.isArray("IS_KINETIC")) {
+				const std::vector<int> vecKin = paramProvider.getIntArray("IS_KINETIC");
+				int numqsReaction = std::count(vecKin.begin(), vecKin.end(), 1);
+				if (vecKin.size() == 1)
+				{
+					// Treat an array with a single element as scalar
+					std::fill(_reactionQuasistationarity.begin(), _reactionQuasistationarity.end(), !static_cast<bool>(vecKin[0]));
+				}
+				else if (vecKin.size() < _reactionQuasistationarity.size())
+				{
+					// Error on too few elements
+					throw InvalidParameterException("IS_KINETIC has to have at least " + std::to_string(_reactionQuasistationarity.size()) + " elements");
+				}
+				else
+				{
+					// Copy what we need (ignore excess values)
+					std::transform(vecKin.begin(), vecKin.begin() + _reactionQuasistationarity.size(), _reactionQuasistationarity.begin(), [](int val) { return !static_cast<bool>(val); });
+				}
 			}
-			else if (vecKin.size() == 1)
-			{	
-				// Treat an array with a single element as scalar
-				std::fill(_reactionQuasistationarity.begin(), _reactionQuasistationarity.end(), !static_cast<bool>(vecKin[0]));
-			}
-			else if (vecKin.size() < _reactionQuasistationarity.size())
+			else
 			{
-				// Error on too few elements
-				throw InvalidParameterException("IS_KINETIC has to have at least " + std::to_string(_reactionQuasistationarity.size()) + " elements");
-			}
-			else if (numqsReaction != 0)
-			{
-				// Copy what we need (ignore excess values)
-				std::transform(vecKin.begin(), vecKin.begin() + _reactionQuasistationarity.size(), _reactionQuasistationarity.begin(), [](int val) { return !static_cast<bool>(val); });
-
+				const bool kineticBinding = paramProvider.getInt("IS_KINETIC");
+				std::fill(_reactionQuasistationarity.begin(), _reactionQuasistationarity.end(), !kineticBinding);
 			}
 		}
+		
 
 		if (!nBound || !boundOffset)
 			return true;
@@ -465,12 +468,15 @@ public:
 		if (_nComp - nQScomp  > 0 ) // if kinetic components exists we can resize QSS
 		{
 			Eigen::MatrixXd QSSCompressed(nQScomp, QSS.cols());
+			int idx = 0;
 			for (std::size_t i = 0; i < _QsCompBulk.size(); ++i)
-			{	
-				if (_QsCompBulk[i] == 1)
-					QSSCompressed.row(i) = QSS.row(i);
+			{
+				if (_QsCompBulk[i] == 0)
+					continue;
+				QSSCompressed.row(idx) = QSS.row(i);
+				idx++;
 			}
-			QSS.swap(QSSCompressed); // Swap the compressed matrix back into QSS
+			QSS.swap(QSSCompressed);
 		}
 
 		//Test if the matrix is full rank
@@ -725,17 +731,17 @@ protected:
[…]
