### 2025-02-19 a154e4d
make residual implementation compatible for AD
