### 2025-02-19 85df024
Add AD functionallity
