+++ src/libcadet/model/StirredTankModel.hpp
@@ -30,6 +30,7 @@
 
 namespace cadet
 {
+	struct ColumnPosition;
 
 namespace model
 {
@@ -77,7 +78,9 @@ public:
 
 	virtual int residual(const SimulationTime& simTime, const ConstSimulationState& simState, double* const res, util::ThreadLocalStorage& threadLocalMem);
 	virtual int residual(const SimulationTime& simTime, const ConstSimulationState& simState, double* const res, const AdJacobianParams& adJac, util::ThreadLocalStorage& threadLocalMem, bool updateJacobian, bool paramSensitivity);
-
+	
+	template <typename StateType, typename ResidualType, typename ParamType, bool wantJac>
+	void applyConservedMoitiesBulk(double t, unsigned int secIdx, const ColumnPosition& colPos, StateType const* const y, double const* const yDot, ResidualType* const resC, LinearBufferAllocator tlmAlloc);
 	virtual int jacobian(const SimulationTime& simTime, const ConstSimulationState& simState, double* const res, const AdJacobianParams& adJac, util::ThreadLocalStorage& threadLocalMem);
 	virtual int residualWithJacobian(const SimulationTime& simTime, const ConstSimulationState& simState, double* const res, const AdJacobianParams& adJac, util::ThreadLocalStorage& threadLocalMem);
 	virtual int residualSensFwdAdOnly(const SimulationTime& simTime, const ConstSimulationState& simState, active* const adRes, util::ThreadLocalStorage& threadLocalMem);
@@ -172,15 +175,11 @@ protected:
 
 	IDynamicReactionModel* _dynReactionBulk; //!< Dynamic reactions in the bulk volume
 
-	active* _temp;
-	double* _temp2;
-
 	Eigen::MatrixXd _MconvMoityBulk; //!<  Matrix with conservation of moieties in the bulk volume
-	int const* _qsReactionBulk; //!< Indices of components that are not conserved in the bulk volume
+	int const* _qsReactionBulk; //!< Indices of reactions that are not conserved in the bulk volume
+	bool _hasQuasiStationaryReactionBulk; //!< Flag that determines whether there are quasi-stationary reactions in the bulk volume
 	std::vector<int> _QsCompBulk; //!< Indices of components that are conserved in the bulk volume
-	unsigned int _nqsReactionBulk;
-	int _nMoitiesBulk;
-	std::vector<std::bitset<3>> stateMap; // !< Bitset that tracks the properties of the state vector (dynamic, moities, algebraic)
+	std::vector<std::bitset<3>> _stateMap; // !< Bitset that tracks the properties of the state vector (dynamic, moities, algebraic)
 
 	class Exporter : public ISolutionExporter
 	{
