### 2025-02-19 b76cc6c
Update AUTHORS.md
