+++ AUTHORS.md
@@ -1,8 +1,8 @@
 # Contributors
 
-The following people have contributed to the CADET-Core software and are listed in historical order. Together they form the "CADET-Core authors" as mentioned in the copyright statements and [LICENSE.md](LICENSE.txt) file.
+The following individuals have made significant contributions to the CADET-Core software and are listed in *chronological order*. Together they form the "CADET-Core authors" as mentioned in the copyright statements and [LICENSE.md](LICENSE.txt) file.
 
-Major additions or modifications to CADET-Core are explicitly stated.
+Major additions or modifications to CADET-Core are explicitly documented here.
 
 ## Special thanks to everyone who has helped with this project
 
@@ -16,7 +16,7 @@ Major additions or modifications to CADET-Core are explicitly stated.
 * [Johannes Schmölder](https://github.com/schmoelder) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany)
 * [Jayghosh Rao](https://github.com/jayghoshter) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany)
 * [Jazib Hassan](https://github.com/jazib-hassan-juelich) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany)
-* [Jan Breuer](https://github.com/jbreue16) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany): Spatial DG discretizations (LRM, LRMP, GRM)
+* [Jan Michael Breuer](https://github.com/jbreue16) (University of Cologne, Faculty of Mathematics and Natural Sciences, Cologne, Germany and Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany): Spatial DG discretizations (LRM, LRMP, GRM), Crystallization module
 * [Ronald Jäpel](https://github.com/ronald-jaepel) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany)
 * [Hannah Lanzrath](https://github.com/hannahlanzrath) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany)
 * [Antonia Berger](https://github.com/AntoniaBerger) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany)
@@ -24,6 +24,7 @@ Major additions or modifications to CADET-Core are explicitly stated.
 
 ## Funding Acknowledgement
 
-* GE Healthcare sponsored development of core-shell beads, multiple particle types, and 2D general rate model
-* Bayer AG contributed to bugfixes, initial C API, initial Python frontend
+* GE Healthcare sponsored the development of core-shell beads, multiple particle types, and 2D general rate model.
+* Bayer AG contributed to bug fixes, initial C API, initial Python frontend.
 * Johannes Schmölder has received support from the IMI2/ EU/EFPIA joint undertaking Inno4Vac (grant no. 101007799).
+* Jan Michael Breuer has received funding by the Deutsche Forschungsgemeinschaft (DFG, German Research Foundation) – 548805630.
