### 2025-02-20 98e41ef
Update AUTHORS.md
Matlab -> MATLAB
