### 2025-02-20 c4ef894
Update AUTHORS.md
Matlab -> MATLAB
