### 2025-02-21 02c1aea
Fix jacobian and residual double initalisation
