+++ src/libcadet/model/StirredTankModel.cpp
@@ -241,17 +241,15 @@ bool CSTRModel::configureModelDiscretization(IParameterProvider& paramProvider, 
 
 		if (_dynReactionBulk->usesParamProviderInDiscretizationConfig())
 			paramProvider.popScope();
-		
-
-			_temp = new active[_nComp];
-			_temp2 = new double[_nComp];
-
+	
 			_MconvMoityBulk = Eigen::MatrixXd::Zero(0,0); // matrix for conserved moities
 			_QsCompBulk.clear();	
 			_nMoitiesBulk = 0;
-			_nqsReactionBulk = 0;
 			_qsReactionBulk = nullptr;
 	}
+	_nqsReactionBulk = 0;
+	_temp = new active[_nComp];
+	_temp2 = new double[_nComp];
 
 	_dynReaction = std::vector<IDynamicReactionModel*>(_nParType, nullptr);
 
@@ -470,23 +468,20 @@ unsigned int CSTRModel::threadLocalMemorySize() const CADET_NOEXCEPT
 
 	if (_nqsReactionBulk > 0)
 	{
-		// Memory for leanConsistentInitialTimeDerivative() 
+		// Memory for ConsistentInitialTimeDerivative() 
 		lms.add<int>(_nComp); // map of kinetic components 
 		lms.add<double>(_nComp); // solution 
 		lms.add<double>(_nComp); // conserved quantities
-		//lms.add<double>(_nComp + _totalBound);
 		lms.add<double>(numDofs()); // fullX 
 		lms.add<double>(numDofs()); // fullRes 
 		lms.add<double>(_nonlinearSolver->workspaceSize(_nComp + _totalBound) * sizeof(double)); //non linear solver workspace
 		lms.addBlock(resImplSize);
 		lms.commit();
 		
-		// Memory for leanConsistentInitialState()
-		//lms.add<int>(_totalBound);
-		lms.add<double>(_nComp); // dReacDt
-		lms.add<double>(_nComp ); //map
+		// Memory for ConsistentInitialTimeDerivative()
+		lms.add<double>(_nComp+1); // dReacDt
+		lms.add<double>(_nComp); //map
 		lms.add<double>(_nComp); //additional puffer
-		//lms.add<double>(_totalBound);
 		lms.commit();
 	}
 	return lms.bufferSize();
@@ -680,6 +675,200 @@ void CSTRModel::consistentInitialState(const SimulationTime& simTime, double* co
 				c[i] -= qFactor * static_cast<double>(_parTypeVolFrac[type]) * qSum;
 			}
 		}
+	}
+	if (_nqsReactionBulk > 0)
+	{
+
+		LinearBufferAllocator tlmAlloc = threadLocalMem.get();
+		BufferedArray<int> qsMask = tlmAlloc.array<int>(_nComp);
+
+		for (int state = 0; state < _nComp; state++)
+		{
+			if (stateMap[state].test(1) || stateMap[state].test(2))
+				qsMask[state] = 1;
+			else
+				qsMask[state] = 0;
+		}
+
+		// Mark components with quasi-stationary reactions partition
+		//qsMask.copyFromVector(_QsCompBulk);
+
+		const linalg::ConstMaskArray mask{ static_cast<int*>(qsMask), static_cast<int>(_nComp) };
+		const int probSize = linalg::numMaskActive(mask);
+
+		// Extract initial values from current state
+		BufferedArray<double> solution = tlmAlloc.array<double>(probSize);
+		linalg::selectVectorSubset(c, mask, static_cast<double*>(solution));
+
+		// Save values of conserved moieties;
+		const unsigned int numActiveComp = numMaskActive(mask, _nComp);
+		BufferedArray<double> conservedQuants = tlmAlloc.array<double>(_nMoitiesBulk);
+
[…]
