### 2025-02-21 8bff4b1
Add timeout for simulations via callback
Adds a timeout for simulations that is implemented as a callback.
Notifications are now also sent whenever a linear system is solved,
which occurs during Newton iteration.

The timeout is exposed via the C-API.
