### 2025-02-21 5f533ae
Add timeout for simulations via callback
Adds a timeout for simulations that is implemented as a callback.
Notifications are now also sent whenever a linear system is solved,
which occurs during Newton iteration.

The timeout is exposed via the C-API.
