### 2025-02-21 7b708ee
Add C-API return code for uninitialized simulator
Adds a dedicated return code for situations where the
underlying simulator has not been initialized before
calling the API method.
