+++ include/cadet/cadet.h
@@ -165,6 +165,10 @@ extern "C"
 	enum cdtErrors
 	{
 		/**
+		 * @brief Simulator not initialized
+		 */
+		cdtSimulatorNotInitialized = -4,
+		/**
 		 * @brief Requested data has not been stored / is not available
 		 */
 		cdtDataNotStored = -3,
@@ -401,6 +405,7 @@ extern "C"
 		 * @details Before this function is called, a simulation has to be run successfully.
 		 * @param [in] drv Driver handle
 		 * @param [out] nUnits Number of unit operations
+		 * @return @c cdtOK on success, a negative value indicating the error otherwise
 		 */
 		cdtResult (*getNumUnitOp)(cdtDriver* drv, int* nUnits);
 
@@ -410,6 +415,7 @@ extern "C"
 		 * @param [in] drv Driver handle
 		 * @param [in] unitOpId ID of the unit operation whose solution is returned
 		 * @param [out] nParTypes Number of particle types
+		 * @return @c cdtOK on success, a negative value indicating the error otherwise
 		 */
 		cdtResult (*getNumParTypes)(cdtDriver* drv, int unitOpId, int* nParTypes);
 
@@ -418,6 +424,7 @@ extern "C"
 		 * @details Before this function is called, a simulation has to be run successfully.
 		 * @param [in] drv Driver handle
 		 * @param [out] nSens Number of parameter sensitivities
+		 * @return @c cdtOK on success, a negative value indicating the error otherwise
 		 */
 		cdtResult (*getNumSensitivities)(cdtDriver* drv, int* nSens);
 
@@ -432,6 +439,7 @@ extern "C"
 		 * @param [out] nTime Number of time points
 		 * @param [out] nPort Number of ports
 		 * @param [out] nComp Number of components
+		 * @return @c cdtOK on success, a negative value indicating the error otherwise
 		 */
 		cdtResult (*getSolutionInlet)(cdtDriver* drv, int unitOpId, double const** time, double const** data, int* nTime, int* nPort, int* nComp);
 
@@ -446,6 +454,7 @@ extern "C"
 		 * @param [out] nTime Number of time points
 		 * @param [out] nPort Number of ports
 		 * @param [out] nComp Number of components
+		 * @return @c cdtOK on success, a negative value indicating the error otherwise
 		 */
 		cdtResult (*getSolutionOutlet)(cdtDriver* drv, int unitOpId, double const** time, double const** data, int* nTime, int* nPort, int* nComp);
 
@@ -462,6 +471,7 @@ extern "C"
 		 * @param [out] nRadialCells Number of radial cells
 		 * @param [out] nComp Number of components
 		 * @param [out] keepAxialSingletonDimension Keep axial singleton dimension
+		 * @return @c cdtOK on success, a negative value indicating the error otherwise
 		 */
 		cdtResult (*getSolutionBulk)(cdtDriver* drv, int unitOpId, double const** time, double const** data, int* nTime, int* nAxialCells, int* nRadialCells, int* nComp, bool* keepAxialSingletonDimension);
 
@@ -481,6 +491,7 @@ extern "C"
 		 * @param [out] nComp Number of components
 		 * @param [out] keepAxialSingletonDimension Keep axial singleton dimension
 		 * @param [out] keepParticleSingletonDimension Keep particle singleton dimension
+		 * @return @c cdtOK on success, a negative value indicating the error otherwise
 		 */
 		cdtResult (*getSolutionParticle)(cdtDriver* drv, int unitOpId, int parType, double const** time, double const** data, int* nTime, int* nAxialCells, int* nRadialCells, int* nParShells, int* nComp, bool* keepAxialSingletonDimension, bool* keepParticleSingletonDimension);
 		/**
[…]
