### 2025-02-21 000c120
Fix stop condition in linear solve callback check
