+++ src/libcadet/SimulatorImpl.cpp
@@ -329,7 +329,7 @@ namespace cadet
 		if (sim->_notification)
 		{
 			const unsigned int secIdx = sim->getCurrentSection(t);
-			if (sim->_notification->timeIntegrationLinearSolve(secIdx, t, NVEC_DATA(y), NVEC_DATA(yDot)))
+			if (!sim->_notification->timeIntegrationLinearSolve(secIdx, t, NVEC_DATA(y), NVEC_DATA(yDot)))
 				return IDA_TOO_MUCH_WORK;
 		}
 
