### 2025-02-21 c4646be
Move callbacks to Driver
Add a MultiCallback that calls multiple sub-callbacks. Moves callback
handling to the Driver to avoid duplication.
