+++ include/cadet/cadet.h
@@ -1058,11 +1058,11 @@ extern "C"
 		 * @brief Set a timeout on the current simulator
 		 * @details The timeout persists for the full lifetime of the simulator.
 		 * @param [in] drv Driver handle
-		 * @param [in] timeOut Timeout for a time integration process in seconds. A value
+		 * @param [in] timeoutSec Timeout for a time integration process in seconds. A value
 		 *                     smaller or equal to @c 0.0 means no timeout (default).
 		 * @return @c cdtOK on success, a negative value indicating the error otherwise
 		 */
-		cdtResult (*setTimeout)(cdtDriver* drv, double timeOut);
+		cdtResult (*setTimeout)(cdtDriver* drv, double timeoutSec);
 
 	} cdtAPIv010000;
 
