### 2025-02-21 748eb59
Move callbacks to Driver
Add a MultiCallback that calls multiple sub-callbacks. Moves callback
handling to the Driver to avoid duplication.
