### 2025-02-21 bf31b83
Add TIMEOUT field to solver config in file format
