### 2025-02-26 #7552 @jedbrown
@ctdegroot @csadorf @bdice 👋 Hi, would you be available to review this submission for JOSS?
