### 2025-02-28 #7552 @bdice
@jedbrown Yes, I could take a review here. (It looks like the paper is short and in pretty good shape already, which makes it easier to say yes. Thanks, authors.)
