### 2025-03-05 8fdee6b
New caclulation version with eigen stare matrix multiplication instead of iteration
simple example works without consistent initalisation

fix conserved moities calculation
