+++ src/libcadet/model/StirredTankModel.cpp
@@ -32,6 +32,7 @@
 #include <functional>
 #include <bitset>
 #include <Eigen/Dense>
+#include <Eigen/Sparse>
 
 #include<iostream>
 
@@ -372,6 +373,7 @@ bool CSTRModel::configure(IParameterProvider& paramProvider)
 		if (_qsReactionBulk != nullptr)
 		{	
 			_dynReactionBulk->fillConservedMoietiesBulk(_MconvMoityBulk, _QsCompBulk); // fill conserved moities matrix
+
 			int nMoitiesBulk = _MconvMoityBulk.rows();
 			if (nMoitiesBulk != 0)
 			{
@@ -1656,6 +1658,110 @@ void CSTRModel::applyConservedMoitiesBulk(double t, unsigned int secIdx, const C
 }
 
 template <typename StateType, typename ResidualType, typename ParamType, bool wantJac>
+void CSTRModel::applyConservedMoitiesBulk2(double t, unsigned int secIdx, const ColumnPosition& colPos, StateType const* const c, double const* const yDot, ResidualType* const resC, LinearBufferAllocator tlmAlloc)
+{
+	// prepare memory
+	LinearBufferAllocator subAlloc = tlmAlloc.manageRemainingMemory();
+
+	// create map to residual vector of concentrations
+	Eigen::Map<Eigen::Vector<ResidualType, Eigen::Dynamic>> mapResC(resC, _nComp);
+
+	// calculate flux for quasi stationary reactions
+	BufferedArray<ResidualType> temp2 = subAlloc.array<ResidualType>(_nComp);
+	Eigen::Map<Eigen::Vector<ResidualType, Eigen::Dynamic>> qsFlux(static_cast<ResidualType*>(temp2), _dynReactionBulk->numReactionsLiquid());
+	qsFlux.setZero();
+
+	_dynReactionBulk->computeQuasiStationaryReactionFlux(t, secIdx, colPos, c, qsFlux, _qsReactionBulk, subAlloc);
+
+	// calculate conserved moities
+	Eigen::Matrix<ResidualType, Eigen::Dynamic, Eigen::Dynamic> M(_nComp, _nComp);
+	_dynReactionBulk->fillConservedMoietiesBulk2(M, _QsCompBulk); // fill conserved moities matrix (alternative method)
+
+
+	// buffer memory for transformed residual
+	BufferedArray<ResidualType> temp = subAlloc.array<ResidualType>(_nComp);
+	Eigen::Map<Eigen::Vector<ResidualType, Eigen::Dynamic>> resCWithMoities(static_cast<ResidualType*>(temp), _nComp);
+	resCWithMoities.setZero();
+
+
+	// multiply conserved moities matrix with residual
+	resCWithMoities = M * mapResC;
+
+	// add quasi stationary reaction to residium
+	const int nQsReac = _dynReactionBulk->numReactionQuasiStationary();
+	int state = M.rows() - 1 ;
+	for (int qsReac = 0; qsReac < nQsReac; qsReac++)
+	{
+		if (state < _nComp)
+		{
+			resCWithMoities[state] = qsFlux[qsReac];
+			state++;
+		}
+		else
+			throw InvalidParameterException(
+				"Residual implementation with conserved moities: Too many quasi stationary reactions detected. "
+				"Please check the implementation of the model."
+			);
+	}
+
+	mapResC = resCWithMoities;
+
+	// multiply conserved moities matrix with jacobian
+	if (wantJac)
+	{
+		// transform _jac into Eigen sparse matrix 
+		Eigen::SparseMatrix<double> jacSparse(_nComp, _nComp);
+
+		// Copy data from original Jacobian
[…]
