+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -346,6 +346,9 @@ namespace cadet
 
 			virtual unsigned int numReactionsLiquid() const CADET_NOEXCEPT { return 1; }
 			virtual unsigned int numReactionsCombined() const CADET_NOEXCEPT { return 1; }
+			virtual unsigned int numReactionQuasiStationary() const CADET_NOEXCEPT { return 0; }
+			template <typename ResidualType>
+			void fillConservedMoietiesBulk21(Eigen::Matrix<ResidualType, Eigen::Dynamic, Eigen::Dynamic>& M, std::vector<int>& QsCompBulk) {}
 
 			CADET_DYNAMICREACTIONMODEL_BOILERPLATE
 

[…]
