### 2025-03-07 #7881 @editorialbot
[REVIEW]: CADET-Core 5.0: High-Performance Solver for Advanced Biotechnology Process Modeling
**Submitting author:** @lieres (<a href="http://orcid.org/0000-0002-0309-8408">Eric von Lieres</a>)
**Repository:** https://github.com/cadet/CADET-Core
**Branch with paper.md** (empty if default branch): paper/JOSS_2024
**Version:** v5.0.4
