### 2025-03-09 cadet/CADET-Core#371 @goxberry
JOSS review meta-issue
This issue is a placeholder meta-issue for use in aggregating issues related to reviewing CADET-Core's JOSS paper (see https://github.com/openjournals/joss-reviews/issues/7881#issue-2902032647)l.
