### 2025-03-09 cadet/CADET-Core#373 @goxberry
JOSS Review Scope Question: Examples require CADET-Process or CADET-Python
Are there plans to include a CADET-Core example in the documentation that does not require at least one other CADET software package?

Part of the JOSS review checklist asks me whether there are examples of using CADET-Core in the documentation, and as far as I can tell, the only examples in the documentation as of commit https://github.com/cadet/CADET-Core/commit/e054d22f59a2835552bb2cb800892f82510769db require CADET-Process or CADET-Python. While these examples use CADET-Core, the JOSS submission as of commit https://github.com/cadet/CADET-Core/commit/ba29484e4d763ceb57084af364f9a06e5f76ac8c makes it clear that both CADET-Process and CADET-Python are frontends of CADET-Core that are not part of the core functionality of CADET-Core. As a reviewer, this situation puts me in an awkward position where I'd need to install at least one of those frontends to review an example, but I can't currently install either in a reasonable amount of effort due to [#372](https://github.com/cadet/CADET-Core/issues/372).

I can build CADET-Core from source, and I can run the CADET-Core CLI on at least one input file from the `test/data` directory, so if there were an example based on one of those files, along with some accompanying Python source code to plot the output (e.g., use `h5py` to read HDF5 output, plot the output with `matplotlib`, or do something similar with XML output), then it would be easier to review an example from CADET-Core.
