### 2025-03-09 #7881 @goxberry
Added https://github.com/cadet/CADET-Core/issues/371 for issue tracking.

To summarize the issues in the link, I'm currently blocked on the remaining tasks because I can't install CADET-Core on macOS via the installation instructions. I can build it from source on macOS and I can run the automated unit tests, but the documented examples require Python packages (CADET-Process or CADET-Python) that link to CADET-Core and I'd rather not have to figure out how to install those from source as well. I may try later to provision a Linux VM for review, but given that the project documentation indicates support for macOS, my preference would be for the project's authors to pursue at least one of the following options:

- update their `conda` package to support arm64 macOS platforms
- add an example that doesn't require CADET-Python or CADET-Process to run, with some code to visualize the output (e.g., using Python with some common data science packages like `h5py` and `matplotlib`)
