### 2025-03-09 4a1f39d
Running version with matrix muliplication for residium and jacobian
