+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -348,7 +348,7 @@ namespace cadet
 			virtual unsigned int numReactionsCombined() const CADET_NOEXCEPT { return 1; }
 			virtual unsigned int numReactionQuasiStationary() const CADET_NOEXCEPT { return 0; }
 			template <typename ResidualType>
-			void fillConservedMoietiesBulk21(Eigen::Matrix<ResidualType, Eigen::Dynamic, Eigen::Dynamic>& M, std::vector<int>& QsCompBulk) {}
+			void fillConservedMoietiesBulk21(Eigen::Matrix<ResidualType, Eigen::Dynamic, Eigen::Dynamic>& M, int& QsCompBulk) {}
 
 			CADET_DYNAMICREACTIONMODEL_BOILERPLATE
 
