+++ src/libcadet/model/reaction/MichaelisMentenReaction.cpp
@@ -154,7 +154,7 @@ public:
 	virtual unsigned int numReactionsCombined() const CADET_NOEXCEPT { return 0; }
 	virtual unsigned int numReactionQuasiStationary() const CADET_NOEXCEPT { return 0; }
 	template <typename ResidualType>
-	void fillConservedMoietiesBulk21(Eigen::Matrix<ResidualType, Eigen::Dynamic, Eigen::Dynamic>& M, std::vector<int>& QsCompBulk) {}
+	void fillConservedMoietiesBulk21(Eigen::Matrix<ResidualType, Eigen::Dynamic, Eigen::Dynamic>& M, int& QsCompBulk) {}
 
 	template <typename RowIterator>
 	void jacobianQuasiSteadyLiquidImpl(double t, unsigned int secIdx, const ColumnPosition& colPos, double const* y, int state, int reaction ,const RowIterator& jac, LinearBufferAllocator workSpace) const { }
