### 2025-03-11 fda6717
Fix PR template
Remove checkboxes for the PR type specification to prevent unchecked
boxes in finalized PRs.
