### 2025-03-12 c263f2c
Update getting started documentation
This commit clarifies the relations and usage of the core simulator,
front ends and the createLWE.
