### 2025-03-12 03f6d26
Update getting started documentation (#376)
Clarify the relations and usage of the core simulator, front ends and
the createLWE.

Update example to CADET-Python v1 interface

Fix link to list of authors, contributors in README.rst
