+++ doc/getting_started/tutorials/breakthrough.rst
@@ -1,24 +1,9 @@
-CADET Introduction
-==================
-
-Performing a forward simulation comprises several steps: 
-  * Setting up the model including all parameters
-  * Defining connectivity and dynamic events
-  * Setting up the simulator and actually running the simulation 
-  * Evaluating results (e.g., plotting)
+CADET User-Introduction
+=======================
 
 In this tutorial, we will build a simple forward simulation with a breakthrough of one component using the following system:
 
 .. figure:: _images/breakthrough_system.svg
-
-For this purpose, we use `CADET-Process <https://cadet-process.readthedocs.io/>`_, an object oriented Python frontend for CADET.
-CADET still must be downloaded (or built from source) as explained in the :ref:`installation guide <installation>`. 
-
-.. figure:: _images/cadet_architecture_overview.png
-
-   Relations between CADET-Core, CADET-Python, and CADET-Process.
-
-For more information on how to set up models using CADET-Python, a plain file based API for CADET, please refer to the :ref:`cadet_python`.
 
 1. Setting Up the Model
 -----------------------
