### 2025-03-12 cadet/CADET-Core#373 @jbreue16
Hi Geoffrey,

Thank you again for taking the time to review CADET.

The general design of CADET-Core is to take a configuration file, as described in our [interface specification](https://cadet.github.io/master/interface/), and perform simulations. This is usually done via `cadet-cli` or our C-API (which is still in beta).

By design, we aim to keep CADET-Core agnostic regarding how the configuration file was created. For this purpose, CADET-Python or CADET-Process can be used. As mentioned in #372, once CADET-Core is installed, you should be able to install these packages via `pip`, also on MacOS with ARM architecture.

For testing purposes, we provide another tool, `createLWE`, which generates a fully configured *Load-Wash-Elute* process configuration. This is a typical setup in protein chromatography.

In summary, to create an example configuration and simulate it, run the following commands:

```sh
createLWE
cadet-cli LWE.h5
```

Some model parameters can be changed, to see the options run ``createLWE --help``.

Accompanying python source code e.g. for plotting can be found in the [tutorial for cadet-python](https://cadet.github.io/master/getting_started/tutorials/breakthrough.html)

We admit that our getting started guide needs refinement and will also update it to better describe this process and the roles of the frontend(s) and createLWE. We realized after your review that this information was not explicitly stated anywhere, so thank you for pointing it out! To track progress, see #376
