+++ AUTHORS.md
@@ -1,12 +1,12 @@
 # Contributors
 
-The following people have contributed to the CADET-Core software and are listed in historical order. Together they form the "CADET-Core authors" as mentioned in the copyright statements and [LICENSE.md](LICENSE.txt) file.
+The following individuals have made significant contributions to the CADET-Core software and are listed in *chronological order*. Together they form the "CADET-Core authors" as mentioned in the copyright statements and [LICENSE.md](LICENSE.txt) file.
 
-Major additions or modifications to CADET-Core are explicitly stated.
+Major additions or modifications to CADET-Core are explicitly documented here.
 
 ## Special thanks to everyone who has helped with this project
 
-* [Eric von Lieres](https://github.com/lieres) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, - Germany): Supervision, concepts, first Matlab implementation, user interface design
+* [Eric von Lieres](https://github.com/lieres) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, - Germany): Supervision, concepts, first MATLAB implementation, user interface design
 * Joel Andersson (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany): Domain decomposition (GRM), linear solver (GRM), first C implementation
 * Sebastian Schnittert (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany): First C++ implementation, first Matlab interface, first file format
 * Andreas Püttmann (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany): Parameter sensitivities, Jacobians via AD, AD techniques (block- & band-compression)
@@ -16,7 +16,7 @@ Major additions or modifications to CADET-Core are explicitly stated.
 * [Johannes Schmölder](https://github.com/schmoelder) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany)
 * [Jayghosh Rao](https://github.com/jayghoshter) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany)
 * [Jazib Hassan](https://github.com/jazib-hassan-juelich) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany)
-* [Jan Breuer](https://github.com/jbreue16) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany): Spatial DG discretizations (LRM, LRMP, GRM)
+* [Jan Michael Breuer](https://github.com/jbreue16) (University of Cologne, Faculty of Mathematics and Natural Sciences, Cologne, Germany and Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany): Spatial DG discretizations (LRM, LRMP, GRM), Crystallization module
 * [Ronald Jäpel](https://github.com/ronald-jaepel) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany)
 * [Hannah Lanzrath](https://github.com/hannahlanzrath) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany)
 * [Antonia Berger](https://github.com/AntoniaBerger) (Forschungszentrum Juelich GmbH, IBG-1: Biotechnology, Juelich, Germany)
@@ -24,6 +24,7 @@ Major additions or modifications to CADET-Core are explicitly stated.
 
 ## Funding Acknowledgement
 
-* GE Healthcare sponsored development of core-shell beads, multiple particle types, and 2D general rate model
-* Bayer AG contributed to bugfixes, initial C API, initial Python frontend
[…]
