### 2025-03-13 ae1d768
Fix cadet::log::SelectiveStdWritePolicy
Fix `cadet::log::SelectiveStdWritePolicy<switchLvl>` to write out objects, not end-of-line sequences
