### 2025-03-17 cadet/CADET-Core#373 @goxberry
> Thank you again for taking the time to review CADET.

You're welcome!

> The general design of CADET-Core is to take a configuration file...

As a professional programmer, I appreciate the decoupled architecture. I can also appreciate that users and developers benefit from the option of interfacing with the CADET-Core library at multiple levels of abstraction. However, as a reviewer tasked with reviewing a paper nominally about CADET-Core, the only way I can assess the functionality of the software without prohibitive effort is to work through a documented example. The only documented examples use CADET-Python or CADET-Process, which effectively increases the scope of my review to include one or more of those packages because research software isn't necessarily straightforward to use. Having been a research software engineer before I left academia, I can appreciate why that tendency happens, but as a reviewer, it means I have to expect that the Python packages are unlikely to "just work", and that using them is likely to involve debugging, setting up a development environment, contributing fixes, reading source code, and so on.

However, having left academia, I'm doing this review in my spare time, and because I agreed to review CADET-Core, expanding the scope of the review in this fashion is a significant increase in effort, and I will have to confer with my corresponding editor about how to proceed further.

> For testing purposes, we provide another tool, `createLWE`, which generates a fully configured _Load-Wash-Elute process_ configuration.

I appreciate this option and I can confirm that the commands provided execute without any apparent errors. However, without understanding the inputs and enough of the underlying modeling, it's difficult to evaluate the output of the software and assess whether that output is physically plausible, as someone with a PhD in chemical engineering who has a computational modeling background but has no background in constitutive models for chromatography, and hasn't worked on unit ops in over a decade. Without documentation of the input file, it would take me a prohibitive amount of effort to understand it as a reviewer.

> Accompanying python source code e.g. for plotting can be found in the [tutorial for cadet-python](https://cadet.github.io/master/getting_started/tutorials/breakthrough.html).

Adapting the tutorial for CADET-Python has not been straightforward because I am not able to run that tutorial to completion. From the perspective of a reviewer, installing CADET-Core in a maintainable way took a significant amount of time from a review perspective because I had to review proper CMake configuration settings and how to use GNU Stow in order to make the installation tidy. Once I refamiliarized myself with those skills, I was able to run the CADET-Python example, but was unable to run it to completion. I fixed some bugs in the example in pull request #378. However, when I run [this corrected portion](https://github.com/cadet/CADET-Core/pull/378/files#diff-b91060d6e5679097dd4c15b10987cc27225ad780ce498beb1c6332b5a4a151feR126-R131) of that example, the output indicates an error has occurred:

```pycon
>>> if data.return_code == 0:   # not data.returncode
...     print("Simulation completed successfully")
...     model.load()
... else:
...     print(data)
...     raise Exception("Simulation failed")
...
ReturnInformation(return_code=2, error_message="IO ERROR: Group '/input/solver/sections' doesn't exist in file\n", log='')
```

Further debugging yielded:

```pycon
>>> import h5py
>>> example = h5py.File("model.h5", "r")
>>> example.keys()
<KeysViewHDF5 ['input']>
>>> example["input"].keys()
<KeysViewHDF5 ['model', 'return', 'solver']>
>>> example["input"]["solver"].keys()
<KeysViewHDF5 ['NTHREADS', 'USER_SOLUTION_TIMES', 'time_integrator']>
```

I went through a similar exercise with the output of the `createLWE` example, and was able to plot the output:

```pycon
>>> import h5py
>>> import numpy
>>> import matplotlib.pyplot
>>> lwe = h5py.File("LWE.h5", "r")
>>> lwe.keys()
<KeysViewHDF5 ['input', 'meta', 'output']>
>>> lwe["output"].keys()
<KeysViewHDF5 ['solution']>
>>> lwe["output"]["solution"].keys()
<KeysViewHDF5 ['SOLUTION_TIMES', 'unit_000']>
>>> lwe["output"]["solution"]["unit_000"].keys()
<KeysViewHDF5 ['SOLUTION_OUTLET_COMP_000', 'SOLUTION_OUTLET_COMP_001', 'SOLUTION_OUTLET_COMP_002', 'SOLUTION_OUTLET_COMP_003']>
>>> t = numpy.asarray(lwe["output"]["solution"]["SOLUTION_TIMES"])
>>> comp_001 = numpy.asarray(lwe["output"]["solution"]["unit_000"]["SOLUTION_OUTLET_COMP_001"])
>>> matplotlib.pyplot.plot(t, comp_001)
>>> matplotlib.pyplot.show()
```

but I don't know if the output is sensible.

I'll also outline a couple other issues I noted during the review because I don't have the time at the moment to document them fully, and I'm not sure I'll be able to do so in the next couple weeks:

- In general, CADET-Python won't locate dynamic libraries on macOS properly. With default build settings, CMake will build CADET-Core such that the dynamic libraries produced have `.dylib` suffixes, whereas CADET-Python assumes dynamic libraries on macOS have `.so` suffixes.
- CADET-Python's platform handling should probably raise exceptions for unsupported platforms (e.g., `'Java'`); see the documentation for the Python standard library's [`platform`](https://docs.python.org/3/library/platform.html) module for details.
- It's unclear to me where I could find online the API documentation for CADET-Core or CADET-Python.
- I'm still unable to install CADET-Process; a reproducer is below:

```console
% cat << EOF > pixi.toml
[project]
authors     = ["Geoffrey M Oxberry <goxberry@gmail.com>"]
channels    = ["conda-forge"]
description = "plotting cadet-python results"
name        = "cadet-plots"
platforms   = ["linux-64", "linux-aarch64", "osx-arm64"]
version     = "0.0.1"

[tasks]
# nothing, intentionally

[dependencies]
h5py       = ">=3.13.0"
oniguruma  = ">=6.9.10"
python     = ">=3.12.0"

[pypi-dependencies]
CADET-Python  = ">=1.0.4"
kiwisolver    = ">=1.4"
matplotlib    = ">=3.10.1"
CADET-Process = ">=0.10.1"
EOF
% pixi shell
  ⠁ default:linux-aarch6400:00:00] resolving numba==0.56.2                    Error:   × failed to solve the pypi requirements of 'default' 'osx-arm64'
  ├─▶ failed to resolve pypi dependencies
  ╰─▶ Because only the following versions of numba are available:
          numba<=0.55.1
          numba==0.55.2
          numba==0.56.0
          numba==0.56.2
          numba==0.56.3
          numba==0.56.4
          numba==0.57.0
          numba==0.57.1
          numba==0.58.0
          numba==0.58.1
          numba==0.59.0
          numba==0.59.1
          numba==0.60.0
          numba==0.61.0
      and numba==0.55.1 depends on numpy>=1.18,<1.22, we can conclude that
      numba>=0.55.1,<0.55.2 depends on numpy>=1.18,<1.22.
      And because numba>=0.55.2,<=0.56.0 depends on numpy>=1.18,<1.23, we can
      conclude that numba>=0.55.1,<0.56.2 depends on numpy>=1.18,<1.23.
      And because numba>=0.56.2,<=0.56.4 depends on numpy>=1.18,<1.24 and
      numpy>=1.21,<1.25, we can conclude that numba>=0.55.1,<0.58.0 depends on
      numpy>=1.18,<1.25.
      And because numba==0.58.0 depends on numpy>=1.21,<1.26 and
      numpy>=1.24,<2.2, we can conclude that all of:
          numba>=0.55.1,<0.58.1
          numba>0.60.0
      depend on numpy>=1.18,<2.2.
      And because numba==0.60.0 depends on numpy>=1.22,<2.1 and
      numpy>=1.22,<1.27, we can conclude that numba>=0.55.1 depends on
      numpy>=1.18,<2.2.
      And because cadet-process==0.10.1 depends on numpy==2.2.3 and
      numba>=0.55.1, we can conclude that cadet-process==0.10.1 cannot be
      used.
      And because only cadet-process<=0.10.1 is available and you require
      cadet-process>=0.10.1, we can conclude that your requirements are
      unsatisfiable.

      hint: Pre-releases are available for `numba` in the requested
      range (e.g., 0.61.0rc2), but pre-releases weren't enabled (try:
      `--prerelease=allow`)
```
