### 2025-03-17 1812522
Update CADET-Python example in documentation (#378)
This commit:
- Updates `returncode` to `return_code` as per CADEt-Python v1.0
- Clarifies slightly how users should configure `Cadet.cadet_path`.
- Fixes a spelling mistake I saw in the comments.
