### 2025-03-17 cadet/CADET-Core#373 @goxberry
Also, if I remove from `pixi.toml` the `linux-aarch64` platform, I get a different error:

```console
% cat << EOF > pixi.toml
[project]
authors     = ["Geoffrey M Oxberry <goxberry@gmail.com>"]
channels    = ["conda-forge"]
description = "plotting cadet-python results"
name        = "cadet-plots"
platforms   = ["linux-64", "osx-arm64"]
version     = "0.0.1"

[tasks]
# nothing, intentionally

[dependencies]
h5py       = ">=3.13.0"
oniguruma  = ">=6.9.10"
python     = ">=3.12.0"

[pypi-dependencies]
CADET-Python  = ">=1.0.4"
kiwisolver    = ">=1.4"
matplotlib    = ">=3.10.1"
CADET-Process = ">=0.10.1"
EOF
% pixi shell
Error:   × failed to solve the pypi requirements of 'default' 'osx-arm64'
  ├─▶ failed to resolve pypi dependencies
  ╰─▶ Because there is no version of betterproto==2.0.0b3 and mcbackend<=[0/286]
      depends on betterproto==2.0.0b3, we can conclude that mcbackend<=0.1.1
      cannot be used. (1)

      Because there is no version of betterproto==2.0.0b6 and mcbackend>=0.5.1
      depends on betterproto==2.0.0b6, we can conclude that mcbackend>=0.5.1
      cannot be used.
      And because we know from (1) that mcbackend<=0.1.1 cannot be used, we
      can conclude that all of:
          mcbackend<=0.1.1
          mcbackend>=0.5.1
       cannot be used. (2)

      Because there is no version of betterproto==2.0.0b5 and
      mcbackend>=0.2.1,<=0.5.0 depends on betterproto==2.0.0b5, we can
      conclude that mcbackend>=0.2.1,<=0.5.0 cannot be used.
      And because we know from (2) that all of:
          mcbackend<=0.1.1
          mcbackend>=0.5.1
       cannot be used, we can conclude that all of:
          mcbackend<=0.1.1
          mcbackend>=0.2.1
       cannot be used. (3)

      Because there is no version of betterproto==2.0.0b4 and
      mcbackend>=0.1.2,<=0.2.0 depends on betterproto==2.0.0b4, we can
      conclude that mcbackend>=0.1.2,<=0.2.0 cannot be used.
      And because we know from (3) that all of:
          mcbackend<=0.1.1
          mcbackend>=0.2.1
       cannot be used, we can conclude that all versions of mcbackend cannot
      be used.
      And because hopsy>=1.4.0 depends on mcbackend, we can conclude that
      hopsy>=1.4.0 cannot be used.
      And because only the following versions of hopsy are available:
          hopsy<=1.4.0
          hopsy==1.4.1
          hopsy==1.5.0
          hopsy==1.5.1
          hopsy==1.5.2
          hopsy==1.5.3
          hopsy==1.6.0
      and cadet-process==0.10.1 depends on hopsy>=1.4.0, we can conclude that
      cadet-process==0.10.1 cannot be used.
      And because only cadet-process<=0.10.1 is available and you require
      cadet-process>=0.10.1, we can conclude that your requirements are
      unsatisfiable.

      hint: `betterproto` was requested with a pre-release marker (e.g.,
      betterproto==2.0.0b3)
```
