### 2025-03-17 #7881 @jedbrown
Hi @lieres. I have to ask you to revisit the project scope and installation hurdles to enable a timely review. There are several possibilities that I think would suffice:
1. full Conda or PyPI support (that has been properly tested on supported platforms) for the entire system that is necessary to review the software through documented examples. I know `cadet-python` is on conda-forge.
2. stand-alone documented example using CADET-Core only, as proposed in https://github.com/cadet/CADET-Core/issues/373
3. a containerized environment (with provenance) that is sufficient to follow documented end-to-end workflows; could use https://mybinder.org/ to deploy
4. a hosted environment that reviewers can use to follow documented end-to-end workflows

We recognize that packaging and distributing software is hard and frustrating, but ensuring that the process is streamlined and really works without caveats for users is important for successful community projects. There is a limit to what JOSS reviewers can be asked to do. It looks like CADET-Python [is only tested](https://github.com/cadet/CADET-Python/blob/master/.github/workflows/pipeline.yml) on `ubuntu-latest`. Ensuring CI coverage of all arches would likely also help catch these issues before they become installation woes for users/reviewers.
