### 2025-03-18 41e7ff3
WIP undependent of old calculation
