+++ src/libcadet/model/reaction/MassActionLawReaction.cpp
@@ -28,7 +28,6 @@
 #include <Eigen/Dense>
 #include <iostream>
 
-using namespace Eigen;
 /*<codegen>
 {
 	"name": "MassActionLawParamHandler",
@@ -315,16 +314,17 @@ public:
 	virtual void setExternalFunctions(IExternalFunction** extFuns, unsigned int size) { _paramHandler.setExternalFunctions(extFuns, size); }
 	virtual bool dependsOnTime() const CADET_NOEXCEPT { return ParamHandler_t::dependsOnTime(); }
 	virtual bool requiresWorkspace() const CADET_NOEXCEPT { return true; }
-	virtual bool hasQuasiStationaryReactionsBulk() const CADET_NOEXCEPT
-	{
-		return std::any_of(_reactionQuasistationarity.begin(), _reactionQuasistationarity.end(), [](int i) -> bool { return i; });
-	}
 	virtual bool hasDynamicReactions() const CADET_NOEXCEPT { return true; }
-	virtual int const* reactionQuasiStationarity() const CADET_NOEXCEPT { return _reactionQuasistationarity.data(); }
 
 	virtual unsigned int workspaceSize(unsigned int nComp, unsigned int totalNumBoundStates, unsigned int const* nBoundStates) const CADET_NOEXCEPT
 	{
 		return _paramHandler.cacheSize(maxNumReactions(), nComp, totalNumBoundStates) + std::max(maxNumReactions() * sizeof(active), 2 * (_nComp + totalNumBoundStates) * sizeof(double));
+	}
+	
+	virtual int const* reactionQuasiStationarity() const CADET_NOEXCEPT { return _reactionQuasiStationarity.data(); }
+	virtual bool hasQuasiStationaryReactionsBulk() const CADET_NOEXCEPT
+	{
+		return std::any_of(_reactionQuasiStationarity.begin(), _reactionQuasiStationarity.end(), [](int i) -> bool { return i; });
 	}
 
 	virtual bool configureModelDiscretization(IParameterProvider& paramProvider, unsigned int nComp, unsigned int const* nBound, unsigned int const* boundOffset)
@@ -343,33 +343,39 @@ public:
 			_expBulkFwd.resize(nComp, nReactions);
 			_expBulkBwd.resize(nComp, nReactions);
 
+			
 			if (paramProvider.exists("IS_KINETIC"))
-			{
-				_reactionQuasistationarity.resize(_stoichiometryBulk.columns(), false);
+			{	// default is kinetic binding
+				_reactionQuasiStationarity.resize(_stoichiometryBulk.columns(), false);
 
 				if (paramProvider.isArray("IS_KINETIC")) {
-					const std::vector<int> vecKin = paramProvider.getIntArray("IS_KINETIC");
-					int numqsReaction = std::count(vecKin.begin(), vecKin.end(), 1);
-					if (vecKin.size() == 1)
-					{
-						// Treat an array with a single element as scalar
-						std::fill(_reactionQuasistationarity.begin(), _reactionQuasistationarity.end(), !static_cast<bool>(vecKin[0]));
+					const std::vector<int> kinetikFlag = paramProvider.getIntArray("IS_KINETIC");
+					if (kinetikFlag.size() == 1)
+					{	// Set all reactions to kinetic binding if IS_KINETIC is a single value
+						std::fill(_reactionQuasiStationarity.begin(), _reactionQuasiStationarity.end(), !static_cast<bool>(kinetikFlag[0]));
 					}
-					else if (vecKin.size() < _reactionQuasistationarity.size())
+					else if (kinetikFlag.size() < _reactionQuasiStationarity.size())
 					{
-						// Error on too few elements
[…]
