### 2025-03-20 ee02816
Update README.rst
Fix link to list of contributors
