### 2025-03-21 0cc51c0
Update documentation landing page (#376)
Add information about CADET platform to landing page
