### 2025-03-21 38eaf06
Update documentation landing page (#376)
Add information about CADET platform to landing page
