### 2025-03-21 6d39653
Update index
Add information about CADET platform to landing page
