### 2025-03-21 0ef26f9
Update installation frontend
Adds links with resources
