+++ doc/getting_started/installation_frontend.rst
@@ -32,6 +32,6 @@ To install ``CADET-Python``, open an `conda shell` and execute:
 Additional Resources
 ^^^^^^^^^^^^^^^^^^^^
 
-`CADET-Process documentation <https://cadet-process.readthedocs.io/>`
-`CADET-Process Workshop <https://github.com/cadet/CADET-Workshop>`
-`CADET-Python Tutorial <https://github.com/cadet/CADET-Python-Tutorial>`
+- `CADET-Process documentation <https://cadet-process.readthedocs.io/>`_
+- `CADET-Process Workshop <https://github.com/cadet/CADET-Workshop>`_
+- `CADET-Python Tutorial <https://github.com/cadet/CADET-Python-Tutorial>`_
