+++ src/libcadet/model/StirredTankModel.cpp
@@ -34,7 +34,7 @@
 #include <Eigen/Dense>
 #include <Eigen/Sparse>
 
-#include<iostream>
+#include<iostream>//todo delete
 
 namespace cadet
 {
@@ -82,8 +82,7 @@ CSTRModel::~CSTRModel() CADET_NOEXCEPT
 	delete[] _offsetParType;
 
 	delete[] _dynReactionBulk;
-	//delete[] _temp;
-	//delete[] _temp2;
+
 }
 
 unsigned int CSTRModel::numDofs() const CADET_NOEXCEPT
@@ -372,9 +371,8 @@ bool CSTRModel::configure(IParameterProvider& paramProvider)
 
 		if (_qsReactionBulk != nullptr)
 		{	
-			//_dynReactionBulk->fillConservedMoietiesBulk(_MconvMoityBulk, _QsCompBulk); // fill conserved moities matrix
 			_QsCompBulk.resize(_nComp);
-			_dynReactionBulk->fillConservedMoietiesBulk2(_MconvMoityBulk2, _nConservedQuants, _QsCompBulk);
+			_dynReactionBulk->fillConservedMoietiesBulk(_MconvMoityBulk, _nConservedQuants, _QsCompBulk);
 		
 
 			bool hasQSBinding = false;
@@ -395,8 +393,10 @@ bool CSTRModel::configure(IParameterProvider& paramProvider)
 	}
 	else
 	{	
-		_MconvMoityBulk = Eigen::MatrixXd::Zero(0, 0); // matrix for conserved moities
+		_MconvMoityBulk = Eigen::Matrix<active, Eigen::Dynamic, Eigen::Dynamic>::Zero(0, 0); // matrix for conserved moities
 		_QsCompBulk.clear();
+		_qsReactionBulk = nullptr;
+		_nConservedQuants = 0;
 	}
 
 	for (unsigned int type = 0; type < _nParType; ++type)
@@ -708,8 +708,8 @@ void CSTRModel::consistentInitialState(const SimulationTime& simTime, double* co
 			if (mIdx >= _nConservedQuants)
 				continue;
 			double dotProduct = 0.0;
-			for (unsigned int j = 0; j < _MconvMoityBulk2.cols(); ++j)
-				dotProduct += static_cast<double>(_MconvMoityBulk2(i, j)) * c[j];
+			for (unsigned int j = 0; j < _MconvMoityBulk.cols(); ++j)
+				dotProduct += static_cast<double>(_MconvMoityBulk(i, j)) * c[j];
 
 			conservedQuants[mIdx] = dotProduct;
 			mIdx++;
@@ -778,12 +778,12 @@ void CSTRModel::consistentInitialState(const SimulationTime& simTime, double* co
 							continue;
 
 						int jIdx = 0;
-						for (unsigned int j = 0; j < _MconvMoityBulk2.cols(); ++j)
+						for (unsigned int j = 0; j < _MconvMoityBulk.cols(); ++j)
 						{
 							if (_QsCompBulk[j] == 0)
 								continue;
 
-							mat.native(mIdx, jIdx) = static_cast<double>(_MconvMoityBulk2(i, j));
+							mat.native(mIdx, jIdx) = static_cast<double>(_MconvMoityBulk(i, j));
 							jIdx++;
 						}
 						mIdx++;
@@ -817,12 +817,12 @@ void CSTRModel::consistentInitialState(const SimulationTime& simTime, double* co
 							continue;
 
 						int jIdx = 0;
-						for (unsigned int j = 0; j < _MconvMoityBulk2.cols(); ++j)
+						for (unsigned int j = 0; j < _MconvMoityBulk.cols(); ++j)
 						{
 							if(_QsCompBulk[j] == 0)
 								continue;
 
[…]
