### 2025-03-24 8e648b2
CI-debug 1
test if bug is at StirrTankModel
