+++ src/libcadet/model/StirredTankModel.hpp
@@ -24,232 +24,218 @@
 #include "linalg/DenseMatrix.hpp"
 #include "model/ModelUtils.hpp"
 #include "Memory.hpp"
-#include <bitset>
+
 #include <array>
 #include <vector>
-#include <Eigen/Dense>
 
 namespace cadet
 {
-struct ColumnPosition;
 
-namespace model
-{
+	namespace model
+	{
 
-/**
- * @brief Continuous stirred tank (reactor) model
- * @details This is a simple CSTR with variable porosity model with variable volume using the ``well mixed assumption''.
- * @f[\begin{align}
-	\frac{\mathrm{d}}{\mathrm{d} t}\left( V^l c_i \right) + V^s \sum_{j=1}^{N_{\text{bnd},i}} q_{i,j} &= F_{\text{in}} c_{\text{in},i} - F_{\text{out}} c_i \\
-	a \frac{\partial q_{i,j}}{\partial t} &= f_{\text{iso},i,j}(c, q) \\
-	\frac{\partial V^l}{\partial t} &= F_{\text{in}} - F_{\text{out}} - F_{\text{filter}}
-\end{align} @f]
- * The model can be used as a plain stir tank without any binding states.
- */
-class CSTRModel : public UnitOperationBase
-{
-public:
+		/**
+		 * @brief Continuous stirred tank (reactor) model
+		 * @details This is a simple CSTR with variable porosity model with variable volume using the ``well mixed assumption''.
+		 * @f[\begin{align}
+			\frac{\mathrm{d}}{\mathrm{d} t}\left( V^l c_i \right) + V^s \sum_{j=1}^{N_{\text{bnd},i}} q_{i,j} &= F_{\text{in}} c_{\text{in},i} - F_{\text{out}} c_i \\
+			a \frac{\partial q_{i,j}}{\partial t} &= f_{\text{iso},i,j}(c, q) \\
+			\frac{\partial V^l}{\partial t} &= F_{\text{in}} - F_{\text{out}} - F_{\text{filter}}
+		\end{align} @f]
+		 * The model can be used as a plain stir tank without any binding states.
+		 */
+		class CSTRModel : public UnitOperationBase
+		{
+		public:
 
-	CSTRModel(UnitOpIdx unitOpIdx);
-	virtual ~CSTRModel() CADET_NOEXCEPT;
+			CSTRModel(UnitOpIdx unitOpIdx);
+			virtual ~CSTRModel() CADET_NOEXCEPT;
 
-	virtual unsigned int numDofs() const CADET_NOEXCEPT;
-	virtual unsigned int numPureDofs() const CADET_NOEXCEPT;
-	virtual bool usesAD() const CADET_NOEXCEPT;
-	virtual unsigned int requiredADdirs() const CADET_NOEXCEPT;
+			virtual unsigned int numDofs() const CADET_NOEXCEPT;
+			virtual unsigned int numPureDofs() const CADET_NOEXCEPT;
+			virtual bool usesAD() const CADET_NOEXCEPT;
+			virtual unsigned int requiredADdirs() const CADET_NOEXCEPT;
 
-	virtual UnitOpIdx unitOperationId() const CADET_NOEXCEPT { return _unitOpIdx; }
-	virtual unsigned int numComponents() const CADET_NOEXCEPT { return _nComp; }
-	virtual void setFlowRates(active const* in, active const* out) CADET_NOEXCEPT;
-	virtual unsigned int numInletPorts() const CADET_NOEXCEPT { return 1; }
-	virtual unsigned int numOutletPorts() const CADET_NOEXCEPT { return 1; }
-	virtual bool canAccumulate() const CADET_NOEXCEPT { return true; }
[…]
