+++ src/libcadet/model/reaction/ReactionModelBase.hpp
@@ -212,25 +212,25 @@ protected:
 	}																																					\
 																																						\
 	virtual int computeQuasiStationaryReactionFlux(double t, unsigned int secIdx, const ColumnPosition& colPos, double const* y,										\
-		Eigen::Map<Eigen::Vector<active, Eigen::Dynamic>> fluxes, int const* mapQSReac, LinearBufferAllocator workSpace)								\
+		Eigen::Map<Eigen::Vector<active, Eigen::Dynamic>> fluxes, LinearBufferAllocator workSpace)								\
 	{                                                                                                                                                   \
 		return quasiStationaryFlux( t, secIdx,   colPos,  y, fluxes, workSpace);																		\
 	}																																					\
 																																						 \
 	 virtual int computeQuasiStationaryReactionFlux(double t, unsigned int secIdx, const ColumnPosition& colPos, active const* y,										\
-		Eigen::Map<Eigen::Vector<double, Eigen::Dynamic>> fluxes, int const* mapQSReac, LinearBufferAllocator workSpace)								\
+		Eigen::Map<Eigen::Vector<double, Eigen::Dynamic>> fluxes, LinearBufferAllocator workSpace)								\
 	{                                                                                                                                                   \
 			return quasiStationaryFlux( t, secIdx,   colPos,  y, fluxes,  workSpace);																		\
 	}																																					\
 																																						\
 	virtual int computeQuasiStationaryReactionFlux(double t, unsigned int secIdx, const ColumnPosition& colPos, double const* y,										\
-		Eigen::Map<Eigen::Vector<double, Eigen::Dynamic>> fluxes, int const* mapQSReac, LinearBufferAllocator workSpace)								\
+		Eigen::Map<Eigen::Vector<double, Eigen::Dynamic>> fluxes, LinearBufferAllocator workSpace)								\
 	{                                                                                                                                                   \
 		return quasiStationaryFlux( t, secIdx,   colPos,  y, fluxes, workSpace);																			\
 	}																																					\
 																																						\
 	virtual int computeQuasiStationaryReactionFlux(double t, unsigned int secIdx, const ColumnPosition& colPos, active const* y,										\
-		Eigen::Map<Eigen::Vector<active, Eigen::Dynamic>> fluxes, int const* mapQSReac, LinearBufferAllocator workSpace)								\
+		Eigen::Map<Eigen::Vector<active, Eigen::Dynamic>> fluxes, LinearBufferAllocator workSpace)								\
 	{                                                                                                                                                   \
 		return quasiStationaryFlux( t, secIdx,   colPos,  y, fluxes,  workSpace);																			\
 	}																																					\
