+++ src/libcadet/model/StirredTankModel - Kopie.cpp
@@ -1,0 +1,2444 @@
+// =============================================================================
+//  CADET
+//
+//  Copyright © 2008-present: The CADET-Core Authors
+//            Please see the AUTHORS.md file.
+//
+//  All rights reserved. This program and the accompanying materials
+//  are made available under the terms of the GNU Public License v3.0 (or, at
+//  your option, any later version) which accompanies this distribution, and
+//  is available at http://www.gnu.org/licenses/gpl.html
+// =============================================================================
+
+#include "model/StirredTankModel.hpp"
+#include "ParamReaderHelper.hpp"
+#include "ParamReaderScopes.hpp"
+#include "cadet/Exceptions.hpp"
+#include "cadet/SolutionRecorder.hpp"
+#include "model/BindingModel.hpp"
+#include "model/ReactionModel.hpp"
+#include "SimulationTypes.hpp"
+#include "ParallelSupport.hpp"
+#include "linalg/Subset.hpp"
+
+#include "ConfigurationHelper.hpp"
+#include "linalg/Norms.hpp"
+#include "AdUtils.hpp"
+
+#include "LoggingUtils.hpp"
+#include "Logging.hpp"
+
+#include <algorithm>
+#include <functional>
+#include <bitset>
+#include <Eigen/Dense>
+#include <Eigen/Sparse>
+
+#include<iostream>//todo delete
+
+namespace cadet
+{
+
+namespace model
+{
+
+namespace
+{
+	inline void bindingFlux(IBindingModel* binding, double t, unsigned int secIdx, const ColumnPosition& colPos, active const* yCp, active const* yQ, active* res, cadet::LinearBufferAllocator buffer, WithParamSensitivity)
+	{
+		binding->flux(t, secIdx, colPos, yQ, yCp, res, buffer, WithParamSensitivity());
+	}
+
+	inline void bindingFlux(IBindingModel* binding, double t, unsigned int secIdx, const ColumnPosition& colPos, active const* yCp, active const* yQ, active* res, cadet::LinearBufferAllocator buffer, WithoutParamSensitivity)
+	{
+		binding->flux(t, secIdx, colPos, yQ, yCp, res, buffer, WithoutParamSensitivity());
+	}
+
+	inline void bindingFlux(IBindingModel* binding, double t, unsigned int secIdx, const ColumnPosition& colPos, double const* yCp, double const* yQ, active* res, cadet::LinearBufferAllocator buffer, WithParamSensitivity)
+	{
+		binding->flux(t, secIdx, colPos, yQ, yCp, res, buffer);
+	}
+
+	inline void bindingFlux(IBindingModel* binding, double t, unsigned int secIdx, const ColumnPosition& colPos, double const* yCp, double const* yQ, double* res, cadet::LinearBufferAllocator buffer, WithoutParamSensitivity)
+	{
+		binding->flux(t, secIdx, colPos, yQ, yCp, res, buffer);
+	}
+}
+
+
+CSTRModel::CSTRModel(UnitOpIdx unitOpIdx) : UnitOperationBase(unitOpIdx), _nComp(0), _nParType(0), _nBound(nullptr), _boundOffset(nullptr), _strideBound(nullptr), _offsetParType(nullptr),
+	_totalBound(0), _analyticJac(true), _jac(), _jacFact(), _factorizeJac(false), _initConditions(0), _initConditionsDot(0), _dynReactionBulk(nullptr)
+{
[…]
