### 2025-03-28 c786ccb
Update equation for accurate allosteric inhibition modeling
adjusted residual and jacobian implementation
New implemented equation is:
\nu_i = \frac{\mu_{\mathrm{max},i} c_S}{k_{\mathrm{MM},i} + c_S} \cdot
\frac{1}{1 + \sum_i \frac{k_{\mathrm{I},i,j}}{k_{\mathrm{I},i,j} +
c_{\mathrm{I},j}}}.
