### 2025-03-28 8d4ffc9
fixup! Update equation for accurate allosteric inhibition modeling
