### 2025-04-01 67881cb
Add simulation reference test between MM with two inhibitor and MAL micro-kinetics
Test compares simulation results form substrat and product once
generated by MM with two inhibitors and once with the induced MAL model.
