### 2025-04-01 bb9af48
Update Michaelis Menten reaction for noncompetitive inhibition modeling (#382)
* Update equation for accurate non-competitive inhibition modeling. adjusted residual and jacobian implementation
New implemented equation is:
$$\nu_i = \frac{\mu_{\mathrm{max},i} c_S}{k_{\mathrm{MM},i} + c_S} \cdot
\frac{1}{1 + \sum_j \frac{k_{\mathrm{I},i,j}}{k_{\mathrm{I},i,j} +
c_{\mathrm{I},j}}}.$$

* Add simulation reference test between MM with two inhibitor and MAL micro-kinetics
