### 2025-04-01 #7881 @lieres
Dear jedbrown,

Thank you very much for getting back to us.

We are sorry to hear, that the setup of CADET for goxberry did not work as seamlessly as expected and apologize for the frustration that may have caused.

**We would like to proceed with option 4) by inviting the reviewers to [our JupyterHub](https://jupyter.cadet-web.de/)**, where they will have access to our workshop materials. These materials include a comprehensive set of examples that can be easily modified for testing purposes, along with visualizations and explanations to support their evaluation.

@bdice, @goxberry, we have sent you a message regarding login to the email addresses listed on your GitHub profiles. Please let us know if you need us to resend the information via a different channel.
Further, we would like to take this opportunity to clarify the relationship between CADET-Core, CADET-Python, and CADET-Process, as well as how they are distributed:

Regarding 1)
The CADET-Core simulator is distributed via `conda` and is fully functional across all operating systems. However, due to a  [conda-forge issue](https://github.com/conda-forge/cadet-feedstock/pull/14), it must be built from source on MacOS-ARM64 systems following the [build instructions for OSX](https://cadet.github.io/master/getting_started/build_osx.html#build-osx). As far as we understand, goxberry was able to successfully build CADET-Core. However, we are continuously looking into this issue and trying to get it resolved ASAP.

CADET-Python and CADET-Process are distributed via `pip` and are tested across operating systems - now also in our CI pipeline, thank you for catching that!

- https://github.com/cadet/CADET-Python/pull/53

We support multiple users on MacOS-ARM64 who use these two packages successfully.

While they were able to successfully install CADET-Python, we understand that the installation of CADET-Process failed for goxberry when using the `Pixi` dependency resolver. We support and recommend using the standard Python package installer `pip` and had no experience with `Pixi` up to this point. However, we found an issue with an upstream dependency which we reported and that has been fixed.

Regarding 2)
CADET-Core can be interfaced directly via HDF5 input/output files. These HDF5 files contain the model specifications as documented on the [CADET-Core Website](https://cadet.github.io/master/interface/index.html) and also contain the simulation output after running a simulation. Additionally, the `create_lwe` method enables users to create and test simple load-wash-elute model configurations as modifiable examples. Options can be accessed via the `--help` command.

However, it is also possible to interface CADET-Core with CADET-Python or CADET-Process.

CADET-Python is a simple wrapper around a generic HDF5 file reader/writer. It can be installed via `pip install CADET-Python`. It enables the user to specify model configurations in Python instead of using another HDF file read/write tool. CADET-Python follows the same API specification as defined on the [CADET-Core Website](https://cadet.github.io/master/interface/index.html). As additional resources for examples we provide a [CADET-Python-Tutorial](https://github.com/cadet/CADET-Python-Tutorial/tree/master).

CADET-Process builds upon CADET-Python and provides tools for easier process modeling, network configuration, evaluation, and optimization. It can be installed via `pip install CADET-Process`.  While we recommend CADET-Process for new users due to its enhanced usability, it is not a requirement. As additional resources, we provide a [CADET-Process workshop repository](https://github.com/cadet/CADET-Workshop) as a helpful introduction.

That said, we are truly grateful for the feedback and discussion, which have been very valuable to us. This exchange highlighted areas where our documentation lacked clarity, and as a result, we have already implemented improvements to better explain how CADET-Core can be interfaced.

- https://github.com/cadet/CADET-Core/pull/376

We recognize that getting started can be challenging, especially for users coming from different fields. We want to emphasize that CADET-Core is a powerful simulator designed to handle complex model networks and a wide range of process and solver parameters. However, we also don't want the reviewers to be burdened with a time-intensive setup.

We offer several support channels, including a community forum with a troubleshooting section, monthly office hours, and once a year a hands-on CADET workshop. As open-source developers in academia, close collaboration with users has always been important for us to get feedback and identify needs and challenges that arise when modeling with CADET. Should more issues arise in the future we would be more than happy to provide support to the reviewers.

We hope these clarifications as well as the recent improvements have addressed your concerns. Please let us know how you would like to proceed.
