### 2025-04-04 e5c748a
Add multi-substrate Michaelis Menten
This commit adds milti-substrate Michaelis Menten.
The Residuum and Jacobi implementations have been adapted accordingly.
A test has been added to compare the implement and AD derivation.
Documentation has been updated.
