+++ src/libcadet/model/reaction/MichaelisMentenReaction.cpp
@@ -142,7 +142,10 @@ public:
 			const unsigned int nReactions = numElements / nComp;
 
 			_stoichiometryBulk.resize(nComp, nReactions);
-			_idxSubstrate = std::vector<int>(nReactions, -1);
+			_idxInhibitor.resize(nReactions);
+			_idxSubstrate.resize(nReactions);
+			_oldInterface = false;
+
 		}
 
 		return true;
@@ -157,7 +160,9 @@ protected:
 	ParamHandler_t _paramHandler; //!< Handles parameters and their dependence on external functions
 
 	linalg::ActiveDenseMatrix _stoichiometryBulk;
-	std::vector<int> _idxSubstrate;
+	std::vector<std::vector<int>> _idxSubstrate;
+	std::vector<std::vector<int>> _idxInhibitor;
+	bool _oldInterface;
 
 	virtual bool configureImpl(IParameterProvider& paramProvider, UnitOpIdx unitOpIdx, ParticleTypeIdx parTypeIdx)
 	{
@@ -167,6 +172,11 @@ protected:
 		if ((_stoichiometryBulk.columns() > 0) && ((_paramHandler.vMax().size() < _stoichiometryBulk.columns()) || (_paramHandler.kMM().size() < _stoichiometryBulk.columns())))
 			throw InvalidParameterException("MM_VMAX and MM_KMM have to have the same size (number of reactions)");
 		
+		if (_stoichiometryBulk.rows() > 1 && (_paramHandler.kMM().size() == _stoichiometryBulk.columns() * _stoichiometryBulk.rows()))
+			_oldInterface = false;
+		else
+			_oldInterface = true;
+
 		if ((_stoichiometryBulk.columns() > 0) && (_paramHandler.kInhibit().size() < _stoichiometryBulk.columns() * _nComp))
 			throw InvalidParameterException("MM_KI have to have the size (number of reactions) x (number of components)");
 
@@ -180,25 +190,43 @@ protected:
 			std::copy(s.begin(), s.end(), _stoichiometryBulk.data());
 
 			// Find substrate index (first educt)
-			for (int i = 0; i < _stoichiometryBulk.columns(); ++i)
+			_idxSubstrate.clear();
+			_idxInhibitor.clear();
+			const std::vector<double> ki = paramProvider.getDoubleArray("MM_KI");
+			for (int r = 0; r < _stoichiometryBulk.columns(); ++r)
 			{
-				_idxSubstrate[i] = -1;
-
-				for (int j = 0; j < _nComp; ++j)
+				std::vector<int> idxSubstrateReaction_r;
+				std::vector<int> idxInhibitorReaction_r;
+				for (int c = 0; c < _nComp; ++c)
 				{
-					if (_stoichiometryBulk.native(j, i) < 0.0)
+					if (_stoichiometryBulk.native(c, r) < 0.0)
+						idxSubstrateReaction_r.push_back(c);
+
+					double kI = ki[_nComp * r + c];
+					if (kI > 0.0)
+						idxInhibitorReaction_r.push_back(c);
+
+				}
+				if(idxSubstrateReaction_r.empty())
+					throw InvalidParameterException("No substrate found in reaction " + std::to_string(r));
+
+				//throw error if inhibitor is substrate
+				if(!idxInhibitorReaction_r.empty())
+				{
+					// check if inhibitor is substrate
+					for (int idxSubstrateReaction_r : idxSubstrateReaction_r)
 					{
-						_idxSubstrate[i] = j;
-						break;
[…]
