### 2025-04-09 cadet/CADET-Core#373 @hannahlanzrath
Thanks for pointing this out! We didn’t have prior experience with `pixi`, so your input was very helpful here. After some investigation, we traced the issue back to an upstream dependency that required a prerelease version of a package. Since `pixi` doesn't resolve prereleases automatically, the installation failed (see [prefix-dev/pixi#1291](https://github.com/prefix-dev/pixi/issues/1291)).

We've reported the issue upstream, and it should be fixed now and installing `CADET-Process` via `pixi` work as expected again.
