### 2025-04-15 76de594
Update windows build instructions
as per a56a4a3b67f3a0d2af2ea863829f32f5f176195e, there are individual
build folders automatically created on windows
