### 2025-04-15 601ff4d
Remove /install folder from build guide
