### 2025-04-15 abedef4
Update build instructions (#390)
* Remove /build and /install folder from guide, as per a56a4a3b67f3a0d2af2ea863829f32f5f176195e, there are individual
build folders automatically created on windows

---------

Co-authored-by: r.jaepel <ronald.jaepel@gmail.com>
