### 2025-04-15 d3588a1
Improve crystallization Jacobian tests to include WENO discretizations
