+++ test/Crystallization.cpp
@@ -101,6 +101,11 @@ TEST_CASE("Crystallization Jacobian verification for a CSTR with initial distrib
 	pp_setup.pushScope("model");
 	pp_setup.pushScope("unit_001");
 
+	// set the discretization scheme
+	pp_setup.pushScope("reaction_bulk");
+	pp_setup.set("CRY_GROWTH_SCHEME_ORDER", 4);
+	pp_setup.popScope();
+
 	cadet::test::column::testJacobianAD(pp_setup, std::numeric_limits<float>::epsilon() * 100.0);
 }
 
@@ -113,6 +118,11 @@ TEST_CASE("Crystallization Jacobian verification for a CSTR with initial distrib
 
 	pp_setup.pushScope("model");
 	pp_setup.pushScope("unit_001");
+
+	// set the discretization scheme
+	pp_setup.pushScope("reaction_bulk");
+	pp_setup.set("CRY_GROWTH_SCHEME_ORDER", 3);
+	pp_setup.popScope();
 
 	cadet::test::column::testJacobianAD(pp_setup, std::numeric_limits<float>::epsilon() * 100.0);
 }
@@ -127,6 +137,11 @@ TEST_CASE("Crystallization Jacobian verification for a CSTR with primary nucleat
 	pp_setup.pushScope("model");
 	pp_setup.pushScope("unit_001");
 
+	// set the discretization scheme
+	pp_setup.pushScope("reaction_bulk");
+	pp_setup.set("CRY_GROWTH_SCHEME_ORDER", 1);
+	pp_setup.popScope();
+
 	cadet::test::column::testJacobianAD(pp_setup, std::numeric_limits<float>::epsilon() * 100.0);
 }
 
@@ -140,6 +155,11 @@ TEST_CASE("Crystallization Jacobian verification for a CSTR with primary nucleat
 	pp_setup.pushScope("model");
 	pp_setup.pushScope("unit_001");
 
+	// set the discretization scheme
+	pp_setup.pushScope("reaction_bulk");
+	pp_setup.set("CRY_GROWTH_SCHEME_ORDER", 1);
+	pp_setup.popScope();
+
 	cadet::test::column::testJacobianAD(pp_setup, std::numeric_limits<float>::epsilon() * 100.0);
 }
 
@@ -152,6 +172,11 @@ TEST_CASE("Crystallization Jacobian verification for a CSTR with primary and sec
 
 	pp_setup.pushScope("model");
 	pp_setup.pushScope("unit_001");
+
+	// set the discretization scheme
+	pp_setup.pushScope("reaction_bulk");
+	pp_setup.set("CRY_GROWTH_SCHEME_ORDER", 3);
+	pp_setup.popScope();
 
 	cadet::test::column::testJacobianAD(pp_setup, std::numeric_limits<float>::epsilon() * 100.0);
 }
