+++ src/libcadet/model/reaction/CrystallizationReaction_fragmentation.cpp
@@ -1,347 +1,0 @@
-// =============================================================================
-//  CADET
-//  
-//  Copyright © 2008-2022: The CADET Authors
-//            Please see the AUTHORS and CONTRIBUTORS file.
-//  
-//  All rights reserved. This program and the accompanying materials
-//  are made available under the terms of the GNU Public License v3.0 (or, at
-//  your option, any later version) which accompanies this distribution, and
-//  is available at http://www.gnu.org/licenses/gpl.html
-// =============================================================================
-
-#include "model/reaction/ReactionModelBase.hpp"
-#include "model/ModelUtils.hpp"
-#include "cadet/Exceptions.hpp"
-#include "model/Parameters.hpp"
-#include "LocalVector.hpp"
-#include "SimulationTypes.hpp"
-#include "linalg/ActiveDenseMatrix.hpp"
-#include "Memory.hpp"
-
-#include <functional>
-#include <algorithm>
-#include <unordered_map>
-#include <string>
-#include <iterator>
-#include <vector>
-
-namespace cadet
-{
-
-	namespace model
-	{
-		namespace detail
-		{
-			struct FragCoefficients
-			{
-				// define breakage-related local parameters
-				std::vector<active> Upsilon_source;
-				std::vector<active> Upsilon_sink;
-
-				// constructor
-				FragCoefficients(const std::vector<active>& binCenters, const std::vector<active>& bins, const active& breakageKernelGamma)
-					: Upsilon_source(binCenters.size()), Upsilon_sink(binCenters.size())
-				{
-					const active N_j = breakageKernelGamma / (breakageKernelGamma - 1.0);
-
-					active Upsilon_birth_sum = 0.0;
-					active Upsilon_death_sum = 0.0;
-					active b_integral_birth = 0.0;
-
-					// calculate upsilon and store it
-					for (int i = 0; i < binCenters.size(); ++i)
-					{
-						// reset the sum for each i
-						Upsilon_birth_sum = 0.0;
-						Upsilon_death_sum = 0.0;
-						const active x_i_3 = binCenters[i] * binCenters[i] * binCenters[i];
-						for (int j = 0; j < i + 1; ++j)
-						{
-							const active x_j_3 = binCenters[j] * binCenters[j] * binCenters[j];
-							if (cadet_likely(i != j))
-							{
-								b_integral_birth = N_j * (pow(bins[j + 1], 3.0 * breakageKernelGamma - 3.0) - pow(bins[j], 3.0 * breakageKernelGamma - 3.0)) / pow(binCenters[i], 3.0 * breakageKernelGamma - 3.0);
-								Upsilon_birth_sum += b_integral_birth * (x_i_3 - x_j_3);
-								Upsilon_death_sum += b_integral_birth * x_j_3;
-							}
-							else
-							{
-								b_integral_birth = N_j * (pow(binCenters[j], 3.0 * breakageKernelGamma - 3.0) - pow(bins[j], 3.0 * breakageKernelGamma - 3.0)) / pow(binCenters[i], 3.0 * breakageKernelGamma - 3.0);
-								Upsilon_death_sum += b_integral_birth * x_j_3;
-							}
-						}
-						if (cadet_likely(i > 0))
-						{
-							Upsilon_source[i] = (N_j - 1.0) * x_i_3 / Upsilon_birth_sum;
-						}
-						else
-						{
-							Upsilon_source[i] = (N_j - 1.0) * x_i_3;
-						}
-						Upsilon_sink[i] = Upsilon_death_sum * Upsilon_source[i] / x_i_3;
-					}
-				}
-			};
-		} // namespace detail
-
-		/**
[…]
