### 2025-04-15 b6650c9
Unify crystallization modules for PBM, aggregation and breakage (#241)
