+++ src/libcadet/model/reaction/CrystallizationReaction_aggregation.cpp
@@ -1,460 +1,0 @@
-#include "model/reaction/ReactionModelBase.hpp"
-#include "model/ModelUtils.hpp"
-#include "cadet/Exceptions.hpp"
-#include "model/Parameters.hpp"
-#include "LocalVector.hpp"
-#include "SimulationTypes.hpp"
-#include "linalg/ActiveDenseMatrix.hpp"
-#include "Memory.hpp"
-
-#include <functional>
-#include <algorithm>
-#include <unordered_map>
-#include <string>
-#include <iterator>
-#include <vector>
-
-namespace cadet
-{
-	namespace model
-	{
-		namespace detail
-		{
-			struct AggCoefficients
-			{
-				std::pair<std::vector<unsigned short int>, std::vector<unsigned short int>> index;  // a pair of vectors to store combo of j and k
-				std::vector<unsigned int> index_tracker;           // the idx can be a large number  
-				std::vector<short int> i_index;
-
-				active sum_volume_cube_root = 0.0;
-				unsigned int count = 0;
-
-				// the sizes of j_index and k_index are not known a priori
-				AggCoefficients(const std::vector<active>& binCenters, const std::vector<active>& bins, const std::vector<active>& binSizes)
-					: index(), index_tracker(binCenters.size() + 1), i_index(binCenters.size() * binCenters.size())
-				{
-					// make the first element of index_tracker 0
-					index_tracker[0] = 0;
-
-					// loop over daughter particle i
-					for (unsigned int i = 0; i < binCenters.size(); ++i)
-					{
-						// source term, find lij in set A
-						for (unsigned short int j = 0; j < i; ++j)
-						{
-							for (unsigned short int k = j; k < i + 1; ++k)
-							{
-								sum_volume_cube_root = pow(binCenters[k] * binCenters[k] * binCenters[k] + binCenters[j] * binCenters[j] * binCenters[j], 1.0 / 3.0);
-								if ((sum_volume_cube_root > bins[i]) && (sum_volume_cube_root <= bins[i + 1]))
-								{
-									index.first.emplace_back(j);
-									index.second.emplace_back(k);
-									++count;
-								}
-							}
-						}
-						index_tracker[i + 1] = count;
-
-						// sink term, loop over another mother particle j
-						for (unsigned int j = 0; j < binCenters.size(); ++j)
-						{
-							sum_volume_cube_root = pow(binCenters[i] * binCenters[i] * binCenters[i] + binCenters[j] * binCenters[j] * binCenters[j], 1.0 / 3.0);
-							unsigned short int k = (i > j) ? i + 1 : j + 1;
-
-							i_index[i * binCenters.size() + j] = -1;        // -1 indicates there is no match
-							for (k; k < binCenters.size(); ++k) 
-							{
-								if ((sum_volume_cube_root > bins[k]) && (sum_volume_cube_root <= bins[k + 1]))
-								{
-									i_index[i * binCenters.size() + j] = k;  // if there is a match, overwrite it
-									break;
-								} 
-							}
-						}
-					}
-				}
-			};
-		} // namespace detail
-
-		/**
-		 * @brief Defines the crystallization reaction model
-		 */
-		class CrystallizationReaction : public IDynamicReactionModel
-		{
-		public:
-
-			CrystallizationReaction() : _nComp(0), _nBins(0), _bins(0), _binCenters(0), _binSizes(0), _agg(nullptr) { }
-			virtual ~CrystallizationReaction() CADET_NOEXCEPT
-			{
-				clearSchemeCoefficients();
-			}
-
-			static const char* identifier() { return "CRYSTALLIZATION"; }
[…]
