### 2025-04-15 4158e35
Unify crystallization modules for PBM, aggregation and breakage (#241)
