+++ test/Crystallization.cpp
@@ -25,10 +25,14 @@
  */
 const char* getTestDirectory();
 
-TEST_CASE("Crystallization in a CSTR with initial distribution and growth", "[Crystallization],[Simulation],[Reference],[CI]")
+
+/*
+ * Pure PBM tests
+*/
+TEST_CASE("Crystallization in a CSTR with initial distribution and growth", "[Crystallization],[PBM],[Simulation],[Reference],[CI]")
 {
-	const std::string& modelFilePath = std::string("/data/configuration_PBM_CSTR_growth_benchmark1.json");
-	const std::string& refFilePath = std::string("/data/ref_PBM_CSTR_growth_benchmark1.h5");
+	const std::string& modelFilePath = std::string("/data/configuration_cry_CSTR_PBM_growth_benchmark1.json");
+	const std::string& refFilePath = std::string("/data/ref_cry_CSTR_PBM_growth_benchmark1.h5");
 	const std::vector<double> absTol = { 1e-10 };
 	const std::vector<double> relTol = { 1e-10 };
 
@@ -36,10 +40,10 @@ TEST_CASE("Crystallization in a CSTR with initial distribution and growth", "[Cr
 	cadet::test::column::testReferenceBenchmark(modelFilePath, refFilePath, "001", absTol, relTol, disc, false);
 }
 
-TEST_CASE("Crystallization in a CSTR with initial distribution and size-dependent growth", "[Crystallization],[Simulation],[Reference],[CI]")
+TEST_CASE("Crystallization in a CSTR with initial distribution and size-dependent growth", "[Crystallization],[PBM],[Simulation],[Reference],[CI]")
 {
-	const std::string& modelFilePath = std::string("/data/configuration_PBM_CSTR_growthSizeDep_benchmark1.json");
-	const std::string& refFilePath = std::string("/data/ref_PBM_CSTR_growthSizeDep_benchmark1.h5");
+	const std::string& modelFilePath = std::string("/data/configuration_cry_CSTR_PBM_growthSizeDep_benchmark1.json");
+	const std::string& refFilePath = std::string("/data/ref_cry_CSTR_PBM_growthSizeDep_benchmark1.h5");
 	const std::vector<double> absTol = { 1e-10 };
 	const std::vector<double> relTol = { 1e-10 };
 
@@ -47,10 +51,10 @@ TEST_CASE("Crystallization in a CSTR with initial distribution and size-dependen
 	cadet::test::column::testReferenceBenchmark(modelFilePath, refFilePath, "001", absTol, relTol, disc, false);
 }
 
-TEST_CASE("Crystallization in a CSTR with primary nucleation and growth", "[Crystallization],[Simulation],[Reference],[CI]")
+TEST_CASE("Crystallization in a CSTR with primary nucleation and growth", "[Crystallization],[PBM],[Simulation],[Reference],[CI]")
 {
-	const std::string& modelFilePath = std::string("/data/configuration_PBM_CSTR_primaryNucleationAndGrowth_benchmark1.json");
-	const std::string& refFilePath = std::string("/data/ref_PBM_CSTR_primaryNucleationAndGrowth_benchmark1.h5");
+	const std::string& modelFilePath = std::string("/data/configuration_cry_CSTR_PBM_primaryNucleationAndGrowth_benchmark1.json");
+	const std::string& refFilePath = std::string("/data/ref_cry_CSTR_PBM_primaryNucleationAndGrowth_benchmark1.h5");
 	const std::vector<double> absTol = { 1e-10 };
 	const std::vector<double> relTol = { 1e-10 };
 
@@ -58,10 +62,10 @@ TEST_CASE("Crystallization in a CSTR with primary nucleation and growth", "[Crys
[…]
