### 2025-04-15 09e5043
Add standalone aggregation/fragmentation code (#241)
This commit adds code for standalone aggregation and fragmentation
crystallization reactions. They can be used as an alternative reaction
by including the respective file. This commit is preparatory for an
actual incorporation and unification into a modularized crystallization
module.
