### 2025-04-17 9a579b2
Improve crystallization cache of reconstruction schemes and Jacobian
