+++ src/libcadet/model/reaction/CrystallizationReaction.cpp
@@ -33,22 +33,22 @@ namespace model
 {
 namespace detail
 {
-			class CrystallizationMode {
-			public:
-				bool hasPBM()    const { return state_[0]; };
-				bool hasAggregation()  const { return state_[1]; };
-				bool hasFragmentation()  const { return state_[2]; };
-				unsigned int getMode() const { return state_.to_ulong(); };
+	class CrystallizationMode {
+	public:
+		bool hasPBM()    const { return state_[0]; };
+		bool hasAggregation()  const { return state_[1]; };
+		bool hasFragmentation()  const { return state_[2]; };
+		unsigned int getMode() const { return state_.to_ulong(); };
 
-				void SetMode(unsigned int mode)
-	{
-					if (mode > 7 || mode < 1)
-						throw InvalidParameterException("Crystallization mode specified as " + std::to_string(mode) + ", which is not defined.");
+		void SetMode(unsigned int mode)
+		{
+			if (mode > 7 || mode < 1)
+				throw InvalidParameterException("Crystallization mode specified as " + std::to_string(mode) + ", which is not defined.");
 
-					state_ = std::bitset<3>(mode);
-				}
-			private:
-				std::bitset<3> state_;
+			state_ = std::bitset<3>(mode);
+		}
+	private:
+		std::bitset<3> state_;
 	};
 
 	/**
@@ -117,154 +117,6 @@ namespace detail
 			r_x_i(r_x_i), phi(phi), F_i(F_i) {}
 	};
 
-	struct HRKorenCoefficients
-	{
-		std::vector<active> A_coeff;
-		std::vector<active> R_coeff;
-
-		HRKorenCoefficients(const std::vector<active>& binSizes)
-			: A_coeff(binSizes.size() - 1), R_coeff(binSizes.size() - 1)
-		{
-			// calculate the coefficients
-			for (int i = 1; i + 1 < binSizes.size(); ++i)
-			{
-				const active delta_xi_xip1 = binSizes[i] + binSizes[i + 1];
-				A_coeff[i] = delta_xi_xip1 / (binSizes[i] + binSizes[i - 1]);
-				R_coeff[i] = delta_xi_xip1 / binSizes[i];
-			}
-		}
-	};
-
-	/**
-	 * Flux reconstruction scheme coefficients.For details see @cite Zhang2024
-	*/
-	struct Weno3Coefficients
-	{
-		std::vector<active> q_1_right_coeff;
-		std::vector<active> q_0_right_coeff;
-		std::vector<active> C_right_coeff;
-		std::vector<active> IS_0_coeff;
-		std::vector<active> IS_1_coeff;
-
-		Weno3Coefficients(const std::vector<active>& binSizes)
-			: q_1_right_coeff(binSizes.size() - 1), q_0_right_coeff(binSizes.size() - 1), C_right_coeff(binSizes.size() - 1), IS_0_coeff(binSizes.size() - 1), IS_1_coeff(binSizes.size() - 1)
-		{
-			// calculate the coefficients and store them, the first element is empty
-			for (int i = 1; i + 1 < binSizes.size(); ++i)
-			{
-				const active delta_sum = binSizes[i] + binSizes[i - 1] + binSizes[i + 1];
-				const active delta_left_sum = binSizes[i] + binSizes[i - 1];
[…]
