### 2025-04-18 ba6358c
Crystallization module: Add exceptions for disallowed configurations (#241)
