### 2025-04-18 ecab878
Add documentation for population balances with aggregation and breakage
