+++ doc/modelling/crystallization/population_balance_model.rst
@@ -3,15 +3,15 @@
 Population Balance Models
 ~~~~~~~~~~~~~~~~~~~~~~~~~
 
-The population balance model (PBM) is given by a particle-number continuity equation which describes the evolution of the number density :math:`n` of the particles over time :math:`t` and with respect to a so-called internal coordinate :math:`x` and external coordinate :math:`z`. 
-The internal coordinate can be chosen as any characteristic property of the particles such as the particle size or volume.
+The population balance model (PBM) is given by a particle-number continuity equation which describes the evolution of the number density :math:`n` of the particles over time :math:`t` and with respect to size, the so-called internal coordinate :math:`x`, and external coordinate :math:`z`. 
 The external coordinate can be a characteristic dimension of the reactor itself, including its axial length. 
 
-In CADET, we chose particle size and axial coordinate as the internal and external coodinate respectively.
 The PBM in CADET is implemented so that it can be used in any unit operation that supports reactions.
 Typical applications consider crystallization in a CSTR or, to model continuous processes, in a Dispersive Plug-Flow Reactor (DPFR), which is described by the LRM without solid phase.
 
-In the following, we give a brief overview on the modelling equations, for more information on the PBM as implemented in CADET, please refer to :cite:`Zhang2024`.
+In the following, we give a brief overview on the PBM modelling equations for growth, nucleation, growth rate dispersion.
+These equations can be combined with :ref:`aggregation_model` and / or :ref:`fragmentation_model`.
+For more information on the PBM as implemented in CADET, please refer to :cite:`Zhang2024` and :cite:`Zhang2025`.
 
 .. figure:: PBM_Part_I.png
     
