### 2025-04-18 d0888cb
Add documentation for population balances with aggregation and breakage
