### 2025-04-18 e36b821
Add documentation for population balances with aggregation and breakage
