### 2025-04-18 30620b8
FD Jacobian tests: Only compare pattern under specific tolerance
For some models with numerically challenging values, the FD Jacobian
pattern comparison seems to be broken / makes no sense due to large
inaccuracies.
