### 2025-04-25 2931a02
Add competitive inhibition and uncompetitive inhibition
