+++ src/libcadet/ReactionModelFactory.cpp
@@ -22,6 +22,8 @@ namespace cadet
 			void registerMassActionLawReaction(std::unordered_map<std::string, std::function<model::IDynamicReactionModel*()>>& reactions);
 			void registerDummyReaction(std::unordered_map<std::string, std::function<model::IDynamicReactionModel*()>>& reactions);
 			void registerMichaelisMentenReaction(std::unordered_map<std::string, std::function<model::IDynamicReactionModel*()>>& reactions);
+			void registerMultiMichaelisMentenInhibitionReaction(std::unordered_map<std::string, std::function<model::IDynamicReactionModel* ()>>& reactions);
+
 
 			void registerCrystallizationReaction(std::unordered_map<std::string, std::function<model::IDynamicReactionModel*()>>& reactions);
 		}
@@ -33,6 +35,7 @@ namespace cadet
 		model::reaction::registerDummyReaction(_dynamicModels);
 		model::reaction::registerMassActionLawReaction(_dynamicModels);
 		model::reaction::registerMichaelisMentenReaction(_dynamicModels);
+		model::reaction::registerMultiMichaelisMentenInhibitionReaction(_dynamicModels);
 
 		model::reaction::registerCrystallizationReaction(_dynamicModels);
 	}
