### 2025-04-25 c4e4b5d
Add competative inhibition and uncompetative inhibition
