### 2025-04-25 145d401
Add competative inhibition and uncompetative inhibition
