+++ src/libcadet/model/reaction/MichaelisMentenInhibitionReaction.cpp
@@ -1,0 +1,514 @@
+// =============================================================================
+//  CADET - The Chromatography Analysis and Design Toolkit
+//
+//  Copyright © 2008-present: The CADET-Core Authors
+//            Please see the AUTHORS.md file.
+//
+//  All rights reserved. This program and the accompanying materials
+//  are made available under the terms of the GNU Public License v3.0 (or, at
+//  your option, any later version) which accompanies this distribution, and
+//  is available at http://www.gnu.org/licenses/gpl.html
+// =============================================================================
+
+#include "model/reaction/ReactionModelBase.hpp"
+#include "model/ExternalFunctionSupport.hpp"
+#include "model/ModelUtils.hpp"
+#include "cadet/Exceptions.hpp"
+#include "model/Parameters.hpp"
+#include "LocalVector.hpp"
+#include "SimulationTypes.hpp"
+#include "linalg/ActiveDenseMatrix.hpp"
+#include "MathUtil.hpp"
+#include "Memory.hpp"
+
+#include <algorithm>
+#include <unordered_map>
+#include <string>
+#include <vector>
+
+/*<codegen>
+{
+    "name": "MultiMichaelisMentenInhibitionParamHandler",
+    "externalName": "ExtMultiMichaelisMentenInhibitionParamHandler",
+    "parameters":
+        [
+            { "type": "ScalarReactionDependentParameter", "varName": "vMax", "confName": "MM_VMAX"},
+            { "type": "ScalarReactionDependentParameter", "varName": "kMM", "confName": "MM_KMM"},
+            
+            { "type": "ComponentDependentReactionDependentParameter", "varName": "kInhibitComp", "confName": "MM_KIC"},
+            { "type": "ComponentDependentReactionDependentParameter", "varName": "kInhibitUnComp", "confName": "MM_KIUC"}
+        ]
+}
+</codegen>*/
+
+/* Parameter description
+ ------------------------
+ MM_VMAX - Maximum reaction rate
+ MM_KMM - Michaelis-Menten constant for each component
+ MM_KIC - Competitive inhibition constants (K̃_{i,j,k})
+ MM_KIUC - Non-competitive/Uncompetitive inhibition constants (K_{i,j,k})
+*/
+
+
+namespace cadet
+{
+namespace model
+{
+
+inline const char* MultiMichaelisMentenInhibitionParamHandler::identifier() CADET_NOEXCEPT { return "MICHAELIS_MENTEN_INHIBITION"; }
+
+inline bool MultiMichaelisMentenInhibitionParamHandler::validateConfig(unsigned int nReactions, unsigned int nComp, unsigned int const* nBoundStates)
+{
+    return true;
+}
+
+inline const char* ExtMultiMichaelisMentenInhibitionParamHandler::identifier() CADET_NOEXCEPT { return "EXT_MICHAELIS_MENTEN_INHIBITION"; }
+
+inline bool ExtMultiMichaelisMentenInhibitionParamHandler::validateConfig(unsigned int nReactions, unsigned int nComp, unsigned int const* nBoundStates)
+{
+    return true;
+}
+
+namespace
+{
+    /**
+        * @brief Registers a matrix-valued parameter (row-major storage) with components as rows
+        * @details The matrix-valued parameter has as many rows as there are components in the system.
+        * @param [in,out] parameters Parameter map
+        * @param [in] unitOpIdx Unit operation id
+        * @param [in] parTypeIdx Particle type index
+        * @param [in] paramName Name of the parameter
+        * @param [in] mat Matrix to register
+        */
[…]
