### 2025-04-25 44ba849
Revert "Add competative inhibition and uncompetative inhibition"
This reverts commit 145d401c6f8a14012b2ddb39337075d3e82f1aca.
