+++ src/libcadet/CMakeLists.txt
@@ -114,7 +114,6 @@ set(LIBCADET_REACTIONMODEL_SOURCES
 	${CMAKE_SOURCE_DIR}/src/libcadet/model/reaction/DummyReaction.cpp
 	${CMAKE_SOURCE_DIR}/src/libcadet/model/reaction/MassActionLawReaction.cpp
 	${CMAKE_SOURCE_DIR}/src/libcadet/model/reaction/MichaelisMentenReaction.cpp
-	${CMAKE_SOURCE_DIR}/src/libcadet/model/reaction/MichaelisMentenInhibitionReaction.cpp
 
 	${CMAKE_SOURCE_DIR}/src/libcadet/model/reaction/CrystallizationReaction.cpp
 )
