+++ doc/modelling/reaction/michaelis_menten_kinetics.rst
@@ -36,7 +36,7 @@ In this case, the flux has the form
         \nu_j = \frac{\mu_{\mathrm{max},j} c_S}{k_{\mathrm{MM},j} + c_S} \cdot \frac{1}{1 + \sum_i c_{Ii}/k_{I,j,i}}.
     \end{aligned}
 
-Here, :math:k_{\mathrm{I},j,i} is the inhibition constant with respect to component :math:i in reaction :math:j.
-If :math:k_{\mathrm{I},j,i} \leq 0, component :math:i does not act as an inhibitor.
+Here, :math:`k_{\mathrm{I},j,i}` is the inhibition constant with respect to component :math:`i` in reaction :math:`j`.
+If :math:`k_{\mathrm{I},j,i} \leq 0`, component :math:`i` does not act as an inhibitor.
 
 Note: Currently, the model does not allow substrates to function as inhibitors.
