### 2025-04-30 091d134
modify CI to handle releases and pre-releases
