+++ .github/workflows/cd_docker.yml
@@ -3,7 +3,8 @@ name: ci
 on:
   push:
     tags:
-      - "v*.*.*.*"
+      - 'v[0-9]+.[0-9]+.[0-9]+'          # Matches v1.2.3
+      - 'v[0-9]+.[0-9]+.[0-9]+-*'        # Matches v1.2.3-alpha, v1.2.3-beta.1, etc.
 
 jobs:
   docker:
@@ -22,6 +23,8 @@ jobs:
             type=semver,pattern={{version}}
             type=semver,pattern={{major}}.{{minor}}
             type=semver,pattern={{major}}
+          # Only set 'latest' if it's NOT a pre-release
+            type=raw,value=latest,enable=${{ !contains(github.ref, '-') }} 
 
       - name: Login to Docker Hub
         if: github.event_name != 'pull_request'
