+++ src/build-tools/CMakeLists.txt
@@ -13,11 +13,13 @@
 # Name of the current project
 project(CadetBuildTools CXX C)
 
-# Add the template code generator
-add_executable(templateCodeGen ${CMAKE_SOURCE_DIR}/src/build-tools/templateCodeGen.cpp)
-target_include_directories(templateCodeGen PRIVATE ${CMAKE_SOURCE_DIR}/ThirdParty/json ${CMAKE_SOURCE_DIR}/ThirdParty/inja)
-target_compile_features(templateCodeGen PRIVATE cxx_std_23)
-set_target_properties(templateCodeGen PROPERTIES CXX_EXTENSIONS OFF)
+if (EXTERNAL_TEMPLATE_CODEGEN STREQUAL "")
+	# Add the template code generator
+	add_executable(templateCodeGen ${CMAKE_SOURCE_DIR}/src/build-tools/templateCodeGen.cpp)
+	target_include_directories(templateCodeGen PRIVATE ${CMAKE_SOURCE_DIR}/ThirdParty/json ${CMAKE_SOURCE_DIR}/ThirdParty/inja)
+	target_compile_features(templateCodeGen PRIVATE cxx_std_23)
+	set_target_properties(templateCodeGen PROPERTIES CXX_EXTENSIONS OFF)
+endif()
 
 # Info message
 message(STATUS "Added build tools")
