### 2025-05-02 8a53632
Add support for external templateCodeGen in CMake
This helps in cross compiled builds where the cross compiler would also
build templateCodeGen. Here, templateCodeGen cannot run on the host and
the build fails.
