### 2025-05-04 4ee38a1
Add test for crystallization with combined PBM, aggregation and fragmentation
