### 2025-05-04 1e36847
Add test for crystallization with combined mode (#401)
This commit adds a Jacobian and numerical reference test for crystallization with growth, aggregation, fragmentation in a CSTR.
