### 2025-05-05 3ef8dc6
Replace CADET with CADET-Core
