+++ BUILD-LINUX.md
@@ -1,6 +1,9 @@
+# Build CADET-Core on Linux
+
 ## Prerequisites
 
 * CMake (>= 3.12.0)
+
 * GCC >= 7.0, Clang >= 3.9, or Intel C++ 18.0
 * Optional: Git
 
