+++ doc/developer_guide/model_expansion.rst
@@ -20,51 +20,22 @@ An extensive description on how to add a new binding model to CADET-Core is give
 A more concise description is given in the following:
 
 1.
-Use the `binding model template <https://github.com/cadet/CADET-Core/tree/master/doc/developer_guide/TemplateBinding.cpp>`_ to create a new binding model file in the CADET binding model source directory
-
-``\src\libcadet\model\binding\YourModelNameBinding.cpp``
-
-and rename the model in the file (i.e. Template to YoureModelName).
-Note that the provided template binding model is an implementation of the multi-component Langmuir binding.
-
+   Use the `binding model template <https://github.com/cadet/CADET-Core/tree/master/doc/developer_guide/TemplateBinding.cpp>`_ to create a new binding model file in the CADET binding model source directory ``\src\libcadet\model\binding\YourModelNameBinding.cpp`` and rename the model in the file (i.e. Template to YoureModelName).
+   Note that the provided template binding model is an implementation of the multi-component Langmuir binding.
 2.
-Register the binding model by adding
-
-``void registerYourModelNameModel(std::unordered_map<std::string, std::function<model::IBindingModel* ()>>& bindings);``
-
-to the ``src\libcadet\BindingModelFactory.cpp`` at the end of all the other registering functions such as ``void registerLinearModel``.
-Then add
-
-``model::binding::registerExampleModel(_bindingModels);``
-
-to the ``BindingModelFactory.cpp`` at the end of all the other registering functions such as ``model::binding::registerLinearModel(_bindingModels)``.
-The final step to register your model is to add your model to the ``\src\libcadet\CMakeLists.txt`` (again look for similar statements for the other binding models) by adding
-
-``${CMAKE_SOURCE_DIR} /src/libcadet/model/binding/ExampleBinding.cpp``
-
-Before continuing with the third step, you should rebuild CADET-Core to verify that the first two steps went well.
-
+   Register the binding model by adding ``void registerYourModelNameModel(std::unordered_map<std::string, std::function<model::IBindingModel* ()>>& bindings);`` to the ``src\libcadet\BindingModelFactory.cpp`` at the end of all the other registering functions such as ``void registerLinearModel``.
+   Then add ``model::binding::registerExampleModel(_bindingModels);`` to the ``BindingModelFactory.cpp`` at the end of all the other registering functions such as ``model::binding::registerLinearModel(_bindingModels)``.
+   The final step to register your model is to add your model to the ``\src\libcadet\CMakeLists.txt`` (again look for similar statements for the other binding models) by adding ``${CMAKE_SOURCE_DIR} /src/libcadet/model/binding/ExampleBinding.cpp``.
+   Before continuing with the third step, you should rebuild CADET-Core to verify that the first two steps went well.
 3.
-The actual implementation of the new binding model follows two main steps: the configuration of the relevant mechanistic parameters and implementation of adsorption flux and Jacobian.
-To set up the configuration of isotherm parameters a macro (.json script) has been included in the code, which generates the relevant code section when the user defines the parameters in the scope of this script.
-To modify the script go to Line 30 in the provided file template and adjust the parameters to your needs, a more detailed description for that is provided in the `aforementioned forum post <https://forum.cadet-web.de/t/registration-implementation-and-testing-of-new-binding-model-in-cadet/533>`_.
-
-Next, the adsorption flux equations need to be implemented int the corresponding function
-
-``int fluxImpl(double t, unsigned int secIdx, const ColumnPosition& colPos, StateType const* y, CpStateType const* yCp, ResidualType* res, LinearBufferAllocator workSpace) const``
-
[…]
