+++ BUILD-WINDOWS.md
@@ -1,3 +1,5 @@
+# Build CADET-Core on Windows
+
 ## Prerequisites
 
 * Microsoft Visual Studio **2022** (Community Edition, or better)
