### 2025-05-05 28a2a0a
Update literature / publication list
