+++ doc/interface/unit_operations/2d_general_rate_model.rst
@@ -12,171 +12,171 @@ For information on model equations, refer to :ref:`2d_general_rate_model_model`.
 ``UNIT_TYPE``
 
    Specifies the type of unit operation model
-   
+
    ================  =================================================  =============
    **Type:** string  **Range:** :math:`\texttt{GENERAL_RATE_MODEL_2D}`  **Length:** 1
    ================  =================================================  =============
-   
+
 ``NCOMP``
 
    Number of chemical components in the chromatographic medium
-   
+
    =============  =========================  =============
    **Type:** int  **Range:** :math:`\geq 1`  **Length:** 1
    =============  =========================  =============
-   
+
 ``ADSORPTION_MODEL``
 
    Specifies the type of binding model of each particle type (or of all particle types if length is 1)
-   
+
    ================  ==========================================  ==========================================
    **Type:** string  **Range:** See Section :ref:`FFAdsorption`  **Length:** :math:`1 / \texttt{NPARTYPE}`
    ================  ==========================================  ==========================================
-   
+
 ``ADSORPTION_MODEL_MULTIPLEX``
 
    Multiplexing mode of :math:`\texttt{ADSORPTION_MODEL}`. If set to 0, each particle type has a different binding model and the length of :math:`\texttt{ADSORPTION_MODEL}` is :math:`\texttt{NPARTYPE}`. If set to 1, all particle types share the same binding model and the length of :math:`\texttt{ADSORPTION_MODEL}` is 1.  This field is optional and inferred from the length of :math:`\texttt{ADSORPTION_MODEL}` if left out.
-   
+
    =============  ===========================  =============
    **Type:** int  **Range:** :math:`\{0, 1\}`  **Length:** 1
    =============  ===========================  =============
-   
+
 ``NBOUND``
 
    Number of bound states for each component in each particle type in type-major ordering
-   
+
    =============  =========================  ==========================================================================
    **Type:** int  **Range:** :math:`\geq 0`  **Length:** :math:`\texttt{NCOMP} / \texttt{NPARTYPE} \cdot \texttt{NCOMP}`
    =============  =========================  ==========================================================================
-   
+
 ``REACTION_MODEL_BULK``
 
    Specifies the type of reaction model of the bulk volume. The model is configured in the subgroup :math:`\texttt{reaction_bulk}`.
-   
+
    ================  ========================================  =============
    **Type:** string  **Range:** See Section :ref:`FFReaction`  **Length:** 1
    ================  ========================================  =============
-   
+
 ``REACTION_MODEL_PARTICLES``
 
    Specifies the type of reaction model of each particle type (or of all particle types if length is 1). The model is configured in the subgroup :math:`\texttt{reaction_particle}`, or :math:`\texttt{reaction_particle_XXX}` in case of disabled multiplexing.
-   
+
    ================  ========================================  ==========================================
    **Type:** string  **Range:** See Section :ref:`FFReaction`  **Length:** :math:`1 / \texttt{NPARTYPE}`
    ================  ========================================  ==========================================
-   
+
 ``REACTION_MODEL_PARTICLES_MULTIPLEX``
 
[…]
