+++ doc/interface/unit_operations/multi_channel_transport_model.rst
@@ -1,6 +1,6 @@
 .. _multi_channel_transport_model_config:
 
-Multichannel Transport model (MCT model) 
+Multichannel Transport model (MCT model)
 ========================================
 
 Group /input/model/unit_XXX - UNIT_TYPE = MULTI_CHANNEL_TRANSPORT
@@ -11,84 +11,84 @@ For information on model equations, refer to :ref:`multi_channel_transport_model
 ``UNIT_TYPE``
 
    Specifies the type of unit operation model
-   
+
    ================  ===================================================  =============
    **Type:** string  **Range:** :math:`\texttt{MULTI_CHANNEL_TRANSPORT}`  **Length:** 1
    ================  ===================================================  =============
-   
+
 ``NCOMP``
 
    Number of chemical components in the chromatographic medium
-   
+
    =============  =========================  =============
    **Type:** int  **Range:** :math:`\geq 1`  **Length:** 1
    =============  =========================  =============
-   
+
 ``REACTION_MODEL_BULK``
 
    Specifies the type of reaction model of the bulk volume. The model is configured in the subgroup :math:`\texttt{reaction_bulk}`.
-   
+
    ================  ========================================  =============
    **Type:** string  **Range:** See Section :ref:`FFReaction`  **Length:** 1
    ================  ========================================  =============
-   
+
 ``INIT_C``
 
    Initial concentrations for each component in all channels of the bulk mobile phase (length :math:`\texttt{NCOMP}`), or for each component in each channel (length :math:`\texttt{NCOMP} \cdot \texttt{NCHANNEL}`, ordering channel-major)
 
    **Unit:** :math:`\mathrm{mol}\,\mathrm{m}^{-3}`
-   
+
    ================  =========================  =========================================================================
    **Type:** double  **Range:** :math:`\geq 0`  **Length:** :math:`\texttt{NCOMP} / \texttt{NCOMP} \cdot \texttt{NCHANNEL}`
    ================  =========================  =========================================================================
-   
+
 ``INIT_STATE``
 
    Full state vector for initialization (optional, :math:`\texttt{INIT_C}` will be ignored; if length is :math:`2\texttt{NDOF}`, then the second half is used for time derivatives)
 
    **Unit:** :math:`various`
-   
+
    ================  =============================  ==================================================
    **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** :math:`\texttt{NDOF} / 2\texttt{NDOF}`
    ================  =============================  ==================================================
-   
+
 ``COL_DISPERSION``
 
    Axial dispersion coefficient.  In case of a spatially inhomogeneous setting, the :math:`\texttt{SENS_PARTYPE}` field is used for indexing the channel when specifying parameter sensitivities.
 
    **Unit:** :math:`\mathrm{m}^{2}\,\mathrm{s}^{-1}`
-   
+
    ================  =========================  =========================================================
    **Type:** double  **Range:** :math:`\geq 0`  **Length:** see :math:`\texttt{COL_DISPERSION_MULTIPLEX}`
    ================  =========================  =========================================================
-   
+
 ``COL_DISPERSION_MULTIPLEX``
 
[…]
