### 2025-05-06 7677c32
Update documentation (#402)
* Formatting fixes

* Add links to interface/modeling sections

* Replace CADET with CADET-Core

* Fix character encoding

* Fix link to CADET-Core installation

* Include reaction models in toc

* Update literature / publication list
