+++ BUILD-WINDOWS.md
@@ -1,3 +1,5 @@
+# Build CADET-Core on Windows
+
 ## Prerequisites
 
 * Microsoft Visual Studio **2022** (Community Edition, or better)
@@ -5,16 +7,16 @@
 * Optional: Git
 * Optional but not generally recommended*: Intel OneAPI TBB
 
-*For most use-cases it is more efficient to parallelize by running multiple CADET simulations instead 
-of parallelizing within one CADET simulation. Including the parallelization code in CADET can lead to performance
-losses, even if parallelization within CADET is not used. 
-Therefore, we recommend not including the parallelization library TBB 
+*For most use-cases it is more efficient to parallelize by running multiple CADET-Core simulations instead
+of parallelizing within one CADET-Core simulation. Including the parallelization code in CADET-Core can lead to performance
+losses, even if parallelization within CADET-Core is not used.
+Therefore, we recommend not including the parallelization library TBB
 unless you know your simulations are large enough to benefit from it.
 
 Assumed directory structure:
 
 <pre>
-|- CADET
+|- CADET-Core
 |    - src
 |    - include
 |    - [...]
@@ -43,14 +45,14 @@ We are using Visual Studio because it is the easiest way to install all required
 - Clink provides text completion, history, and line-editing to Windows Command Prompt
 - Download the [clink installer exe](https://github.com/chrisant996/clink) and install clink.
 
-## Obtain CADET code
+## Obtain CADET-Core code
 
-- Clone the CADET source code into a `CADET` folder: 
-  - `git clone https://github.com/cadet/cadet-core.git CADET`
+- Clone the CADET-Core source code into a `CADET-Core` folder:
+  - `git clone https://github.com/cadet/cadet-core.git CADET-Core`
 
-## Build CADET in Visual Studio
+## Build CADET-Core in Visual Studio
 
-- Open Visual Studio and open the `CADET` folder
+- Open Visual Studio and open the `CADET-Core` folder
 - Navigate to "Tools" - "Command Line" and open either a "Developer Command Prompt" or "Developer PowerShell"
 - Execute the following command:
     - `vcpkg integrate install` (this only needs to be run _once_ per PC and will require admin privileges)
@@ -59,17 +61,17 @@ We are using Visual Studio because it is the easiest way to install all required
 - Wait for `cmake generation` to finish (see `output` window)
 - From the status bar at the top select `Build`, `Build all`
 - Once that finishes, select `Build`, `Install CadetFramework`
-- The binaries will be located in `CADET\out\install\aRELEASE\bin`
+- The binaries will be located in `CADET-Core\out\install\aRELEASE\bin`
 
 
-## Build CADET from the command line
+## Build CADET-Core from the command line
 
 - Open Visual Studio and select "continue without code"
 - Navigate to "Tools" - "Command Line" and open either a "Developer Command Prompt" or "Developer PowerShell"
 - Execute the following commands:
 - For Command Prompt:
-    - `if not exist CADET\build mkdir CADET\build`
[…]
