+++ doc/modelling/reaction/michaelis_menten_kinetics.rst
@@ -1,12 +1,11 @@
-.. _michaelis_menten_kinetics_model:
+.. _michaelis_menten_kinetics_model_with_inhibition:
 
-Michaelis Menten kinetics
+Michaelis Menten kinetics with inhibition
 -------------------------
 
 Implements liquid phase Michaelis-Menten reaction kinetics of the form
 
 .. math::
-
     \begin{aligned}
         f_\text{react} = S \mathbf{\nu},
     \end{aligned}
@@ -16,27 +15,74 @@ where :math:`S` is the stoichiometric matrix and the fluxes are given by
 .. math::
 
     \begin{aligned}
-        \nu_j = \frac{\mu_{\mathrm{max},j} \, c_S}{k_{\mathrm{MM},j} + c_S},
+        \nu_j = \prod_i^{N_{sub}} \frac{\mu_{\mathrm{max},j} \, c_i}{k_{\mathrm{MM},j} + c_i},
     \end{aligned}
 
 where
 
 - :math:`j` is the reaction index,
-- :math:`c_S` is the substrate component,
+- :math:`N_{sub}` is the number of substrates in the reaction,
+- :math:`c_i` is the substrate component,
 - :math:`\mu_{\mathrm{max},j}`, is the limiting rate approached by the system at saturation,
-- :math:`k_{\mathrm{MM},j}` is the Michaelis constant, which is defined as the concentration of substrate at which the reaction rate is half ov :math:`\mu_{\mathrm{max},j}`.
+- :math:`k_{\mathrm{MM},j}` is the Michaelis constant, which is defined as the concentration of substrate at which the reaction rate is half of :math:`\mu_{\mathrm{max},j}`.
 
 
-In addition, the reaction might be inhibited by other components.
-In this case, the flux has the form
+In addition, a substrate :math:`i` might be inhibited by other components.
+There are three types of inhibition implemented.
+
+Competitive inhibition
+^^^^^^^^^^^^^^^^^^^^^^^
 
 .. math::
 
     \begin{aligned}
-        \nu_j = \frac{\mu_{\mathrm{max},j} c_S}{k_{\mathrm{MM},j} + c_S} \cdot \frac{1}{1 + \sum_i c_{Ii}/k_{I,j,i}}.
+        \nu_{i,j} =  \frac{\mu_{\mathrm{max},j} \cdot c_i}{k_{\mathrm{MM},j}(1+\sum_k^{N_{inh}} \frac{c_{i,k}}{K_{i,j,k}}) + c_i},
     \end{aligned}
 
-Here, :math:`k_{\mathrm{I},j,i}` is the inhibition constant with respect to component :math:`i` in reaction :math:`j`.
-If :math:`k_{\mathrm{I},j,i} \leq 0`, component :math:`i` does not act as an inhibitor.
+where
+ - :math:`c_i` is the substrate component and :math:`c_{i,k}` is the inhibitor  acting on substrate :math:`c_i`.
+ - :math:`N_{inh}` is the number of inhibitors in the reaction,
+ - :math:`K_{i,j,k}` is the inhibition constant with respect to component :math:`c_{i,k}` in reaction :math:`j`.
+        If :math:`K_{i,j,k} \leq 0`, component :math:`c_{i,k}` does not act as an inhibitor.
 
-Note: Currently, the model does not allow substrates to function as inhibitors.
+
+Uncompetitive inhibition
+^^^^^^^^^^^^^^^^^^^^^^^
+
+.. math::
+
+    \begin{aligned}
[…]
