+++ doc/modelling/reaction/index.rst
@@ -10,7 +10,6 @@ CADET features one reaction type:
 
  - :ref:`mass_action_law_model` (most commonly used)
  - :ref:`michaelis_menten_kinetics_model`
- - :ref:`michaelis_menten_kinetics_model_with_inhibition`
 
 Historically, a chromatography system is modeled as a reaction system without considering any transport phenomenon. We also introduce some reaction-based models that can be solved in CADET:
 
