### 2025-05-06 26756ce
Add piece-wise constant radial dependence of parameters to 2D DG discretization (#337)
* Introduce element-wise radial zone dependence of parameters: velocity,
dispersion, porosity, particle type volume fractions

* Move radial dispersion into numerical flux

* Component dependent radial/axial dispersion

* Fix computation of velocity from flow rates and add consistency check
for element wise constant velocities

* Add numerical reference test
