+++ test/Crystallization.cpp
@@ -31,7 +31,7 @@ const char* getTestDirectory();
 */
 TEST_CASE("Crystallization in a CSTR with initial distribution and growth", "[Crystallization],[PBM],[Simulation],[Reference],[CI]")
 {
-	const std::string& modelFilePath = std::string("/data/configuration_cry_CSTR_PBM_growth_benchmark1.json");
+	const std::string& modelFilePath = std::string("/data/model_cry_CSTR_PBM_growth_benchmark1.json");
 	const std::string& refFilePath = std::string("/data/ref_cry_CSTR_PBM_growth_benchmark1.h5");
 	const std::vector<double> absTol = { 1e-10 };
 	const std::vector<double> relTol = { 1e-10 };
@@ -42,7 +42,7 @@ TEST_CASE("Crystallization in a CSTR with initial distribution and growth", "[Cr
 
 TEST_CASE("Crystallization in a CSTR with initial distribution and size-dependent growth", "[Crystallization],[PBM],[Simulation],[Reference],[CI]")
 {
-	const std::string& modelFilePath = std::string("/data/configuration_cry_CSTR_PBM_growthSizeDep_benchmark1.json");
+	const std::string& modelFilePath = std::string("/data/model_cry_CSTR_PBM_growthSizeDep_benchmark1.json");
 	const std::string& refFilePath = std::string("/data/ref_cry_CSTR_PBM_growthSizeDep_benchmark1.h5");
 	const std::vector<double> absTol = { 1e-10 };
 	const std::vector<double> relTol = { 1e-10 };
@@ -53,7 +53,7 @@ TEST_CASE("Crystallization in a CSTR with initial distribution and size-dependen
 
 TEST_CASE("Crystallization in a CSTR with primary nucleation and growth", "[Crystallization],[PBM],[Simulation],[Reference],[CI]")
 {
-	const std::string& modelFilePath = std::string("/data/configuration_cry_CSTR_PBM_primaryNucleationAndGrowth_benchmark1.json");
+	const std::string& modelFilePath = std::string("/data/model_cry_CSTR_PBM_primaryNucleationAndGrowth_benchmark1.json");
 	const std::string& refFilePath = std::string("/data/ref_cry_CSTR_PBM_primaryNucleationAndGrowth_benchmark1.h5");
 	const std::vector<double> absTol = { 1e-10 };
 	const std::vector<double> relTol = { 1e-10 };
@@ -64,7 +64,7 @@ TEST_CASE("Crystallization in a CSTR with primary nucleation and growth", "[Crys
 
 TEST_CASE("Crystallization in a CSTR with primary nucleation, growth and growth rate dispersion", "[Crystallization],[PBM],[Simulation],[Reference],[CI]")
 {
-	const std::string& modelFilePath = std::string("/data/configuration_cry_CSTR_PBM_primaryNucleationGrowthGrowthRateDispersion_benchmark1.json");
+	const std::string& modelFilePath = std::string("/data/model_cry_CSTR_PBM_primaryNucleationGrowthGrowthRateDispersion_benchmark1.json");
 	const std::string& refFilePath = std::string("/data/ref_cry_CSTR_PBM_primaryNucleationGrowthGrowthRateDispersion_benchmark1.h5");
 	const std::vector<double> absTol = { 1e-10 };
 	const std::vector<double> relTol = { 1e-10 };
@@ -75,7 +75,7 @@ TEST_CASE("Crystallization in a CSTR with primary nucleation, growth and growth 
 
 TEST_CASE("Crystallization in a CSTR with primary and secondary nucleation and growth", "[Crystallization],[PBM],[Simulation],[Reference],[CI]")
 {
[…]
