### 2025-05-06 d4c6796
Add numerical reference test for 2DLRMP to CI
