+++ src/libcadet/model/parts/ConvectionDispersionOperatorDG.hpp
@@ -21,7 +21,6 @@
 #include "ParamIdUtil.hpp"
 #include "AutoDiff.hpp"
 #include "Memory.hpp"
-//#include "Weno.hpp" // todo weno DG
 #include "SimulationTypes.hpp"
 #include <ParamReaderHelper.hpp>
 #include "linalg/BandedEigenSparseRowIterator.hpp"
@@ -64,10 +63,10 @@ namespace cadet
 			u c_{\text{in},i}(t) &= u c_i(t,0) - D_{\text{ax},i} \frac{\partial c_i}{\partial z}(t,0) \\
 			\frac{\partial c_i}{\partial z}(t,L) &= 0
 			\end{align} @f]
-			 * Methods are described in @cite todo, and @cite Puttmann2013, @cite Puttmann2016 (forward sensitivities, AD, band compression)
+			 * Methods are described in @cite Breuer2023, and @cite Puttmann2013, @cite Puttmann2016 (forward sensitivities, AD, band compression)
 			 *
 			 * This class does not store the Jacobian. It only fills existing matrices given to its residual() functions.
-			 * It assumes that there is no offset to the inlet in the local state vector and that the firsts cell is placed
+			 * It assumes that there is no offset to the inlet in the local state vector and that the firsts element is placed
 			 * directly after the inlet DOFs.
 			 */
 			class AxialConvectionDispersionOperatorBaseDG
@@ -79,7 +78,7 @@ namespace cadet
 
 				void setFlowRates(const active& in, const active& out, const active& colPorosity) CADET_NOEXCEPT;
 
-				bool configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, unsigned int nComp, int polynomial_integration_mode, unsigned int nCells, unsigned int polyDeg, unsigned int strideNode);
+				bool configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, unsigned int nComp, int polynomial_integration_mode, unsigned int nelements, unsigned int polyDeg, unsigned int strideNode);
 				bool configure(UnitOpIdx unitOpIdx, IParameterProvider& paramProvider, std::unordered_map<ParameterId, active*>& parameters);
 				bool notifyDiscontinuousSectionTransition(double t, unsigned int secIdx, Eigen::MatrixXd& jacInlet);
 
@@ -101,10 +100,10 @@ namespace cadet
 				inline const active& currentVelocity(double pos) const CADET_NOEXCEPT { return _curVelocity; }
 				inline bool forwardFlow() const CADET_NOEXCEPT { return _curVelocity >= 0.0; }
 
-				inline double cellLeftBound(unsigned int idx) const CADET_NOEXCEPT { return idx * static_cast<double>(_deltaZ); }
+				inline double elemLeftBound(unsigned int idx) const CADET_NOEXCEPT { return idx * static_cast<double>(_deltaZ); }
 				double relativeCoordinate(unsigned int idx) const CADET_NOEXCEPT
 				{
-					//const unsigned int cell = floor(idx / _nNodes);
+					//const unsigned int element = floor(idx / _nNodes);
 					//const unsigned int node = idx % _nNodes;
 					// divide by column length to get relative position
[…]
