+++ src/libcadet/model/LumpedRateModelWithPoresDG-LinearSolver.cpp
@@ -128,7 +128,7 @@ int LumpedRateModelWithPoresDG::linearSolve(double t, double alpha, double outer
 
 	// Handle inlet DOFs:
 	// Inlet at z = 0 for forward flow, at z = L for backward flow.
-	unsigned int offInlet = (_convDispOp.forwardFlow()) ? 0 : (_disc.nCol - 1u) * idxr.strideColCell();
+	unsigned int offInlet = (_convDispOp.forwardFlow()) ? 0 : (_disc.nElem - 1u) * idxr.strideColCell();
 
 	for (int comp = 0; comp < _disc.nComp; comp++) {
 		for (int node = 0; node < (_disc.exactInt ? _disc.nNodes : 1); node++) {
