+++ src/libcadet/model/parts/ConvectionDispersionOperatorDG.cpp
@@ -13,7 +13,6 @@
 #include "model/parts/ConvectionDispersionOperatorDG.hpp"
 #include "cadet/Exceptions.hpp"
 
-#include "Stencil.hpp"
 #include "ParamReaderHelper.hpp"
 #include "AdUtils.hpp"
 #include "SimulationTypes.hpp"
@@ -65,23 +64,23 @@ AxialConvectionDispersionOperatorBaseDG::~AxialConvectionDispersionOperatorBaseD
  * @param [in] helper Configures parameter dependencies
  * @param [in] nComp Number of components
  * @param [in] exact_integration DG volume integral computation
- * @param [in] nCells Number of axial cells
+ * @param [in] nElem Number of axial elements
  * @param [in] polyDeg Polynomial degree of DG approach
  * @param [in] strideNode node stride in state vector
  * @return @c true if configuration went fine, @c false otherwise
  */
-bool AxialConvectionDispersionOperatorBaseDG::configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, unsigned int nComp, int polynomial_integration_mode, unsigned int nCells, unsigned int polyDeg, unsigned int strideNode)
+bool AxialConvectionDispersionOperatorBaseDG::configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, unsigned int nComp, int polynomial_integration_mode, unsigned int nElem, unsigned int polyDeg, unsigned int strideNode)
 {
 	const bool firstConfigCall = _auxState == nullptr; // used to not multiply allocate memory
 
 	_nComp = nComp;
 	_exactInt = static_cast<bool>(polynomial_integration_mode); // only integration mode 0 applies the inexact collocated diagonal LGL mass matrix
-	_nCells = nCells;
+	_nElem = nElem;
 	_polyDeg = polyDeg;
 	_nNodes = _polyDeg + 1u;
-	_nPoints = _nNodes * _nCells;
+	_nPoints = _nNodes * _nElem;
 	_strideNode = strideNode;
-	_strideCell = _nNodes * strideNode;
+	_strideElem = _nNodes * strideNode;
 
 	/* Allocate space for DG discretization operations */
 	_nodes.resize(_nNodes);
@@ -102,7 +101,7 @@ bool AxialConvectionDispersionOperatorBaseDG::configureModelDiscretization(IPara
 		_subsState[i] = 0.0;
 	}
 	_boundary.setZero();
-	_surfaceFlux.resize(_nCells + 1u);
+	_surfaceFlux.resize(_nElem + 1u);
 	_surfaceFlux.setZero();
 
 	_newStaticJac = true;
@@ -143,23 +142,23 @@ bool AxialConvectionDispersionOperatorBaseDG::configure(UnitOpIdx unitOpIdx, IPa
 
 	// Read geometry parameters
 	_colLength = paramProvider.getDouble("COL_LENGTH");
-	_deltaZ = _colLength / _nCells;
+	_deltaZ = _colLength / _nElem;
 
 	/* compute dispersion jacobian blocks(without parameters except element spacing, i.e. static entries) */
-	// we only need unique dispersion blocks, which are given by cells 1, 2, nCol for inexact integration DG and by cells 1, 2, 3, nCol-1, nCol for eaxct integration DG
+	// we only need unique dispersion blocks, which are given by elements 1, 2, nElem for inexact integration DG and by elements 1, 2, 3, nElem-1, nElem for eaxct integration DG
 	if (firstConfigCall)
[…]
