### 2025-05-07 82b1776
WIP: outsource particle discretization code
