+++ src/libcadet/model/parts/ParticleDispersionOperatorDG.hpp
@@ -1,0 +1,217 @@
+// =============================================================================
+//  CADET
+//  
+//  Copyright © 2008-present: The CADET-Core Authors
+//            Please see the AUTHORS.md file.
+//  
+//  All rights reserved. This program and the accompanying materials
+//  are made available under the terms of the GNU Public License v3.0 (or, at
+//  your option, any later version) which accompanies this distribution, and
+//  is available at http://www.gnu.org/licenses/gpl.html
+// =============================================================================
+
+/**
+ * @file
+ * Defines the particle dispersion transport operator according to the discontinuous Galerkin discretization.
+ */
+
+#ifndef LIBCADET_PARTICLEDISPERSIONOPERATORDG_HPP_
+#define LIBCADET_PARTICLEDISPERSIONOPERATORDG_HPP_
+
+#include "ParamIdUtil.hpp"
+#include "AutoDiff.hpp"
+#include "Memory.hpp"
+#include "SimulationTypes.hpp"
+#include "ParamReaderHelper.hpp"
+#include "linalg/BandedEigenSparseRowIterator.hpp"
+#include "model/parts/DGToolbox.hpp"
+#include "model/ParameterMultiplexing.hpp"
+
+#include <unordered_map>
+#include <unordered_set>
+#include <Eigen/Dense>
+#include <Eigen/Sparse>
+#include <vector>
+
+using namespace Eigen;
+
+namespace cadet
+{
+
+class IParameterProvider;
+class IConfigHelper;
+struct AdJacobianParams;
+struct SimulationTime;
+class IModel;
+
+namespace model
+{
+
+class IParameterStateDependence;
+
+namespace parts
+{
+
+	constexpr double _SurfVolRatioSphere = 3.0; //!< Surface to volume ratio for a spherical particle
+	constexpr double _SurfVolRatioCylinder = 2.0; //!< Surface to volume ratio for a cylindrical particle
+	constexpr double _SurfVolRatioSlab = 1.0; //!< Surface to volume ratio for a slab-shaped particle
+
+	/**
+	 * @brief Particle dispersion transport operator
+	 * @details Implements the equation
+	 *
+	 * @f[\begin{align}
+	 \frac{\partial c^{\mathrm{p}}_{i}}{\partial t} + \frac{1 - \varepsilon^{\mathrm{p}}}{\varepsilon^{\mathrm{p}}} \frac{\partial c^{\mathrm{s}}_{i}}{\partial t} &=
+	 \frac{1}{r^2} \frac{\partial }{\partial r} \left( r^2 D_{i}^{\mathrm{p}} \frac{\partial c^{\mathrm{p}}_{i}}{\partial r} \right)
+	 - \frac{1 - \varepsilon^{\mathrm{p}}}{\varepsilon^{\mathrm{p}}} \frac{1}{r^2} \frac{\partial }{\partial r} \left( r^2 D_{i}^{\mathrm{s}} \frac{\partial c^{\mathrm{s}}_{i}}{\partial r} \right) \\
+	 \end{align} @f]
+	 * with Danckwerts boundary conditions (see @cite Danckwerts1953)
+	 @f[ \begin{align}
[…]
