+++ src/libcadet/model/parts/ParticleDispersionOperatorDG.cpp
@@ -1,0 +1,1568 @@
+// =============================================================================
+//  CADET
+//
+//  Copyright © 2008-present: The CADET-Core Authors
+//            Please see the AUTHORS.md file.
+//
+//  All rights reserved. This program and the accompanying materials
+//  are made available under the terms of the GNU Public License v3.0 (or, at
+//  your option, any later version) which accompanies this distribution, and
+//  is available at http://www.gnu.org/licenses/gpl.html
+// =============================================================================
+
+#include "model/parts/ParticleDispersionOperatorDG.hpp"
+#include "cadet/Exceptions.hpp"
+
+#include "ParamReaderHelper.hpp"
+#include "AdUtils.hpp"
+#include "SimulationTypes.hpp"
+#include "model/ParameterDependence.hpp"
+#include "SensParamUtil.hpp"
+#include "ConfigurationHelper.hpp"
+
+#include "LoggingUtils.hpp"
+#include "Logging.hpp"
+
+#include <Eigen/Dense>
+#include <Eigen/Sparse>
+
+using namespace Eigen;
+
+namespace cadet
+{
+
+namespace model
+{
+
+namespace parts
+{
+
+	/**
+	 * @brief Creates a ParticleDispersionOperatorBaseDG
+	 */
+	ParticleDispersionOperatorDG::ParticleDispersionOperatorDG() : _nParElem(nullptr), _nParPointsBeforeType(nullptr),
+		_parPolyDeg(nullptr), _nParNode(nullptr), _nParPoints(nullptr), _parExactInt(nullptr), _parGSM(nullptr), //_parTypeOffset(nullptr),
+		_nBound(nullptr), _boundOffset(nullptr), _strideBound(nullptr), _nBoundBeforeType(nullptr), _offsetSurfDiff(nullptr), _deltaR(nullptr), _parNodes(nullptr),
+		_parPolyDerM(nullptr), _minus_InvMM_ST(nullptr), _minus_parInvMM_Ar(nullptr), _parInvWeights(nullptr), _parInvMM(nullptr), _parInvMM_Leg(nullptr),
+		_Ir(nullptr), _Dr(nullptr), _secondOrderStiffnessM(nullptr), _DGjacParDispBlocks(nullptr), _g_p(nullptr), _g_pSum(nullptr),
+		_surfaceFluxParticle(nullptr), _localFlux(nullptr)
+	{
+	}
+
+	ParticleDispersionOperatorDG::~ParticleDispersionOperatorDG() CADET_NOEXCEPT
+	{
+		delete[] _nParElem;
+		delete[] _nParPointsBeforeType;
+		delete[] _parPolyDeg;
+		delete[] _nParNode;
+		delete[] _nParPoints;
+		delete[] _parExactInt;
+		delete[] _parGSM;
+		//delete[] _parTypeOffset;
+		delete[] _nBound;
+		delete[] _boundOffset;
+		delete[] _strideBound;
+		delete[] _nBoundBeforeType;
+
+		delete[] _offsetSurfDiff;
+
+		delete[] _deltaR;
+		delete[] _parNodes;
+		delete[] _parPolyDerM;
+		delete[] _minus_InvMM_ST;
+		delete[] _minus_parInvMM_Ar;
+		delete[] _parInvWeights;
+		delete[] _parInvMM;
+		delete[] _parInvMM_Leg;
+		delete[] _Ir;
+		delete[] _Dr;
+		delete[] _secondOrderStiffnessM;
+
+		delete[] _DGjacParDispBlocks;
+
+		delete[] _g_p;
+		delete[] _g_pSum;
+		delete[] _surfaceFluxParticle;
[…]
