+++ src/libcadet/model/parts/ParticleDispersionOperatorDG.cpp
@@ -102,11 +102,14 @@ namespace parts
 		_parDepSurfDiffusion.clear();
 	}
 
-	bool ParticleDispersionOperatorDG::configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, int nComp, int nParType)
+	bool ParticleDispersionOperatorDG::configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, const int nComp, const int nParType, const int strideBulkComp)
 	{
+		_strideBulkComp = strideBulkComp;
 
 		_nParType = nParType;
 		_nComp = nComp;
+
+		_filmDiffusion.resize(_nComp * _nParType); // filled in notifyDiscontinuousSectionTransition
 
 		std::vector<int> nBound;
 		const bool newNBoundInterface = paramProvider.exists("NBOUND");
@@ -1168,13 +1171,17 @@ namespace parts
 	 * @param [in] secIdx Index of the new section that is about to be integrated
 	 * @return @c true if flow direction has changed, otherwise @c false
 	 */
-	bool ParticleDispersionOperatorDG::notifyDiscontinuousSectionTransition(double t, unsigned int secIdx, MatrixXd& jacInlet)
+	bool ParticleDispersionOperatorDG::notifyDiscontinuousSectionTransition(double t, unsigned int secIdx, active const* const filmDiff)
 	{
+		for (int fd = 0; fd < _nComp * _nParType; fd++)
+			_filmDiffusion[fd] = filmDiff[fd];
 
 		//_curSection = secIdx;
 		//_newStaticJac = true;
 
 		// todo update operators and Jacobian blocks
+
+		return true;
 	}
 
 	/**
@@ -1188,7 +1195,7 @@ namespace parts
 	 * @param [in] jac Matrix that holds the Jacobian
 	 * @return @c 0 on success, @c -1 on non-recoverable error, and @c +1 on recoverable error
 	 */
-	int ParticleDispersionOperatorDG::residual(const IModel& model, double t, unsigned int secIdx, double const* y, double const* yDot, double* res, Eigen::SparseMatrix<double, Eigen::RowMajor>& jac)
+	int ParticleDispersionOperatorDG::residual(const IModel& model, double t, unsigned int secIdx, double const* y, double const* yBulk, double const* yDot, double* res, Eigen::SparseMatrix<double, Eigen::RowMajor>& jac)
 	{
 		//// Reset Jacobian but keep pattern
 		//double* val = jac.valuePtr();
@@ -1197,57 +1204,55 @@ namespace parts
 
 		linalg::BandedEigenSparseRowIterator jacIt(jac, 0);
 
-		return residualImpl<double, double, double, linalg::BandedEigenSparseRowIterator, true>(model, t, secIdx, y, yDot, res, jacIt);
+		return residualImpl<double, double, double, linalg::BandedEigenSparseRowIterator, true>(model, t, secIdx, y, yBulk, yDot, res, jacIt);
 	}
 
-	int ParticleDispersionOperatorDG::residual(const IModel& model, double t, unsigned int secIdx, double const* y, double const* yDot, double* res, WithoutParamSensitivity)
+	int ParticleDispersionOperatorDG::residual(const IModel& model, double t, unsigned int secIdx, double const* y, double const* yBulk, double const* yDot, double* res, WithoutParamSensitivity)
 	{
[…]
