+++ src/libcadet/model/parts/ParticleDispersionOperatorDG.hpp
@@ -18,6 +18,7 @@
 #ifndef LIBCADET_PARTICLEDISPERSIONOPERATORDG_HPP_
 #define LIBCADET_PARTICLEDISPERSIONOPERATORDG_HPP_
 
+#include "cadet/StrongTypes.hpp"
 #include "ParamIdUtil.hpp"
 #include "AutoDiff.hpp"
 #include "Memory.hpp"
@@ -85,7 +86,7 @@ namespace parts
 		ParticleDispersionOperatorDG();
 		~ParticleDispersionOperatorDG() CADET_NOEXCEPT;
 
-		bool configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, int nComp, int nParType);
+		bool configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, const int nComp, const int nParType, const int strideBulkComp);
 		bool configure(UnitOpIdx unitOpIdx, IParameterProvider& paramProvider, std::unordered_map<ParameterId, active*>& parameters);
 		
 		void setEquidistantRadialDisc(unsigned int parType);
@@ -95,14 +96,14 @@ namespace parts
 
 		void clearParDepSurfDiffusion();
 
-		bool notifyDiscontinuousSectionTransition(double t, unsigned int secIdx, Eigen::MatrixXd& jacInlet);
+		bool notifyDiscontinuousSectionTransition(double t, unsigned int secIdx, active const* const filmDiff);
 
-		int residual(const IModel& model, double t, unsigned int secIdx, double const* y, double const* yDot, double* res, Eigen::SparseMatrix<double, Eigen::RowMajor>& jac);
-		int residual(const IModel& model, double t, unsigned int secIdx, double const* y, double const* yDot, active* res, Eigen::SparseMatrix<double, Eigen::RowMajor>& jac);
-		int residual(const IModel& model, double t, unsigned int secIdx, double const* y, double const* yDot, double* res, WithoutParamSensitivity);
-		int residual(const IModel& model, double t, unsigned int secIdx, double const* y, double const* yDot, active* res, WithParamSensitivity);
-		int residual(const IModel& model, double t, unsigned int secIdx, active const* y, double const* yDot, active* res, WithParamSensitivity);
-		int residual(const IModel& model, double t, unsigned int secIdx, active const* y, double const* yDot, active* res, WithoutParamSensitivity);
+		int residual(const IModel& model, double t, unsigned int secIdx, double const* y, double const* yBulk, double const* yDot, double* res, Eigen::SparseMatrix<double, Eigen::RowMajor>& jac);
+		int residual(const IModel& model, double t, unsigned int secIdx, double const* y, double const* yBulk, double const* yDot, active* res, Eigen::SparseMatrix<double, Eigen::RowMajor>& jac);
+		int residual(const IModel& model, double t, unsigned int secIdx, double const* y, double const* yBulk, double const* yDot, double* res, WithoutParamSensitivity);
+		int residual(const IModel& model, double t, unsigned int secIdx, double const* y, double const* yBulk, double const* yDot, active* res, WithParamSensitivity);
+		int residual(const IModel& model, double t, unsigned int secIdx, active const* y, active const* yBulk, double const* yDot, active* res, WithParamSensitivity);
+		int residual(const IModel& model, double t, unsigned int secIdx, active const* y, active const* yBulk, double const* yDot, active* res, WithoutParamSensitivity);
 
 
 		/* Physical model parameters */
@@ -117,6 +118,8 @@ namespace parts
 		std::vector<active> _parOuterSurfAreaPerVolume; //!< Particle shell outer sphere surface to volume ratio
[…]
