+++ src/libcadet/model/parts/ParticleDiffusionOperatorDG.cpp
@@ -10,7 +10,7 @@
 //  is available at http://www.gnu.org/licenses/gpl.html
 // =============================================================================
 
-#include "model/parts/ParticleDispersionOperatorDG.hpp"
+#include "model/parts/ParticleDiffusionOperatorDG.hpp"
 #include "cadet/Exceptions.hpp"
 
 #include "ParamReaderHelper.hpp"
@@ -38,9 +38,9 @@ namespace parts
 {
 
 	/**
-	 * @brief Creates a ParticleDispersionOperatorBaseDG
+	 * @brief Creates a ParticleDiffusionOperatorDG
 	 */
-	ParticleDispersionOperatorDG::ParticleDispersionOperatorDG() : _nParElem(nullptr), _nParPointsBeforeType(nullptr),
+	ParticleDiffusionOperatorDG::ParticleDiffusionOperatorDG() : _nParElem(nullptr), _nParPointsBeforeType(nullptr),
 		_parPolyDeg(nullptr), _nParNode(nullptr), _nParPoints(nullptr), _parExactInt(nullptr), _parGSM(nullptr), //_parTypeOffset(nullptr),
 		_nBound(nullptr), _boundOffset(nullptr), _strideBound(nullptr), _nBoundBeforeType(nullptr), _offsetSurfDiff(nullptr), _deltaR(nullptr), _parNodes(nullptr),
 		_parPolyDerM(nullptr), _minus_InvMM_ST(nullptr), _minus_parInvMM_Ar(nullptr), _parInvWeights(nullptr), _parInvMM(nullptr), _parInvMM_Leg(nullptr),
@@ -49,7 +49,7 @@ namespace parts
 	{
 	}
 
-	ParticleDispersionOperatorDG::~ParticleDispersionOperatorDG() CADET_NOEXCEPT
+	ParticleDiffusionOperatorDG::~ParticleDiffusionOperatorDG() CADET_NOEXCEPT
 	{
 		delete[] _nParElem;
 		delete[] _nParPointsBeforeType;
@@ -86,7 +86,7 @@ namespace parts
 		delete[] _localFlux;
 	}
 
-	void ParticleDispersionOperatorDG::clearParDepSurfDiffusion()
+	void ParticleDiffusionOperatorDG::clearParDepSurfDiffusion()
 	{
 		if (_singleParDepSurfDiffusion)
 		{
@@ -102,7 +102,7 @@ namespace parts
 		_parDepSurfDiffusion.clear();
 	}
 
-	bool ParticleDispersionOperatorDG::configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, const int nComp, const int nParType, const int strideBulkComp)
+	bool ParticleDiffusionOperatorDG::configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, const int nComp, const int nParType, const int strideBulkComp)
 	{
 		_strideBulkComp = strideBulkComp;
 
@@ -110,6 +110,7 @@ namespace parts
 		_nComp = nComp;
 
 		_filmDiffusion.resize(_nComp * _nParType); // filled in notifyDiscontinuousSectionTransition
+		_poreAccessFactor.resize(_nComp * _nParType); // filled in notifyDiscontinuousSectionTransition
 
 		std::vector<int> nBound;
 		const bool newNBoundInterface = paramProvider.exists("NBOUND");
@@ -160,7 +161,7 @@ namespace parts
 
 			if (parPolyDegs.size() == 1)
 			{
-				// Multiplex number of particle shells to all particle types
+				// Multiplex number of particle elements to all particle types
[…]

[…]
