### 2025-05-08 97f4c1a
Update dev guide: test flags
