+++ src/libcadet/model/GeneralRateModelDG.hpp
@@ -817,239 +817,6 @@ protected:
 
 	parts::ParticleDiffusionOperatorDG _parDiffOp; //!< Particle dispersion operator
 
-// ===========================================================================================================================================================  //
-// ========================================			DG functions to compute particle discretization			==================================================  //
-// ===========================================================================================================================================================  //
-
-	/**
-	 * @brief promotes doubles to actives
-	 * @detail promotes consecutive doubles to consecutive actives (with zero gradients) based on input double pointer
-	 */
-	void vectorPromoter(double* state, const unsigned int nVals) {
-
-		const int nDirs = ad::getDirections();
-		const int stride = (1 + nDirs);
-		const int ADsize = stride * nVals;
-		double buff = 0.0;
-
-		for (int val = 1; val <= nVals; val++) // start with last entry to avoid overwriting
-		{
-			buff = state[nVals - val];
-			std::fill(state + ADsize - val * stride, state + ADsize - (val - 1) * stride, 0.0);
-			state[ADsize - val * stride] = buff;
-		}
-	}
-
-	template<typename ResidualType, typename ParamType>
-	void applyParInvMap(Eigen::Map<Vector<ResidualType, Dynamic>, 0, InnerStride<>>& state, unsigned int parType) {
-		for (int cell = 0; cell < _disc.nParCell[parType]; cell++) {
-			state.segment(cell * _disc.nParNode[parType], _disc.nParNode[parType]) *= 2.0 / static_cast<ParamType>(_disc.deltaR[_disc.offsetMetric[parType] + cell]) * 2.0 / static_cast<ParamType>(_disc.deltaR[_disc.offsetMetric[parType] + cell]);
-		}
-	}
-
-	template<typename StateType, typename ResidualType>
-	void parGSMVolumeIntegral(const int parType, Eigen::Map<const Vector<StateType, Dynamic>, 0, InnerStride<Dynamic>>& state, Eigen::Map<Vector<ResidualType, Dynamic>, 0, InnerStride<Dynamic>>& stateDer) {
-
-		int nNodes = _disc.nParNode[parType];
-
-		stateDer.segment(0, nNodes)
-			-= (_disc.minus_parInvMM_Ar[parType].template cast<StateType>() * state.segment(0, nNodes)).template cast<ResidualType>();
-	}
-
-	template<typename StateType, typename ResidualType>
-	void parVolumeIntegral(const int parType, const bool aux, Eigen::Map<const Vector<StateType, Dynamic>, 0, InnerStride<Dynamic>>& state, Eigen::Map<Vector<ResidualType, Dynamic>, 0, InnerStride<Dynamic>>& stateDer) {
-
-		int nNodes = _disc.nParNode[parType];
-
-		/* no additional metric term for auxiliary equation or particle equation with exact integration scheme
-		   -> res = - D * (d_p * c^p + invBeta_p sum_mi d_s c^s) */
-		if (aux || (_disc.parExactInt[parType] && _parGeomSurfToVol[parType] == _disc.SurfVolRatioSlab)) {
-			// comp-cell-node state vector: use of Eigen lib performance
-			for (unsigned int Cell = 0; Cell < _disc.nParCell[parType]; Cell++) {
-				stateDer.segment(Cell * nNodes, nNodes)
-					-= (_parDiffOp._parPolyDerM[parType].template cast<StateType>() * state.segment(Cell * nNodes, nNodes)).template cast<ResidualType>();
-			}
-		}
[…]
