+++ src/libcadet/model/parts/ParticleDiffusionOperatorDG.cpp
@@ -1202,34 +1202,34 @@ namespace parts
 	 * @param [in] jac Matrix that holds the Jacobian
 	 * @return @c 0 on success, @c -1 on non-recoverable error, and @c +1 on recoverable error
 	 */
-	int ParticleDiffusionOperatorDG::residual(double t, unsigned int parType, unsigned int colNode, unsigned int secIdx, double const* yPar, double const* yBulk, double const* yDotPar, double* resPar, int const* const qsBinding, WithoutParamSensitivity)
+	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int parType, unsigned int colNode, double const* yPar, double const* yBulk, double const* yDotPar, double* resPar, int const* const qsBinding, WithoutParamSensitivity)
 	{
-		return residualImpl<double, double, double>(t, parType, colNode, secIdx, yPar, yBulk, yDotPar, resPar, qsBinding);
+		return residualImpl<double, double, double>(t, secIdx, parType, colNode, yPar, yBulk, yDotPar, resPar, qsBinding);
 	}
 
-	int ParticleDiffusionOperatorDG::residual(double t, unsigned int parType, unsigned int colNode, unsigned int secIdx, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithoutParamSensitivity)
+	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int parType, unsigned int colNode, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithoutParamSensitivity)
 	{
-		return residualImpl<active, active, double>(t, parType, colNode, secIdx, yPar, yBulk, yDotPar, resPar, qsBinding);
+		return residualImpl<active, active, double>(t, secIdx, parType, colNode, yPar, yBulk, yDotPar, resPar, qsBinding);
 	}
 
-	int ParticleDiffusionOperatorDG::residual(double t, unsigned int parType, unsigned int colNode, unsigned int secIdx, double const* yPar, double const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithParamSensitivity)
+	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int parType, unsigned int colNode, double const* yPar, double const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithParamSensitivity)
 	{
-		return residualImpl<double, active, active>(t, parType, colNode, secIdx, yPar, yBulk, yDotPar, resPar, qsBinding);
+		return residualImpl<double, active, active>(t, secIdx, parType, colNode, yPar, yBulk, yDotPar, resPar, qsBinding);
 	}
 
-	int ParticleDiffusionOperatorDG::residual(double t, unsigned int parType, unsigned int colNode, unsigned int secIdx, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithParamSensitivity)
+	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int parType, unsigned int colNode, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithParamSensitivity)
 	{
-		return residualImpl<active, active, active>(t, parType, colNode, secIdx, yPar, yBulk, yDotPar, resPar, qsBinding);
+		return residualImpl<active, active, active>(t, secIdx, parType, colNode, yPar, yBulk, yDotPar, resPar, qsBinding);
 	}
 
[…]

[…]
