+++ src/libcadet/model/GeneralRateModelDG.hpp
@@ -239,11 +239,6 @@ protected:
 
 	void assembleDiscretizedGlobalJacobian(double alpha, Indexer idxr);
 
-	void setEquidistantRadialDisc(unsigned int parType);
-	void setEquivolumeRadialDisc(unsigned int parType);
-	void setUserdefinedRadialDisc(unsigned int parType);
-	void updateRadialDisc();
-
 	void addTimeDerivativeToJacobianParticleShell(linalg::BandedEigenSparseRowIterator& jac, const Indexer& idxr, double alpha, unsigned int parType);
 
 	unsigned int numAdDirsForJacobian() const CADET_NOEXCEPT;
@@ -291,34 +286,9 @@ protected:
 		const double SurfVolRatioCylinder = 2.0; //!< Surface to volume ratio for a cylindrical particle
 		const double SurfVolRatioSphere = 3.0; //!< Surface to volume ratio for a spherical particle
 
-		/*					DG specifics				*/
-		
-		active* deltaR; //!< equidistant particle element spacing for each particle type
-		Eigen::VectorXd* parNodes; //!< Array with positions of nodes in radial reference element for each particle
-		Eigen::MatrixXd* parPolyDerM; //!< Array with polynomial derivative Matrix for each particle
-		Eigen::MatrixXd* minus_InvMM_ST; //!< equals minus inverse mass matrix times transposed stiffness matrix. Required solely for exact integration DG discretization of particle equation
-		Eigen::VectorXd* parInvWeights; //!< Array with weights for LGL quadrature of size nNodes for each particle
-		Eigen::MatrixXd* parInvMM; //!< dense inverse mass matrix for exact integration of integrals with metrics, for each particle
-		Eigen::MatrixXd* parInvMM_Leg; //!< dense inverse mass matrix (Legendre) for exact integration of integral without metric, for each particle
-		Eigen::MatrixXd* secondOrderStiffnessM; //!< specific second order stiffness matrix
-		Eigen::MatrixXd* minus_parInvMM_Ar; //!< inverse mass matrix times specific second order stiffness matrix
-		Eigen::Vector<active, Dynamic>* Ir; //!< metric part for each particle type and cell, particle type major ordering
-		Eigen::MatrixXd* Dr; //!< derivative matrices including metrics for each particle type and cell, particle type major ordering
-		Eigen::VectorXi offsetMetric; //!< offset required to access metric dependent DG operator storage of Ir, Dr -> summed up nCells of all previous parTypes
-
-		Eigen::MatrixXd* DGjacParDispBlocks; //!< particle dispersion blocks of DG jacobian
-
-		Eigen::Vector<active, Dynamic>* g_p; //!< auxiliary variable g = dc_p / dr
-		Eigen::Vector<active, Dynamic>* g_pSum; //!< auxiliary variable g = sum_{k \in p, s_i} dc_k / dr
-		Eigen::Vector<active, Dynamic>* surfaceFluxParticle; //!< stores the surface flux values for each particle
-		active* localFlux; //!< stores the local (at respective particle) film diffusion flux
-
 		Discretization() : nParCell(nullptr), nParPointsBeforeType(nullptr), parPolyDeg(nullptr), nParNode(nullptr),
 			nParPoints(nullptr), parExactInt(nullptr), parTypeOffset(nullptr), nBound(nullptr), boundOffset(nullptr),
[…]
