+++ src/libcadet/model/GeneralRateModelDG-LinearSolver.cpp
@@ -205,7 +205,7 @@ void GeneralRateModelDG::assembleDiscretizedGlobalJacobian(double alpha, Indexer
 
 						// surface diffusion + kinetic binding -> additional DG-discretized mass balance equation for solid, for which the (inexact integration) discretization special case also holds
 						else if (_hasSurfaceDiffusion[parType]
-							&& static_cast<double>(getSectionDependentSlice(_parSurfDiffusion, _disc.strideBound[_disc.nParType], _disc.curSection)[_disc.nBoundBeforeType[parType] + getOffsetSurfDiff(parType, comp, bnd)]) != 0.0)
+							&& static_cast<double>(getSectionDependentSlice(_parSurfDiffusion, _disc.strideBound[_disc.nParType], _disc.curSection)[_disc.nBoundBeforeType[parType] + _parDiffOp.getOffsetSurfDiff(parType, comp, bnd)]) != 0.0)
 							continue;
 
 						// Add derivative with respect to dq / dt to Jacobian
