+++ src/libcadet/model/parts/ParticleDiffusionOperatorDG.hpp
@@ -242,7 +242,7 @@ namespace parts
 		Eigen::MatrixXd* _minus_parInvMM_Ar; //!< inverse mass matrix times specific second order stiffness matrix
 		Eigen::Vector<active, Dynamic>* _Ir; //!< metric part for each particle type and element, particle type major ordering
 		Eigen::MatrixXd* _Dr; //!< derivative matrices including metrics for each particle type and element, particle type major ordering
-		Eigen::VectorXi _offsetMetric; //!< offset required to access metric dependent DG operator storage of Ir, Dr -> summed up nCells of all previous parTypes
+		Eigen::VectorXi _offsetMetric; //!< offset required to access metric dependent DG operator storage of Ir, Dr -> summed up nelems of all previous parTypes
 
 		Eigen::MatrixXd* _DGjacParDispBlocks; //!< particle dispersion blocks of DG jacobian
 
@@ -255,12 +255,90 @@ namespace parts
 		void initializeDG();
 
 		void initializeDGjac(std::vector<double> parGeomSurfToVol);
-		Eigen::MatrixXd DGjacobianParDispBlock(unsigned int elemIdx, unsigned int parType, double parGeomSurfToVol);
-		Eigen::MatrixXd GSMjacobianParDispBlock(unsigned int parType, double parGeomSurfToVol);
-		Eigen::MatrixXd getParBMatrix(int parType, int element, double parGeomSurfToVol);
-		Eigen::MatrixXd parAuxBlockGstar(unsigned int elemIdx, unsigned int parType, MatrixXd leftG, MatrixXd middleG, MatrixXd rightG);
-		Eigen::MatrixXd getParGBlock(unsigned int elemIdx, unsigned int parType);
 
+		/**
+		 * @brief analytically calculates the static (per section) particle jacobian
+		 * @return 1 if jacobain calculation fits the predefined pattern of the jacobian, 0 if not.
+		 */
+		int calcStaticAnaParticleDiffJacobian(unsigned int parType, unsigned int colNode, const active* const parDiff, const active* const parSurfDiff, const double* const invBetaP, int const* const qsBinding,
+			unsigned int offsetCp, SparseMatrix<double, RowMajor>& globalJac) {
+
+			// DG particle dispersion Jacobian
+			if (_parExactInt[parType])
+				calcParticleDGSEMJacobian(parType, colNode, parDiff, parSurfDiff, invBetaP, qsBinding, offsetCp, globalJac);
+			else // deprecated
+				calcParticleCollocationDGSEMJacobian(parType, colNode, parDiff, parSurfDiff, invBetaP, qsBinding, offsetCp, globalJac);
+
+			return true;
+		}
+
+		/**
+		 * @brief calculates, writes the physical radial/particle coordinates of the DG discretization with double! interface nodes
+		 */
+		virtual int writeParticleCoordinates(unsigned int parType, double* coords) const
+		{
+			active const* const pcr = _parCenterRadius.data() + _offsetMetric[parType];
+
+			// Note that the DG particle shells are oppositely ordered compared to the FV particle shells
+			for (unsigned int par = 0; par < _nParPoints[parType]; par++) {
+
+				unsigned int elem = std::floor(par / _nParNode[parType]);
+
+				double r_L = static_cast<double>(pcr[elem]) - 0.5 * static_cast<double>(_deltaR[_offsetMetric[parType] + elem]);
[…]
