### 2025-05-12 #7881 @goxberry
As a status update: my review is "accept with minor revisions"; the suggested revisions are at the end of this comment.

> We would like to proceed with option 4) by inviting the reviewers to [our JupyterHub](https://jupyter.cadet-web.de/), where they will have access to our workshop materials.

Thank you for providing this option! I was able to log into CADET's JupyterHub instance and confirm the functionality of CADET by executing all of the Python code in the [CADET 5.0.3 Introduction](https://cadet.github.io/v5.0.3/getting_started/tutorials/breakthrough.html).

> While they were able to successfully install CADET-Python, we understand that the installation of CADET-Process failed for goxberry when using the Pixi dependency resolver. We support and recommend using the standard Python package installer pip and had no experience with Pixi up to this point. However, we found an issue with an upstream dependency which we reported and that has been fixed.

I appreciate Hannah's bug report to the `pixi` project. The motivation for using `pixi` is that it keeps [consistent state](https://prefix.dev/blog/pixi_a_fast_conda_alternative#reason-2:-better-integration-with-pypl) for both `conda` and `pip` package managers, avoiding [well-known problems](https://www.anaconda.com/blog/using-pip-in-a-conda-environment) that can occur when using both package managers.

> CADET-Process builds upon CADET-Python and provides tools for easier process modeling, network configuration, evaluation, and optimization. It can be installed via `pip install CADET-Process`.

I was able to `pip install CADET-Process` locally and execute all of the Python code in the [CADET 5.0.3 Introduction](https://cadet.github.io/v5.0.3/getting_started/tutorials/breakthrough.html).

> That said, we are truly grateful for the feedback and discussion, which have been very valuable to us. This exchange highlighted areas where our documentation lacked clarity, and as a result, we have already implemented improvements to better explain how CADET-Core can be interfaced.

I appreciate the clarifications made in https://github.com/cadet/CADET-Core/pull/376.

## Suggested revisions

> We recognize that getting started can be challenging, especially for users coming from different fields. [...] However, we also don't want the reviewers to be burdened with a time-intensive setup.

I appreciate not wanting to burden users with time-intensive setup. For clarity, I suggest the [top-level CADET documentation](https://cadet.github.io/master/getting_started/installation_core.html) mention the authors' observation that

> [CADET-Core] must be built from source on MacOS-ARM64 systems following the [build instructions for OSX](https://cadet.github.io/master/getting_started/build_osx.html#build-osx)

so that users on macOS ARM systems do not attempt to use the `conda`-based installation instructions.
