### 2025-05-12 913d625
Deprecate collocation variant of DGSEM particle discretization (#411)
