+++ src/libcadet/model/GeneralRateModelDG.hpp
@@ -272,7 +272,6 @@ protected:
 		unsigned int* parPolyDeg; //!< polynomial degree of particle elements
 		unsigned int* nParNode; //!< Array with number of radial nodes per cell in each particle type
 		unsigned int* nParPoints; //!< Array with number of radial nodes per cell in each particle type
-		bool* parExactInt; //!< 1 for exact integration, 0 for inexact LGL quadrature for each particle type
 		bool* parGSM; //!< specifies whether (single element) Galerkin spectral method should be used in particles
 		unsigned int* parTypeOffset; //!< Array with offsets (in particle block) to particle type, additional last element contains total number of particle DOFs
 		unsigned int* nBound; //!< Array with number of bound states for each component and particle type (particle type major ordering)
@@ -313,7 +312,7 @@ protected:
 		active* localFlux; //!< stores the local (at respective particle) film diffusion flux
 
 		Discretization() : nParCell(nullptr), nParPointsBeforeType(nullptr), parPolyDeg(nullptr), nParNode(nullptr),
-			nParPoints(nullptr), parExactInt(nullptr), parTypeOffset(nullptr), nBound(nullptr), boundOffset(nullptr),
+			nParPoints(nullptr), parTypeOffset(nullptr), nBound(nullptr), boundOffset(nullptr),
 			strideBound(nullptr), nBoundBeforeType(nullptr), offsetSurfDiff(nullptr), deltaR(nullptr), parNodes(nullptr),
 			parPolyDerM(nullptr), minus_InvMM_ST(nullptr), parInvWeights(nullptr), parInvMM(nullptr), parInvMM_Leg(nullptr),
 			Ir(nullptr), Dr(nullptr), DGjacParDispBlocks(nullptr), g_p(nullptr), g_pSum(nullptr),
@@ -328,7 +327,6 @@ protected:
 			delete[] parPolyDeg;
 			delete[] nParNode;
 			delete[] nParPoints;
-			delete[] parExactInt;
 			delete[] parTypeOffset;
 			delete[] nBound;
 			delete[] boundOffset;
@@ -467,31 +465,27 @@ protected:
 			// Auxiliary Block [ d g(c) / d c ], additionally depends on boundary entries of neighbouring cells
 			MatrixXd gBlock = MatrixXd::Zero(nParNode[parType], nParNode[parType] + 2);
 			gBlock.block(0, 1, nParNode[parType], nParNode[parType]) = parPolyDerM[parType];
-			if (parExactInt[parType]) {
-				if (cellIdx == 0 || cellIdx == nParCell[parType] + 1) { // cellIdx out of bounds
-					return MatrixXd::Zero(nParNode[parType], nParNode[parType] + 2);
-				}
-				if (cellIdx != 1 && cellIdx != nParCell[parType]) { // inner cell
-					gBlock.block(0, 0, nParNode[parType], 1) -= 0.5 * parInvMM_Leg[parType].block(0, 0, nParNode[parType], 1);
-					gBlock.block(0, 1, nParNode[parType], 1) += 0.5 * parInvMM_Leg[parType].block(0, 0, nParNode[parType], 1);
-					gBlock.block(0, nParNode[parType], nParNode[parType], 1) -= 0.5 * parInvMM_Leg[parType].block(0, nParNode[parType] - 1, nParNode[parType], 1);
-					gBlock.block(0, nParNode[parType] + 1, nParNode[parType], 1) += 0.5 * parInvMM_Leg[parType].block(0, nParNode[parType] - 1, nParNode[parType], 1);
-				}
-				else if (cellIdx == 1u) { // left boundary cell
[…]
