+++ src/libcadet/model/GeneralRateModelDG.cpp
@@ -186,12 +186,10 @@ bool GeneralRateModelDG::configureModelDiscretization(IParameterProvider& paramP
 
 	std::vector<int> parPolyDeg(_disc.nParType);
 	std::vector<int> ParNelem(_disc.nParType);
-	std::vector<bool> parExactInt(_disc.nParType, true);
 	if (firstConfigCall)
 	{
 		_disc.parPolyDeg = new unsigned int[_disc.nParType];
 		_disc.nParCell = new unsigned int[_disc.nParType];
-		_disc.parExactInt = new bool[_disc.nParType];
 		_disc.parGSM = new bool[_disc.nParType];
 	}
 
@@ -205,12 +203,6 @@ bool GeneralRateModelDG::configureModelDiscretization(IParameterProvider& paramP
 
 		if ((std::any_of(ParNelem.begin(), ParNelem.end(), [](int value) { return value < 1; })))
 			throw InvalidParameterException("Particle number of elements must be at least 1!");
-
-		if(paramProvider.exists("PAR_EXACT_INTEGRATION"))
-			parExactInt = paramProvider.getBoolArray("PAR_EXACT_INTEGRATION");
-
-		if ((std::any_of(parExactInt.begin(), parExactInt.end(), [](bool value) { return !value; })))
-			LOG(Warning) << "Inexact integration method (cf. PAR_EXACT_INTEGRATION) in particles might add severe! stiffness to the system and disables consistent initialization!";
 
 		if (parPolyDeg.size() == 1)
 		{
@@ -232,16 +224,6 @@ bool GeneralRateModelDG::configureModelDiscretization(IParameterProvider& paramP
 			throw InvalidParameterException("Field PAR_NELEM must have 1 or NPARTYPE (" + std::to_string(_disc.nParType) + ") entries");
 		else
 			std::copy_n(ParNelem.begin(), _disc.nParType, _disc.nParCell);
-		if (parExactInt.size() == 1)
-		{
-			// Multiplex number of particle shells to all particle types
-			for (unsigned int i = 0; i < _disc.nParType; ++i)
-				std::fill(_disc.parExactInt, _disc.parExactInt + _disc.nParType, parExactInt[0]);
-		}
-		else if (parExactInt.size() < _disc.nParType)
-			throw InvalidParameterException("Field PAR_EXACT_INTEGRATION must have 1 or NPARTYPE (" + std::to_string(_disc.nParType) + ") entries");
-		else
-			std::copy_n(parExactInt.begin(), _disc.nParType, _disc.parExactInt);
 	}
 	else if (paramProvider.exists("NPAR"))
 	{
@@ -1366,9 +1348,6 @@ int GeneralRateModelDG::residualParticle(double t, unsigned int parType, unsigne
 
 	LinearBufferAllocator tlmAlloc = threadLocalMem.get();
 
-	// Special case: individual treatment of time derivatives in particle mass balance at inner particle boundary node
-	bool specialCase = !_disc.parExactInt[parType] && (_parGeomSurfToVol[parType] != _disc.SurfVolRatioSlab && _parCoreRadius[parType] == 0.0);
-
 	// Prepare parameters
 	active const* const parDiff = getSectionDependentSlice(_parDiffusion, _disc.nComp * _disc.nParType, secIdx) + parType * _disc.nComp;
 
@@ -1407,62 +1386,14 @@ int GeneralRateModelDG::residualParticle(double t, unsigned int parType, unsigne
 			);
 		const ColumnPosition colPos{ z, 0.0, r };
 
-		// Handle time derivatives, binding, dynamic reactions.
-		// Special case: Dont add time derivatives to inner boundary node for DG discretized mass balance equations.
[…]
