+++ src/libcadet/model/reaction/ActivatedSludgeModelDrei.cpp
@@ -1,0 +1,298 @@
+// =============================================================================
+//  CADET - The Chromatography Analysis and Design Toolkit
+//
+//  Copyright © 2008-present: The CADET-Core Authors
+//            Please see the AUTHORS.md file.
+//
+//  All rights reserved. This program and the accompanying materials
+//  are made available under the terms of the GNU Public License v3.0 (or, at
+//  your option, any later version) which accompanies this distribution, and
+//  is available at http://www.gnu.org/licenses/gpl.html
+// =============================================================================
+
+#include "model/reaction/ReactionModelBase.hpp"
+#include "model/ExternalFunctionSupport.hpp"
+#include "model/ModelUtils.hpp"
+#include "cadet/Exceptions.hpp"
+#include "model/Parameters.hpp"
+#include "LocalVector.hpp"
+#include "SimulationTypes.hpp"
+#include "linalg/ActiveDenseMatrix.hpp"
+#include "MathUtil.hpp"
+#include "Memory.hpp"
+
+#include <algorithm>
+#include <unordered_map>
+#include <string>
+#include <vector>
+
+/*<codegen>
+{
+	"name": "MichaelisMentenParamHandler",
+	"externalName": "ExtMichaelisMentenParamHandler",
+	"parameters":
+		[
+			{ "type": "ScalarReactionDependentParameter", "varName": "vMax", "confName": "MM_VMAX"},
+			{ "type": "ScalarReactionDependentParameter", "varName": "kMM", "confName": "MM_KMM"},
+			{ "type": "ComponentDependentReactionDependentParameter", "varName": "kInhibit", "confName": "MM_KI"}
+		]
+}
+</codegen>*/
+
+/* Parameter description
+ ------------------------
+*/
+
+
+namespace cadet
+{
+
+namespace model
+{
+
+inline const char* MichaelisMentenParamHandler::identifier() CADET_NOEXCEPT { return "MICHAELIS_MENTEN"; }
+
+inline bool MichaelisMentenParamHandler::validateConfig(unsigned int nReactions, unsigned int nComp, unsigned int const* nBoundStates)
+{
+	return true;
+}
+
+inline const char* ExtMichaelisMentenParamHandler::identifier() CADET_NOEXCEPT { return "EXT_MICHAELIS_MENTEN"; }
+
+inline bool ExtMichaelisMentenParamHandler::validateConfig(unsigned int nReactions, unsigned int nComp, unsigned int const* nBoundStates)
+{
+	return true;
+}
+
+namespace
+{
+	/**
+	 * @brief Registers a matrix-valued parameter (row-major storage) with components as rows
+	 * @details The matrix-valued parameter has as many rows as there are components in the system.
+	 * @param [in,out] parameters Parameter map
+	 * @param [in] unitOpIdx Unit operation id
+	 * @param [in] parTypeIdx Particle type index
+	 * @param [in] paramName Name of the parameter
+	 * @param [in] mat Matrix to register
+	 */
+	inline void registerCompRowMatrix(std::unordered_map<ParameterId, active*>& parameters, UnitOpIdx unitOpIdx, ParticleTypeIdx parTypeIdx, const std::string& paramName, cadet::linalg::ActiveDenseMatrix& mat)
+	{
+		const cadet::StringHash hashName = cadet::hashStringRuntime(paramName);
[…]
