### 2025-05-13 cd09ef8
ASM3: compute p2-5
