### 2025-05-13 cd8d388
ASM3: compute r6-12
