### 2025-05-13 3c56f72
ASM3: set up parameters and derived variables
