### 2025-05-13 aeed290
Improve documentation of sensitivity interface
