### 2025-05-13 ec6f394
Add sensitivity tests to CI (#410)
