### 2025-05-13 aa0abd4
Add sensitivity tests to CI (#410)
