### 2025-05-14 9971358
register model and jacobian equation as pseudo code
