+++ src/libcadet/model/reaction/ActivatedSludgeModelThree.cpp
@@ -28,8 +28,8 @@
 
 /*<codegen>
 {
-	"name": "MichaelisMentenParamHandler",
-	"externalName": "ExtMichaelisMentenParamHandler",
+	"name": "ActivatedSludgeModelThreeParamHandler",
+	"externalName": "ExtActivatedSludgeModelThreeParamHandler",
 	"parameters":
 		[
 			{ "type": "ScalarReactionDependentParameter", "varName": "vMax", "confName": "MM_VMAX"},
@@ -50,16 +50,16 @@ namespace cadet
 namespace model
 {
 
-inline const char* MichaelisMentenParamHandler::identifier() CADET_NOEXCEPT { return "MICHAELIS_MENTEN"; }
+inline const char* ActivatedSludgeModelThreeParamHandler::identifier() CADET_NOEXCEPT { return "ACTIVATED_SLUDGE_MODEL3"; }
 
-inline bool MichaelisMentenParamHandler::validateConfig(unsigned int nReactions, unsigned int nComp, unsigned int const* nBoundStates)
+inline bool ActivatedSludgeModelThreeParamHandler::validateConfig(unsigned int nReactions, unsigned int nComp, unsigned int const* nBoundStates)
 {
 	return true;
 }
 
-inline const char* ExtMichaelisMentenParamHandler::identifier() CADET_NOEXCEPT { return "EXT_MICHAELIS_MENTEN"; }
+inline const char* ExtActivatedSludgeModelThreeParamHandler::identifier() CADET_NOEXCEPT { return "EXT_ACTIVATED_SLUDGE_MODEL3"; }
 
-inline bool ExtMichaelisMentenParamHandler::validateConfig(unsigned int nReactions, unsigned int nComp, unsigned int const* nBoundStates)
+inline bool ExtActivatedSludgeModelThreeParamHandler::validateConfig(unsigned int nReactions, unsigned int nComp, unsigned int const* nBoundStates)
 {
 	return true;
 }
@@ -111,12 +111,12 @@ namespace
  * @tparam ParamHandler_t Type that can add support for external function dependence
  */
 template <class ParamHandler_t>
-class MichaelisMentenReactionBase : public DynamicReactionModelBase
+class ActivatedSludgeModelThreeBase : public DynamicReactionModelBase
 {
 public:
 
-	MichaelisMentenReactionBase() : _idxSubstrate(0) { }
-	virtual ~MichaelisMentenReactionBase() CADET_NOEXCEPT { }
+	ActivatedSludgeModelThreeBase() : _idxSubstrate(0) { }
+	virtual ~ActivatedSludgeModelThreeBase() CADET_NOEXCEPT { }
 
 	static const char* identifier() { return ParamHandler_t::identifier(); }
 	virtual const char* name() const CADET_NOEXCEPT { return ParamHandler_t::identifier(); }
@@ -211,7 +211,6 @@ protected:
 		ParamType etaNend = 0.5;
 
 		// derived parameters
[…]
