### 2025-05-14 12a5b67
ASM3: add Aeration
