### 2025-05-14 5e65317
ASM3: basic stoichiometry definition
