### 2025-05-14 26d0ff1
ASM3: parameter set required for stoichiometry
