### 2025-05-14 d07851e
ASM3: differentiate p1-5
