+++ src/libcadet/model/reaction/ActivatedSludgeModelThree.cpp
@@ -312,8 +312,6 @@ protected:
 		//_paramHandler.configure(paramProvider, _stoichiometry.columns(), _nComp, _nBoundStates);
 		_paramHandler.registerParameters(_parameters, unitOpIdx, parTypeIdx, _nComp, _nBoundStates);
 
-		// configure stoichiometric matrix from given parameters
-		//configureStoich(paramProvider,unitOpIdx, parTypeIdx)
 
 		//registerCompRowMatrix(_parameters, unitOpIdx, parTypeIdx, "MM_STOICHIOMETRY_BULK", _stoichiometryBulk); todo
 
@@ -416,7 +414,8 @@ protected:
 		fluxes[11] = bAUT * etaNend * SNO / (SNO + KHNO3) * KHO2 / (SO + KHO2) * XA;
 
 		// r13: Aeration
-		fluxes[12] = iO2 / V * 1000;
+		// TODO: is V in litres?
+		fluxes[12] = iO2 / V;
 			
 		// Add reaction terms to residual
 		_stoichiometry.multiplyVector(static_cast<flux_t*>(fluxes), factor, res);
@@ -442,12 +441,63 @@ protected:
 	void jacobianLiquidImpl(double t, unsigned int secIdx, const ColumnPosition& colPos, double const* y, double factor, const RowIterator& jac, LinearBufferAllocator workSpace) const
 	{
 		typename ParamHandler_t::ParamsHandle const p = _paramHandler.update(t, secIdx, colPos, _nComp, _nBoundStates, workSpace);
-		curJac = jac;
-			
+		RowIterator curJac = jac;
+
+		double Kh20 = 10.2;
+		double T = 20;
+		double iO2 = 0;
+		double V = 1;
+		double k_sto20 = 13.68;
+		double KX = 1;
+		double KHO2 = 0.2;
+		double KHSS = 3;
+		double KHNO3 = 0.5;
+		double etaHNO3 = 0.5;
+		double KHNH4 = 0.01;
+		double KHALK = 0.1;
+		double KHSTO = 0.11;
+		double muH20 = 3;
+		double etaHend = 0.5;
+		double bH20 = 0.33;
+		double muAUT20 = 1.12;
+		double KNO2 = 0.5;
+		double KNNH4 = 0.7;
+		double KNALK = 0.5;
+		double bAUT20 = 0.18;
+		double etaNend = 0.5;
+
+		// derived parameters
+		double ft04 = exp(-0.04 * (20.0 - static_cast<double>(T)));
+		double ft07 = exp(-0.06952 * (20 - static_cast<double>(T)));
+		double ft105 = exp(-0.105 * (20 - static_cast<double>(T)));
+		double k_sto = static_cast<double>(k_sto20) * ft07;
+		double muH = static_cast<double>(muH20) * ft07;
+		double bH = static_cast<double>(bH20) * ft07;
[…]
