### 2025-05-14 8bb2dff
ASM3: differentiate r6-12
