+++ src/libcadet/model/reaction/ActivatedSludgeModelThree.cpp
@@ -491,110 +491,88 @@ protected:
 		double XA = y[11];
 		double XMI = y[12]; // unused
 
+		/*
+		size_t idxSO = 0;
+		size_t idxSS = 1;
+		size_t idxSNH = 2;
+		size_t idxSNO = 3;
+		size_t idxSN2 = 4;
+		size_t idxSALK = 5;
+		size_t idxSI = 6;
+		size_t idxXI = 7;
+		size_t idxXS = 8;
+		size_t idxXH = 9;
+		size_t idxXSTO = 10;
+		size_t idxXA = 11;
+		size_t idxXMI = 12;
+		*/
+
 		// reaction 1: Kh20 * ft04 * XS/XH_S / (XS/XH_S + KX) * XH;
-		// dp1/XS = Kh20 * ft04 * XH / (XS/XH_S + KX) - Kh20 * ft04 * XS/XH_S / (XS/XH_S + KX)^2 * XH;
 		double dp1_XS = Kh20 * ft04 * XH / (XS/XH_S + KX) - Kh20 * ft04 * XS/XH_S / ((XS/XH_S + KX) * (XS/XH_S + KX)) * XH;
-		// dp1/XH_S = -Kh20 * ft04 * XS / (XS/XH_S + KX) / XH_S^2 * XH;
 		double dp1_XH_S = -Kh20 * ft04 * XS / (XS/XH_S + KX) / (XH_S * XH_S) * XH;
-		// dp1/XH = Kh20 * ft04 * XS/XH_S / (XS/XH_S + KX);
 		double dp1_XH = Kh20 * ft04 * XS/XH_S / (XS/XH_S + KX);
 
-		//curJac = jac;
-		// curJac[8] =  dr1/XS;
-		// curJac[X] =  dr1/XH_S; ??
-		// curJac[9] = dr1/XH;
-
-		// curjac++;
-
 		// reaction2: k_sto * SO / (SO + KHO2) * SS / (SS + KHSS) * SNO / (SNO + KHNO3) * XH;
-		// dp2/SO = k_sto * ft07 * SS / (SS + KHSS) * SNO / (SNO + KHNO3) / (SO + KHO2)^2 * XH;
 		double dp2_SO = k_sto * ft07 * SS / (SS + KHSS) * SNO / (SNO + KHNO3) / ((SO + KHO2) * (SO + KHO2)) * XH;
-		// dp2/SS = k_sto * ft07 * SO / (SO + KHO2) * SNO / (SNO + KHNO3) / (SS + KHSS)^2 * XH;
 		double dp2_SS = k_sto * ft07 * SO / (SO + KHO2) * SNO / (SNO + KHNO3) / ((SS + KHSS) * (SS + KHSS)) * XH;
-		// dp2/SNO = k_sto * ft07 * SO / (SO + KHO2) * SS / (SS + KHSS) / (SNO + KHNO3)^2 * XH;
 		double dp2_SNO = k_sto * ft07 * SO / (SO + KHO2) * SS / (SS + KHSS) / ((SNO + KHNO3) * (SNO + KHNO3)) * XH;
-		// dp2/XH = k_sto * ft07 * SO / (SO + KHO2) * SS / (SS + KHSS) * SNO / (SNO + KHNO3);
 		double dp2_XH = k_sto * ft07 * SO / (SO + KHO2) * SS / (SS + KHSS) * SNO / (SNO + KHNO3);
 
-		//curJac[-1] = dr2/S0;
-		//curJac[0] = dr2/SS;
-		//curJac[1] = dr2/SNO;
-		//curJac[8] = dr2/XH;
-
[…]
