### 2025-05-14 acabe81
ASM3: compute jacobian
