+++ src/libcadet/model/reaction/ActivatedSludgeModelThree.cpp
@@ -491,7 +491,6 @@ protected:
 		double XA = y[11];
 		double XMI = y[12]; // unused
 
-		/*
 		size_t idxSO = 0;
 		size_t idxSS = 1;
 		size_t idxSNH = 2;
@@ -505,82 +504,96 @@ protected:
 		size_t idxXSTO = 10;
 		size_t idxXA = 11;
 		size_t idxXMI = 12;
+
+		double d[13][13] = {};
+		//memset(d, 0, 13 * 13 * sizeof(double));
+		// p1: Kh20 * ft04 * XS/XH_S / (XS/XH_S + KX) * XH;
+		d[0][idxXS] = Kh20 * ft04 * XH / (XS/XH_S + KX) - Kh20 * ft04 * XS/XH_S / ((XS/XH_S + KX) * (XS/XH_S + KX)) * XH;
+		//double dp1_XH_S = -Kh20 * ft04 * XS / (XS/XH_S + KX) / (XH_S * XH_S) * XH;
+		d[0][idxXH] = Kh20 * ft04 * XS/XH_S / (XS/XH_S + KX);
+
+		/*
+		curJac = jac;
+		size_t curIdxXS = idxXS;
+		for (size_t i = 0; i < _stoichiometry.rows(); i++) {
+			curJac[idxXS] += _stoichiometry.native(i, 0) * dp1_XS;
+			curJac[idxXH] += _stoichiometry.native(i, 0) * dp1_XH;
+			curJac++;
+			curIdxXS--;
+		}
 		*/
 
-		// reaction 1: Kh20 * ft04 * XS/XH_S / (XS/XH_S + KX) * XH;
-		double dp1_XS = Kh20 * ft04 * XH / (XS/XH_S + KX) - Kh20 * ft04 * XS/XH_S / ((XS/XH_S + KX) * (XS/XH_S + KX)) * XH;
-		double dp1_XH_S = -Kh20 * ft04 * XS / (XS/XH_S + KX) / (XH_S * XH_S) * XH;
-		double dp1_XH = Kh20 * ft04 * XS/XH_S / (XS/XH_S + KX);
-
 		// reaction2: k_sto * SO / (SO + KHO2) * SS / (SS + KHSS) * SNO / (SNO + KHNO3) * XH;
-		double dp2_SO = k_sto * ft07 * SS / (SS + KHSS) * SNO / (SNO + KHNO3) / ((SO + KHO2) * (SO + KHO2)) * XH;
-		double dp2_SS = k_sto * ft07 * SO / (SO + KHO2) * SNO / (SNO + KHNO3) / ((SS + KHSS) * (SS + KHSS)) * XH;
-		double dp2_SNO = k_sto * ft07 * SO / (SO + KHO2) * SS / (SS + KHSS) / ((SNO + KHNO3) * (SNO + KHNO3)) * XH;
-		double dp2_XH = k_sto * ft07 * SO / (SO + KHO2) * SS / (SS + KHSS) * SNO / (SNO + KHNO3);
+		d[1][idxSO] = k_sto * ft07 * SS / (SS + KHSS) * SNO / (SNO + KHNO3) / ((SO + KHO2) * (SO + KHO2)) * XH;
+		d[1][idxSS] = k_sto * ft07 * SO / (SO + KHO2) * SNO / (SNO + KHNO3) / ((SS + KHSS) * (SS + KHSS)) * XH;
+		d[1][idxSNO] = k_sto * ft07 * SO / (SO + KHO2) * SS / (SS + KHSS) / ((SNO + KHNO3) * (SNO + KHNO3)) * XH;
[…]
