### 2025-05-14 ff508ae
ASM3: set up params for codegen
