### 2025-05-14 564fc8c
add ASM3 to factory
