### 2025-05-14 b3660e7
ASM3: fix codegen json
