### 2025-05-14 f704613
ASM3: use scalar external params
