### 2025-05-14 49dd73c
ASM3: fix jacobian computation
