### 2025-05-14 e78df34
overload ScalarParameter methods for reaction model compatibility
