### 2025-05-15 507847c
parameter declariation and jacobian fix
