+++ src/libcadet/model/reaction/ActivatedSludgeModelThree.cpp
@@ -32,11 +32,11 @@
 	"externalName": "ExtActivatedSludgeModelThreeParamHandler",
 	"parameters":
 		[
-			{ "type": "ScalarParameter", "varName": "Kh20", "confName": "ASM3_KH_20"},
+			{ "type": "ScalarParameter", "varName": "Kh20", "confName": "ASM3_KH20"},
 			{ "type": "ScalarParameter", "varName": "T", "confName": "ASM3_T"},
-			{ "type": "ScalarParameter", "varName": "iO2", "confName": "ASM3_I_O2"},
+			{ "type": "ScalarParameter", "varName": "iO2", "confName": "ASM3_IO2"},
 			{ "type": "ScalarParameter", "varName": "V", "confName": "ASM3_V"},
-			{ "type": "ScalarParameter", "varName": "k_sto20", "confName": "ASM3_K_STO_20"},
+			{ "type": "ScalarParameter", "varName": "k_sto20", "confName": "ASM3_KSTO20"},
 			{ "type": "ScalarParameter", "varName": "KX", "confName": "ASM3_KX"},
 			{ "type": "ScalarParameter", "varName": "KHO2", "confName": "ASM3_KHO2"},
 			{ "type": "ScalarParameter", "varName": "KHSS", "confName": "ASM3_KHSS"},
@@ -45,21 +45,23 @@
 			{ "type": "ScalarParameter", "varName": "KHNH4", "confName": "ASM3_KHNH4"},
 			{ "type": "ScalarParameter", "varName": "KHALK", "confName": "ASM3_KHALK"},
 			{ "type": "ScalarParameter", "varName": "KHSTO", "confName": "ASM3_KHSTO"},
-			{ "type": "ScalarParameter", "varName": "muH20", "confName": "ASM3_MU_H_20"},
-			{ "type": "ScalarParameter", "varName": "etaHend", "confName": "ASM3_ETA_H_END"},
-			{ "type": "ScalarParameter", "varName": "bH20", "confName": "ASM3_B_H_20"},
-			{ "type": "ScalarParameter", "varName": "muAUT20", "confName": "ASM3_MU_AUT_20"},
+			{ "type": "ScalarParameter", "varName": "muH20", "confName": "ASM3_MU_H20"},
+			{ "type": "ScalarParameter", "varName": "etaHend", "confName": "ASM3_ETAH_END"},
+			{ "type": "ScalarParameter", "varName": "bH20", "confName": "ASM3_BH20"},
+			{ "type": "ScalarParameter", "varName": "muAUT20", "confName": "ASM3_MU_AUT20"},
 			{ "type": "ScalarParameter", "varName": "KNO2", "confName": "ASM3_KNO2"},
 			{ "type": "ScalarParameter", "varName": "KNNH4", "confName": "ASM3_KNNH4"},
 			{ "type": "ScalarParameter", "varName": "KNALK", "confName": "ASM3_KNALK"},
-			{ "type": "ScalarParameter", "varName": "bAUT20", "confName": "ASM3_B_AUT_20"},
-			{ "type": "ScalarParameter", "varName": "etaNend", "confName": "ASM3_ETA_N_END"}
+			{ "type": "ScalarParameter", "varName": "bAUT20", "confName": "ASM3_BAUT20"},
+			{ "type": "ScalarParameter", "varName": "etaNend", "confName": "ASM3_ETAN_END"}
 		]
 }
 </codegen>*/
 
 /* Parameter description
  ------------------------
+
+
 */
 
[…]
