### 2025-05-15 0750d08
fic internal variable determination numbered value int-> double
