### 2025-05-15 ff9cd64
Expand sensitivity tests (#410)
