### 2025-05-15 dc79d18
Modularize particle DG discretization (#406)
