+++ src/libcadet/model/GeneralRateModelDG.cpp
@@ -55,8 +55,7 @@ constexpr double SurfVolRatioSlab = 1.0;
 
 
 GeneralRateModelDG::GeneralRateModelDG(UnitOpIdx unitOpIdx) : UnitOperationBase(unitOpIdx),
-	_hasSurfaceDiffusion(0, false), _singleParDepSurfDiffusion(false), _dynReactionBulk(nullptr),
-	_globalJac(), _globalJacDisc(), _jacInlet(), _hasParDepSurfDiffusion(false),
+	_globalJac(), _globalJacDisc(), _jacInlet(), _dynReactionBulk(nullptr),
 	_analyticJac(true), _jacobianAdDirs(0), _factorizeJacobian(false), _tempState(nullptr),
 	_initC(0), _initCp(0), _initQ(0), _initState(0), _initStateDot(0)
 {
@@ -67,8 +66,6 @@ GeneralRateModelDG::~GeneralRateModelDG() CADET_NOEXCEPT
 	delete[] _tempState;
 
 	delete _dynReactionBulk;
-
-	clearParDepSurfDiffusion();
 
 	delete _linearSolver;
 }
@@ -100,22 +97,6 @@ bool GeneralRateModelDG::usesAD() const CADET_NOEXCEPT
 	// We only need AD if we are not computing the Jacobian analytically
 	return !_analyticJac;
 #endif
-}
-
-void GeneralRateModelDG::clearParDepSurfDiffusion()
-{
-	if (_singleParDepSurfDiffusion)
-	{
-		if (!_parDepSurfDiffusion.empty())
-			delete _parDepSurfDiffusion[0];
-	}
-	else
-	{
-		for (IParameterStateDependence* pd : _parDepSurfDiffusion)
-			delete pd;
-	}
-
-	_parDepSurfDiffusion.clear();
 }
 
 bool GeneralRateModelDG::configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper)
@@ -184,78 +165,21 @@ bool GeneralRateModelDG::configureModelDiscretization(IParameterProvider& paramP
 	else // Infer number of particle types
 		_disc.nParType = nBound.size() / _disc.nComp;
 
-	std::vector<int> parPolyDeg(_disc.nParType);
-	std::vector<int> ParNelem(_disc.nParType);
+	paramProvider.popScope();
+	Indexer idxr(_disc);
+	const bool particleConfSuccess = _parDiffOp.configureModelDiscretization(paramProvider, helper, _disc.nComp, _disc.nParType, idxr.strideColComp());
+	paramProvider.pushScope("discretization");
+
 	if (firstConfigCall)
 	{
-		_disc.parPolyDeg = new unsigned int[_disc.nParType];
-		_disc.nParCell = new unsigned int[_disc.nParType];
-		_disc.parGSM = new bool[_disc.nParType];
-	}
-
-	if (paramProvider.exists("PAR_POLYDEG"))
-	{
-		parPolyDeg = paramProvider.getIntArray("PAR_POLYDEG");
-
-		if ((std::any_of(parPolyDeg.begin(), parPolyDeg.end(), [](int value) { return value < 1; })))
-			throw InvalidParameterException("Particle polynomial degrees must be at least 1!");
-		ParNelem = paramProvider.getIntArray("PAR_NELEM");
-
-		if ((std::any_of(ParNelem.begin(), ParNelem.end(), [](int value) { return value < 1; })))
-			throw InvalidParameterException("Particle number of elements must be at least 1!");
-
-		if (parPolyDeg.size() == 1)
-		{
-			// Multiplex number of particle shells to all particle types
-			for (unsigned int i = 0; i < _disc.nParType; ++i)
[…]
