+++ src/libcadet/model/parts/ParticleDiffusionOperatorDG.cpp
@@ -1,0 +1,2404 @@
+// =============================================================================
+//  CADET
+//
+//  Copyright © 2008-present: The CADET-Core Authors
+//            Please see the AUTHORS.md file.
+//
+//  All rights reserved. This program and the accompanying materials
+//  are made available under the terms of the GNU Public License v3.0 (or, at
+//  your option, any later version) which accompanies this distribution, and
+//  is available at http://www.gnu.org/licenses/gpl.html
+// =============================================================================
+
+#include "model/parts/ParticleDiffusionOperatorDG.hpp"
+#include "cadet/Exceptions.hpp"
+
+#include "ParamReaderHelper.hpp"
+#include "AdUtils.hpp"
+#include "SimulationTypes.hpp"
+#include "model/ParameterDependence.hpp"
+#include "SensParamUtil.hpp"
+#include "ConfigurationHelper.hpp"
+
+#include "LoggingUtils.hpp"
+#include "Logging.hpp"
+
+#include <Eigen/Dense>
+#include <Eigen/Sparse>
+
+using namespace Eigen;
+
+namespace cadet
+{
+
+namespace model
+{
+
+namespace parts
+{
+
+	/**
+	 * @brief Creates a ParticleDiffusionOperatorDG
+	 */
+	ParticleDiffusionOperatorDG::ParticleDiffusionOperatorDG() : _nParElem(nullptr), _nParPointsBeforeType(nullptr),
+		_parPolyDeg(nullptr), _nParNode(nullptr), _nParPoints(nullptr), _parGSM(nullptr), //_parTypeOffset(nullptr),
+		_nBound(nullptr), _boundOffset(nullptr), _strideBound(nullptr), _nBoundBeforeType(nullptr), _deltaR(nullptr), _parNodes(nullptr),
+		_parPolyDerM(nullptr), _minus_InvMM_ST(nullptr), _minus_parInvMM_Ar(nullptr), _parInvWeights(nullptr), _parInvMM(nullptr), _parInvMM_Leg(nullptr),
+		_Ir(nullptr), _secondOrderStiffnessM(nullptr), _DGjacParDispBlocks(nullptr), _g_p(nullptr), _g_pSum(nullptr),
+		_surfaceFluxParticle(nullptr), _localFlux(nullptr)
+	{
+	}
+
+	ParticleDiffusionOperatorDG::~ParticleDiffusionOperatorDG() CADET_NOEXCEPT
+	{
+		delete[] _nParElem;
+		delete[] _nParPointsBeforeType;
+		delete[] _parPolyDeg;
+		delete[] _nParNode;
+		delete[] _nParPoints;
+		delete[] _parGSM;
+		//delete[] _parTypeOffset;
+		delete[] _nBound;
+		delete[] _boundOffset;
+		delete[] _strideBound;
+		delete[] _nBoundBeforeType;
+
+		delete[] _deltaR;
+		delete[] _parNodes;
+		delete[] _parPolyDerM;
+		delete[] _minus_InvMM_ST;
+		delete[] _minus_parInvMM_Ar;
+		delete[] _parInvWeights;
+		delete[] _parInvMM;
+		delete[] _parInvMM_Leg;
+		delete[] _Ir;
+		delete[] _secondOrderStiffnessM;
+
+		delete[] _DGjacParDispBlocks;
+
+		delete[] _g_p;
+		delete[] _g_pSum;
+		delete[] _surfaceFluxParticle;
+		delete[] _localFlux;
+
+		clearParDepSurfDiffusion();
+	}
+
+	void ParticleDiffusionOperatorDG::clearParDepSurfDiffusion()
+	{
+		if (_singleParDepSurfDiffusion)
[…]
