### 2025-05-15 f0ef933
Modularize particle DG discretization (#406)
