### 2025-05-15 8e6629e
Modularize particle DG discretization (#406)
