+++ src/libcadet/model/GeneralRateModelDG.hpp
@@ -23,6 +23,7 @@
 #include "cadet/StrongTypes.hpp"
 #include "cadet/SolutionExporter.hpp"
 #include "model/parts/ConvectionDispersionOperatorDG.hpp"
+#include "model/parts/ParticleDiffusionOperatorDG.hpp"
 #include "AutoDiff.hpp"
 #include "linalg/BandedEigenSparseRowIterator.hpp"
 #include "linalg/EigenSolverWrapper.hpp"
@@ -250,8 +251,6 @@ protected:
 	int multiplexInitialConditions(const cadet::ParameterId& pId, unsigned int adDirection, double adValue);
 	int multiplexInitialConditions(const cadet::ParameterId& pId, double val, bool checkSens);
 
-	void clearParDepSurfDiffusion();
-
 	parts::cell::CellParameters makeCellResidualParams(unsigned int parType, int const* qsReaction) const;
 
 #ifdef CADET_CHECK_ANALYTIC_JACOBIAN
@@ -267,12 +266,7 @@ protected:
 		unsigned int nPoints; //!< Number of discrete column Points
 		bool exactInt;	//!< 1 for exact integration, 0 for inexact LGL quadrature
 		unsigned int nParType; //!< Number of particle types
-		unsigned int* nParCell; //!< Array with number of radial cells in each particle type
-		unsigned int* nParPointsBeforeType; //!< Array with total number of radial points before a particle type (cumulative sum of nParPoints), additional last element contains total number of particle shells
-		unsigned int* parPolyDeg; //!< polynomial degree of particle elements
-		unsigned int* nParNode; //!< Array with number of radial nodes per cell in each particle type
 		unsigned int* nParPoints; //!< Array with number of radial nodes per cell in each particle type
-		bool* parGSM; //!< specifies whether (single element) Galerkin spectral method should be used in particles
 		unsigned int* parTypeOffset; //!< Array with offsets (in particle block) to particle type, additional last element contains total number of particle DOFs
 		unsigned int* nBound; //!< Array with number of bound states for each component and particle type (particle type major ordering)
 		unsigned int* boundOffset; //!< Array with offset to the first bound state of each component in the solid phase (particle type major ordering)
@@ -284,315 +278,23 @@ protected:
 		// parameter
 		int curSection; //!< current time section index
 
-		const double SurfVolRatioSlab = 1.0; //!< Surface to volume ratio for a slab-shaped particle
-		const double SurfVolRatioCylinder = 2.0; //!< Surface to volume ratio for a cylindrical particle
-		const double SurfVolRatioSphere = 3.0; //!< Surface to volume ratio for a spherical particle
-
-		/*					DG specifics				*/
-		
-		active* deltaR; //!< equidistant particle element spacing for each particle type
-		Eigen::VectorXd* parNodes; //!< Array with positions of nodes in radial reference element for each particle
-		Eigen::MatrixXd* parPolyDerM; //!< Array with polynomial derivative Matrix for each particle
-		Eigen::MatrixXd* minus_InvMM_ST; //!< equals minus inverse mass matrix times transposed stiffness matrix. Required solely for exact integration DG discretization of particle equation
[…]
