### 2025-05-15 d93381a
Modularize particle DG discretization (#406)
