### 2025-05-15 d201fab
Modularize particle DG discretization (#406)
