### 2025-05-15 93e5db5
Modularize particle DG discretization (#406)
