+++ doc/interface/reaction/activated_sludge_model.rst
@@ -1,0 +1,184 @@
+.. _activated_sludge_model_config:
+
+Activated Sludge Model (ASM3h)
+~~~~~~~~~~~~~~~~~~~~~~~~~
+
+**Group /input/model/unit_XXX/reaction - REACTION_MODEL = ACTIVATED_SLUDGE_MODEL3**
+
+For information on model equations, refer to :ref:`activated_sludge_model`.
+
+``ASM3_KH20``
+
+   Hydrolysis rate constant :math:`k_H` at 20 °C.
+   
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_T``
+
+   Temperature :math:`T`
+   
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_IO2``
+
+   Aeration rate :math:`iO_2`
+   
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\ge 0`       **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_V``
+
+   Aeration volume :math:`V`
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\gt 0`       **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KSTO20``
+
+   Saturation constant :math:`K_{STO}` at 20 °C
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\gt 0`       **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KX``
+
+   
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KHO2``
+
+   
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KHSS``
+
+   
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KHNO3``
+
+   
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_ETA_HNO3``
+
+   
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KHNH4``
+
+   
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KHALK``
+
+   
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KHSTO``
+
+   
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_MU_H20``
+
+   
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_ETAH_END``
+
+   
+
[…]
