+++ doc/interface/reaction/activated_sludge_model.rst
@@ -1,0 +1,184 @@
+.. _activated_sludge_model_config:
+
+Activated Sludge Model (ASM3h)
+~~~~~~~~~~~~~~~~~~~~~~~~~
+
+**Group /input/model/unit_XXX/reaction - REACTION_MODEL = ACTIVATED_SLUDGE_MODEL3**
+
+For information on model equations, refer to :ref:`activated_sludge_model`.
+
+``ASM3_KH20``
+
+   Hydrolysis rate constant :math:`k_H` at 20 °C.
+   
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_T``
+
+   Temperature :math:`T`.
+   
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_IO2``
+
+   Aeration rate :math:`iO_2`.
+   
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\ge 0`       **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_V``
+
+   Aeration volume :math:`V`.
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\gt 0`       **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KSTO20``
+
+   Saturation constant :math:`K_{STO}` at 20 °C.
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\gt 0`       **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KX``
+
+   Hydrolysis saturation constant :math:`KX`.
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KHO2``
+
+   :math:`SO_2` saturation constant :math:`KHO_2`.
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KHSS``
+
+   Substrate :math:`SS` saturation constant :math:`KHSS`.
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KHNO3``
+
+   :math:`SNO_X` saturation constant :math:`KHNO_3`.
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_ETA_HNO3``
+
+   Anoxic reduction factor :math:`\eta_{HNO_3}`.
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KHNH4``
+
+   Ammonium saturation constant :math:`KHNH_4`.
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_KHALK``
+
+   :math:`XH` alkalinity saturation constant :math:`KH_{ALK}`.
+
+   ================  =============================  ========================================================
[…]
