### 2025-05-19 ee896db
ASM3: document remaining params from SIMBA#
