### 2025-05-19 ce294c2
fix lost of jacobi equaitons
