+++ src/libcadet/model/reaction/ActivatedSludgeModelThree.cpp
@@ -239,7 +239,7 @@ protected:
 		_stoichiometry.native(0, 3) = 1 - 1 / YH_aer;
 		_stoichiometry.native(0, 5) = -1 * (1 - fXI);
 		_stoichiometry.native(0, 7) = -1;
-		_stoichiometry.native(0, 9) = -64.0/14.0 * 1/YA + 1;
+		_stoichiometry.native(0, 9) = -(64.0/14.0) * 1/YA + 1;
 		_stoichiometry.native(0, 10) = -1 * (1 - fXI);
 		_stoichiometry.native(0, 12) = 1;
 
@@ -387,26 +387,32 @@ protected:
 		// StateType XI = y[7]; // unused
 		StateType XS = y[8];
 		StateType XH = y[9];
-		StateType XH_S = XH; // ASM3hC: XH_S = max(XH, 0.1)
 		StateType XSTO = y[10];
 		StateType XA = y[11];
 		// StateType XMI = y[12]; // unused
 
 		
 		// p1: Hydrolysis of organic structures
-		fluxes[0] = Kh20 * ft04 * XS/XH_S / (XS/XH_S + KX) * XH;
+		fluxes[0] = Kh20 * ft04 * (XS/XH) / ((XS/XH) + KX) * XH;
+		if (XH < 0.1)
+			fluxes[0] = Kh20 * ft04 * (XS/0.1) / ((XS/0.1) + KX) * XH;
+
 		
 		// p2: Aerobic storage of SS
-		fluxes[1] = k_sto * SO / (SO + KHO2) * SS / (SS + KHSS) * SNO / (SNO + KHNO3) * XH;
+		fluxes[1] = k_sto * SO / (SO + KHO2) * SS / (SS + KHSS)  * XH;
 
 		// p3: Anoxic storage of SS
 		fluxes[2] = k_sto * etaHNO3 * KHO2 / (SO + KHO2) * SS / (SS + KHSS) * SNO / (SNO + KHNO3) * XH;
 
 		// p4: Aerobic growth of heterotrophic biomass (XH)
-		fluxes[3] = muH * SO / (SO + KHO2) * SNH / (SNH + KHNH4) * SALK / (SALK + KHALK) * XSTO/XH_S / (XSTO/XH_S + KHSTO) * XH;
+		fluxes[3] = muH * SO / (SO + KHO2) * SNH / (SNH + KHNH4) * SALK / (SALK + KHALK) * ((XSTO/XH)) / (((XSTO/XH)) + KHSTO) * XH;
+		if (XH < 0.1)
+			fluxes[3] = muH * SO / (SO + KHO2) * SNH / (SNH + KHNH4) * SALK / (SALK + KHALK) * (XSTO / 0.1) / ((XSTO / 0.1) + KHSTO) * XH;
 
 		// p5: Anoxic growth of heterotrophic biomass (XH, denitrification)
-		fluxes[4] = muH * etaHNO3 * KHO2 / (KHO2 + SO) * SNH / (KHNH4 + SNH) * SALK / (KHALK + SALK) * XSTO / XH_S * 1 / (KHSTO + XSTO/XH_S) * SNO / (KHNO3 + SNO) * XH;
+		fluxes[4] = muH * etaHNO3 * KHO2 / (KHO2 + SO) * SNH / (KHNH4 + SNH) * SALK / (KHALK + SALK) * (XSTO / XH) * 1 / (KHSTO + (XSTO / XH)) * SNO / (KHNO3 + SNO) * XH;
+		if (XH < 0.1)
[…]
