### 2025-05-19 3c2ce35
Fix ScalarParameter method argument description
