### 2025-05-19 63d5652
Overload ScalarParameter memory reservation functions (#412)
The functions are overloaded to allow different parameters needed e.g. in a planned extension of the reaction module.
