### 2025-05-19 1773de2
ASM3: use ScalarParameter fix from #412
