+++ src/libcadet/model/binding/AffinityComplexTitrationBinding.cpp
@@ -1,0 +1,253 @@
+// =============================================================================
+//  CADET
+//  
+//  Copyright © 2008-present: The CADET-Core Authors
+//            Please see the AUTHORS.md file.
+//  
+//  All rights reserved. This program and the accompanying materials
+//  are made available under the terms of the GNU Public License v3.0 (or, at
+//  your option, any later version) which accompanies this distribution, and
+//  is available at http://www.gnu.org/licenses/gpl.html
+// =============================================================================
+
+#include "model/binding/BindingModelBase.hpp"
+#include "model/ExternalFunctionSupport.hpp"
+#include "model/ModelUtils.hpp"
+#include "cadet/Exceptions.hpp"
+#include "model/Parameters.hpp"
+#include "LocalVector.hpp"
+#include "SimulationTypes.hpp"
+
+#include <functional>
+#include <unordered_map>
+#include <string>
+#include <vector>
+
+/*<codegen>
+{
+	"name": "AffinityComplexTitrationParamHandler",
+	"externalName": "ExtAffinityComplexTitrationParamHandler",
+	"parameters":
+		[
+			{ "type": "ScalarComponentDependentParameter", "varName": "kA", "confName": "ACT_KA"},
+			{ "type": "ScalarComponentDependentParameter", "varName": "kD", "confName": "ACT_KD"},
+			{ "type": "ScalarComponentDependentParameter", "varName": "qMax", "confName": "ACT_QMAX"},
+			{ "type": "ScalarComponentDependentParameter", "varName": "etaA", "confName": "ACT_ETAA"},
+			{ "type": "ScalarComponentDependentParameter", "varName": "etaG", "confName": "ACT_ETAG"},
+			{ "type": "ScalarComponentDependentParameter", "varName": "pKaA", "confName": "ACT_PKAA"},
+			{ "type": "ScalarComponentDependentParameter", "varName": "pKaG", "confName": "ACT_PKAG"}
+		]
+}
+</codegen>*/
+
+/* Parameter description
+ ------------------------
+ All these parameters also need to be defined for pH
+ kA = Adsorption rate
+ kD = Desorption rate
+ qMax = Binding apacity
+ etaA = Slope for the binding capacity changes against pH
+ pKaA = Center for the binding capacity changes against pH
+ etaG = Slope for the equilibrium constant changes against pH
+ pKaG = Center for the equilibrium constant changes against pH
+*/
+
+namespace cadet
+{
+
+	namespace model
+	{
+
+		inline const char* AffinityComplexTitrationParamHandler::identifier() CADET_NOEXCEPT { return "AFFINITY_COMPLEX_TITRATION"; }
+
+		inline bool AffinityComplexTitrationParamHandler::validateConfig(unsigned int nComp, unsigned int const* nBoundStates)
+		{
+			if ((_kA.size() != _kD.size()) || (_kA.size() != _qMax.size()) || (_kA.size() < nComp) || (_kA.size() != _etaA.size()) || (_kA.size() != _pKaA.size()) || (_kA.size() != _etaG.size()) || (_kA.size() != _pKaG.size()))
+				throw InvalidParameterException("ACT_KA, ACT_KD, ACT_QMAX, ACT_ETAA, ACT_PKAA, ACT_ETAG and ACT_PKAG have to have the same size");
+
+			return true;
+		}
+
+		inline const char* ExtAffinityComplexTitrationParamHandler::identifier() CADET_NOEXCEPT { return "EXT_AFFINITY_COMPLEX_TITRATION"; }
+
[…]
