+++ doc/interface/binding/index.rst
@@ -74,4 +74,4 @@ This group also takes precedence over a possibly existing ``/input/model/unit_XX
     hic_constant_water_activity
     hic_unified
     multi_component_colloidal
-
+    affinity_complex_titration
