### 2025-05-20 fdd9af4
Add base for adding ACT isotherm
