+++ test/BindingModels.cpp
@@ -1852,3 +1852,79 @@ CADET_BINDINGTEST("HIC_CONSTANT_WATER_ACTIVITY", "EXT_HIC_CONSTANT_WATER_ACTIVIT
 			"EXT_HICUNI_REFC1": 7.0
 	)json", \
 		1e-10, 1e-10, CADET_NONBINDING_LIQUIDPHASE_COMP_USED, CADET_COMPARE_BINDING_VS_NONBINDING)
+CADET_BINDINGTEST("AFFINITY_COMPLEX_TITRATION", "EXT_AFFINITY_COMPLEX_TITRATION", (0, 1), (0, 0, 1), (5.5, 0.06666667, 0.0), (5.5, 0.1, 0.06666667, 0.0), \
+	R"json( "ACT_KA": [1.0, 1386.9],
+	        "ACT_KD": [1.0, 1.0],
+	        "ACT_QMAX": [0.0000000001, 10.7439614],
+	        "ACT_ETAA": [0.0, 1.81],
+	        "ACT_ETAG": [0.0, 2.28],
+	        "ACT_PKAA": [0.0, 2.07],
+            "ACT_PKAG": [0.0, 5.29]
+	)json", \
+	R"json( "ACT_KA": [1.0, 1.0, 1386.9],
+	        "ACT_KD": [1.0, 1.0, 1.0],
+	        "ACT_QMAX": [0.0000000001, 0.0000000001, 10.7439614],
+	        "ACT_ETAA": [0.0, 0.0, 1.81],
+	        "ACT_ETAG": [0.0, 0.0, 2.28],
+	        "ACT_PKAA": [0.0, 0.0, 2.07],
+            "ACT_PKAG": [0.0, 0.0, 5.29]
+	)json", \
+	R"json( "EXT_ACT_KA": [0.0, 0.0],
+	        "EXT_ACT_KA_T": [1.0, 1386.9],
+	        "EXT_ACT_KA_TT": [0.0, 0.0],
+	        "EXT_ACT_KA_TTT": [0.0, 0.0],
+	        "EXT_ACT_KD": [0.0, 0.0],
+	        "EXT_ACT_KD_T": [1.0, 1.0],
+	        "EXT_ACT_KD_TT": [0.0, 0.0],
+	        "EXT_ACT_KD_TTT": [0.0, 0.0],
+	        "EXT_ACT_QMAX": [0.0, 0.0],
+	        "EXT_ACT_QMAX_T": [0.0000000001, 10.7439614],
+	        "EXT_ACT_QMAX_TT": [0.0, 0.0],
+	        "EXT_ACT_QMAX_TTT": [0.0, 0.0],
+	        "EXT_ACT_ETAA": [0.0, 0.0],
+	        "EXT_ACT_ETAA_T": [0.0, 1.81],
+	        "EXT_ACT_ETAA_TT": [0.0, 0.0],
+	        "EXT_ACT_ETAA_TTT": [0.0, 0.0],
+	        "EXT_ACT_ETAG": [0.0, 0.0],
+	        "EXT_ACT_ETAG_T": [0.0, 2.28],
+	        "EXT_ACT_ETAG_TT": [0.0, 0.0],
+	        "EXT_ACT_ETAG_TTT": [0.0, 0.0],
+	        "EXT_ACT_PKAA": [0.0, 0.0],
+	        "EXT_ACT_PKAA_T": [0.0, 2.07],
+	        "EXT_ACT_PKAA_TT": [0.0, 0.0],
+	        "EXT_ACT_PKAA_TTT": [0.0, 0.0],
[…]
