+++ doc/literature.bib
@@ -662,3 +662,14 @@
   title        = {On the {Structure} of a {Catalyst} {Surface}},
   volume       = {16},
 }
+
+@article{Zhang2025ACT,
+  author       = {Zhang, Wendi and DiNenna, Virginia and Przybycien, Todd},
+  year         = {2025},
+  doi          = {10.1002/bit.29035},
+  issn         = {0098-1354},
+  journal = {Biotechnology and Bioengineering},
+  pages        = {108860},
+  title        = {An Affinity Complex Titration Isotherm for Mechanistic Modeling in Protein A Chromatography},
+  volume       = {192},
+}
