### 2025-05-20 9903bd2
Add ACT isotherm (#415)
