+++ src/libcadet/CMakeLists.txt
@@ -103,6 +103,7 @@ set(LIBCADET_BINDINGMODEL_SOURCES
 	${CMAKE_SOURCE_DIR}/src/libcadet/model/binding/HICUnifiedBinding.cpp
 	${CMAKE_SOURCE_DIR}/src/libcadet/model/binding/SipsBinding.cpp
 	${CMAKE_SOURCE_DIR}/src/libcadet/model/binding/MultiComponentLDFFreundlichBinding.cpp
+        ${CMAKE_SOURCE_DIR}/src/libcadet/model/binding/AffinityComplexTitrationBinding.cpp
 )
 set(LIBCADET_EXCHANGEMODEL_SOURCES
 	${CMAKE_SOURCE_DIR}/src/libcadet/model/exchange/LinearExchange.cpp
