### 2025-05-20 69a968f
add jacobian test and fix jacobian
