+++ test/ReactionModels.cpp
@@ -131,7 +131,7 @@ TEST_CASE("MichaelisMenten kinetic analytic Jacobian vs AD without inhibition", 
 
 TEST_CASE("MichaelisMenten kinetic analytic Jacobian vs AD with inhibition", "[MichaelisMenten],[ReactionModel],[Jacobian],[AD]")
 {
-	const unsigned int nBound[] = {1, 2, 1};
+	const unsigned int nBound[] = {1.0, 2.0, 1.0};
 	const double point[] = {1.0, 2.0, 1.4, 2.1, 0.2, 1.1, 1.8};
 	cadet::test::reaction::testDynamicJacobianAD("MICHAELIS_MENTEN", 3, nBound,
 		R"json({
@@ -145,17 +145,54 @@ TEST_CASE("MichaelisMenten kinetic analytic Jacobian vs AD with inhibition", "[M
 		point, 1e-15, 1e-15
 	);
 }
-TEST_CASE("ASM3 analytic Jacobian vs AD", "[ASM3],[ReactionModel],[Jacobian],[AD]")
+TEST_CASE("ASM3 analytic Jacobian vs AD", "[ASM3],[ReactionModel],[Jacobian],[AD],[testHere]")
 {
-	const unsigned int nBound[] = {0};
+	const unsigned int nBound[] = { 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,0.0, 0.0, 0.0 };
 	unsigned int ncomp = 13;
-	const double point[] = { 1.0, 2.0, 1.4, 2.1, 0.2, 1.1, 1.8 };
-	cadet::test::reaction::testDynamicJacobianAD("ACTIVATES_SLUDGE_MODEL3", ncomp, nBound,
+	const double point[] = { 1.0, 2.0, 1.4, 2.1, 0.2, 1.1, 1.8, 1.5, 1.0, 4.2, 1.4, 0.3, 1.4};
+	cadet::test::reaction::testDynamicJacobianAD("ACTIVATED_SLUDGE_MODEL3", ncomp, nBound,
 		R"json({
-			"ACSM3_p1": [1.0, 2.0, 0.4],
-			"ACSM3_p2": [-1.0, 1.0, -1.0, -1.0, -1.0, -1.0, 3.0, 2.0, -1.0]
+			"ASM3_FISS_BM_PROD": 1.0,
+			"ASM3_FSI": 0.0,
+			"ASM3_YH_AER": 0.8,
+			"ASM3_YH_ANOX": 0.65,
+			"ASM3_YSTO_AER": 0.8375,
+			"ASM3_YSTO_ANOX": 0.7,
+			"ASM3_FXI": 0.2,
+			"ASM3_YA": 0.24,
+			"ASM3_KH20": 9.0,
+			"ASM3_KX": 1.0,
+			"ASM3_KSTO20": 12.0,
+			"ASM3_MU_H20": 3.0,
+			"ASM3_BH20": 0.33,
+			"ASM3_ETA_HNO3": 0.5,
+			"ASM3_KHO2": 0.2,
+			"ASM3_KHSS": 10.0,
+			"ASM3_KHNO3": 0.5,
+			"ASM3_KHNH4": 0.01,
+			"ASM3_KHALK": 0.1,
+			"ASM3_KHSTO": 0.1,
+			"ASM3_MU_AUT20": 1.12,
+			"ASM3_BAUT20": 0.18,
+			"ASM3_ETAH_END": 0.5,
+			"ASM3_ETAN_END": 0.5,
+			"ASM3_KNO2": 0.5,
[…]
