+++ src/libcadet/model/reaction/ActivatedSludgeModelThree.cpp
@@ -185,54 +185,54 @@ protected:
 		_stoichiometry.setAll(0);
 
 		// parameter set ASM3h
-		double iNSI = paramProvider.getDouble("ASM3_INSI");
-		double iNSS = paramProvider.getDouble("ASM3_INSS");
-		double iNXI = paramProvider.getDouble("ASM3_INXI");
-		double iNXS = paramProvider.getDouble("ASM3_INXS");
-		double iNBM = paramProvider.getDouble("ASM3_INBM");
+		const double iNSI = paramProvider.getDouble("ASM3_INSI");
+		const double iNSS = paramProvider.getDouble("ASM3_INSS");
+		const double iNXI = paramProvider.getDouble("ASM3_INXI");
+		const double iNXS = paramProvider.getDouble("ASM3_INXS");
+		const double iNBM = paramProvider.getDouble("ASM3_INBM");
 
-		double fSI = paramProvider.getDouble("ASM3_FSI");
-		double fXI = paramProvider.getDouble("ASM3_FXI");
+		const double fSI = paramProvider.getDouble("ASM3_FSI");
+		const double fXI = paramProvider.getDouble("ASM3_FXI");
 
-		double YH_aer = paramProvider.getDouble("ASM3_YH_AER");
-		double YH_anox = paramProvider.getDouble("ASM3_YH_ANOX");
-		double YSTO_aer = paramProvider.getDouble("ASM3_YSTO_AER");
-		double YSTO_anox = paramProvider.getDouble("ASM3_YSTO_ANOX");
-		double YA = paramProvider.getDouble("ASM3_YA");
+		const double YH_aer = paramProvider.getDouble("ASM3_YH_AER");
+		const double YH_anox = paramProvider.getDouble("ASM3_YH_ANOX");
+		const double YSTO_aer = paramProvider.getDouble("ASM3_YSTO_AER");
+		const double YSTO_anox = paramProvider.getDouble("ASM3_YSTO_ANOX");
+		const double YA = paramProvider.getDouble("ASM3_YA");
 
-		double fiSS_BM_prod = paramProvider.getDouble("ASM3_FISS_BM_PROD");
-		double iVSS_BM = paramProvider.getDouble("ASM3_IVSS_BM");
-		double iTSS_VSS_BM = paramProvider.getDouble("ASM3_ITSS_VSS_BM");
+		const double fiSS_BM_prod = paramProvider.getDouble("ASM3_FISS_BM_PROD");
+		const double iVSS_BM = paramProvider.getDouble("ASM3_IVSS_BM");
+		const double iTSS_VSS_BM = paramProvider.getDouble("ASM3_ITSS_VSS_BM");
 
 		// internal variables
-		double fXMI_BM = fiSS_BM_prod * fXI * iVSS_BM * (iTSS_VSS_BM - 1);
-		double c1n = iNXS - iNSI * fSI - (1 - fSI) * iNSS;
-		double c2n = iNSS;
-		double c3n = iNSS;
-		double c4n = -iNBM;
-		double c5n = c4n;
-		double c6n = -fXI * iNXI + iNBM;
-		double c7n = c6n;
-		double c10n = -1 / YA - iNBM;
-		double c11n = -fXI * iNXI + iNBM;
-		double c12n = c11n;
-		double c3no = (YSTO_anox - 1.0) / (40.0 / 14.0);
-		double c5no = (1.0 - 1 / YH_anox) / (40.0 / 14.0);
-		double c7no = (fXI - 1) / (40.0 / 14.0);
-		double c9no = -14.0 / 40.0;
-		double c10no = 1 / YA;
-		double c12no = c7no;
-		double c1a = c1n / 14.0;
[…]
