### 2025-05-20 2ce03de
add AD test only for reactions in liquit phase
