+++ test/ReactionModels.cpp
@@ -145,12 +145,12 @@ TEST_CASE("MichaelisMenten kinetic analytic Jacobian vs AD with inhibition", "[M
 		point, 1e-15, 1e-15
 	);
 }
-TEST_CASE("ASM3 analytic Jacobian vs AD", "[ASM3],[ReactionModel],[Jacobian],[AD],[testHere]")
+TEST_CASE("ASM3 analytic Jacobian vs AD", "[ASM3],[ReactionModel],[Jacobian],[AD]")
 {
-	const unsigned int nBound[] = { 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,0.0, 0.0, 0.0 };
+	const unsigned int nBound[] = { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ,0, 0, 0 };
 	unsigned int ncomp = 13;
 	const double point[] = { 1.0, 2.0, 1.4, 2.1, 0.2, 1.1, 1.8, 1.5, 1.0, 4.2, 1.4, 0.3, 1.4};
-	cadet::test::reaction::testDynamicJacobianAD("ACTIVATED_SLUDGE_MODEL3", ncomp, nBound,
+	cadet::test::reaction::testLiquidReactionJacobianAD("ACTIVATED_SLUDGE_MODEL3", ncomp, nBound,
 		R"json({
 			"ASM3_FISS_BM_PROD": 1.0,
 			"ASM3_FSI": 0.0,
