### 2025-05-20 5e1c124
ASM3: group parameter documentation
