+++ doc/interface/reaction/activated_sludge_model.rst
@@ -7,13 +7,8 @@ Activated Sludge Model (ASM3h)
 
 For information on model equations, refer to :ref:`activated_sludge_model`.
 
-``ASM3_KH20``
-
-   Hydrolysis rate constant :math:`k_H` at 20 °C.
-   
-   ================  =============================  ========================================================
-   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
-   ================  =============================  ========================================================
+Environment/process parameters
+----
 
 ``ASM3_T``
 
@@ -21,6 +16,14 @@ For information on model equations, refer to :ref:`activated_sludge_model`.
    
    ================  =============================  ========================================================
    **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
+   ================  =============================  ========================================================
+
+``ASM3_V``
+
+   Reactor volume :math:`V`. Set this to the column volume since the model can't compute it itself.
+
+   ================  =============================  ========================================================
+   **Type:** double  **Range:** :math:`\gt 0`       **Length:** 1
    ================  =============================  ========================================================
 
 ``ASM3_IO2``
@@ -31,12 +34,15 @@ For information on model equations, refer to :ref:`activated_sludge_model`.
    **Type:** double  **Range:** :math:`\ge 0`       **Length:** 1
    ================  =============================  ========================================================
 
-``ASM3_V``
+Maximum rates
+----
 
-   Aeration volume :math:`V`.
+``ASM3_KH20``
 
+   Hydrolysis rate constant :math:`k_H` at 20 °C.
+   
    ================  =============================  ========================================================
-   **Type:** double  **Range:** :math:`\gt 0`       **Length:** 1
+   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
    ================  =============================  ========================================================
 
 ``ASM3_KSTO20``
@@ -47,81 +53,9 @@ For information on model equations, refer to :ref:`activated_sludge_model`.
    **Type:** double  **Range:** :math:`\gt 0`       **Length:** 1
    ================  =============================  ========================================================
 
-``ASM3_KX``
-
-   Saturation/inhibition coefficient for particulate COD :math:`KX`.
-
-   ================  =============================  ========================================================
-   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
-   ================  =============================  ========================================================
-
-``ASM3_KHO2``
-
-   Saturation/inhibition coefficient for oxygen, heterotrophic growth :math:`KHO_2`.
-
-   ================  =============================  ========================================================
-   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
-   ================  =============================  ========================================================
-
-``ASM3_KHSS``
-
-   Substrate :math:`SS` saturation constant :math:`KHSS`.
-
-   ================  =============================  ========================================================
-   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
-   ================  =============================  ========================================================
-
-``ASM3_KHNO3``
-
-   :math:`SNO_X` saturation constant :math:`KHNO_3`.
-
-   ================  =============================  ========================================================
-   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
-   ================  =============================  ========================================================
-
-``ASM3_ETA_HNO3``
-
-   Anoxic reduction factor :math:`\eta_{HNO_3}`.
-
-   ================  =============================  ========================================================
-   **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** 1
-   ================  =============================  ========================================================
-
-``ASM3_KHNH4``
-
-   Ammonium saturation constant :math:`KHNH_4`.
-
[…]
