### 2025-05-21 1c44058
Fix reordering of surface diffusion in GRM_DG unit
