+++ src/libcadet/model/GeneralRateModelDG.hpp
@@ -282,7 +282,6 @@ protected:
 		bool newStaticJac; //!< determines wether static analytical jacobian is to be computed
 
 		// parameter
-		unsigned int* offsetSurfDiff; //!< particle surface diffusion (may be section and component dependent)
 		int curSection; //!< current time section index
 
 		const double SurfVolRatioSlab = 1.0; //!< Surface to volume ratio for a slab-shaped particle
@@ -313,7 +312,7 @@ protected:
 
 		Discretization() : nParCell(nullptr), nParPointsBeforeType(nullptr), parPolyDeg(nullptr), nParNode(nullptr),
 			nParPoints(nullptr), parTypeOffset(nullptr), nBound(nullptr), boundOffset(nullptr),
-			strideBound(nullptr), nBoundBeforeType(nullptr), offsetSurfDiff(nullptr), deltaR(nullptr), parNodes(nullptr),
+			strideBound(nullptr), nBoundBeforeType(nullptr), deltaR(nullptr), parNodes(nullptr),
 			parPolyDerM(nullptr), minus_InvMM_ST(nullptr), parInvWeights(nullptr), parInvMM(nullptr), parInvMM_Leg(nullptr),
 			Ir(nullptr), Dr(nullptr), DGjacParDispBlocks(nullptr), g_p(nullptr), g_pSum(nullptr),
 			surfaceFluxParticle(nullptr), localFlux(nullptr)
@@ -332,8 +331,6 @@ protected:
 			delete[] boundOffset;
 			delete[] strideBound;
 			delete[] nBoundBeforeType;
-
-			delete[] offsetSurfDiff;
 
 			delete[] deltaR;
 			delete[] parNodes;
@@ -1028,7 +1025,7 @@ protected:
 							int const* const qsReaction = _binding[parType]->reactionQuasiStationarity();
 
 							for (unsigned int bnd = 0; bnd < _disc.nBound[parType * _disc.nComp + comp]; bnd++) {
-								if (_parSurfDiff[getOffsetSurfDiff(parType, comp, bnd)] != 0.0) {
+								if (_parSurfDiff[_disc.nBoundBeforeType[parType] + _disc.boundOffset[parType * _disc.nComp + comp] + bnd] != 0.0) {
 									// row: add current component offset and go node strides from there for each dispersion block entry
 									// col: jump oover liquid states, add current bound state offset and go node strides from there for each dispersion block entry
 									tripletList.push_back(T(offset + comp * sComp + i * sNode,
@@ -1073,7 +1070,7 @@ protected:
 							int const* const qsReaction = _binding[parType]->reactionQuasiStationarity();
 
 							for (unsigned int bnd = 0; bnd < _disc.nBound[parType * _disc.nComp + comp]; bnd++) {
-								if (_parSurfDiff[getOffsetSurfDiff(parType, comp, bnd)] != 0.0) {
+								if (_parSurfDiff[_disc.nBoundBeforeType[parType] + _disc.boundOffset[parType * _disc.nComp + comp] + bnd] != 0.0) {
 									// row: add current component offset and go node strides from there for each dispersion block entry
 									// col: jump over liquid states, add current bound state offset and go node strides from there for each dispersion block entry. adjust for j start
 									tripletList.push_back(T(offset + comp * sComp + i * sNode,
@@ -1116,7 +1113,7 @@ protected:
 							int const* const qsReaction = _binding[parType]->reactionQuasiStationarity();
 
 							for (unsigned int bnd = 0; bnd < _disc.nBound[parType * _disc.nComp + comp]; bnd++) {
-								if (_parSurfDiff[getOffsetSurfDiff(parType, comp, bnd)] != 0.0) {
[…]
