+++ src/libcadet/model/GeneralRateModelDG.hpp
@@ -282,7 +282,6 @@ protected:
 		bool newStaticJac; //!< determines wether static analytical jacobian is to be computed
 
 		// parameter
-		unsigned int* offsetSurfDiff; //!< particle surface diffusion (may be section and component dependent)
 		int curSection; //!< current time section index
 
 		const double SurfVolRatioSlab = 1.0; //!< Surface to volume ratio for a slab-shaped particle
@@ -313,7 +312,7 @@ protected:
 
 		Discretization() : nParCell(nullptr), nParPointsBeforeType(nullptr), parPolyDeg(nullptr), nParNode(nullptr),
 			nParPoints(nullptr), parTypeOffset(nullptr), nBound(nullptr), boundOffset(nullptr),
-			strideBound(nullptr), nBoundBeforeType(nullptr), offsetSurfDiff(nullptr), deltaR(nullptr), parNodes(nullptr),
+			strideBound(nullptr), nBoundBeforeType(nullptr), deltaR(nullptr), parNodes(nullptr),
 			parPolyDerM(nullptr), minus_InvMM_ST(nullptr), parInvWeights(nullptr), parInvMM(nullptr), parInvMM_Leg(nullptr),
 			Ir(nullptr), Dr(nullptr), DGjacParDispBlocks(nullptr), g_p(nullptr), g_pSum(nullptr),
 			surfaceFluxParticle(nullptr), localFlux(nullptr)
@@ -332,8 +331,6 @@ protected:
 			delete[] boundOffset;
 			delete[] strideBound;
 			delete[] nBoundBeforeType;
-
-			delete[] offsetSurfDiff;
 
 			delete[] deltaR;
 			delete[] parNodes;
@@ -997,7 +994,7 @@ protected:
 	void calcParticleJacobianPattern(std::vector<T>& tripletList, unsigned int parType, unsigned int colNode, unsigned int secIdx) {
 
 		// Ordering of particle surface diffusion:
-		// bnd0comp0, bnd0comp1, bnd0comp2, bnd1comp0, bnd1comp1, bnd1comp2
+		// bnd0comp0, bnd1comp0, bnd0comp1, bnd1comp1, bnd0comp2, bnd1comp2
 		active const* const _parSurfDiff = getSectionDependentSlice(_parSurfDiffusion, _disc.strideBound[_disc.nParType], secIdx) + _disc.nBoundBeforeType[parType];
 
 		Indexer idxr(_disc);
@@ -1028,7 +1025,7 @@ protected:
 							int const* const qsReaction = _binding[parType]->reactionQuasiStationarity();
 
 							for (unsigned int bnd = 0; bnd < _disc.nBound[parType * _disc.nComp + comp]; bnd++) {
-								if (_parSurfDiff[getOffsetSurfDiff(parType, comp, bnd)] != 0.0) {
+								if (_parSurfDiff[_disc.nBoundBeforeType[parType] + _disc.boundOffset[parType * _disc.nComp + comp] + bnd] != 0.0) {
 									// row: add current component offset and go node strides from there for each dispersion block entry
 									// col: jump oover liquid states, add current bound state offset and go node strides from there for each dispersion block entry
 									tripletList.push_back(T(offset + comp * sComp + i * sNode,
@@ -1073,7 +1070,7 @@ protected:
 							int const* const qsReaction = _binding[parType]->reactionQuasiStationarity();
 
 							for (unsigned int bnd = 0; bnd < _disc.nBound[parType * _disc.nComp + comp]; bnd++) {
-								if (_parSurfDiff[getOffsetSurfDiff(parType, comp, bnd)] != 0.0) {
[…]
