+++ src/libcadet/model/ParameterMultiplexing.cpp
@@ -143,7 +143,7 @@ MultiplexMode readAndRegisterMultiplexCompTypeSecParam(IParameterProvider& param
 
 				values = std::move(p);
 
-				for (std::size_t s = 0; s < values.size() / nComp; ++s)
+				for (std::size_t s = 0; s < values.size() / nComp / nParType; ++s)
 				{
 					for (unsigned int i = 0; i < nComp; ++i)
 						parameters[makeParamId(nameHash, uoi, i, ParTypeIndep, BoundStateIndep, ReactionIndep, s)] = &values[s * nParType * nComp + i];
@@ -206,7 +206,32 @@ bool multiplexCompTypeSecParameterValue(const ParameterId& pId, StringHash nameH
 				return true;
 			}
 		case MultiplexMode::ComponentType:
+		{
+			if ((pId.component == CompIndep) || (pId.particleType == ParTypeIndep) || (pId.boundState != BoundStateIndep)
+				|| (pId.reaction != ReactionIndep) || (pId.section != SectionIndep))
+				return false;
+
+			if (sensParams && !contains(*sensParams, &data[pId.particleType * nComp + pId.component]))
+				return false;
+
+			for (unsigned int s = 0; s < data.size() / nComp / nParType; ++s)
+				data[s * nComp * nParType + pId.particleType * nComp + pId.component].setValue(value);
+
+			return true;
+		}
 		case MultiplexMode::ComponentSectionType:
+		{
+			if ((pId.component == CompIndep) || (pId.particleType == ParTypeIndep) || (pId.boundState != BoundStateIndep)
+				|| (pId.reaction != ReactionIndep) || (pId.section == SectionIndep))
+				return false;
+
+			if (sensParams && !contains(*sensParams, &data[pId.section * nComp * nParType + pId.particleType * nComp + pId.component]))
+				return false;
+
+			data[pId.section * nComp * nParType + pId.particleType * nComp + pId.component].setValue(value);
+
+			return true;
+		}
 		case MultiplexMode::RadialSection:
 		case MultiplexMode::Independent:
 		case MultiplexMode::ComponentRadial:
@@ -231,33 +256,56 @@ bool multiplexCompTypeSecParameterAD(const ParameterId& pId, StringHash nameHash
 	switch (mode)
 	{
 		case MultiplexMode::Component:
-			{
-				if ((pId.component == CompIndep) || (pId.particleType != ParTypeIndep) || (pId.boundState != BoundStateIndep)
-					|| (pId.reaction != ReactionIndep) || (pId.section != SectionIndep))
-					return false;
+		{
+			if ((pId.component == CompIndep) || (pId.particleType != ParTypeIndep) || (pId.boundState != BoundStateIndep)
+				|| (pId.reaction != ReactionIndep) || (pId.section != SectionIndep))
+				return false;
 
-				sensParams.insert(&data[pId.component]);
+			sensParams.insert(&data[pId.component]);
 
-				for (unsigned int i = 0; i < nParType; ++i)
-					data[i * nComp + pId.component].setADValue(adDirection, adValue);
+			for (unsigned int i = 0; i < nParType; ++i)
+				data[i * nComp + pId.component].setADValue(adDirection, adValue);
 
-				return true;
-			}
+			return true;
+		}
 		case MultiplexMode::ComponentSection:
-			{
-				if ((pId.component == CompIndep) || (pId.particleType != ParTypeIndep) || (pId.boundState != BoundStateIndep)
-					|| (pId.reaction != ReactionIndep) || (pId.section == SectionIndep))
-					return false;
+		{
+			if ((pId.component == CompIndep) || (pId.particleType != ParTypeIndep) || (pId.boundState != BoundStateIndep)
[…]
