### 2025-05-21 4800a71
Add tests for multiplexing with sensitivities (#419)
