### 2025-05-21 ed009c3
Add tests for multiplexing with sensitivities (#419)
