+++ src/libcadet/model/parts/ParticleDiffusionOperatorDG.hpp
@@ -88,8 +88,8 @@ namespace parts
 		ParticleDiffusionOperatorDG();
 		~ParticleDiffusionOperatorDG() CADET_NOEXCEPT;
 
-		bool configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, const int nComp, const int nParType, const int strideBulkComp);
-		bool configure(UnitOpIdx unitOpIdx, IParameterProvider& paramProvider, std::unordered_map<ParameterId, active*>& parameters);
+		bool configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, const int nComp, const int parTypeIdx, const int nParType, const int strideBulkComp);
+		bool configure(UnitOpIdx unitOpIdx, IParameterProvider& paramProvider, std::unordered_map<ParameterId, active*>& parameters, const int nParType);
 
 		void setEquidistantRadialDisc(unsigned int parType);
 		void setEquivolumeRadialDisc(unsigned int parType);
@@ -105,19 +105,20 @@ namespace parts
 		int residual(double t, unsigned int secIdx, unsigned int parType, unsigned int colNode, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithParamSensitivity);
 		int residual(double t, unsigned int secIdx, unsigned int parType, unsigned int colNode, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithoutParamSensitivity);
 
+		int _parTypeIdx; //!< Particle type index (wrt the unit operation that owns this particle model)
 
 		/* Physical model parameters */
 
-		std::vector<active> _parRadius; //!< Particle radius \f$ r_p \f$
+		active _parRadius; //!< Particle radius \f$ r_p \f$
 		bool _singleParRadius;
-		std::vector<active> _parCoreRadius; //!< Particle core radius \f$ r_c \f$
+		active _parCoreRadius; //!< Particle core radius \f$ r_c \f$
 		bool _singleParCoreRadius;
-		std::vector<active> _parPorosity; //!< Particle porosity (internal porosity) \f$ \varepsilon_p \f$
+		active _parPorosity; //!< Particle porosity (internal porosity) \f$ \varepsilon_p \f$
 		bool _singleParPorosity;
 		std::vector<active> _invBetaP; //!< Ratio of solid to liquid particle volume
-		std::vector<double> _parGeomSurfToVol; //!< Particle surface to volume ratio factor (i.e., 3.0 for spherical, 2.0 for cylindrical, 1.0 for hexahedral)
-		std::vector<active> _parOuterSurfAreaPerVolume; //!< Particle element outer sphere surface to volume ratio
-		std::vector<active> _parInnerSurfAreaPerVolume; //!< Particle element inner sphere surface to volume ratio
+		double _parGeomSurfToVol; //!< Particle surface to volume ratio factor (i.e., 3.0 for spherical, 2.0 for cylindrical, 1.0 for hexahedral)
+		active _parOuterSurfAreaPerVolume; //!< Particle element outer sphere surface to volume ratio
+		active _parInnerSurfAreaPerVolume; //!< Particle element inner sphere surface to volume ratio
 
 		std::vector<active> _filmDiffusion; //!< Particle diffusion coefficient \f$ D_p \f$
 		//MultiplexMode _filmDiffusionMode;
@@ -127,13 +128,12 @@ namespace parts
 		MultiplexMode _parDiffusionMode;
 		std::vector<active> _parSurfDiffusion; //!< Particle surface diffusion coefficient \f$ D_s \f$
 		MultiplexMode _parSurfDiffusionMode;
-		std::vector<IParameterStateDependence*> _parDepSurfDiffusion; //!< Parameter dependencies for particle surface diffusion
[…]
