+++ src/libcadet/model/parts/ParticleDiffusionOperatorDG.cpp
@@ -40,83 +40,32 @@ namespace parts
 	/**
 	 * @brief Creates a ParticleDiffusionOperatorDG
 	 */
-	ParticleDiffusionOperatorDG::ParticleDiffusionOperatorDG() : _nParElem(nullptr), _nParPointsBeforeType(nullptr),
-		_parPolyDeg(nullptr), _nParNode(nullptr), _nParPoints(nullptr), _parGSM(nullptr), //_parTypeOffset(nullptr),
-		_nBound(nullptr), _boundOffset(nullptr), _strideBound(nullptr), _nBoundBeforeType(nullptr), _deltaR(nullptr), _parNodes(nullptr),
-		_parPolyDerM(nullptr), _minus_InvMM_ST(nullptr), _minus_parInvMM_Ar(nullptr), _parInvWeights(nullptr), _parInvMM(nullptr), _parInvMM_Leg(nullptr),
-		_Ir(nullptr), _secondOrderStiffnessM(nullptr), _DGjacParDispBlocks(nullptr), _g_p(nullptr), _g_pSum(nullptr),
-		_surfaceFluxParticle(nullptr), _localFlux(nullptr)
+	ParticleDiffusionOperatorDG::ParticleDiffusionOperatorDG() : _boundOffset(nullptr), _localFlux(nullptr)
 	{
 	}
 
 	ParticleDiffusionOperatorDG::~ParticleDiffusionOperatorDG() CADET_NOEXCEPT
 	{
-		delete[] _nParElem;
-		delete[] _nParPointsBeforeType;
-		delete[] _parPolyDeg;
-		delete[] _nParNode;
-		delete[] _nParPoints;
-		delete[] _parGSM;
-		//delete[] _parTypeOffset;
-		delete[] _nBound;
-		delete[] _boundOffset;
-		delete[] _strideBound;
-		delete[] _nBoundBeforeType;
-
-		delete[] _deltaR;
-		delete[] _parNodes;
-		delete[] _parPolyDerM;
-		delete[] _minus_InvMM_ST;
-		delete[] _minus_parInvMM_Ar;
-		delete[] _parInvWeights;
-		delete[] _parInvMM;
-		delete[] _parInvMM_Leg;
-		delete[] _Ir;
-		delete[] _secondOrderStiffnessM;
-
 		delete[] _DGjacParDispBlocks;
 
-		delete[] _g_p;
-		delete[] _g_pSum;
-		delete[] _surfaceFluxParticle;
 		delete[] _localFlux;
-
-		clearParDepSurfDiffusion();
 	}
 
-	void ParticleDiffusionOperatorDG::clearParDepSurfDiffusion()
-	{
-		if (_singleParDepSurfDiffusion)
-		{
-			if (!_parDepSurfDiffusion.empty())
-				delete _parDepSurfDiffusion[0];
-		}
-		else
-		{
-			for (IParameterStateDependence* pd : _parDepSurfDiffusion)
-				delete pd;
-		}
-
-		_parDepSurfDiffusion.clear();
-	}
-
-	bool ParticleDiffusionOperatorDG::configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, const int nComp, const int nParType, const int strideBulkComp)
+	bool ParticleDiffusionOperatorDG::configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, const int nComp, const int parTypeIdx, const int nParType, const int strideBulkComp)
 	{
 		_strideBulkComp = strideBulkComp;
 
-		_nParType = nParType;
+		_parTypeIdx;
 		_nComp = nComp;
 
[…]

[…]
