### 2025-05-22 0070c21
Add binding model tests to CI
