### 2025-05-22 7be73be
delete moity member from cstr and make them member of MAL
This way we are more flexible to use conserved moities in for different
units.
