+++ src/libcadet/model/reaction/MassActionLawReaction.cpp
@@ -57,1124 +57,1272 @@
 namespace cadet
 {
 
-namespace model
-{
-
-inline const char* MassActionLawParamHandler::identifier() CADET_NOEXCEPT { return "MASS_ACTION_LAW"; }
-
-inline bool MassActionLawParamHandler::validateConfig(unsigned int nReactions, unsigned int nComp, unsigned int const* nBoundStates)
-{
-	return true;
-}
-
-inline const char* ExtMassActionLawParamHandler::identifier() CADET_NOEXCEPT { return "EXT_MASS_ACTION_LAW"; }
-
-inline bool ExtMassActionLawParamHandler::validateConfig(unsigned int nReactions, unsigned int nComp, unsigned int const* nBoundStates)
-{
-	return true;
-}
-
-
-namespace
-{
-	/**
-	 * @brief Registers a matrix-valued parameter (row-major storage) with bound states as rows
-	 * @details The matrix-valued parameter has as many rows as there are bound states in the system.
-	 * @param [in,out] parameters Parameter map
-	 * @param [in] unitOpIdx Unit operation id
-	 * @param [in] parTypeIdx Particle type index
-	 * @param [in] paramName Name of the parameter
-	 * @param [in] mat Matrix to register
-	 * @param [in] nComp Number of components
-	 * @param [in] boundOffset Array with offsets to bound states of a specific component
-	 */
-	inline void registerBoundStateRowMatrix(std::unordered_map<ParameterId, active*>& parameters, UnitOpIdx unitOpIdx, ParticleTypeIdx parTypeIdx, const std::string& paramName, cadet::linalg::ActiveDenseMatrix& mat,
-		unsigned int nComp, unsigned int const* boundOffset)
+	namespace model
 	{
-		const cadet::StringHash hashName = cadet::hashStringRuntime(paramName);
-		cadet::registerParam2DArray(parameters, mat.data(), mat.elements(), [=](bool multi, unsigned int row, unsigned int col)
-			{
-				const unsigned int comp = std::lower_bound(boundOffset, boundOffset + nComp, row) - boundOffset;
-				const unsigned int bnd = row - boundOffset[comp];
-				return cadet::makeParamId(hashName, unitOpIdx, comp, parTypeIdx, bnd, col, cadet::SectionIndep);
-			},
-			mat.columns()
-		);
-	}
 
-	/**
-	 * @brief Registers a matrix-valued parameter (row-major storage) with components as rows
-	 * @details The matrix-valued parameter has as many rows as there are components in the system.
-	 * @param [in,out] parameters Parameter map
-	 * @param [in] unitOpIdx Unit operation id
-	 * @param [in] parTypeIdx Particle type index
-	 * @param [in] paramName Name of the parameter
-	 * @param [in] mat Matrix to register
-	 */
-	inline void registerCompRowMatrix(std::unordered_map<ParameterId, active*>& parameters, UnitOpIdx unitOpIdx, ParticleTypeIdx parTypeIdx, const std::string& paramName, cadet::linalg::ActiveDenseMatrix& mat)
-	{
-		const cadet::StringHash hashName = cadet::hashStringRuntime(paramName);
-		cadet::registerParam2DArray(parameters, mat.data(), mat.elements(), [=](bool multi, unsigned int row, unsigned int col)
-			{
-				return cadet::makeParamId(hashName, unitOpIdx, row, parTypeIdx, cadet::BoundStateIndep, col, cadet::SectionIndep);
-			},
-			mat.columns()
-		);
-	}
[…]
