+++ src/libcadet/model/StirredTankModel.cpp
@@ -357,25 +357,19 @@ namespace cadet
 
 			// Reconfigure reaction model
 			bool dynReactionConfSuccess = true;
+			bool conservedMoitiesSuccess = true;
 			_hasQuasiStationaryReactionBulk = false;
-			_MconvMoityBulk = Eigen::Matrix<active, Eigen::Dynamic, Eigen::Dynamic>::Zero(0, 0); // matrix for conserved moities
-			_QsCompBulk.clear();
-			_nConservedQuants = 0;
 			if (_dynReactionBulk && _dynReactionBulk->requiresConfiguration())
 			{
 				paramProvider.pushScope("reaction_bulk");
 				dynReactionConfSuccess = _dynReactionBulk->configure(paramProvider, _unitOpIdx, ParTypeIndep);
 				paramProvider.popScope();
 
-				// Get quasi stationary reactions information vector
-				_dynReactionBulk->hasQuasiStationaryReactionsBulk();
 
 				if (_dynReactionBulk->hasQuasiStationaryReactionsBulk())
 				{
-					_QsCompBulk.resize(_nComp);
-					_dynReactionBulk->fillConservedMoietiesBulk(_MconvMoityBulk, _nConservedQuants, _QsCompBulk);
-
-
+					conservedMoitiesSuccess =  _dynReactionBulk->configureConservedMoity();
+					
 					bool hasQSBinding = false;
 					for (int i = 0; i < _nParType; i++)
 					{
@@ -403,7 +397,7 @@ namespace cadet
 				dynReactionConfSuccess = _dynReaction[type]->configure(paramProvider, _unitOpIdx, type) && dynReactionConfSuccess;
 			}
 
-			return bindingConfSuccess && dynReactionConfSuccess;
+			return bindingConfSuccess && dynReactionConfSuccess && conservedMoitiesSuccess;
 		}
 
 		unsigned int CSTRModel::threadLocalMemorySize() const CADET_NOEXCEPT
@@ -654,9 +648,10 @@ namespace cadet
 				BufferedArray<int> qsMask = tlmAlloc.array<int>(_nComp);
 
 				// mark which concentrations need to be reacalculated
-				qsMask.copyFromVector(_QsCompBulk);
+				std::vector<int> qsCompBulk = _dynReactionBulk->quasiStationaryComponentMap();
+				qsMask.copyFromVector(qsCompBulk);
 
-				int nActiveStates = std::count(_QsCompBulk.begin(), _QsCompBulk.end(), 1);
+				int nActiveStates = std::count(qsCompBulk.begin(), qsCompBulk.end(), 1);
 
 				const linalg::ConstMaskArray mask{ static_cast<int*>(qsMask), static_cast<int>(_nComp) };
 				const int probSize = linalg::numMaskActive(mask);
@@ -667,20 +662,20 @@ namespace cadet
 
 				// Save values of conserved moieties;
 				const unsigned int numActiveComp = numMaskActive(mask, _nComp);
-				BufferedArray<double> conservedQuants = tlmAlloc.array<double>(nActiveStates); // number of conserved quantities
+				std::unique_ptr<double[]> conservedQuants = std::make_unique<double[]>(numActiveComp);
 
 				// Calculate conserved quantities for the inital state 
-
+				int numConsQuants = _dynReactionBulk->numConservedMoities();
 				int mIdx = 0;
 				for (unsigned int i = 0; i < _nComp; ++i)
 				{
-					if (_QsCompBulk[i] == 0)
+					if (qsCompBulk[i] == 0)
 						continue;
-					if (mIdx >= _nConservedQuants)
[…]
