### 2025-05-22 f5e3286
Fix 953597e4: surface diffusion reordering
