+++ src/libcadet/model/GeneralRateModelDG.cpp
@@ -1436,11 +1436,11 @@ int GeneralRateModelDG::residualParticle(double t, unsigned int parType, unsigne
 
 			for (int bnd = 0; bnd < _disc.nBound[parType * _disc.nComp + comp]; bnd++)
 			{
-				if (parSurfDiff[_disc.nBoundBeforeType[parType] + _disc.boundOffset[parType * _disc.nComp + comp] + bnd] != 0.0) // some bound states might still not be effected by surface diffusion
+				if (parSurfDiff[_disc.boundOffset[parType * _disc.nComp + comp] + bnd] != 0.0) // some bound states might still not be effected by surface diffusion
 				{
 					Eigen::Map<const Vector<StateType, Dynamic>, 0, InnerStride<Dynamic>> c_s(c_p + _disc.nComp + idxr.offsetBoundComp(ParticleTypeIndex{ parType }, ComponentIndex{ comp }) + bnd, nPoints, InnerStride<Dynamic>(idxr.strideParNode(parType)));
 					ParamType invBetaP = (1.0 - static_cast<ParamType>(_parPorosity[parType])) / (static_cast<ParamType>(_poreAccessFactor[_disc.nComp * parType + comp]) * static_cast<ParamType>(_parPorosity[parType]));
-					sum_cp_cs += invBetaP * static_cast<ParamType>(parSurfDiff[_disc.nBoundBeforeType[parType] + _disc.boundOffset[parType * _disc.nComp + comp] + bnd]) * c_s;
+					sum_cp_cs += invBetaP * static_cast<ParamType>(parSurfDiff[_disc.boundOffset[parType * _disc.nComp + comp] + bnd]) * c_s;
 
 					/* For kinetic bindings with surface diffusion: add the additional DG-discretized particle mass balance equations to residual */
 
@@ -1455,7 +1455,7 @@ int GeneralRateModelDG::residualParticle(double t, unsigned int parType, unsigne
 
 						// Apply squared inverse mapping and surface diffusion
 						c_s_modified = 2.0 / static_cast<ParamType>(_disc.deltaR[_disc.offsetMetric[parType]]) * 2.0 / static_cast<ParamType>(_disc.deltaR[_disc.offsetMetric[parType]]) *
-							static_cast<ParamType>(parSurfDiff[_disc.nBoundBeforeType[parType] + _disc.boundOffset[parType * _disc.nComp + comp] + bnd]) * c_s;
+							static_cast<ParamType>(parSurfDiff[_disc.boundOffset[parType * _disc.nComp + comp] + bnd]) * c_s;
 
 						Eigen::Map<const Vector<ResidualType, Dynamic>, 0, InnerStride<Dynamic>> c_s_modified_const(&c_s_modified[0], nPoints, InnerStride<Dynamic>(1));
 						parGSMVolumeIntegral<ResidualType, ResidualType>(parType, c_s_modified_const, resCs);
@@ -1502,7 +1502,7 @@ int GeneralRateModelDG::residualParticle(double t, unsigned int parType, unsigne
 			{
 				for (int bnd = 0; bnd < _disc.nBound[parType * _disc.nComp + comp]; bnd++)
 				{
-					if (parSurfDiff[_disc.nBoundBeforeType[parType] + _disc.boundOffset[parType * _disc.nComp + comp] + bnd] != 0.0) // some bound states might still not be effected by surface diffusion
+					if (parSurfDiff[_disc.boundOffset[parType * _disc.nComp + comp] + bnd] != 0.0) // some bound states might still not be effected by surface diffusion
 					{
 						// Get solid phase vector
[…]
