### 2025-05-22 3b2f47a
Fix 953597e4: surface diffusion reordering
