### 2025-05-22 928538d
CI: move crystallization and sensitivity tests to own testrunner call
jo
