+++ .github/workflows/ci.yml
@@ -163,18 +163,50 @@ jobs:
           ${INSTALL_PREFIX}/bin/cadet-cli --version || true
           ${INSTALL_PREFIX}/bin/createLWE
           ${INSTALL_PREFIX}/bin/cadet-cli LWE.h5 || true
-      - name: Run CI test set I
-        run: |
-          ${BUILD_DIR}/test/testRunner [CI]
-      - name: Run CI test set II
+      - name: Run CI test set I - bindings
         run: |
           ${BUILD_DIR}/test/testRunner [CI_binding]
-      - name: Run CI test set III
+      - name: Run CI test set II - sensitivities
         run: |
-          ${BUILD_DIR}/test/testRunner [CI_sensitivity1]
+          ${BUILD_DIR}/test/testRunner [CI_sens1]
+          ${BUILD_DIR}/test/testRunner [CI_sens2]
+          ${BUILD_DIR}/test/testRunner [CI_sens3]
+          ${BUILD_DIR}/test/testRunner [CI_sens4]
+          ${BUILD_DIR}/test/testRunner [CI_sens5]
+          ${BUILD_DIR}/test/testRunner [CI_sens6]
+          ${BUILD_DIR}/test/testRunner [CI_sens7]
+          ${BUILD_DIR}/test/testRunner [CI_sens8]
+          ${BUILD_DIR}/test/testRunner [CI_sens9]
+          ${BUILD_DIR}/test/testRunner [CI_sens11]
+          ${BUILD_DIR}/test/testRunner [CI_sens12]
+          ${BUILD_DIR}/test/testRunner [CI_sens13]
+          ${BUILD_DIR}/test/testRunner [CI_sens14]
+          ${BUILD_DIR}/test/testRunner [CI_sens15]
+      - name: Run CI test set III - crystallization
+        run: |
+          ${BUILD_DIR}/test/testRunner [CI_cry1]
+          ${BUILD_DIR}/test/testRunner [CI_cry2]
+          ${BUILD_DIR}/test/testRunner [CI_cry3]
+          ${BUILD_DIR}/test/testRunner [CI_cry4]
+          ${BUILD_DIR}/test/testRunner [CI_cry5]
+          ${BUILD_DIR}/test/testRunner [CI_cry6]
+          ${BUILD_DIR}/test/testRunner [CI_cry7]
+          ${BUILD_DIR}/test/testRunner [CI_cry8]
+          ${BUILD_DIR}/test/testRunner [CI_cry9]
+          ${BUILD_DIR}/test/testRunner [CI_cry10]
+          ${BUILD_DIR}/test/testRunner [CI_cry11]
+          ${BUILD_DIR}/test/testRunner [CI_cry12]
+          ${BUILD_DIR}/test/testRunner [CI_cry13]
+          ${BUILD_DIR}/test/testRunner [CI_cry14]
+          ${BUILD_DIR}/test/testRunner [CI_cry15]
+          ${BUILD_DIR}/test/testRunner [CI_cry16]
+          ${BUILD_DIR}/test/testRunner [CI_cry17]
+          ${BUILD_DIR}/test/testRunner [CI_cry18]
+          ${BUILD_DIR}/test/testRunner [CI_cry19]
+          ${BUILD_DIR}/test/testRunner [CI_cry20]
       - name: Run CI test set IV
         run: |
-          ${BUILD_DIR}/test/testRunner [CI_sensitivity2]
+          ${BUILD_DIR}/test/testRunner [CI]
   MacOS:
     runs-on: macos-latest
     strategy:
