### 2025-05-22 37c75ee
Docs: add note for MacOS ARM64 users to build from source
