### 2025-05-22 5c4fc79
Docs: add note for MacOS ARM64 users to build from source (#420)
