### 2025-05-22 #7881 @lieres
Thank you very much for your thoughtful review and the constructive suggestions.

We have addressed the suggested revision in https://github.com/cadet/CADET-Core/pull/420 by adding an emphasized note to the top-level CADET-Core Installation documentation.

As a further update, we expect the next release to be available on conda-forge with native macOS ARM64 support.
To that end, we have added support for cross-compilation: 1) https://github.com/conda-forge/cadet-feedstock/pull/29 2) https://github.com/cadet/CADET-Core/commit/b55349ee91e9062da42d6d3b7de61b1df92fa2c8

Thanks again for your helpful feedback!
