+++ doc/modelling/binding/index.rst
@@ -207,6 +207,11 @@ The models also differ in whether a mobile phase modifier (e.g., salt) is suppor
      - ×
      - ✓
      - ✓
+   * - :ref:`affinity_complex_titration`
+     - ✓
+     - ×
+     - ✓
+     - ×
    * - :ref:`multi_component_bi_langmuir_ldf_model`
      - ✓
      - ×
