### 2025-05-23 50bc670
use index iteration
