+++ src/libcadet/model/reaction/MassActionLawReaction.cpp
@@ -437,7 +437,7 @@ namespace cadet
 
 			auto configureConservedMoity() -> bool override
 			{
-				CalcConservedMoietiesBulk();
+				ConservedMoietiesBulk();
 				return true;
 			}
 			auto matrixMoietiesBulk() -> Eigen::Matrix<active, Eigen::Dynamic, Eigen::Dynamic> override
@@ -448,6 +448,16 @@ namespace cadet
 			auto quasiStationaryComponentMap() const -> std::vector<int> override
 			{
 				return _quasiStationaryComponentBulkMap;
+			}
+			
+			auto algIdx() -> std::vector<int> override
+			{
+				return _algIdxBulk;
+			}
+
+			auto consMoityIdx() -> std::vector<int> override
+			{
+				return _consMoityIdxBulk;
 			}
 
 
@@ -476,10 +486,10 @@ namespace cadet
 			//todo add these to workspace
 			std::vector<int> _reactionQuasiStationarity;
 			std::vector<int> _quasiStationaryReactionIndices;
-			std::vector<int> _quasiStationaryComponentBulkIndices; //!< Indices of components that are conserved in the bulk volume
+			std::vector<int> _algIdxBulk; //!< Indices of components that are conserved in the bulk volume
+			std::vector<int> _consMoityIdxBulk; //!< Indices of components that are conserved in the bulk volume
 			std::vector<int> _quasiStationaryComponentBulkMap;
-
-
+			
 			Eigen::Matrix<active, Eigen::Dynamic, Eigen::Dynamic> _MconvMoityBulk; //!<  Matrix with conservation of moieties in the bulk volume
 			unsigned int _nConservedQuants;
 
@@ -1082,122 +1092,7 @@ namespace cadet
 			 * @param [in,out] componentsInQssReactions Vector marking components that participate in quasi-stationary reactions (1 if active, 0 otherwise)
 			 * @tparam ResidualType Type used for the residual calculations
 			 */
-			template<typename ResidualType>
-			void ConservedMoietiesBulk(Eigen::Matrix<ResidualType, Eigen::Dynamic, Eigen::Dynamic>& conservationMatrix, int& numConservationLaws, std::vector<int>& componentsInQssReactions)
-			{
-				// Count the number of quasi-stationary reactions
-				int numQuasiStationaryReactions = std::count(_reactionQuasiStationarity.begin(), _reactionQuasiStationarity.end(), true);
-				std::fill(componentsInQssReactions.begin(), componentsInQssReactions.end(), 0);
-
-				if (numQuasiStationaryReactions == 0)
-				{
-					numConservationLaws = 0;
-					return;
-				}
-				// Create quasiStationaryStoichiometry matrix with stoichiometry of quasi-stationary reactions
-				Eigen::MatrixXd quasiStationaryStoichiometry(_stoichiometryBulk.rows(), numQuasiStationaryReactions);
-
-				int reactionIndex = 0;
-				for (int j = 0; j < _stoichiometryBulk.columns(); j++)
-				{
-					if (_reactionQuasiStationarity[j] == 0)
-						continue;
-
-					for (int i = 0; i < _stoichiometryBulk.rows(); ++i)
-					{
-						quasiStationaryStoichiometry(i, reactionIndex) = static_cast<double>(_stoichiometryBulk.native(i, j));
-					}
-					reactionIndex++;
-				}
-
-				// Count active components and mark them in componentsInQssReactions
-				std::vector<int> quasiStationaryCompIdx;
-				quasiStationaryCompIdx.reserve(_nComp);
-
-				std::vector<int> notQuasiStationaryCompIdx;
-				notQuasiStationaryCompIdx.reserve(_nComp);
-
-				if (componentsInQssReactions.size() != _nComp)
-				{
[…]
