+++ src/libcadet/model/StirredTankModel.cpp
@@ -667,12 +667,8 @@ namespace cadet
 				// Calculate conserved quantities for the inital state 
 				int numConsQuants = _dynReactionBulk->numConservedMoities();
 				int mIdx = 0;
-				for (unsigned int i = 0; i < _nComp; ++i)
+				for (int i: _dynReactionBulk->consMoityIdx())
 				{
-					if (qsCompBulk[i] == 0)
-						continue;
-					if (mIdx >= numConsQuants)
-						continue;
 					double dotProduct = 0.0;
 					for (unsigned int j = 0; j < _dynReactionBulk->matrixMoietiesBulk().cols(); ++j)
 						dotProduct += static_cast<double>(_dynReactionBulk->matrixMoietiesBulk()(i, j)) * c[j];
@@ -736,13 +732,8 @@ namespace cadet
 
 							// Replace upper part with conservation relations
 							int  mIdx = 0;
-							int numConsQuants = _dynReactionBulk->numConservedMoities();
-							for (unsigned int i = 0; i < _nComp; ++i)
+							for (int i: _dynReactionBulk->consMoityIdx())
 							{
-								if (qsCompBulk[i] == 0)
-									continue;
-								if (mIdx >= numConsQuants)
-									continue;
 
 								int jIdx = 0;
 								for (unsigned int j = 0; j < _dynReactionBulk->matrixMoietiesBulk().cols(); ++j)
@@ -775,20 +766,13 @@ namespace cadet
 							linalg::copyMatrixSubset(_jac, mask, mask, mat);
 							// Replace upper part with conservation relations
 							int  mIdx = 0;
-							int numConsQuants = _dynReactionBulk->numConservedMoities();
-							for (unsigned int i = 0; i < _nComp; ++i)
+							for (int i : _dynReactionBulk->consMoityIdx())
 							{
-								if (qsCompBulk[i] == 0)
-									continue;
-								if (mIdx >= numConsQuants)
-									continue;
-
 								int jIdx = 0;
 								for (unsigned int j = 0; j < _dynReactionBulk->matrixMoietiesBulk().cols(); ++j)
 								{
-									if (qsCompBulk[j] == 0)
+									if (_dynReactionBulk->quasiStationaryComponentMap()[j] == 0)
 										continue;
-
 									mat.native(mIdx, jIdx) = static_cast<double>(_dynReactionBulk->matrixMoietiesBulk()(i, j));
 									jIdx++;
 								}
@@ -812,19 +796,14 @@ namespace cadet
 						// Extract values from residual
 						linalg::selectVectorSubset(fullResidual + _nComp, mask, r);
 
-						int numConsQuants = _dynReactionBulk->numConservedMoities();
 						int  mIdx = 0;
-						for (unsigned int i = 0; i < _nComp; ++i)
+						for (int i : _dynReactionBulk->consMoityIdx())
 						{
-							if (qsCompBulk[i] == 0) // 
-								continue;
-							if (mIdx >= numConsQuants)
-								continue;
 							int jIdx = 0;
 							double dotProduct = 0.0;
 							for (unsigned int j = 0; j < _dynReactionBulk->matrixMoietiesBulk().cols(); ++j)
 							{
-								if (qsCompBulk[j] == 0)
+								if (_dynReactionBulk->quasiStationaryComponentMap()[j] == 0)
 									continue;
 								dotProduct += static_cast<double>(_dynReactionBulk->matrixMoietiesBulk()(i, j)) * x[jIdx];
[…]
