### 2025-05-23 55d20f2
Crystallization test: reduce discretization in Jacobian AD test to fit the max AD directions
