### 2025-05-23 271e774
Increase number of maximum allowed AD directions in CI and default testing settings
