### 2025-05-23 #7881 @bdice
Thanks for the updates here -- I will do a review pass soon!
