### 2025-05-25 cadet/CADET-Core#422 @bdice
JOSS review - bdice
Hi! Thank you for your patience on my JOSS review for CADET-core (https://github.com/openjournals/joss-reviews/issues/7881).

I have reviewed the software and found everything in tip-top shape: docs, licenses, contributing guides, installation instructions, and the JupyterHub server you shared with me containing a tutorial all worked as expected. I was able to reproduce the results of https://github.com/cadet/CADET-Workshop on my local machine as well (running Ubuntu).

At first I had some confusion around the same issues listed in #373 - but I think some of the revisions that have already been made have helped in this area. The discussion from @lieres here (https://github.com/openjournals/joss-reviews/issues/7881#issuecomment-2770251498) was very helpful for me as a reviewer.

My only significant request for improvements would be here:

> **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

I have enough knowledge of chromatography to understand the scope and purpose of CADET, but I don't know much about other software in this field. It seems like other software I could find in this area is commercial / non-OSS, but it may be worth mentioning to give the reader a better understanding of the current landscape.

Let's discuss on this issue - feel free to make some changes in the paper, or explain if you think no changes are warranted, and then I can approve the JOSS review.
