### 2025-05-26 b205d13
use index iteration for hess matrix
