### 2025-05-26 aa087f3
Refactor particle operator to be particle type specific (#423)
