### 2025-05-27 61a0f94
Refactor particle operator to be particle type specific (#423)
