### 2025-05-27 160043c
Refactor particle operator to be particle type specific (#423)
