+++ src/libcadet/model/ParameterMultiplexing.hpp
@@ -277,6 +277,228 @@ namespace model
 	bool multiplexBndCompTypeSecParameterAD(const ParameterId& pId, StringHash nameHash, MultiplexMode mode, std::vector<active>& data,
 		unsigned int nParType, unsigned int nComp, unsigned int const* strideBound, unsigned int const* nBound, unsigned int const* boundOffset, unsigned int const* nBoundBeforeType, unsigned int adDirection, double adValue, std::unordered_set<active*>& sensParams);
 
+
+/* For particle type specific particle models: only read/register one particle type */
+
+	/**
+	 * @brief For a specific particle type: Reads, multiplexes, and registers a parameter that may depend on particle type
+	 * @details Reads the fiels NAME from the IParameterProvider. The multiplexing behavior is inferred from
+	 *          the length of the field NAME.
+	 *
+	 *          Depending on the (inferred) mode, the read parameter values are multiplexed to a size of nParType.
+	 *          In case of multiplexing, only the first instance of the parameter is registered.
+	 *
+	 * @param [in] paramProvider ParameterProvider
+	 * @param [in] parameters Map to register the parameters in
+	 * @param [out] values Array to store the read parameters in
+	 * @param [in] name Name of the parameter
+	 * @param [in] nParType Number of particle types
+	 * @param [in] parTypeIndex index of the considered particle type
+	 * @param [in] uoi Unit operation index
+	 * @return @c true if the parameter is multiplexed (single value), otherwise @c false
+	 */
+	bool readAndRegisterSingleTypeMultiplexTypeParam(IParameterProvider& paramProvider, std::unordered_map<ParameterId, active*>& parameters, std::vector<active>& values, const std::string& name, unsigned int nParType, unsigned int parTypeIdx, UnitOpIdx uoi);
+
+	/**
+	 * @brief For a specific particle type: Sets the value of a multiplexed parameter that may depend on particle type
+	 * @details Sets the value of a parameter and multiplexes the value onto all parameter instances.
+	 *          The parameter array is expected to have a size of nParType.
+	 *
+	 *          In case of multiplexing, the given value is applied to all parameter instances.
+	 *
+	 *          This function optionally checks whether the specified parameter is listed as sensitive. It uses the first
+	 *          parameter group / instance to check the @p sensParams set.
+	 *
+	 * @param [in] pId ParameterID
+	 * @param [in] nameHash Hash of the parameter name
+	 * @param [in] mode Multiplexing mode as obtained by readAndRegisterMultiplexTypeParam()
+	 * @param [in,out] data Array with parameters whose values are updated
+	 * @param [in] parTypeIdx index of the current particle type
+	 * @param [in] val Value to apply to the parameter(s)
+	 * @param [in] sensParams If not @c nullptr, the set is checked for the specified parameter.
+	 *                        If it is not contained in the set, the value is not applied to the parameter.
+	 * @return @c true if the value has been applied, or @c false otherwise
+	 */
+	bool singleTypeMultiplexTypeParameterValue(const ParameterId& pId, StringHash nameHash, bool mode, active& data, unsigned int parTypeIdx, double value, std::unordered_set<active*> const* sensParams);
+
+	/**
+	 * @brief For a specific particle type: Sets AD info of a multiplexed parameter that may depend on particle type
+	 * @details Sets the AD direction and seed value of a parameter and multiplexes the info onto all parameter instances.
+	 *          The parameter array is expected to have a size of nParType.
+	 *
+	 *          In case of multiplexing, the given AD info is applied to all parameter instances.
+	 *
+	 *          The first parameter group / instance is added to the @p sensParams set in order to mark the parameter as sensitive.
+	 *
+	 * @param [in] pId ParameterID
[…]
