+++ src/libcadet/model/GeneralRateModelDG.cpp
@@ -167,7 +167,12 @@ bool GeneralRateModelDG::configureModelDiscretization(IParameterProvider& paramP
 
 	paramProvider.popScope();
 	Indexer idxr(_disc);
-	const bool particleConfSuccess = _parDiffOp.configureModelDiscretization(paramProvider, helper, _disc.nComp, _disc.nParType, idxr.strideColComp());
+	_parDiffOp = new parts::ParticleDiffusionOperatorDG[_disc.nParType];
+	bool particleConfSuccess = true;
+	for (int parType = 0; parType < _disc.nParType; parType++)
+	{
+		particleConfSuccess = particleConfSuccess && _parDiffOp[parType].configureModelDiscretization(paramProvider, helper, _disc.nComp, parType, _disc.nParType, idxr.strideColComp());
+	}
 	paramProvider.pushScope("discretization");
 
 	if (firstConfigCall)
@@ -175,7 +180,7 @@ bool GeneralRateModelDG::configureModelDiscretization(IParameterProvider& paramP
 		_disc.nParPoints = new unsigned int[_disc.nParType];
 		for (int type = 0; type < _disc.nParType; type++)
 		{
-			_disc.nParPoints[type] = _parDiffOp._nParPoints[type];
+			_disc.nParPoints[type] = _parDiffOp[type]._nParPoints;
 		}
 	}
 
@@ -380,8 +385,6 @@ bool GeneralRateModelDG::configure(IParameterProvider& paramProvider)
 {
 	_parameters.clear();
 
-	const bool firstConfigCall = _parDiffOp._deltaR == nullptr; // used to not multiply allocate memory
-
 	const bool transportSuccess = _convDispOp.configure(_unitOpIdx, paramProvider, _parameters);
 
 	// Read geometry parameters handled by unit operation
@@ -547,8 +550,11 @@ bool GeneralRateModelDG::configure(IParameterProvider& paramProvider)
 		}
 	}
 
-	const bool particleConfSuccess = _parDiffOp.configure(_unitOpIdx, paramProvider, _parameters);
-
+	bool particleConfSuccess = true;
+	for (int parType = 0; parType < _disc.nParType; parType++)
+	{
+		particleConfSuccess = particleConfSuccess && _parDiffOp[parType].configure(_unitOpIdx, paramProvider, _parameters, _disc.nParType, _disc.nBoundBeforeType, _disc.strideBound[_disc.nParType]);
+	}
 	// jaobian pattern set after binding and particle surface diffusion are configured
 	setJacobianPattern_GRM(_globalJac, 0, _dynReactionBulk);
 	_globalJacDisc = _globalJac;
@@ -642,7 +648,10 @@ void GeneralRateModelDG::notifyDiscontinuousSectionTransition(double t, unsigned
 
 	_convDispOp.notifyDiscontinuousSectionTransition(t, secIdx, _jacInlet);
 
-	_parDiffOp.notifyDiscontinuousSectionTransition(t, secIdx, getSectionDependentSlice(_filmDiffusion, _disc.nComp * _disc.nParType, secIdx), &_poreAccessFactor[0]);
+	for (int parType = 0; parType < _disc.nParType; parType++)
+	{
+		_parDiffOp[parType].notifyDiscontinuousSectionTransition(t, secIdx, getSectionDependentSlice(_filmDiffusion, _disc.nComp * _disc.nParType, secIdx), &_poreAccessFactor[0]);
+	}
 
 	_disc.curSection = secIdx;
 	_disc.newStaticJac = true;
@@ -771,7 +780,7 @@ void GeneralRateModelDG::extractJacobianFromAD(active const* const adRes, unsign
[…]
