### 2025-05-27 649e4fb
Refactor particle operator to be particle type specific (#423)
