### 2025-05-27 22fe2fa
Refactor particle operator to be particle type specific (#423)
