+++ src/libcadet/model/ParameterMultiplexing.cpp
@@ -611,6 +611,635 @@ bool multiplexBndCompTypeSecParameterAD(const ParameterId& pId, StringHash nameH
 	return false;
 }
 
+
+/* For particle type specific particle models: only read/register one particle type */
+
+bool readAndRegisterSingleTypeMultiplexTypeParam(IParameterProvider& paramProvider, std::unordered_map<ParameterId, active*>& parameters, active value, const std::string& name, unsigned int nParType, unsigned int parTypeIdx, UnitOpIdx uoi)
+{
+	std::vector<active> tmpVals;
+	const bool singleValue = readScalarParameterOrArray(tmpVals, paramProvider, name, nParType);
+	const StringHash nameHash = hashStringRuntime(name);
+
+	if (singleValue)
+	{
+		value = tmpVals[0];
+		if(parTypeIdx == 0) // only register once, for first parType
+			parameters[makeParamId(nameHash, uoi, CompIndep, ParTypeIndep, BoundStateIndep, ReactionIndep, SectionIndep)] = &value;
+	}
+	else
+	{
+		value = tmpVals[parTypeIdx];
+		parameters[makeParamId(nameHash, uoi, CompIndep, parTypeIdx, BoundStateIndep, ReactionIndep, SectionIndep)] = &value;
+	}
+
+	return singleValue;
+}
+
+bool singleTypeMultiplexTypeParameterValue(const ParameterId& pId, StringHash nameHash, bool mode, active& data, unsigned int parTypeIdx, double value, std::unordered_set<active*> const* sensParams)
+{
+	if (!mode || (pId.name != nameHash) || (pId.particleType != ParTypeIndep && pId.particleType != parTypeIdx))
+		return false;
+
+	if ((pId.section != SectionIndep) || (pId.component != CompIndep) || (pId.boundState != BoundStateIndep) || (pId.reaction != ReactionIndep) || (pId.particleType != ParTypeIndep))
+		return false;
+	if (sensParams && !contains(*sensParams, &data))
+		return false;
+
+	data.setValue(value);
+
+	return true;
+}
+
+bool singleTypeMultiplexTypeParameterAD(const ParameterId& pId, StringHash nameHash, bool mode, active& data, unsigned int parTypeIdx, unsigned int adDirection, double adValue, std::unordered_set<active*>& sensParams)
+{
+	if (!mode || (pId.name != nameHash) || (pId.particleType != ParTypeIndep && pId.particleType != parTypeIdx))
+		return false;
+
+	if ((pId.section != SectionIndep) || (pId.component != CompIndep) || (pId.boundState != BoundStateIndep) || (pId.reaction != ReactionIndep) || (pId.particleType != ParTypeIndep))
+		return false;
+
+	sensParams.insert(&data);
+	data.setADValue(adDirection, adValue);
+
+	return true;
+}
+
+
+MultiplexMode readAndRegisterSingleTypeMultiplexCompTypeSecParam(IParameterProvider& paramProvider, std::unordered_map<ParameterId, active*>& parameters, std::vector<active>& values, const std::string& name, unsigned int nParType, unsigned int nComp, unsigned int parTypeIdx, UnitOpIdx uoi)
+{
+	MultiplexMode mode = MultiplexMode::Independent;
+	readScalarParameterOrArray(values, paramProvider, name, 1);
+	if (paramProvider.exists(name + "_MULTIPLEX"))
+	{
+		const int modeConfig = paramProvider.getInt(name + "_MULTIPLEX");
+		if (modeConfig == 0)
+		{
+			mode = MultiplexMode::Component;
+			if (values.size() != nComp)
[…]
