### 2025-05-27 cca940d
Add binding and reactions to particle operator
