+++ src/libcadet/model/parts/ParticleDiffusionOperatorDG.hpp
@@ -18,6 +18,7 @@
 #ifndef LIBCADET_PARTCICLEDIFFUSIONOPERATORDG_HPP_
 #define LIBCADET_PARTCICLEDIFFUSIONOPERATORDG_HPP_
 
+#include "model/BindingModel.hpp"
 #include "cadet/StrongTypes.hpp"
 #include "ParamIdUtil.hpp"
 #include "AutoDiff.hpp"
@@ -41,15 +42,12 @@ using namespace Eigen;
 namespace cadet
 {
 
-class IParameterProvider;
 class IConfigHelper;
-struct AdJacobianParams;
-struct SimulationTime;
-class IModel;
 
 namespace model
 {
 
+class IDynamicReactionModel;
 class IParameterStateDependence;
 
 namespace parts
@@ -96,14 +94,16 @@ namespace parts
 		void setUserdefinedRadialDisc();
 		void updateRadialDisc();
 
+		//parts::cell::CellParameters makeCellResidualParams(int const* qsReaction) const;
+
 		void clearParDepSurfDiffusion();
 
 		bool notifyDiscontinuousSectionTransition(double t, unsigned int secIdx, active const* const filmDiff, active const* const poreAccessFactor);
 
-		int residual(double t, unsigned int secIdx, unsigned int colNode, double const* yPar, double const* yBulk, double const* yDotPar, double* resPar, int const* const qsBinding, WithoutParamSensitivity);
-		int residual(double t, unsigned int secIdx, unsigned int colNode, double const* yPar, double const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithParamSensitivity);
-		int residual(double t, unsigned int secIdx, unsigned int colNode, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithParamSensitivity);
-		int residual(double t, unsigned int secIdx, unsigned int colNode, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithoutParamSensitivity);
+		int residual(double t, unsigned int secIdx, double const* yPar, double const* yBulk, double const* yDotPar, double* resPar, ColumnPosition colPos, const bool wantRes, linalg::BandedEigenSparseRowIterator& jacIt, WithoutParamSensitivity);
+		int residual(double t, unsigned int secIdx, double const* yPar, double const* yBulk, double const* yDotPar, active* resPar, ColumnPosition colPos, const bool wantRes, linalg::BandedEigenSparseRowIterator& jacIt, WithParamSensitivity);
+		int residual(double t, unsigned int secIdx, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, ColumnPosition colPos, const bool wantRes, linalg::BandedEigenSparseRowIterator& jacIt, WithParamSensitivity);
+		int residual(double t, unsigned int secIdx, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, ColumnPosition colPos, const bool wantRes, linalg::BandedEigenSparseRowIterator& jacIt, WithoutParamSensitivity);
 
 		unsigned int _parTypeIdx; //!< Particle type index (wrt the unit operation that owns this particle model)
 
@@ -134,6 +134,11 @@ namespace parts
 		bool _hasSurfaceDiffusion; //!< Determines whether surface diffusion is present
 
 		unsigned int _nComp; //!< Number of components
+
+		IBindingModel* _binding; //!< Binding Model
+		bool _singleBinding; //!< Determines whether only a single binding model is present (still only one per parType is supported)
[…]
