+++ src/libcadet/model/parts/ParticleDiffusionOperatorDG.cpp
@@ -13,12 +13,17 @@
 #include "model/parts/ParticleDiffusionOperatorDG.hpp"
 #include "cadet/Exceptions.hpp"
 
+#include "BindingModelFactory.hpp"
+#include "ReactionModelFactory.hpp"
 #include "ParamReaderHelper.hpp"
+#include "ParamReaderScopes.hpp"
 #include "AdUtils.hpp"
 #include "SimulationTypes.hpp"
 #include "model/ParameterDependence.hpp"
 #include "SensParamUtil.hpp"
 #include "ConfigurationHelper.hpp"
+#include "model/BindingModel.hpp"
+#include "model/ReactionModel.hpp"
 
 #include "LoggingUtils.hpp"
 #include "Logging.hpp"
@@ -40,12 +45,13 @@ namespace parts
 	/**
 	 * @brief Creates a ParticleDiffusionOperatorDG
 	 */
-	ParticleDiffusionOperatorDG::ParticleDiffusionOperatorDG() : _boundOffset(nullptr), _nBound(nullptr), _localFlux(nullptr), _deltaR(nullptr), _Ir(nullptr), _DGjacParDispBlocks(nullptr), _minus_InvMM_ST(nullptr), _parInvMM(nullptr), _parDepSurfDiffusion(nullptr)
+	ParticleDiffusionOperatorDG::ParticleDiffusionOperatorDG() : _binding(nullptr), _boundOffset(nullptr), _nBound(nullptr), _localFlux(nullptr), _deltaR(nullptr), _Ir(nullptr), _DGjacParDispBlocks(nullptr), _minus_InvMM_ST(nullptr), _parInvMM(nullptr), _parDepSurfDiffusion(nullptr)
 	{
 	}
 
 	ParticleDiffusionOperatorDG::~ParticleDiffusionOperatorDG() CADET_NOEXCEPT
 	{
+		delete[] _binding;
 		delete[] _boundOffset;
 		delete[] _nBound;
 		delete[] _localFlux;
@@ -55,6 +61,18 @@ namespace parts
 		delete[] _minus_InvMM_ST;
 		delete[] _parInvMM;
 		delete[] _parDepSurfDiffusion;
+	}
+
+	// note: currently only single bindings are supported
+	void ParticleDiffusionOperatorDG::clearBindingModels() CADET_NOEXCEPT
+	{
+		if (_binding)
+			delete _binding;
+	}
+	void ParticleDiffusionOperatorDG::clearDynamicReactionModels() CADET_NOEXCEPT
+	{
+		if (_dynReaction)
+			delete _dynReaction;
 	}
 
 	bool ParticleDiffusionOperatorDG::configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, const int nComp, const int parTypeIdx, const int nParType, const int strideBulkComp)
@@ -332,108 +350,77 @@ namespace parts
 			}
 		}
 
-		//_curSection = -1;
+		// ==== Construct and configure binding model
+		clearBindingModels();
+		_binding = nullptr;
 
-		//// ==== Construct and configure binding model
-		//clearBindingModels();
-		//_binding = std::vector<IBindingModel*>(_nParType, nullptr);
+		std::vector<std::string> bindModelNames = { "NONE" };
+		if (paramProvider.exists("ADSORPTION_MODEL"))
+			bindModelNames = paramProvider.getStringArray("ADSORPTION_MODEL");
 
-		//std::vector<std::string> bindModelNames = { "NONE" };
-		//if (paramProvider.exists("ADSORPTION_MODEL"))
-		//	bindModelNames = paramProvider.getStringArray("ADSORPTION_MODEL");
[…]
