### 2025-05-28 6b7ac33
Update src/libcadet/model/reaction/MichaelisMentenReaction.cpp
Co-authored-by: Johannes Schmölder <20299934+schmoelder@users.noreply.github.com>
