+++ doc/modelling/reaction/index.rst
@@ -6,7 +6,7 @@ Reaction models
 
 Reaction models describe the (net) fluxes :math:`f_{\mathrm{react}}` of a
 reaction mechanism.
-CADET features one reaction type:
+CADET features multiple reaction types:
 
  - :ref:`mass_action_law_model` (most commonly used)
  - :ref:`michaelis_menten_kinetics_model` (for enzyme kinetics)
