+++ doc/interface/reaction/michaelis_menten_kinetics.rst
@@ -38,15 +38,15 @@ For information on model equations, refer to :ref:`michaelis_menten_kinetics_mod
 	Inhibition constant for competitive inhibition :math:`K_{i,j,k}` w.r.t substrate :math:`i`, inhibitor :math:`j` and reaction :math:`j`. If :math:`k_{i,j,k} \leq 0`, the component does not inhibit the reaction.
 	Input as reaction index major.
 
-   ================  =============================  ========================================================
+   ================  =============================  =============================================================================
    **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** :math:`\texttt{NREACT} \cdot \texttt{NCOMP} \cdot \texttt{NCOMP}`
-   ================  =============================  ========================================================
+   ================  =============================  =============================================================================
 
 ``MM_KI_NC``
 
 	Inhibition constant non competitive inhibition:math:`\tilde{K}_{i,j,k}` w.r.t substrate :math:`i`, inhibitor :math:`j` and reaction :math:`j`. If :math:`k_{i,j,k} \leq 0`, the component does not inhibit the reaction.
 	Input as reaction index major.
 
-   ================  =============================  ========================================================
+   ================  =============================  =============================================================================
    **Type:** double  **Range:** :math:`\mathbb{R}`  **Length:** :math:`\texttt{NREACT} \cdot \texttt{NCOMP} \cdot \texttt{NCOMP}`
-   ================  =============================  ========================================================
+   ================  =============================  =============================================================================
