### 2025-05-28 7142e2e
Docs: add build type to cmake options
