### 2025-05-28 9e5f2c6
Add sensitivity support for bulk reactions (#426)
