+++ src/libcadet/model/GeneralRateModel2D.cpp
@@ -2358,6 +2358,9 @@ bool GeneralRateModel2D::setParameter(const ParameterId& pId, double value)
 
 		if (_convDispOp.setParameter(pId, value))
 			return true;
+
+		if (model::setParameter(pId, value, std::vector<IDynamicReactionModel*>{ _dynReactionBulk }, true))
+			return true;
 	}
 
 	const bool result = UnitOperationBase::setParameter(pId, value);
@@ -2367,6 +2370,34 @@ bool GeneralRateModel2D::setParameter(const ParameterId& pId, double value)
 		updateRadialDisc();
 
 	return result;
+}
+
+bool GeneralRateModel2D::setParameter(const ParameterId& pId, int value)
+{
+	if ((pId.unitOperation != _unitOpIdx) && (pId.unitOperation != UnitOpIndep))
+		return false;
+
+	if (pId.unitOperation == _unitOpIdx)
+	{
+		if (model::setParameter(pId, value, std::vector<IDynamicReactionModel*>{ _dynReactionBulk }, true))
+			return true;
+	}
+
+	return UnitOperationBase::setParameter(pId, value);
+}
+
+bool GeneralRateModel2D::setParameter(const ParameterId& pId, bool value)
+{
+	if ((pId.unitOperation != _unitOpIdx) && (pId.unitOperation != UnitOpIndep))
+		return false;
+
+	if (pId.unitOperation == _unitOpIdx)
+	{
+		if (model::setParameter(pId, value, std::vector<IDynamicReactionModel*>{ _dynReactionBulk }, true))
+			return true;
+	}
+
+	return UnitOperationBase::setParameter(pId, value);
 }
 
 void GeneralRateModel2D::setSensitiveParameterValue(const ParameterId& pId, double value)
@@ -2394,6 +2425,9 @@ void GeneralRateModel2D::setSensitiveParameterValue(const ParameterId& pId, doub
 			return;
 
 		if (_convDispOp.setSensitiveParameterValue(_sensParams, pId, value))
+			return;
+
+		if (model::setSensitiveParameterValue(pId, value, _sensParams, std::vector<IDynamicReactionModel*>{ _dynReactionBulk }, true))
 			return;
 	}
 
@@ -2468,6 +2502,12 @@ bool GeneralRateModel2D::setSensitiveParameter(const ParameterId& pId, unsigned 
 		if (_convDispOp.setSensitiveParameter(_sensParams, pId, adDirection, adValue))
 		{
 			LOG(Debug) << "Found parameter " << pId << ": Dir " << adDirection << " is set to " << adValue;
+			return true;
+		}
+
+		if (model::setSensitiveParameter(pId, adDirection, adValue, _sensParams, std::vector<IDynamicReactionModel*> { _dynReactionBulk }, true))
+		{
+			LOG(Debug) << "Found parameter " << pId << " in DynamicBulkReactionModel: Dir " << adDirection << " is set to " << adValue;
 			return true;
 		}
 	}
