### 2025-05-28 ce58c10
Docs: add build type to cmake options (#425)
