### 2025-05-28 aed7d49
Docs: add build type to cmake options (#425)
