### 2025-05-28 fcd606c
Docs: add build type to cmake options (#425)
