### 2025-05-28 8273903
Add gcovr-based code coverage workflow with Codecov upload
