+++ .github/workflows/coverage.yml
@@ -106,13 +106,13 @@ jobs:
 
       - name: Generate Coverage Report
         run: |
-          gcovr --root "${SRC_DIR}" --filter "${SRC_DIR}" --xml -o coverage.xml
-          gcovr --root "${SRC_DIR}" --filter "${SRC_DIR}" --html --html-details -o coverage.html
+            gcovr --root "${SRC_DIR}" --filter "${SRC_DIR}" --exclude "ThirdParty" --xml -o coverage.xml
+            gcovr --root "${SRC_DIR}" --filter "${SRC_DIR}" --exclude "ThirdParty" --html --html-details -o coverage.html
       
       - name: Upload Coverage to Codecov
         uses: codecov/codecov-action@v5
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
-          slug: cadet/cadet-core
+          slug: cadet/CADET-Core
           files: coverage.xml
 
