### 2025-05-28 0603ff2
Add gcovr-based code coverage workflow with Codecov upload (#428)
