### 2025-05-28 2622afd
Coverage workflow: exclude third party libraries (#429)
jo
