+++ src/libcadet/model/parts/ParticleDiffusionOperatorDG.cpp
@@ -20,6 +20,7 @@
 #include "AdUtils.hpp"
 #include "SimulationTypes.hpp"
 #include "model/ParameterDependence.hpp"
+#include "model/parts/BindingCellKernel.hpp"
 #include "SensParamUtil.hpp"
 #include "ConfigurationHelper.hpp"
 #include "model/BindingModel.hpp"
@@ -52,6 +53,7 @@ namespace parts
 	ParticleDiffusionOperatorDG::~ParticleDiffusionOperatorDG() CADET_NOEXCEPT
 	{
 		delete[] _binding;
+		delete[] _dynReaction;
 		delete[] _boundOffset;
 		delete[] _nBound;
 		delete[] _localFlux;
@@ -61,18 +63,6 @@ namespace parts
 		delete[] _minus_InvMM_ST;
 		delete[] _parInvMM;
 		delete[] _parDepSurfDiffusion;
-	}
-
-	// note: currently only single bindings are supported
-	void ParticleDiffusionOperatorDG::clearBindingModels() CADET_NOEXCEPT
-	{
-		if (_binding)
-			delete _binding;
-	}
-	void ParticleDiffusionOperatorDG::clearDynamicReactionModels() CADET_NOEXCEPT
-	{
-		if (_dynReaction)
-			delete _dynReaction;
 	}
 
 	bool ParticleDiffusionOperatorDG::configureModelDiscretization(IParameterProvider& paramProvider, const IConfigHelper& helper, const int nComp, const int parTypeIdx, const int nParType, const int strideBulkComp)
@@ -351,7 +341,6 @@ namespace parts
 		}
 
 		// ==== Construct and configure binding model
-		clearBindingModels();
 		_binding = nullptr;
 
 		std::vector<std::string> bindModelNames = { "NONE" };
@@ -382,12 +371,11 @@ namespace parts
 
 		// the following scopeguard enters the correct scope in the parameterprovider
 		MultiplexedScopeSelector scopeGuard(paramProvider, "adsorption", _singleBinding, _parTypeIdx, nParType == 1, _binding->usesParamProviderInDiscretizationConfig());
-		bindingConfSuccess = _binding->configureModelDiscretization(paramProvider, _nComp, _nBound + _parTypeIdx * _nComp, _boundOffset + _parTypeIdx * _nComp) && bindingConfSuccess;
+		bindingConfSuccess = _binding->configureModelDiscretization(paramProvider, _nComp, _nBound, _boundOffset) && bindingConfSuccess;
 
 		// ==== Construct and configure dynamic reaction model
 		bool reactionConfSuccess = true;
 
-		clearDynamicReactionModels();
 		_dynReaction = nullptr;
 
 		if (paramProvider.exists("REACTION_MODEL_PARTICLES"))
@@ -417,7 +405,7 @@ namespace parts
 				throw InvalidParameterException("Unknown dynamic reaction model " + dynReactModelNames[_singleDynReaction ? 0 : _parTypeIdx]);
 
 			MultiplexedScopeSelector scopeGuard(paramProvider, "reaction_particle", _singleDynReaction, _parTypeIdx, nParType == 1, _dynReaction->usesParamProviderInDiscretizationConfig());
-			reactionConfSuccess = _dynReaction->configureModelDiscretization(paramProvider, _nComp, _nBound + _parTypeIdx * _nComp, _boundOffset + _parTypeIdx * _nComp) && reactionConfSuccess;
+			reactionConfSuccess = _dynReaction->configureModelDiscretization(paramProvider, _nComp, _nBound, _boundOffset) && reactionConfSuccess;
 		}
 
[…]

[…]
