### 2025-05-31 ababb7c
Add MAL reaction test with sensitivities (#426)
