### 2025-06-01 d20617c
Improve test coverage
