### 2025-06-01 d3c1829
Add MAL reaction test with sensitivity for GRM DG
