### 2025-06-01 d59b596
add external reactiondependentcomponentmatrixdependent type
no register and configure function yet
