### 2025-06-02 8a133a2
update doc: add example configuration
