### 2025-06-02 b29e992
CI fix: fail and exit on first testRunner error (#430)
This commit adds a command to the CI workflow, telling the shell to immediately exit if any command returns a non-zero exit status
