### 2025-06-02 a97fda2
add configure and registration function
