### 2025-06-02 121d032
Add binding and reaction to particle module
