### 2025-06-02 f9613ec
fixup! configureModelDisc
