+++ src/libcadet/model/parts/ParticleDiffusionOperatorDG.cpp
@@ -339,108 +339,66 @@ namespace parts
 			}
 		}
 
-		//_curSection = -1;
+		// ==== Construct and configure binding model
 
-		//// ==== Construct and configure binding model
-		//clearBindingModels();
-		//_binding = std::vector<IBindingModel*>(_nParType, nullptr);
+		std::vector<std::string> bindModelNames = { "NONE" };
+		if (paramProvider.exists("ADSORPTION_MODEL"))
+			bindModelNames = paramProvider.getStringArray("ADSORPTION_MODEL");
 
-		//std::vector<std::string> bindModelNames = { "NONE" };
-		//if (paramProvider.exists("ADSORPTION_MODEL"))
-		//	bindModelNames = paramProvider.getStringArray("ADSORPTION_MODEL");
+		if (paramProvider.exists("ADSORPTION_MODEL_MULTIPLEX"))
+			_singleBinding = (paramProvider.getInt("ADSORPTION_MODEL_MULTIPLEX") == 1);
+		else
+		{
+			// Infer multiplex mode
+			_singleBinding = (bindModelNames.size() == 1);
+		}
 
-		//if (paramProvider.exists("ADSORPTION_MODEL_MULTIPLEX"))
-		//	_singleBinding = (paramProvider.getInt("ADSORPTION_MODEL_MULTIPLEX") == 1);
-		//else
-		//{
-		//	// Infer multiplex mode
-		//	_singleBinding = (bindModelNames.size() == 1);
-		//}
+		if (!_singleBinding && (bindModelNames.size() < nParType))
+			throw InvalidParameterException("Field ADSORPTION_MODEL contains too few elements (" + std::to_string(nParType) + " required)");
+		else if (_singleBinding && (bindModelNames.size() != 1))
+			throw InvalidParameterException("Field ADSORPTION_MODEL requires (only) 1 element");
 
-		//if (!_singleBinding && (bindModelNames.size() < _nParType))
-		//	throw InvalidParameterException("Field ADSORPTION_MODEL contains too few elements (" + std::to_string(_nParType) + " required)");
-		//else if (_singleBinding && (bindModelNames.size() != 1))
-		//	throw InvalidParameterException("Field ADSORPTION_MODEL requires (only) 1 element");
+		bool bindingConfSuccess = true;
 
-		//bool bindingConfSuccess = true;
-		//for (unsigned int i = 0; i < _nParType; ++i)
-		//{
-		//	if (_singleBinding && (i > 0))
-		//	{
-		//		// Reuse first binding model
-		//		_binding[i] = _binding[0];
-		//	}
-		//	else
-		//	{
-		//		_binding[i] = helper.createBindingModel(bindModelNames[i]);
-		//		if (!_binding[i])
-		//			throw InvalidParameterException("Unknown binding model " + bindModelNames[i]);
+		_binding = helper.createBindingModel(bindModelNames[_singleBinding ? 0 : _parTypeIdx]);
+		if (!_binding)
+			throw InvalidParameterException("Unknown binding model " + bindModelNames[_singleBinding ? 0 : _parTypeIdx]);
 
-		//		MultiplexedScopeSelector scopeGuard(paramProvider, "adsorption", _singleBinding, i, _nParType == 1, _binding[i]->usesParamProviderInDiscretizationConfig());
-		//		bindingConfSuccess = _binding[i]->configureModelDiscretization(paramProvider, _nComp, _nBound + i * _nComp, _boundOffset + i * _nComp) && bindingConfSuccess;
-		//	}
-		//}
[…]
