### 2025-06-02 65262ef
fixup! configure()
