### 2025-06-02 c96d631
fixup! use internal qsBinding pointer
