+++ src/libcadet/model/parts/ParticleDiffusionOperatorDG.cpp
@@ -1031,24 +1031,24 @@ namespace parts
 	 * @param [in] jac Matrix that holds the Jacobian
 	 * @return @c 0 on success, @c -1 on non-recoverable error, and @c +1 on recoverable error
 	 */
-	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int colNode, double const* yPar, double const* yBulk, double const* yDotPar, double* resPar, int const* const qsBinding, WithoutParamSensitivity)
+	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int colNode, double const* yPar, double const* yBulk, double const* yDotPar, double* resPar, WithoutParamSensitivity)
 	{
-		return residualImpl<double, double, double>(t, secIdx, colNode, yPar, yBulk, yDotPar, resPar, qsBinding);
+		return residualImpl<double, double, double>(t, secIdx, colNode, yPar, yBulk, yDotPar, resPar);
 	}
 
-	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int colNode, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithoutParamSensitivity)
+	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int colNode, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, WithoutParamSensitivity)
 	{
-		return residualImpl<active, active, double>(t, secIdx, colNode, yPar, yBulk, yDotPar, resPar, qsBinding);
+		return residualImpl<active, active, double>(t, secIdx, colNode, yPar, yBulk, yDotPar, resPar);
 	}
 
-	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int colNode, double const* yPar, double const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithParamSensitivity)
+	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int colNode, double const* yPar, double const* yBulk, double const* yDotPar, active* resPar, WithParamSensitivity)
 	{
-		return residualImpl<double, active, active>(t, secIdx, colNode, yPar, yBulk, yDotPar, resPar, qsBinding);
+		return residualImpl<double, active, active>(t, secIdx, colNode, yPar, yBulk, yDotPar, resPar);
 	}
 
-	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int colNode, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, int const* const qsBinding, WithParamSensitivity)
+	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int colNode, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, WithParamSensitivity)
 	{
-		return residualImpl<active, active, active>(t, secIdx, colNode, yPar, yBulk, yDotPar, resPar, qsBinding);
+		return residualImpl<active, active, active>(t, secIdx, colNode, yPar, yBulk, yDotPar, resPar);
 	}
 
 	/**
@@ -1071,8 +1071,10 @@ namespace parts
 	}
 
 	template <typename StateType, typename ResidualType, typename ParamType>
-	int ParticleDiffusionOperatorDG::residualImpl(double t, unsigned int secIdx, unsigned int colNode, StateType const* yPar, StateType const* yBulk, double const* yDotPar, ResidualType* resPar, int const* const qsBinding)
[…]
