### 2025-06-02 993f1ba
fixup! remove configuration from GRM.cpp
