### 2025-06-02 4293fe9
fixup! particle model residual templating
