+++ src/libcadet/model/parts/ParticleDiffusionOperatorDG.cpp
@@ -1025,31 +1025,50 @@ namespace parts
 	 * @param [in] model Model that owns the operator
 	 * @param [in] t Current time point
 	 * @param [in] secIdx Index of the current section
-	 * @param [in] y Pointer to unit operation's state vector
-	 * @param [in] yDot Pointer to unit operation's time derivative state vector
-	 * @param [out] res Pointer to unit operation's residual vector
-	 * @param [in] jac Matrix that holds the Jacobian
+	 * @param [in] yPar Pointer to particle phase entry in unit state vector
+	 * @param [in] yBulk Pointer to corresponding bulk phase entry in unit state vector
+	 * @param [in] yDotPar Pointer to particle phase derivative entry in unit state vector
+	 * @param [out] resPar Pointer Pointer to particle phase entry in unit residual vector
+	 * @param [out] colPos column position of the particle (particle coordinate zero)
+	 * @param [in] jacIt Matrix iterator pointing to the particle phase entry in the unit Jacobian
 	 * @return @c 0 on success, @c -1 on non-recoverable error, and @c +1 on recoverable error
 	 */
-	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int colNode, double const* yPar, double const* yBulk, double const* yDotPar, double* resPar, WithoutParamSensitivity)
+	template<bool wantJac, bool wantRes>
+	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, double const* yPar, double const* yBulk, double const* yDotPar, double* resPar, ColumnPosition colPos, linalg::BandedEigenSparseRowIterator& jacIt, LinearBufferAllocator tlmAlloc, WithoutParamSensitivity)
 	{
-		return residualImpl<double, double, double>(t, secIdx, colNode, yPar, yBulk, yDotPar, resPar);
+		return residualImpl<double, double, double, wantJac, wantRes>(t, secIdx, yPar, yBulk, yDotPar, resPar, colPos, jacIt, tlmAlloc);
 	}
 
-	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int colNode, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, WithoutParamSensitivity)
+	template<bool wantJac, bool wantRes>
+	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, active const* yPar, active const* yBulk, double const* yDotPar, active* resPar, ColumnPosition colPos, linalg::BandedEigenSparseRowIterator& jacIt, LinearBufferAllocator tlmAlloc, WithoutParamSensitivity)
 	{
-		return residualImpl<active, active, double>(t, secIdx, colNode, yPar, yBulk, yDotPar, resPar);
+		return residualImpl<active, active, double, wantJac, wantRes>(t, secIdx, yPar, yBulk, yDotPar, resPar, colPos, jacIt, tlmAlloc);
 	}
 
-	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, unsigned int colNode, double const* yPar, double const* yBulk, double const* yDotPar, active* resPar, WithParamSensitivity)
+	template<bool wantJac, bool wantRes>
+	int ParticleDiffusionOperatorDG::residual(double t, unsigned int secIdx, double const* yPar, double const* yBulk, double const* yDotPar, active* resPar, ColumnPosition colPos, linalg::BandedEigenSparseRowIterator& jacIt, LinearBufferAllocator tlmAlloc, WithParamSensitivity)
 	{
[…]
