### 2025-06-02 477b769
fixup! move jacobian diagonal recomputation to particle module
