### 2025-06-02 df46c3a
fixup! move bindingCellKernel
