### 2025-06-02 a7f7469
Move particle diffusion computation into individual function
