### 2025-06-02 38a0246
fixup! remove staggered residual call in UO
