### 2025-06-02 a4dec54
fixup! add exception for failed singleBinding, singleReaction configuration
