### 2025-06-02 caf6e51
Add binding and reactions to particle operator (#424)
This commit adds binding and reactions to particle operator, outsourcing that functionality from the GRM_DG
