### 2025-06-02 424cdac
Add binding and reactions to particle operator (#424)
This commit adds binding and reactions to particle operator, outsourcing that functionality from the GRM_DG
