### 2025-06-03 87d679f
Modularize particle discretization in particle operator (#431)
