+++ src/libcadet/model/GeneralRateModelDG.cpp
@@ -67,7 +67,7 @@ GeneralRateModelDG::~GeneralRateModelDG() CADET_NOEXCEPT
 
 	_binding.clear(); // binding models are deleted in the respective particle model
 	_dynReaction.clear(); // particle reaction models are deleted in the respective particle model
-	delete[] _parDiffOp;
+	delete[] _particle;
 
 	delete _dynReactionBulk;
 
@@ -171,11 +171,11 @@ bool GeneralRateModelDG::configureModelDiscretization(IParameterProvider& paramP
 
 	paramProvider.popScope();
 	Indexer idxr(_disc);
-	_parDiffOp = new parts::ParticleDiffusionOperatorDG[_disc.nParType];
+	_particle = new parts::GeneralRateParticle[_disc.nParType];
 	bool particleConfSuccess = true;
 	for (int parType = 0; parType < _disc.nParType; parType++)
 	{
-		particleConfSuccess = particleConfSuccess && _parDiffOp[parType].configureModelDiscretization(paramProvider, helper, _disc.nComp, parType, _disc.nParType, idxr.strideColComp());
+		particleConfSuccess = particleConfSuccess && _particle[parType].configureModelDiscretization(paramProvider, helper, _disc.nComp, parType, _disc.nParType, idxr.strideColComp());
 	}
 	paramProvider.pushScope("discretization");
 
@@ -184,7 +184,7 @@ bool GeneralRateModelDG::configureModelDiscretization(IParameterProvider& paramP
 		_disc.nParPoints = new unsigned int[_disc.nParType];
 		for (int type = 0; type < _disc.nParType; type++)
 		{
-			_disc.nParPoints[type] = _parDiffOp[type]._nParPoints;
+			_disc.nParPoints[type] = _particle[type].nDiscPoints();
 		}
 	}
 
@@ -292,14 +292,14 @@ bool GeneralRateModelDG::configureModelDiscretization(IParameterProvider& paramP
 
 	for (unsigned int parType = 0; parType < _disc.nParType; ++parType)
 	{
-		_binding[parType] = _parDiffOp[parType]._binding;
-		_singleBinding = _parDiffOp[parType]._singleBinding;
-		if (parType > 0 && _singleBinding != _parDiffOp[parType]._singleBinding)
+		_binding[parType] = _particle[parType].getBinding();
+		_singleBinding = _particle[parType].singleBinding();
+		if (parType > 0 && _singleBinding != _particle[parType].singleBinding())
 			throw InvalidParameterException("Configuration of binding went wrong");
 
-		_dynReaction[parType] = _parDiffOp[parType]._dynReaction;
-		_singleDynReaction = _parDiffOp[parType]._singleDynReaction;
-		if (parType > 0 && _singleDynReaction != _parDiffOp[parType]._singleDynReaction)
+		_dynReaction[parType] = _particle[parType].getReaction();
+		_singleDynReaction = _particle[parType].singleReaction();
+		if (parType > 0 && _singleDynReaction != _particle[parType].singleReaction())
 			throw InvalidParameterException("Configuration of particle reaction went wrong");
 	}
 
@@ -442,7 +442,7 @@ bool GeneralRateModelDG::configure(IParameterProvider& paramProvider)
 	bool particleConfSuccess = true;
 	for (int parType = 0; parType < _disc.nParType; parType++)
 	{
[…]
