### 2025-06-06 cadet/CADET-Core#422 @hannahlanzrath
Hello Bradley Dice,

Thank you very much for your review and the positive feedback. We really appreciate the time and effort you put into evaluating CADET-Core!

Regarding your comment on the State of the field, we've expanded on this point in a dedicated section of the paper. The changes are included in the following commit: [8427323](https://github.com/cadet/CADET-Core/commit/8427323c96e862342795a24b39a3b3a221595606).

You can also find the updated PDF here: [10.21105.joss.07881.pdf](https://github.com/openjournals/joss-papers/blob/joss.07881/joss.07881/10.21105.joss.07881.pdf)

In this revision, we briefly describe the current landscape of chromatography simulation software, emphasizing that most alternatives are commercial and closed-source, as you already noted. We hope this provides useful context for readers.

Please let us know if this addition addresses your concerns.

Thanks again!
