### 2025-06-07 #7881 @bdice
Thank you @hannahlanzrath for addressing my requests in https://github.com/cadet/CADET-Core/issues/422. I checked the remaining boxes on my review and I accept the paper with no further changes requested.

Special thanks to @goxberry @jedbrown @lieres for the insightful conversations earlier in this issue, which helped me in completing my review. 👍
