### 2025-06-09 #7881 @jedbrown
Thank you all for your diligence through this review process. The paper looks good to me. @hannahlanzrath could you please confirm whether you intend to make a new release (perhaps 5.0.4) to incorporate the results of this review? Please also report the archival URL (Zenodo or your choice).
