### 2025-06-13 #7881 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
