### 2025-06-25 cadet/CADET-Core#373 @jbreue16
fixed as per our discussion (https://github.com/openjournals/joss-reviews/issues/7881#issuecomment-2770251498)
