### 2025-06-27 #7881 @lieres
Dear Rachel,

We have made the requested changes to the Zenodo title and US affiliation.

However, we would prefer to keep "spatial discontinuous" instead of "spatially discontinuous". While "spatially discontinuous" may appear to be a grammatically improved formulation, "spatial discontinuous" is a standard term in the field. In this context, "spatial" refers to the type of discretization method, distinguishing it from temporal or spatio-temporal variants, rather than describing the nature of the discontinuities. As such, "spatial discontinuous Galerkin method" is a commonly used and accepted expression in the literature.
