### 2025-07-03 #7881 @rkurchin
(take 2)
