### 2025-07-03 #7881 @lieres
While my last name appears correctly in the PDF, it seems that the non-breaking space included in the .tex file has made its way into the webpage display as "&#160;". Could this be corrected?
