### 2025-07-03 #7881 @lieres
The webpage also displays a [MATLAB](https://joss.theoj.org/papers/in/MATLAB) badge, which should be removed since the MATLAB interface of CADET was deprecated some time ago.
