### 2025-07-05 #7881 @Kevin-Mattheus-Moerman
@openjournals/dev can you help fix the MATLAB label above? Also why is that added since we do not have the label here?

@lieres I think you should edit the author section for you to:
```
  - given-names: Eric
    dropping-particle: von
    surname: Lieres
    corresponding: true # (This is how to denote the corresponding author)
    orcid: 0000-0002-0309-8408
    affiliation: "1, 2"
```
Can you try this and let me know when you've implemented it?

Once we've fixed these issues we can "re-accept" to update/fix the paper.
