### 2025-07-05 #7881 @hannahlanzrath
Dear @Kevin-Mattheus-Moerman,

Thank you for your support.

I’ve updated the naming format as you suggested, but unfortunately, the line break is now incorrect again, as shown in the [draft PDF](https://github.com/cadet/CADET-Core/actions/runs/16092394401) generated by our workflow. I’ve previously tried setting "von" as a non-dropping particle and also defining "von Lieres" as the full surname. However, in both cases, the name still breaks between "von" and "Lieres".

Since "von Lieres" is the complete surname, it should not be split across lines. I was able to achieve the correct formatting using a non-breaking space (&#160;), but unfortunately, this appears rendered literally on the JOSS website.

Could you advise on how to correctly format the name in JOSS so that "von Lieres" is kept together?

Bestwishes and thank you in advance,
Hannah

---

**Correct:**

**Incorrect:**
