### 2025-07-11 #7881 @Kevin-Mattheus-Moerman
@xuanxu @openjournals/dev can you check this issue too? ☝️  thanks
