### 2025-07-11 #7881 @xuanxu
The non-breaking space has been removed from the paper-md file so re-accepting should updated that on the website.
