### 2025-07-24 #7881 @hannahlanzrath
Hello JOSS Team,

I'm a bit unclear about why the paper has been reaccepted, but I noticed that the MATLAB label has reappeared again. Could you please clarify?

From our side, everything is in order, as we’ve decided to accept the name break as it is. So, if the MATLAB label is removed, I believe this issue can be closed.
